% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addBiometricsToConfig}
\alias{addBiometricsToConfig}
\title{Add biometrics information to config}
\usage{
addBiometricsToConfig(projectConfiguration, dataDT, overwrite = FALSE)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and file names}

\item{dataDT}{A `data.table` with observed data.}

\item{overwrite}{If TRUE, existing rows will be overwritten.}
}
\value{
Returns invisibly.
}
\description{
Add biometrics information to config
}
\seealso{
Other observed data processing:
\code{\link[=aggregateObservedDataGroups]{aggregateObservedDataGroups()}},
\code{\link[=convertDataCombinedToDataTable]{convertDataCombinedToDataTable()}},
\code{\link[=convertDataTableToDataCombined]{convertDataTableToDataCombined()}},
\code{\link[=readObservedDataByDictionary]{readObservedDataByDictionary()}},
\code{\link[=updateDataGroupId]{updateDataGroupId()}},
\code{\link[=updateOutputPathId]{updateOutputPathId()}},
\code{\link[=validateObservedData]{validateObservedData()}}
}
\concept{observed data processing}
