% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{addCaptionTextAddon}
\alias{addCaptionTextAddon}
\title{Add Caption Text Addon}
\usage{
addCaptionTextAddon(captiontext, plotCaptionAddon)
}
\arguments{
\item{captiontext}{A character string representing the main caption text.}

\item{plotCaptionAddon}{A character string representing the additional caption
text to be added. If NULL or NA, it will not be added.}
}
\value{
A modified caption text with the additional caption added if applicable.
}
\description{
This function modifies a caption text by ensuring it ends with a period
and appending an additional caption if provided. Ensure the additional caption
ends also with a period before adding the additional text.
}
\keyword{internal}
