% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{addConfigToTemplate}
\alias{addConfigToTemplate}
\title{adds the default configuration to the template configuration}
\usage{
addConfigToTemplate(wb, templateSheet, sheetName, dtNewConfig)
}
\arguments{
\item{wb}{Plotconfiguration file}

\item{templateSheet}{name of the template sheet}

\item{sheetName}{name of new sheet}

\item{dtNewConfig}{`data.table` with default configuration}
}
\value{
Plotconfiguration file
}
\description{
if the template does not exist in the plot-configuration file in the project directory
it is taken from the plot-configuration file of the package installation.
In this case formats are not preserved
}
