% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{addDefaultConfigForHistograms}
\alias{addDefaultConfigForHistograms}
\title{Add Default Configuration for Histograms}
\usage{
addDefaultConfigForHistograms(
  projectConfiguration,
  pkParameterDT = NULL,
  sheetName = "Histograms",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration class object containing configuration details, including:
- `addOns$reportsFile`: A string representing the file path to the Excel workbook containing the plot configurations.}

\item{pkParameterDT}{Optional. A data object containing pkParameter.}

\item{sheetName}{A character string specifying the name of the sheet in the plot configuration table.
Default is "Histograms".}

\item{overwrite}{A boolean indicating whether existing configurations should be overwritten.
Default is FALSE.}
}
\value{
NULL This function updates the Excel workbook in place and does not return a value.
}
\description{
This function adds a default configuration sheet for histograms to the plot configuration table.
It can either create a new sheet or overwrite an existing one based on the specified parameters.
}
\details{
The function retrieves scenario definitions, output path IDs, and data groups from the project configuration.
It checks if the specified sheet already exists and whether to overwrite it. If not, it creates a new header
and fills in the default configuration values for the histograms.

Additionally, the function performs a validity check to ensure that it is not executed during a context
where helper functions are prohibited (validRun). If such a context is detected, an error is raised to prevent execution.
}
\seealso{
Other plot configuration helper function:
\code{\link[=addDefaultConfigForDistributionsVsDemographics]{addDefaultConfigForDistributionsVsDemographics()}},
\code{\link[=addDefaultConfigForPKBoxwhsikerPlots]{addDefaultConfigForPKBoxwhsikerPlots()}},
\code{\link[=addDefaultConfigForPKForestPlots]{addDefaultConfigForPKForestPlots()}},
\code{\link[=addDefaultConfigForTimeProfilePlots]{addDefaultConfigForTimeProfilePlots()}}

Other functions to generate plots displaying distribution vs demographics:
\code{\link[=plotDistributionVsDemographics]{plotDistributionVsDemographics()}},
\code{\link[=validateDistributionVsDemographicsConfig]{validateDistributionVsDemographicsConfig()}}
}
\concept{functions to generate plots displaying distribution vs demographics}
\concept{plot configuration helper function}
