% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKBoxwhisker.R
\name{addDefaultConfigForPKBoxwhsikerPlots}
\alias{addDefaultConfigForPKBoxwhsikerPlots}
\title{Add Default Configuration for Box-and-Whisker Plots}
\usage{
addDefaultConfigForPKBoxwhsikerPlots(
  projectConfiguration,
  pkParameterDT,
  sheetName = "PKParameter_Boxplot",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration class object containing configuration details, including:
- `addOns$reportsFile`: A string representing the file path to the Excel workbook containing the plot configurations.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{sheetName}{A character string specifying the name of the sheet in the plot configuration table.
Default is "PKParameter_Boxplot".}

\item{overwrite}{A boolean indicating whether existing configurations should be overwritten.
Default is FALSE.}
}
\value{
NULL This function updates the Excel workbook in place and does not return a value.
It is called for its side effects.
}
\description{
This function adds a default configuration sheet for box-and-whisker plots to the plot configuration table.
It can either create a new sheet or overwrite an existing one based on the specified parameters.
}
\details{
The function retrieves scenario definitions, output path IDs, and data groups from the project configuration.
It checks if the specified sheet already exists and whether to overwrite it. If not, it creates a new header
and fills in the default configuration values for the box-and-whisker plots.

Additionally, the function performs a validity check to ensure that it is not executed during a context
where helper functions are prohibited (`validRun`). If such a context is detected, an error is raised to prevent execution.
}
\seealso{
Other plot configuration helper function:
\code{\link[=addDefaultConfigForDistributionsVsDemographics]{addDefaultConfigForDistributionsVsDemographics()}},
\code{\link[=addDefaultConfigForHistograms]{addDefaultConfigForHistograms()}},
\code{\link[=addDefaultConfigForPKForestPlots]{addDefaultConfigForPKForestPlots()}},
\code{\link[=addDefaultConfigForTimeProfilePlots]{addDefaultConfigForTimeProfilePlots()}}

Other functions to generate box whisker plots:
\code{\link[=plotPKBoxwhisker]{plotPKBoxwhisker()}},
\code{\link[=validatePKBoxwhiskerConfig]{validatePKBoxwhiskerConfig()}}
}
\concept{functions to generate box whisker plots}
\concept{plot configuration helper function}
