% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addDefaultConfigForPKForestPlots}
\alias{addDefaultConfigForPKForestPlots}
\title{Add Default Configuration for PK Forest Plots}
\usage{
addDefaultConfigForPKForestPlots(
  projectConfiguration,
  pkParameterDT,
  sheetName = "PKParameter_Forest",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{sheetName}{Name of the sheet to create.}

\item{overwrite}{Logical indicating if existing data should be overwritten.}
}
\value{
NULL (invisible).
}
\description{
Adds default configurations for forest plots to the `Plots.xlsx` configuration file.
}
\seealso{
\code{\link{plotPKForestAggregatedAbsoluteValues}}, \code{\link{plotPKForestPointEstimateOfAbsoluteValues}},
\code{\link{plotPKForestAggregatedRatios}}, \code{\link{plotPKForestPointEstimateOfRatios}},

Other plot functions:
\code{\link[=addDefaultConfigForTimeProfilePlots]{addDefaultConfigForTimeProfilePlots()}},
\code{\link[=plotDistributionVsDemographics]{plotDistributionVsDemographics()}},
\code{\link[=plotHistograms]{plotHistograms()}},
\code{\link[=plotPKBoxwhisker]{plotPKBoxwhisker()}},
\code{\link[=plotPKForestAggregatedAbsoluteValues]{plotPKForestAggregatedAbsoluteValues()}},
\code{\link[=plotPKForestAggregatedRatios]{plotPKForestAggregatedRatios()}},
\code{\link[=plotPKForestPointEstimateOfAbsoluteValues]{plotPKForestPointEstimateOfAbsoluteValues()}},
\code{\link[=plotPKForestPointEstimateOfRatios]{plotPKForestPointEstimateOfRatios()}},
\code{\link[=plotSensitivity]{plotSensitivity()}},
\code{\link[=plotTimeProfiles]{plotTimeProfiles()}}

Other plot configuration helper function:
\code{\link[=addDefaultConfigForDistributionsVsDemographics]{addDefaultConfigForDistributionsVsDemographics()}},
\code{\link[=addDefaultConfigForHistograms]{addDefaultConfigForHistograms()}},
\code{\link[=addDefaultConfigForPKBoxwhsikerPlots]{addDefaultConfigForPKBoxwhsikerPlots()}},
\code{\link[=addDefaultConfigForTimeProfilePlots]{addDefaultConfigForTimeProfilePlots()}}

Other functions to generate forest plots:
\code{\link[=plotPKForestAggregatedAbsoluteValues]{plotPKForestAggregatedAbsoluteValues()}},
\code{\link[=plotPKForestAggregatedRatios]{plotPKForestAggregatedRatios()}},
\code{\link[=plotPKForestPointEstimateOfAbsoluteValues]{plotPKForestPointEstimateOfAbsoluteValues()}},
\code{\link[=plotPKForestPointEstimateOfRatios]{plotPKForestPointEstimateOfRatios()}},
\code{\link[=validatePKForestAggregatedAbsoluteValuesConfig]{validatePKForestAggregatedAbsoluteValuesConfig()}},
\code{\link[=validatePKForestAggregatedRatiosConfig]{validatePKForestAggregatedRatiosConfig()}},
\code{\link[=validatePKForestPointEstimateOfAbsoluteValuesConfig]{validatePKForestPointEstimateOfAbsoluteValuesConfig()}},
\code{\link[=validatePKForestPointEstimateOfRatiosConfig]{validatePKForestPointEstimateOfRatiosConfig()}}
}
\concept{functions to generate forest plots}
\concept{plot configuration helper function}
\concept{plot functions}
