% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{addDefaultConfigForTimeProfilePlots}
\alias{addDefaultConfigForTimeProfilePlots}
\title{Add Default Configuration for Time Profile Plots}
\usage{
addDefaultConfigForTimeProfilePlots(
  projectConfiguration,
  dataObserved = NULL,
  sheetName = "TimeProfiles",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A `ProjectConfiguration` class object containing configuration details, including:
- `addOns$reportsFile`: A string representing the file path to the Excel workbook containing the plot configurations.}

\item{dataObserved}{Optional. A data object containing observed data, if available.}

\item{sheetName}{A character string specifying the name of the sheet in the plot configuration table.
Default is `TimeProfiles`.}

\item{overwrite}{A boolean indicating whether existing configurations should be overwritten.
Default is FALSE.}
}
\value{
invisible(NULL). Updates the Excel workbook in place.
}
\description{
This function adds a default configuration sheet for time profile plots to the plot configuration table.
It can either create a new sheet or overwrite an existing one based on the specified parameters.
}
\details{
The function retrieves scenario definitions, output path IDs, and data groups from the project configuration.
It checks if the specified sheet already exists and whether to overwrite it. If not, it creates a new header
and fills in the default configuration values for the time profile plots.

Additionally, the function performs a validity check to ensure that it is not executed during a context
where helper functions are prohibited (`validRun`). If such a context is detected, an error is raised to prevent execution.
}
\examples{
\dontrun{
addDefaultConfigForTimeProfilePlots(
  projectConfiguration = projectConfiguration,
  dataObserved = dataObserved,
  sheetName = "TimeProfiles"
)
}
}
\seealso{
Other plot functions:
\code{\link[=addDefaultConfigForPKForestPlots]{addDefaultConfigForPKForestPlots()}},
\code{\link[=plotDistributionVsDemographics]{plotDistributionVsDemographics()}},
\code{\link[=plotHistograms]{plotHistograms()}},
\code{\link[=plotPKBoxwhisker]{plotPKBoxwhisker()}},
\code{\link[=plotPKForestAggregatedAbsoluteValues]{plotPKForestAggregatedAbsoluteValues()}},
\code{\link[=plotPKForestAggregatedRatios]{plotPKForestAggregatedRatios()}},
\code{\link[=plotPKForestPointEstimateOfAbsoluteValues]{plotPKForestPointEstimateOfAbsoluteValues()}},
\code{\link[=plotPKForestPointEstimateOfRatios]{plotPKForestPointEstimateOfRatios()}},
\code{\link[=plotSensitivity]{plotSensitivity()}},
\code{\link[=plotTimeProfiles]{plotTimeProfiles()}}

Other plot configuration helper function:
\code{\link[=addDefaultConfigForDistributionsVsDemographics]{addDefaultConfigForDistributionsVsDemographics()}},
\code{\link[=addDefaultConfigForHistograms]{addDefaultConfigForHistograms()}},
\code{\link[=addDefaultConfigForPKBoxwhsikerPlots]{addDefaultConfigForPKBoxwhsikerPlots()}},
\code{\link[=addDefaultConfigForPKForestPlots]{addDefaultConfigForPKForestPlots()}}
}
\concept{plot configuration helper function}
\concept{plot functions}
