% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{addDemographicsToBin}
\alias{addDemographicsToBin}
\title{Add Demographics to Bin}
\usage{
addDemographicsToBin(plotData, onePlotConfig, scenarioList)
}
\arguments{
\item{plotData}{A data.table containing the plot data.}

\item{onePlotConfig}{A configuration for the specific plot.}

\item{scenarioList}{A list of scenarios to consider.}
}
\value{
A data.table containing the updated plot data with demographic bins.
}
\description{
Adds demographic data to bins for plotting.
}
\keyword{internal}
