% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{addFacets}
\alias{addFacets}
\title{Add Facets to a ggplot Object}
\usage{
addFacets(plotObject, facetScale, facetAspectRatio = 0.5, nFacetColumns)
}
\arguments{
\item{plotObject}{A ggplot object to which the facets should be added.}

\item{facetScale}{A character string indicating the scale of the facets. Options are "free", "fixed", "free_x", or "free_y".}

\item{facetAspectRatio}{A numeric value specifying the aspect ratio of the facets. Default is 0.5 (landscape half-height).}

\item{nFacetColumns}{An integer specifying the number of columns to use for the facet layout. If NULL, no faceting is done.}
}
\value{
An updated ggplot object with facets added.
}
\description{
This function adds facets to a given ggplot object, allowing for better visualization of data subsets.
Faceting is done by the variable `plotTag`.
}
\keyword{internal}
