% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addInterOccasionalVariability}
\alias{addInterOccasionalVariability}
\title{Add Inter-Occasional Variability}
\usage{
addInterOccasionalVariability(
  value,
  scenarioName,
  pkParameter,
  outputPathId,
  sigma
)
}
\arguments{
\item{value}{Numeric vector of values.}

\item{scenarioName}{Name of the scenario.}

\item{pkParameter}{Name of the PK parameter.}

\item{outputPathId}{ID for the output path.}

\item{sigma}{Standard deviation for the variability.}
}
\value{
Numeric vector with added variability.
}
\description{
Adds inter-occasional variability to the given values.
}
\keyword{internal}
