% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{addMessageToLog}
\alias{addMessageToLog}
\title{Add a message to the log file}
\usage{
addMessageToLog(messageText)
}
\arguments{
\item{messageText}{character with message text}
}
\description{
Add a message to the log file
}
\seealso{
Other log file management:
\code{\link[=captureLog]{captureLog()}},
\code{\link[=initLogfunction]{initLogfunction()}},
\code{\link[=saveSessionInfo]{saveSessionInfo()}},
\code{\link[=setShowLogMessages]{setShowLogMessages()}},
\code{\link[=writeTableToLog]{writeTableToLog()}}
}
\concept{log file management}
