% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addMetaDataToDataSet}
\alias{addMetaDataToDataSet}
\title{add meta data to a data set}
\usage{
addMetaDataToDataSet(dataSet, groupData)
}
\arguments{
\item{dataSet}{`DataSet` with observed data}

\item{groupData}{corresponding datatable with metadat}
}
\value{
`DataSet` with observed data with added metadata
}
\description{
add meta data to a data set
}
