% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addNObservations}
\alias{addNObservations}
\title{Add Observations to PK Parameter Data}
\usage{
addNObservations(pkParameterDT, pkParameterObserved, nObservationDefault, asCI)
}
\arguments{
\item{pkParameterDT}{Data table containing PK parameter data.}

\item{pkParameterObserved}{Optional data table for observed PK parameters.}

\item{nObservationDefault}{Default number of observations.}

\item{asCI}{Logical indicating if confidence intervals should be calculated.}
}
\value{
Updated data table with observations added.
}
\description{
This function adds the number of observations to the PK parameter data table.
}
\keyword{internal}
