% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addObservedData}
\alias{addObservedData}
\title{Add Observed Data to Plot Data}
\usage{
addObservedData(plotData, pkParameterObserved, onePlotConfig)
}
\arguments{
\item{plotData}{Data table with calculated plot data.}

\item{pkParameterObserved}{Optional data table for observed PK parameters.}

\item{onePlotConfig}{Configuration for the plot.}
}
\value{
Updated plot data with observed data added.
}
\description{
This function adds the observed data to the plot data.
}
\keyword{internal}
