% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{addPlotIdTable}
\alias{addPlotIdTable}
\title{Creates PlotId for each scenario outputs group}
\usage{
addPlotIdTable(plotDataList, onePlotConfig, projectConfiguration)
}
\arguments{
\item{plotDataList}{list which serves as bracket for data related objects needed for plots
plotData `data.table` with data to plot
dtPlotId `data.table`  with caption information and tags for each plot
scaleVectors, scaling vectors to adjust aesthetics like color}

\item{onePlotConfig}{}

\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
`data.table` with columns 'Scenario','OutputPathId','PlotId'
}
\description{
Creates PlotId for each scenario outputs group
}
