% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addPrecisisonRemark}
\alias{addPrecisisonRemark}
\title{Add Precision Remark to Plot}
\usage{
addPrecisisonRemark(combinedObject, asPointeEstimate, plotData)
}
\arguments{
\item{combinedObject}{The plot object to adjust.}

\item{asPointeEstimate}{Logical indicating if confidence intervals should be calculated.}

\item{plotData}{Data table containing plot data.}
}
\value{
The adjusted plot object with precision remark if applicable.
}
\description{
Add Precision Remark to Plot
}
\keyword{internal}
