% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{addRandomPKDataToTestProject}
\alias{addRandomPKDataToTestProject}
\title{Add Random PK Data to Test Project}
\usage{
addRandomPKDataToTestProject(projectConfiguration, pkParameterDT)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}

\item{pkParameterDT}{A data.table containing PK parameters and values.}
}
\value{
Invisible NULL.
}
\description{
This function adds random pharmacokinetic (PK) data to the test project,
including shifting values and calculating statistics for selected individuals.
}
\keyword{internal}
