% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{addRandomTPDataToTestProject}
\alias{addRandomTPDataToTestProject}
\title{Add Random Time Profile Data to Test Project}
\usage{
addRandomTPDataToTestProject(
  projectConfiguration,
  scenarioResult,
  csvfile,
  ids
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}

\item{scenarioResult}{A scenario result object containing simulation results.}

\item{csvfile}{A character string specifying the name of the CSV file to save the data.}

\item{ids}{A vector of individual IDs to include in the random data.}
}
\value{
Invisible NULL.
}
\description{
This function adds random time profile data to the test project for selected individuals.
}
\keyword{internal}
