% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{addScenariosToEPackage}
\alias{addScenariosToEPackage}
\title{Add Scenarios to Electronic Package}
\usage{
addScenariosToEPackage(projectConfiguration, configTable, subfolder)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project-specific configurations,
including paths for electronic package folders and
output directories.}

\item{configTable}{A data.table containing configuration details for scenarios,
including scenario names, long names, and data group IDs.}

\item{subfolder}{A character string indicating the subfolder filled by the config table}
}
\value{
This function does not return a value. It performs file operations and
        generates log files based on the configuration provided.
}
\description{
This function processes a configuration table of scenarios, validates filenames,
and organizes the necessary files for an electronic package export. It also
generates logs and summary files for review.
}
