% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity.R
\name{addSensitivityTable}
\alias{addSensitivityTable}
\title{Add Sensitivity Table to Project Configuration}
\usage{
addSensitivityTable(
  projectConfiguration,
  scenarioList = NULL,
  scenarioName,
  sheetName = scenarioName
)
}
\arguments{
\item{projectConfiguration}{An object representing the project configuration.}

\item{scenarioList}{A list of scenarios from which to extract parameter paths. Defaults to NULL.}

\item{scenarioName}{The name of the scenario to extract parameter paths from.}

\item{sheetName}{The name of the sheet in the Excel file to populate. Defaults to the value of scenarioName.}
}
\value{
The modified project configuration object.
}
\description{
This function adds a sensitivity table to the provided project configuration.
It loads a template Excel file and populates it with parameter paths from the specified scenario.
}
