% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotForrest.R
\name{addTableToForest}
\alias{addTableToForest}
\title{Add Table to Forest Plot}
\usage{
addTableToForest(
  plotObject,
  xgridTable,
  digitsToRound,
  digitsToShow,
  withoutTable
)
}
\arguments{
\item{plotObject}{A ggplot object representing the forest plot.}

\item{xgridTable}{Numeric vector for x-axis grid table positions.}

\item{digitsToRound}{Integer indicating the number of digits to round to.}

\item{digitsToShow}{Integer indicating the number of digits to show.}

\item{withoutTable}{Logical indicating whether to display the table.}
}
\value{
A ggplot object with the table added.
}
\description{
This function adds a table to the forest plot if specified.
}
\keyword{internal}
