% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{addTimeRangeTagsToData}
\alias{addTimeRangeTagsToData}
\title{Add Time Range Tags to Data}
\usage{
addTimeRangeTagsToData(
  timeRangeColumns,
  dataOld,
  configTable,
  applicationTimes,
  timeTags
)
}
\arguments{
\item{timeRangeColumns}{A vector of time range column names.}

\item{dataOld}{A data table containing the old data.}

\item{configTable}{A data frame containing the configuration data.}

\item{applicationTimes}{A list of application times.}
}
\value{
A data table with added time range tags.
}
\description{
This function adds time range tags to the provided data based on the configuration table.
}
\keyword{internal}
