% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{addTimeTagsToCaption}
\alias{addTimeTagsToCaption}
\title{Add Time Tags to Caption}
\usage{
addTimeTagsToCaption(dtCaption, timeTags, splitPlotsPerTimeRange)
}
\arguments{
\item{dtCaption}{A data table containing the caption data.}

\item{timeTags}{A vector of time tags to be added.}

\item{splitPlotsPerTimeRange}{Boolean, if TRUE for each Time-range a plot is created}
}
\value{
A data table with added time tags.
}
\description{
This function adds time tags to the caption data based on the specified time tags.
}
\keyword{internal}
