% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addUniqueColumns}
\alias{addUniqueColumns}
\title{add all columns of observedData which are uniuqe for a group}
\usage{
addUniqueColumns(dataObserved, aggregatedData)
}
\arguments{
\item{aggregatedData}{}
}
\description{
add all columns of observedData which are uniuqe for a group
}
