% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{addUserDefinedParameters}
\alias{addUserDefinedParameters}
\title{Add User Defined Parameters}
\usage{
addUserDefinedParameters(userdefinedParameters, dtUserdefPKParameter)
}
\arguments{
\item{userdefinedParameters}{A vector of user-defined PK parameter names.}

\item{dtUserdefPKParameter}{A data frame of user-defined PK parameters.}
}
\value{
NULL (invisible).
throws Error if user-defined parameters are not defined or are not unique.
}
\description{
Adds user-defined PK parameters to the OSPSuite based on the provided definitions.
}
