% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{adjustDataTableDimensions}
\alias{adjustDataTableDimensions}
\title{Adjust the dimensions of a data.table to match existing data}
\usage{
adjustDataTableDimensions(dt, existingData)
}
\arguments{
\item{dt}{A data.table containing the new data to be written.}

\item{existingData}{A data.table containing the existing data in the specified sheet.}
}
\value{
A data.table with adjusted dimensions.
}
\description{
This function adjusts the dimensions of a new data.table to match the existing data.
If the existing data has more rows, it appends NA rows to the new data.
}
\keyword{internal}
