% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{adjustForestDataPerGroup}
\alias{adjustForestDataPerGroup}
\title{Adjust Forest Data Per Group}
\usage{
adjustForestDataPerGroup(dataGroup, onePlotConfig)
}
\arguments{
\item{dataGroup}{A data table representing a group of plot data.}

\item{onePlotConfig}{Configuration for the plot, containing metadata like plot names.}
}
\value{
A modified data table with updated `outputPathId` as a factor and generated `plotTag`.
}
\description{
Processes a data group to ensure unique scenarios and generates plot tags.
}
\details{
This function checks the uniqueness of scenarios within the provided data group,
modifies the `outputPathId` to be a factor with specified levels, and generates
plot tags based on the `outputPathId`.
}
\keyword{internal}
