% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{adjustPkForestPlotObject}
\alias{adjustPkForestPlotObject}
\title{Adjust PK Forest Plot Object}
\usage{
adjustPkForestPlotObject(plotObject, scaleVectors, vlineIntercept)
}
\arguments{
\item{plotObject}{The plot object to adjust.}

\item{scaleVectors}{A list defining colors, fills, and shapes for the plot.}

\item{vlineIntercept}{Optional vertical line intercept.}
}
\value{
The adjusted plot object.
}
\description{
Adjusts aesthetics of the PK forest plot object.
}
\keyword{internal}
