% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{aggregatedObservedDataGroups}
\alias{aggregatedObservedDataGroups}
\title{Aggregate Observed Data Groups}
\usage{
aggregatedObservedDataGroups(
  dataObserved,
  groups = NULL,
  aggregationFlag = c("GeometricStdDev", "ArithmeticStdDev", "Percentiles", "Custom"),
  percentiles = c(5, 50, 95),
  groupSuffix = "aggregated",
  customFunction = NULL,
  lloq3Columns = NULL,
  lloq2Columns = NULL
)
}
\arguments{
\item{dataObserved}{A data.table containing observed data.}

\item{groups}{A character vector specifying the groups to aggregate.
If NULL, all available groups are used.}

\item{aggregationFlag}{A character string indicating the aggregation method.
Options include "GeometricStdDev", "ArithmeticStdDev", "Percentiles", or "Custom".}

\item{percentiles}{A numeric vector of percentiles to calculate if aggregationFlag is "Percentiles".
Default is c(5, 50, 95).}

\item{groupSuffix}{A character string to append to group names in the aggregated output.
Default is 'aggregated'.}

\item{customFunction}{A custom function for aggregation if aggregationFlag is "Custom".
Default is NULL.}

\item{lloq3Columns}{A character vector specifying columns to check for LLOQ (Lower Limit of Quantification) for 3/3 data points. Default is NULL.}

\item{lloq2Columns}{A character vector specifying columns to check for LLOQ for 2/3 data points. Default is NULL.}
}
\value{
A data.table containing aggregated observed data.
}
\description{
This function aggregates observed data based on specified groups and an aggregation method.
It allows for different aggregation techniques, including geometric and arithmetic standard deviations,
percentiles, or a user-defined custom function.
}
\details{
The function also checks for values below the Lower Limit of Quantification (LLOQ) and adjusts the
aggregated results accordingly.

For custom functions, the `lloq3Columns` and `lloq2Columns` parameters are required to specify
which columns should be checked against LLOQ thresholds. For other aggregation methods,
these parameters can be set to NULL, as the function will handle LLOQ checks internally based
on the aggregation type.

A custom function should take a numeric vector `y` as input and return a list containing:
- `yValues`: The aggregated value (e.g., mean).
- `yMin`: The lower value of the aggregated data, (e.g. mean - sd).
- `yMax`: The upper value of the aggregated data, (e.g. mean + sd).
- `yErrorType`: A string indicating the type of error associated with the aggregation,
it is used in plot legends and captions.
It must be a concatenation of the descriptor of yValues and the descriptor of yMin - yMax range
separated by "|" (e.g., "mean | standard deviation" or "median | 5th - 95th percentile").
}
