% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{aggregatePointEstimate}
\alias{aggregatePointEstimate}
\title{Aggregate Point Estimates}
\usage{
aggregatePointEstimate(
  pkParameterFiltered,
  onePlotConfig,
  ratioMode,
  nBootstrap,
  confLevel,
  aggregationFun,
  aggregationFlag
)
}
\arguments{
\item{pkParameterFiltered}{A list of data tables containing filtered PK parameter data.}

\item{onePlotConfig}{Configuration for the plot.}

\item{ratioMode}{A character string indicating the mode for ratio calculations.}

\item{nBootstrap}{Integer specifying the number of bootstrap samples.}

\item{confLevel}{Numeric value for the confidence level of the intervals.}

\item{aggregationFun}{Function used for aggregation.}

\item{aggregationFlag}{Optional aggregation method.}
}
\value{
A data table containing the aggregated point estimates and confidence intervals.
}
\description{
Aggregates point estimates from the filtered PK parameter data based on the specified ratio mode.
}
\keyword{internal}
