% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aggregation.R
\name{aggregateRatiosByBootstrapping}
\alias{aggregateRatiosByBootstrapping}
\title{Aggregate Ratios by Bootstrapping}
\usage{
aggregateRatiosByBootstrapping(
  nBootstrap,
  value,
  valueReference,
  aggregationFun,
  confLevel = NULL,
  sampleSize = NULL
)
}
\arguments{
\item{nBootstrap}{An integer specifying the number of bootstrap samples to generate.}

\item{value}{A numeric vector containing the values to aggregate.}

\item{valueReference}{A numeric vector containing the reference values for aggregation.}

\item{confLevel}{A numeric value indicating the confidence level for the confidence intervals (optional).}

\item{sampleSize}{number of samples to draw per bootstrap}

\item{aggregationFunction}{A function to aggregate the bootstrapped samples.}

\item{percentiles}{additional parameter passed to aggregationFunction}
}
\value{
A data frame containing the aggregated ratios and their confidence intervals.
}
\description{
This function calculates statistics the results based on bootstrapping.
}
\keyword{internal}
