% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{alignColumnNames}
\alias{alignColumnNames}
\title{Align column names of a data.table with existing data}
\usage{
alignColumnNames(dt, existingData)
}
\arguments{
\item{dt}{A data.table containing the new data to be written.}

\item{existingData}{A data.table containing the existing data in the specified sheet.}
}
\value{
A data.table with aligned column names.
}
\description{
This function aligns the column names of a new data.table with those of the existing data.
It ensures that the new data has the same column names as the existing data, handling
potential ambiguities.
}
\keyword{internal}
