% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reporting.R
\name{appendDocToTemplate}
\alias{appendDocToTemplate}
\title{Place the content of the newly create word document to the template document.}
\usage{
appendDocToTemplate(wordConversionTemplate, docReportPath)
}
\arguments{
\item{wordConversionTemplate}{path of template (should contain a bookmark REPORTBLOCK
to mark the place for inclusion)}

\item{docReportPath}{path of word document to append}
}
\description{
cursor is set to Bookmark, file will be overwritten
}
