% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\name{BIOMETRICUNITS}
\alias{BIOMETRICUNITS}
\title{enumeration biometric units}
\usage{
BIOMETRICUNITS
}
\description{
enumeration biometric units
}
\seealso{
Other enumerations:
\code{\link{DATACLASS}},
\code{\link{EXPORTDIR}},
\code{\link{TIMERANGE}}
}
\concept{enumerations}
