% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aggregation.R
\name{calculateAggregationWithCIBYGroup}
\alias{calculateAggregationWithCIBYGroup}
\title{Calculate Aggregation and Confidence Interval by Group}
\usage{
calculateAggregationWithCIBYGroup(
  dt,
  aggregationFun,
  confLevel = 0.9,
  identifier,
  nBootstrap = 100,
  valueColumn = "value",
  direction = "y"
)
}
\arguments{
\item{dt}{A data.table containing the data. It must include the columns specified in
`identifier` and the `valueColumn`.}

\item{aggregationFun}{A function to calculate the aggregation (e.g., geometric mean).
This function should accept a numeric vector and return a single numeric value.}

\item{confLevel}{A numeric value representing the confidence level for the confidence interval.
The default value is 0.9, corresponding to a 90% confidence interval. Must be between 0 and 1.}

\item{identifier}{A character vector of column names in the data.table to group by.
The function will calculate the aggregation and confidence interval for each unique combination
of these identifiers.}

\item{nBootstrap}{An integer specifying the number of bootstrap samples to use. The default is 100.}

\item{valueColumn}{The name of the column containing the values to aggregate. The default is 'value'.}

\item{direction}{A character string indicating the direction of the results. It can be either
'y' (default) or 'x'. This affects how the results are named in the output data.table.}
}
\value{
A data.table containing the following columns:
  - `yValues`: The estimated value from the aggregation function.
  - `yMin`: The lower bound of the confidence interval.
  - `yMax`: The upper bound of the confidence interval.
  - `seed`: The seed used for bootstrapping, derived from the identifiers.
  - `yErrorType`: A descriptive string indicating the aggregation function and confidence interval bounds.
}
\description{
This function calculates a specified aggregation function (e.g., geometric mean)
and the associated confidence interval using bootstrapping, grouped by specified identifiers.
The function allows for flexible definitions of the aggregation function and can handle
different value columns and output formats based on the specified direction.
A unique seed for bootstrapping is generated for each group of identifiers to ensure
reproducibility of the bootstrap samples.
}
\examples{
\dontrun{
# Example usage
pkParameterDT <- data.table(
  pkParameter = c("A", "A", "B", "B"),
  outputPathId = c(1, 1, 2, 2),
  scenario = c("S1", "S2", "S1", "S2"),
  value = c(10, 20, 15, 25)
)

result <- calculateAggregationWithCIBYGroup(
  dt = pkParameterDT,
  aggregationFun = function(y) exp(mean(log(y[y > 0]))),
  confLevel = 0.9,
  identifier = c("pkParameter", "outputPathId", "scenario"),
  nBootstrap = 100,
  valueColumn = "value",
  direction = "y"
)
}

}
