% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{calculateAndLoadPKParameter}
\alias{calculateAndLoadPKParameter}
\title{Calculate and load PK Parameters}
\usage{
calculateAndLoadPKParameter(
  projectConfiguration,
  scenarioResults,
  pkParameterSheets,
  withRecalculation = TRUE
)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration, including output folder and PK parameter file.}

\item{scenarioResults}{A list of scenario results containing simulation data.}

\item{pkParameterSheets}{A vector with name of the sheets in the PK parameter file to read.}

\item{withRecalculation}{A boolean indicating whether to recalculate PK parameters if results already exist.}
}
\value{
A data.table containing combined PK analyses for all scenarios.
}
\description{
This function calculates pharmacokinetic (PK) parameters for given scenarios based on project configuration and results.
}
