% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{calculatePKParameterForScenarios}
\alias{calculatePKParameterForScenarios}
\title{Calculate  Pharmacokinetic (PK) Parameters}
\usage{
calculatePKParameterForScenarios(projectConfiguration, scenarioResults)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration settings,
including the output folder path and the PK parameter file.}

\item{scenarioResults}{A named list of scenario results, each containing
simulation data for PK analysis.}
}
\value{
This function is called for its side effects and does not return a value.
}
\description{
This function calculates pharmacokinetic (PK) parameters for specified scenarios
based on project configuration and simulation results. It generates output files
for each scenario in the designated output folder.
}
\seealso{
Other scenario management:
\code{\link[=createScenariosWrapped]{createScenariosWrapped()}},
\code{\link[=loadPKParameter]{loadPKParameter()}},
\code{\link[=loadScenarioResultsToFramework]{loadScenarioResultsToFramework()}},
\code{\link[=runAndSaveScenarios]{runAndSaveScenarios()}},
\code{\link[=runOrLoadScenarios]{runOrLoadScenarios()}}
}
\concept{scenario management}
