% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{calculatePlotDataNone}
\alias{calculatePlotDataNone}
\title{Calculate Plot Data for No Ratios}
\usage{
calculatePlotDataNone(
  pkParameterDT,
  asCI,
  aggregationFun,
  nBootstrap,
  confLevel
)
}
\arguments{
\item{pkParameterDT}{Data table containing PK parameter data.}

\item{asCI}{Logical indicating if confidence intervals should be calculated.}

\item{aggregationFun}{Function used for aggregation.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}
}
\value{
Data table with calculated plot data for no ratios.
}
\description{
This function calculates the plot data when no ratios are used.
}
\keyword{internal}
