% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{calculatePlotDataRatioOfPopulation}
\alias{calculatePlotDataRatioOfPopulation}
\title{Calculate Plot Data for Ratio of Population}
\usage{
calculatePlotDataRatioOfPopulation(
  onePlotConfig,
  pkParameterDT,
  asCI,
  aggregationFun,
  aggregationFlag,
  nBootstrap,
  confLevel
)
}
\arguments{
\item{onePlotConfig}{Configuration for the plot.}

\item{pkParameterDT}{Data table containing PK parameter data.}

\item{asCI}{Logical indicating if confidence intervals should be calculated.}

\item{aggregationFun}{Function used for aggregation.}

\item{aggregationFlag}{Optional aggregation method.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}
}
\value{
Data table with calculated plot data for ratio of population.
}
\description{
This function calculates the plot data for the ratio of population.
}
\keyword{internal}
