% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{calculateRatiosByAggregation}
\alias{calculateRatiosByAggregation}
\title{Calculate Ratios for Workflow}
\usage{
calculateRatiosByAggregation(configList, pkParameterDT, aggregationFun)
}
\arguments{
\item{configList}{A data.table containing the configuration for the current workflow.}

\item{pkParameterDT}{A data.table with PK parameters.}

\item{coefficientOfVariation}{Coefficient of variation for simulations.}

\item{nObservations}{Number of observations to simulate.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}

\item{statFun}{Function to compute statistics.}

\item{seed}{Random seed for reproducibility.}
}
\value{
A data.table containing the bootstrapped statistics.
}
\description{
Calculates ratios for the specified workflow based on the provided configuration.
}
\keyword{internal}
