% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{calculateRatiosByBoostrapping}
\alias{calculateRatiosByBoostrapping}
\title{Calculate Ratios for Workflow}
\usage{
calculateRatiosByBoostrapping(
  configList,
  pkParameterDT,
  coefficientOfVariation,
  nObservations,
  nBootstrap,
  confLevel,
  statFun,
  seed
)
}
\arguments{
\item{configList}{A data.table containing the configuration for the current workflow, including
reference and scenario details.}

\item{pkParameterDT}{A data.table with PK parameters that will be used for ratio calculations.}

\item{coefficientOfVariation}{A numeric value representing the coefficient of variation for simulations.}

\item{nObservations}{An integer representing the number of observations to simulate.}

\item{nBootstrap}{An integer specifying the number of bootstrap samples to generate.}

\item{confLevel}{A numeric value representing the confidence level for intervals.}

\item{statFun}{A function to compute statistics from the bootstrap samples.}

\item{seed}{An integer for setting the random seed to ensure reproducibility of results.}
}
\value{
A data.table containing the calculated bootstrapped statistics, including the ratio and
        confidence intervals.
}
\description{
Calculates ratios for the specified workflow based on the provided configuration. This function
handles the merging of reference and scenario data and computes the ratios using the specified
statistical method.
}
\keyword{internal}
