% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{calculateUnitFactors}
\alias{calculateUnitFactors}
\title{Calculate Unit Factors and Adjust Observed Data}
\usage{
calculateUnitFactors(dataObserved, dtOutputs, scenarioList)
}
\arguments{
\item{dataObserved}{A data.table containing observed data. It must include columns for 'scenario', 'outputPathId', 'yUnit', and 'yValues'.}

\item{dtOutputs}{A data.table containing output path information, including 'outputPathId' and 'displayUnit'.}

\item{scenarioList}{A list of scenarios, each containing simulation details relevant to the observed data.}
}
\value{
A data.table containing the updated observed data with adjusted yValues, unit factors, and units.
}
\description{
This function calculates unit conversion factors for observed data and adjusts the observed data accordingly.
}
\details{
The function merges unit conversion factors into the observed data, adjusts the yValues based on the calculated data factors,
and computes the unit factor for time.
}
\keyword{internal}
