% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{captureLog}
\alias{captureLog}
\title{Catch messages, warnings, and errors and route them to the log}
\usage{
captureLog(expr, finallyExpression = invisible())
}
\arguments{
\item{expr}{The expression to evaluate.}

\item{finallyExpression}{The expression to evaluate finally.}
}
\description{
Wraps `ospsuite.utils::logCatch` and preserves an optional `finally` expression
for backwards compatibility.
}
\seealso{
Other log file management:
\code{\link[=addMessageToLog]{addMessageToLog()}},
\code{\link[=initLogfunction]{initLogfunction()}},
\code{\link[=saveSessionInfo]{saveSessionInfo()}},
\code{\link[=setShowLogMessages]{setShowLogMessages()}},
\code{\link[=writeTableToLog]{writeTableToLog()}}
}
\concept{log file management}
