% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{checkAlphanumericalIDs}
\alias{checkAlphanumericalIDs}
\title{check if a vector of ids is alphanumerical}
\usage{
checkAlphanumericalIDs(vector, failureMessage)
}
\arguments{
\item{vector}{vector of ids}

\item{failureMessage}{message to plot on failure}
}
\description{
check if a vector of ids is alphanumerical
}
