% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{checkAndAdjustYlimits}
\alias{checkAndAdjustYlimits}
\title{Check and adjust Y limits for plots}
\usage{
checkAndAdjustYlimits(plotData, yScale, timeRangeFilter, plotType)
}
\arguments{
\item{plotData}{Object of class `PlotDataTimeProfile`.}

\item{yScale}{Y scale of the plot.}

\item{timeRangeFilter}{Name of the time range filter.}

\item{plotType}{Type of the plot.}
}
\value{
Numeric vector of Y limits.
}
\description{
Check and adjust Y limits for plots
}
\keyword{internal}
