% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{checkFileNameValidity}
\alias{checkFileNameValidity}
\title{Check File Name Validity}
\usage{
checkFileNameValidity(fileName, isDataSet = FALSE)
}
\arguments{
\item{fileName}{A character string representing the filename to be checked.}

\item{isDataSet}{A logical value indicating whether the filename is for a dataset.
If TRUE, only "txt" is allowed as a file extension with a maximum
length of 32 characters. If FALSE, "xml", "txt", and "csv" are allowed
with a maximum length of 64 characters.}
}
\value{
A logical value: TRUE if the filename is valid, FALSE otherwise.
        If invalid, a warning is issued detailing the specific reason(s).
}
\description{
This function checks if a given filename is valid based on specified criteria,
including allowed file extensions, maximum length, and that it does not start
with a numerical digit.
}
\examples{
checkFileNameValidity("valid_file.txt")
checkFileNameValidity("1invalid_file.txt")
checkFileNameValidity("too_long_filename_that_exceeds_the_limit.txt", isDataSet = TRUE)

}
