% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{checkFileNameValidityForDirectory}
\alias{checkFileNameValidityForDirectory}
\title{Check on directory for valid file names}
\usage{
checkFileNameValidityForDirectory(directoryPath)
}
\arguments{
\item{directoryPath}{path to the directory where the electronic package is filed}
}
\description{
Checks if all files and all files of data sets of a directory fulfill conditions for file names for electronic package
}
