% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{checkLLOQ}
\alias{checkLLOQ}
\title{Sets values which do not match the LLOQ criteria to NA}
\usage{
checkLLOQ(
  aggregatedData,
  lloqCheckColumns2of3,
  lloqCheckColumns1of2,
  aggregationFlag
)
}
\arguments{
\item{aggregatedData}{A `data.table` with aggregated data.}

\item{lloqCheckColumns2of3}{A character vector specifying columns to check for LLOQ (Lower Limit of Quantification) for 1/3 data points.
Default is NULL, is used only for `aggregationFlag` "Custom".}

\item{lloqCheckColumns1of2}{A character vector specifying columns to check for LLOQ for 2/3 data points.
Default is NULL, is used only for `aggregationFlag` "Custom".}

\item{aggregationFlag}{A character string indicating the aggregation method.
Options include "GeometricStdDev", "ArithmeticStdDev", "Percentiles", or "Custom".}
}
\value{
Updated aggregated data `data.table`.
}
\description{
Sets values which do not match the LLOQ criteria to NA
}
\keyword{internal}
