% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{checkPrecision}
\alias{checkPrecision}
\title{Check Precision of xValues}
\usage{
checkPrecision(dt)
}
\arguments{
\item{dt}{A data.table containing the columns xValues, xMin, and xMax.}
}
\value{
A data.table with updated xMin and xMax values, and an additional column
        estimatedN if the required precision is not reached.
}
\description{
This function calculates the precision of xValues in a data.table based on the provided
xMin and xMax columns. If the precision is below a specified threshold, it sets xMin and
xMax to NA and calculates an estimated N value. It also logs a message if the required
precision is not reached.
}
\details{
The precision is calculated as xValues / (xMax - xMin). The precision threshold
         is retrieved from the options with a default value of 0.01. If any xMin values
         remain, a message is logged indicating that the required precision was not reached.
}
\keyword{internal}
