% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{checkScenarioNameValidity}
\alias{checkScenarioNameValidity}
\title{Checks if scenarioName + extension .xml fulfills conditions for file names for electronic package}
\usage{
checkScenarioNameValidity(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\description{
Checks if scenarioName + extension .xml fulfills conditions for file names for electronic package
}
