% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{checkSheetExists}
\alias{checkSheetExists}
\title{Check if the specified sheet exists in the workbook}
\usage{
checkSheetExists(wb, sheetName)
}
\arguments{
\item{wb}{A workbook object created by `openxlsx::loadWorkbook()`.}

\item{sheetName}{A character string specifying the name of the sheet to check.}
}
\value{
NULL. If the sheet does not exist, an error is thrown.
}
\description{
This function checks if the given sheet name exists in the workbook.
}
\keyword{internal}
