% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{cleanCharacterColumns}
\alias{cleanCharacterColumns}
\title{Clean character columns by trimming whitespace and replacing curly quotes}
\usage{
cleanCharacterColumns(dt, emptyAsNA)
}
\arguments{
\item{dt}{A `data.table` to process.}

\item{emptyAsNA}{A logical value. If TRUE, empty strings in character columns are converted to NA.
If FALSE NA is returned as empty string.}
}
\value{
A `data.table` with cleaned character columns.
}
\description{
This helper function trims whitespace from character columns and replaces curly quotes with straight quotes.
}
\keyword{internal}
