% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{collectPlotData}
\alias{collectPlotData}
\title{COllects all data neede for this plots}
\usage{
collectPlotData(projectConfiguration, panelConfig, observedData, facetType)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{observedData}{}
}
\value{
`data.table` with plotData and attribute metaData
}
\description{
add colums for facet
}
