% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{concatWithAnd}
\alias{concatWithAnd}
\title{Concatenate Text Vector with Commas and "and"}
\usage{
concatWithAnd(textVector)
}
\arguments{
\item{textVector}{A character vector of text strings to be concatenated.}
}
\value{
A single character string representing the concatenated text.
}
\description{
This function takes a vector of text strings and concatenates them into a single string.
For vectors with more than two elements, it separates them with commas and uses "and"
before the last element.
}
\keyword{internal}
