% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{convertAndShiftTimeUnits}
\alias{convertAndShiftTimeUnits}
\title{Convert and Shift Time Units}
\usage{
convertAndShiftTimeUnits(timeprofile, targetTimeUnit, timeOffset = 0)
}
\arguments{
\item{timeprofile}{A data.table containing time profile data.}

\item{targetTimeUnit}{The target time unit for conversion.}

\item{timeOffset}{An optional time offset to be applied after conversion (default is 0).}
}
\value{
A data.table with the time values converted and shifted.
}
\description{
This function converts time values to a target unit and applies an optional time offset.
}
