% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{convertColumnsToNumeric}
\alias{convertColumnsToNumeric}
\title{Convert convertible columns to numeric}
\usage{
convertColumnsToNumeric(dt, alwaysCharacter)
}
\arguments{
\item{dt}{A `data.table` to process.}

\item{alwaysCharacter}{A character vector of column names to remain as character.}
}
\value{
A `data.table` with numeric conversions applied where appropriate.
}
\description{
This helper function converts columns in a data.table to numeric if possible.
}
\keyword{internal}
