% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertDataByDictionary}
\alias{convertDataByDictionary}
\title{Convert data by dictionary}
\usage{
convertDataByDictionary(
  data,
  dataFilter,
  dict,
  dictionaryName,
  debugMode = FALSE
)
}
\arguments{
\item{data}{The data to be converted.}

\item{dataFilter}{The filter to be applied to the data.}

\item{dict}{The dictionary to be used for conversion.}

\item{dictionaryName}{The name of the dictionary.}

\item{debugMode}{If TRUE, enables debug mode with additional warnings. Default is FALSE.}
}
\value{
A `data.table` containing the converted data.
}
\description{
This function converts the data based on the provided dictionary and filters.
}
\keyword{internal}
