% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertIdentifierColumns}
\alias{convertIdentifierColumns}
\title{Update identifier columns in a data.table}
\usage{
convertIdentifierColumns(dt, identifierCols)
}
\arguments{
\item{dt}{The input data.table}

\item{identifierCols}{A character vector specifying the columns to be updated}
}
\value{
The updated data.table
}
\description{
This function updates the specified identifier columns in a data.table by replacing commas with underscores.
It also checks for the presence of commas in the columns and generates a warning message if found.
}
\examples{
dt <- data.table(col1 = c("a,b,c", "d,e,f"), col2 = c("x,y,z", "1,2,3"))
identifierCols <- c("col1", "col2")
updatedDt <- updateIdentifierColumns(dt, identifierCols)
}
