% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{convertSheet}
\alias{convertSheet}
\title{Convert Sheet Data to Data Frame}
\usage{
convertSheet(sheetsData)
}
\arguments{
\item{sheetsData}{A list containing the sheet data in a structured format.}
}
\value{
A list of data frames representing the sheets ready for Excel writing.
}
\description{
This function converts the sheet data from a JSON structure into a data frame format suitable for Excel.
}
\keyword{internal}
