% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{copyConfigSheet}
\alias{copyConfigSheet}
\title{Copy a configuration sheet from source to destination}
\usage{
copyConfigSheet(
  sourceSheetName,
  destinationSheetName,
  sourceFile,
  destinationFile
)
}
\arguments{
\item{sourceSheetName}{A character string specifying the name of the sheet to copy.}

\item{destinationSheetName}{A character string specifying the name of the new sheet.}

\item{sourceFile}{A character string specifying the path to the source Excel file.}

\item{destinationFile}{A character string specifying the path to the destination Excel file.}
}
\value{
Invisibly returns NULL. The function performs a side effect (copying a sheet between workbooks).
}
\description{
This function copies a specified sheet from a source Excel file to a destination Excel file,
including the data and styles.
}
\seealso{
Other function to read from and write to xlsx: 
\code{\link{separateAndTrimColumn}()},
\code{\link{setHeadersToLowerCase}()},
\code{\link{splitInputs}()},
\code{\link{synchronizeScenariosOutputsWithPlots}()},
\code{\link{synchronizeScenariosWithPlots}()},
\code{\link{xlsxAddDataUsingTemplate}()},
\code{\link{xlsxAddSheet}()},
\code{\link{xlsxCloneAndSet}()},
\code{\link{xlsxReadData}()},
\code{\link{xlsxWriteData}()}
}
\concept{function to read from and write to xlsx}
