% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{copyFilesToAnalysisProgram}
\alias{copyFilesToAnalysisProgram}
\title{Copy Files to Analysis Program}
\usage{
copyFilesToAnalysisProgram(configTableToExport, dirAnalysisProgram)
}
\arguments{
\item{configTableToExport}{A data.table containing the configuration table to export.}

\item{dirAnalysisProgram}{The directory where files should be copied.}
}
\description{
This function copies valid files to the designated analysis program directory#'
}
\keyword{internal}
