% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{createAnalysisDirectory}
\alias{createAnalysisDirectory}
\title{Create Analysis Directory}
\usage{
createAnalysisDirectory(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project-specific configurations.}
}
\value{
The path to the created analysis directory.
}
\description{
This function creates the analysis directory if it does not already exist.
}
\keyword{internal}
