% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPkBoxplot.R
\name{createBaseBoxWhisker}
\alias{createBaseBoxWhisker}
\title{Create a Box and Whisker Plot Object}
\usage{
createBaseBoxWhisker(
  plotDataPk,
  yScale,
  asRatio,
  colorVector,
  onePlotConfig,
  ...
)
}
\arguments{
\item{plotDataPk}{A data frame containing the plot data for a specific PK parameter.}

\item{yScale}{A character string indicating the scale for the y-axis (e.g., "linear", "log").}

\item{colorVector}{A vector of colors to be used for filling the plot.}

\item{onePlotConfig}{A list containing configuration settings for the plot.}

\item{...}{Additional arguments to be passed to the plotting functions.}

\item{ratioMode}{A character string indicating the mode of ratio to be used for plotting
(e.g., "individualRatios", "ratioOfPopulation", "none").}
}
\value{
A ggplot object representing the box and whisker plot.
}
\description{
This helper function generates a box and whisker plot object based on the provided
plot data and configuration. It supports different ratio modes and applies the
appropriate aesthetics and scales to the plot.
}
\keyword{internal}
