% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{createDefaultProjectConfiguration.wrapped}
\alias{createDefaultProjectConfiguration.wrapped}
\title{Create a default `ProjectConfiguration`}
\usage{
createDefaultProjectConfiguration.wrapped(path)
}
\arguments{
\item{path}{Full path of an XLS/XLSX file}
}
\value{
Object of type ProjectConfiguration
}
\description{
wrap of (esqlabsR::createDefaultProjectConfiguration())
}
