% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{createExportLogFile}
\alias{createExportLogFile}
\title{Create Export Log File}
\usage{
createExportLogFile(configTableToExport, projectConfiguration)
}
\arguments{
\item{configTableToExport}{A data.table containing the configuration table to export.}

\item{projectConfiguration}{A list containing project-specific configurations.}
}
\value{
The path to the created export log file.
}
\description{
This function creates the export log file based on the configuration table.
}
\keyword{internal}
