% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{createNewConfig}
\alias{createNewConfig}
\title{Create New Configuration for Time Profile Plots}
\usage{
createNewConfig(scenarios, dataObserved)
}
\arguments{
\item{scenarios}{A data.table containing scenario definitions.}

\item{dataObserved}{Optional. A data object containing observed data, if available.}
}
\value{
A data.table containing the new configuration for time profile plots.
}
\description{
This internal function generates a new configuration data table for time profile plots based on the provided scenarios and project configuration.
}
\keyword{internal}
