% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{createPanelPlotsForPlotName}
\alias{createPanelPlotsForPlotName}
\title{Generate one plot as facet panels}
\usage{
createPanelPlotsForPlotName(
  projectConfiguration,
  onePlotConfig,
  dataObserved,
  nFacetColumns,
  rmdContainer,
  facetAspectRatio,
  aggregationFun
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{rmdContainer}{Object of class `RmdContainer`.}

\item{aggregationFun}{Function to aggregate simulated data.}

\item{observedData}{`data.table` format corresponds to tables  generated by \code{\link{readDataByDictionary}}}
}
\value{
Object of class `rmdContainer` with added figures.
}
\description{
Generate one plot as facet panels
}
