% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPkBoxplot.R
\name{createPKBoxPlotForPlotName}
\alias{createPKBoxPlotForPlotName}
\title{Read PK Ratio Configuration Table}
\usage{
createPKBoxPlotForPlotName(onePlotConfig, rmdContainer, pkParameterDT)
}
\arguments{
\item{pkParameterDT}{A data.table with PK parameters.}

\item{projectConfiguration}{A ProjectConfiguration object.}

\item{sheetName}{Name of the sheet to read from.}
}
\value{
A validated configuration table as a data.table.
}
\description{
Reads and validates the PK ratio configuration table from the specified sheet.
}
\keyword{internal}
