% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKRatioForestPlot.R
\name{createPKDDIRatioForPlotName}
\alias{createPKDDIRatioForPlotName}
\title{Create PK DDI Ratio Plot}
\usage{
createPKDDIRatioForPlotName(
  onePlotConfig,
  rmdContainer,
  pkParameterDT,
  coefficientOfVariation,
  nObservationDefault,
  nBootstrap,
  confLevel,
  statFun,
  seed,
  vlineIntercept,
  scaleVactors,
  digitsToRound,
  digitsToShow,
  xlabel
)
}
\arguments{
\item{onePlotConfig}{A data.table containing the plot configuration.}

\item{rmdContainer}{An RmdContainer object.}

\item{pkParameterDT}{A data.table with PK parameters.}

\item{coefficientOfVariation}{Coefficient of variation for simulations.}

\item{nObservationDefault}{Default number of observations.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}

\item{statFun}{Function to compute statistics.}

\item{seed}{Random seed for reproducibility.}

\item{vlineIntercept}{Intercept for vertical line on plots.}

\item{scaleVactors}{A list of aesthetic parameters for plots.}

\item{digitsToRound}{Number of digits to round in output.}

\item{digitsToShow}{Number of digits to display.}

\item{xlabel}{Label for x-axis.}
}
\value{
An updated RmdContainer object with the added plot.
}
\description{
Creates a plot for the PK DDI ratio based on the provided configuration.
}
