% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{createPkPlotForPlotName}
\alias{createPkPlotForPlotName}
\title{Create PK Ratio Plot}
\usage{
createPkPlotForPlotName(
  onePlotConfig,
  rmdContainer,
  pkParameterDT,
  pkParameterObserved,
  ratioCalculationMethod,
  ratioCalculationInputs,
  vlineIntercept,
  scaleVectors,
  digitsToRound,
  digitsToShow,
  xlabel,
  tableLabels
)
}
\arguments{
\item{onePlotConfig}{A data.table containing the plot configuration, including details such as
scenario names and parameters to be plotted.}

\item{rmdContainer}{An RmdContainer object that will hold the generated plot and related information.}

\item{pkParameterDT}{A data.table with PK parameters, including the necessary columns for analysis.}

\item{pkParameterObserved}{A data.table with observed PK parameters. This parameter is optional.}

\item{ratioCalculationMethod}{A string specifying the method used for calculating ratios (e.g., 'byBootsTrapping').}

\item{ratioCalculationInputs}{A list containing inputs specific to the method used for calculating ratios.}

\item{vlineIntercept}{A numeric value for the intercept of the vertical line on plots.}

\item{scaleVectors}{A list of aesthetic parameters for the plots, including colors and shapes.}

\item{digitsToRound}{An integer specifying the number of digits to round in output.}

\item{digitsToShow}{An integer specifying the number of digits to display in plot annotations.}

\item{xlabel}{A string representing the label for the x-axis in the plots.}

\item{tableLabels}{A list of labels for tables generated in the plot.}
}
\value{
An updated RmdContainer object with the added plot and related information.
}
\description{
Creates a plot for the PK ratio based on the provided configuration. This function handles the
preparation of data, plotting, and exporting of the generated plot.
}
