% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{createReviewerAidTable}
\alias{createReviewerAidTable}
\title{Create Reviewer Aid Table}
\usage{
createReviewerAidTable(configTableToExport, projectConfiguration)
}
\arguments{
\item{configTableToExport}{A data.table containing the configuration table to export.}

\item{projectConfiguration}{A ProjectConfiguration object containing project-specific configurations.}
}
\description{
This function creates a reviewer aid table based on the configuration table.
}
\keyword{internal}
