% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{createScenariosWrapped}
\alias{createScenariosWrapped}
\title{Create Scenario objects from `ScenarioConfiguration` objects}
\usage{
createScenariosWrapped(projectConfiguration, scenarioNames = NULL)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and file names}

\item{scenarioNames}{Names of the scenarios that are defined in the excel file.
If NULL (default), all scenarios specified in the excel file will be created.}
}
\value{
Named list of Scenario objects.
}
\description{
wrap of `esqlabsR::createDefaultProjectConfiguration()` with `esqlabsR::createScenarios()` as input
}
\seealso{
Other scenario management:
\code{\link[=calculatePKParameterForScenarios]{calculatePKParameterForScenarios()}},
\code{\link[=loadPKParameter]{loadPKParameter()}},
\code{\link[=loadScenarioResultsToFramework]{loadScenarioResultsToFramework()}},
\code{\link[=runAndSaveScenarios]{runAndSaveScenarios()}},
\code{\link[=runOrLoadScenarios]{runOrLoadScenarios()}}
}
\concept{scenario management}
