% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{createSensitivityPlotForPlotName}
\alias{createSensitivityPlotForPlotName}
\title{Create Sensitivity Plot for a Given Plot Name}
\usage{
createSensitivityPlotForPlotName(
  onePlotConfig,
  rmdPlotManager,
  projectConfiguration,
  scenarioList
)
}
\arguments{
\item{onePlotConfig}{A data frame containing the configuration for a single plot.}

\item{rmdPlotManager}{An RmdPlotManager object to which the plot will be added.}

\item{projectConfiguration}{A list containing project configuration settings.}

\item{scenarioList}{A list of scenarios for which the sensitivity analysis will be performed.}
}
\value{
The updated RmdPlotManager object with the added plot.
}
\description{
This function creates a sensitivity plot based on the provided plot configuration and adds it to the Rmd container.
}
\keyword{internal}
