% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{createSummaryOverviewTable}
\alias{createSummaryOverviewTable}
\title{Create Summary Overview Table}
\usage{
createSummaryOverviewTable(configTableToExport, projectConfiguration)
}
\arguments{
\item{configTableToExport}{A data.table containing the configuration table to export.}

\item{projectConfiguration}{A ProjectConfiguration object containing project-specific configurations.}
}
\description{
This function creates a summary overview table based on the configuration table.
}
\keyword{internal}
