% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\name{DATACLASS}
\alias{DATACLASS}
\title{enumeration keys for `DataClass`}
\usage{
DATACLASS
}
\description{
enumeration keys for `DataClass`
}
\seealso{
Other enumerations:
\code{\link{BIOMETRICUNITS}},
\code{\link{EXPORTDIR}},
\code{\link{TIMERANGE}}
}
\concept{enumerations}
