% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_enum.R
\docType{data}
\name{DATATYPE}
\alias{DATATYPE}
\title{enumeration keys for data type}
\format{
An object of class \code{list} of length 2.
}
\usage{
DATATYPE
}
\description{
enumeration keys for data type
}
\keyword{datasets}
