% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{determineFacetColumns}
\alias{determineFacetColumns}
\title{Determine Facet Columns}
\usage{
determineFacetColumns(dtCaption, nFacetColumns, facetType, plotName)
}
\arguments{
\item{dtCaption}{A data table containing the caption data.}

\item{nFacetColumns}{The number of facet columns.}

\item{facetType}{The type of facet being used.}

\item{plotName}{The name of the plot.}
}
\value{
The number of facet columns to be used.
}
\description{
This function determines the number of facet columns based on the caption data and facet type.
}
\keyword{internal}
