% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{.addDimensionColumns}
\alias{.addDimensionColumns}
\title{Add xDimension and yDimension columns required by ospsuite::plotTimeProfile.}
\usage{
.addDimensionColumns(dt)
}
\arguments{
\item{dt}{A data.table with at least `xUnit` and `yUnit` columns.}
}
\value{
The same data.table with `xDimension` and `yDimension` added in-place.
}
\description{
Add xDimension and yDimension columns required by ospsuite::plotTimeProfile.
}
\keyword{internal}
