% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.buildVirtualTwinPopulation}
\alias{.buildVirtualTwinPopulation}
\title{Build Virtual Twin Population}
\usage{
.buildVirtualTwinPopulation(projectConfiguration, params, dPop)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration details.}

\item{params}{A list of parameters for the virtual twin population.}

\item{dPop}{A data.table containing population data.}
}
\value{
A data.table representing the virtual twin population.
}
\description{
This function builds an virtual twin population table based on the provided parameters and population data.
}
\keyword{internal}
