% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.cleanUpSheetList}
\alias{.cleanUpSheetList}
\title{Clean Up Sheet List}
\usage{
.cleanUpSheetList(sheets)
}
\arguments{
\item{sheets}{A character vector of sheet names.}
}
\value{
A cleaned character vector of sheet names.
}
\description{
This function sanitizes a list of sheet names by removing duplicates and whitespace.
}
\keyword{internal}
