% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{.constructMetDataForTimeProfile}
\alias{.constructMetDataForTimeProfile}
\title{construct metaData out of Data}
\usage{
.constructMetDataForTimeProfile(plotData)
}
\arguments{
\item{plotData}{}
}
\value{
list with metaData
}
\description{
construct metaData out of Data
}
