% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.convertBiomForIndStatics}
\alias{.convertBiomForIndStatics}
\title{Convert Biometrics for Individual Statistics}
\usage{
.convertBiomForIndStatics(biomForInd, sim)
}
\arguments{
\item{biomForInd}{A data.table containing biometrics for an individual.}

\item{sim}{A simulation object.}
}
\value{
A data.table with converted biometrics.
}
\description{
This function converts biometrics data for individual statistics into a specific format.
}
\keyword{internal}
