% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.createIndividualCharacteristics}
\alias{.createIndividualCharacteristics}
\title{Create Individual Characteristics}
\usage{
.createIndividualCharacteristics(biomForInd)
}
\arguments{
\item{biomForInd}{A data.table containing biometrics for an individual.}
}
\value{
An individual characteristics object.
}
\description{
This function creates individual characteristics from biometrics data.
}
\keyword{internal}
