% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.generatePopulationFiles}
\alias{.generatePopulationFiles}
\title{Generate Population Files}
\usage{
.generatePopulationFiles(
  dtTwinPops,
  params,
  dtIndividualBiometrics,
  projectConfiguration,
  sim
)
}
\arguments{
\item{dtTwinPops}{A data.table containing virtual twin population data.}

\item{params}{A list of parameters for the virtual twin population.}

\item{dtIndividualBiometrics}{A data.table containing individual biometrics.}

\item{projectConfiguration}{A list containing project configuration details.}

\item{sim}{A simulation object.}
}
\description{
This function generates population files based on individual biometrics and parameters.
}
\keyword{internal}
