% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.getAllParameterForSheets}
\alias{.getAllParameterForSheets}
\title{Get All Parameters for Sheets}
\usage{
.getAllParameterForSheets(projectConfiguration, sheets, paramsXLSpath, sim)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration details.}

\item{sheets}{A character vector of sheet names to read parameters from.}

\item{paramsXLSpath}{A string representing the path to the parameters Excel file.}

\item{sim}{A simulation object.}
}
\value{
A list of parameters for the specified sheets.
}
\description{
This function retrieves parameters from specified sheets in an Excel file.
}
\keyword{internal}
