% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{.getMappingForTimeprofiles}
\alias{.getMappingForTimeprofiles}
\title{Creates mapping for plotData.}
\usage{
.getMappingForTimeprofiles(plotData, metaData, userMapping)
}
\arguments{
\item{plotData}{Data to map.}

\item{metaData}{A list with metadata for plotData.}

\item{userMapping}{Mapping provided by the user; this will update the internal mapping.}
}
\value{
A mapping object for ggplot2.
}
\description{
This function generates a mapping for the plotting based on the provided plot data and metadata.
}
\keyword{internal}
