% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.processIndividual}
\alias{.processIndividual}
\title{Process Individual}
\usage{
.processIndividual(individual, biomForInd, params, projectConfiguration, sim)
}
\arguments{
\item{individual}{An individual object.}

\item{biomForInd}{A data.table containing biometrics for an individual.}

\item{params}{A list of parameters for the virtual twin population.}

\item{projectConfiguration}{A list containing project configuration details.}

\item{sim}{A simulation object.}
}
\value{
A list of results for the individual.
}
\description{
This function processes individual data and generates as results a list of all parameters transferred to population csv
}
\keyword{internal}
