% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.readParameterSheetList}
\alias{.readParameterSheetList}
\title{Read Parameter Sheet List}
\usage{
.readParameterSheetList(projectConfiguration, dtTwinPops, sim)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration details.}

\item{dtTwinPops}{A data.table containing virtual twin population data.}

\item{sim}{A simulation object.}
}
\value{
A list of parameters for the specified sheets.
}
\description{
This function reads parameters from specified sheets in an Excel file.
}
\keyword{internal}
