% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{.savePopulationFile}
\alias{.savePopulationFile}
\title{Save Population File}
\usage{
.savePopulationFile(poptable, populationName, projectConfiguration)
}
\arguments{
\item{poptable}{A data.table representing the population data.}

\item{populationName}{name of poulation.}

\item{projectConfiguration}{A list containing project configuration details.}
}
\description{
This function saves the population data to a CSV file.
}
\keyword{internal}
