% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{.validateAndConvertData}
\alias{.validateAndConvertData}
\title{validates observed data and converts it to appropriate format}
\usage{
.validateAndConvertData(plotData, predictedIsNeeded)
}
\arguments{
\item{plotData}{either data.table with columns 'xValues','yValues','group' or
object of class 'DataCombined'}

\item{predictedIsNeeded}{if TRUE only observed data are returned,
if not column "predicted" exists, predicted values are caluclated}
}
\value{
`data.table` with data to plot
}
\description{
validates observed data and converts it to appropriate format
}
