% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{evaluateRatioByBootstrapping}
\alias{evaluateRatioByBootstrapping}
\title{Evaluate Ratio by Bootstrapping}
\usage{
evaluateRatioByBootstrapping(
  onePlotConfig,
  pkParameterDT,
  aggregationFun,
  aggregationFlag,
  nBootstrap,
  confLevel
)
}
\arguments{
\item{onePlotConfig}{Configuration for the plot.}

\item{pkParameterDT}{Data table containing PK parameter data.}

\item{aggregationFun}{Function used for aggregation.}

\item{aggregationFlag}{Optional aggregation method.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}

\item{nObservations}{Number of samples to draw per bootstrap.}
}
\value{
A data.table containing the evaluated ratios.
}
\description{
Evaluates the ratio using bootstrapping.
}
\keyword{internal}
