% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{excelToListStructure}
\alias{excelToListStructure}
\title{Convert Excel file to List Structure}
\usage{
excelToListStructure(df)
}
\arguments{
\item{df}{A data frame representing the content of an Excel sheet.}
}
\value{
A list structure ready for JSON serialization, containing column names and rows.
}
\description{
This function converts the content of an Excel sheet into a list structure suitable for JSON serialization.
}
\keyword{internal}
