% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_workflow.R
\name{executeAsValidRun}
\alias{executeAsValidRun}
\title{Sets options to the values they should have for a valid run}
\usage{
executeAsValidRun(isValidRun)
}
\arguments{
\item{isValidRun}{if TTRUE}
}
\description{
relevant options are
  - watermark for figures (see vignette(package = 'ospsuite.plots',topic = 'ospsuite_plots'))
  - OSPSuite.RF.skipFailingPlots
}
