% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\name{EXPORTDIR}
\alias{EXPORTDIR}
\title{enumeration keys for exportDirectories}
\usage{
EXPORTDIR
}
\description{
enumeration keys for exportDirectories
}
\seealso{
Other enumerations:
\code{\link{BIOMETRICUNITS}},
\code{\link{DATACLASS}},
\code{\link{TIMERANGE}}
}
\concept{enumerations}
