% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{exportPKWithIOVForEPackage}
\alias{exportPKWithIOVForEPackage}
\title{Export PK Parameters with IOV for Electronic Package}
\usage{
exportPKWithIOVForEPackage(projectConfiguration, pkParameterDT)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project-specific configurations.}

\item{pkParameterDT}{A data.table containing PK parameter data.}
}
\value{
This function does not return a value. It performs file operations to save the exported PK parameters.
}
\description{
This function exports pharmacokinetic (PK) parameters with inter-observation variability (IOV)
for the electronic package, saving them in separate CSV files based on scenarios.
}
\keyword{internal}
