% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{exportTimeProfileDataForEPackage}
\alias{exportTimeProfileDataForEPackage}
\title{Export Time Profile Data for Electronic Package}
\usage{
exportTimeProfileDataForEPackage(projectConfiguration, dataToExport)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project-specific configurations.}

\item{dataToExport}{A data.table containing the data to be exported.}
}
\value{
This function does not return a value. It performs file operations to save the exported data.
}
\description{
This function exports time profile data for the electronic package by merging
the provided data with configuration details and saving it in the specified format.
}
\keyword{internal}
