% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{extendPopulationByUserDefinedParams_RF}
\alias{extendPopulationByUserDefinedParams_RF}
\title{Add user defined variability on parameters to a population.}
\usage{
extendPopulationByUserDefinedParams_RF(
  population,
  parameterPaths,
  meanValues,
  sdValues,
  distributions = Distributions$Normal
)
}
\arguments{
\item{population}{Object of type `Population`}

\item{parameterPaths}{Vector of parameter path for which the variability is to be added.}

\item{meanValues}{Vector of mean values of the parameters. Must have the same
length as `parameterPaths`. The type of mean (arithmetic, geometric)
depends on the selected `distribution`. The values must be in the base
units of the parameters.}

\item{sdValues}{Vector of standard deviation values of the parameters. Must
have the same length as `parameterPaths`. The type of standard deviation
depends on the selected `distribution`.}

\item{distributions}{Type of distribution from which the random values will
  be sampled. Must have the same length as `parameterPaths`.
A list of supported distributions is defined in `Distributions`. Default is `"Normal"`.}
}
\description{
Add user defined variability on parameters to a population.
}
\keyword{internal}
