% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{FACETTYPE}
\alias{FACETTYPE}
\title{enumeration keys for type of facetting}
\format{
An object of class \code{list} of length 3.
}
\usage{
FACETTYPE
}
\description{
enumeration keys for type of facetting
}
\keyword{datasets}
