% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{filterParameterObserved}
\alias{filterParameterObserved}
\title{Filter Observed Parameters}
\usage{
filterParameterObserved(dataObservedPK, onePlotConfig)
}
\arguments{
\item{dataObservedPK}{Optional data.table containing observed PK parameter
data for comparison, which can include columns for observed values and associated
metadata.}

\item{onePlotConfig}{A data.table containing configuration settings for a
single plot, including plot name, x-axis scale, and other aesthetic settings.}
}
\value{
Filtered data table of observed PK parameters.
}
\description{
Filters observed PK parameters based on the provided configuration.
}
\keyword{internal}
