% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{formatPercentiles}
\alias{formatPercentiles}
\title{This function takes a numeric vector of percentiles and maps specific values to their corresponding labels.
It also formats other numeric values based on whether they are integers or not.}
\usage{
formatPercentiles(percentiles, suffix = "", allAsPercentiles = FALSE)
}
\arguments{
\item{percentiles}{A numeric vector of percentiles (0, 50, 100, and other values).}

\item{suffix}{A character string to append to formatted percentile values.}

\item{allAsPercentiles}{boolean, if FALSE for percentiles = 0,50,100 min, median and max is returned
otherwise 0th 50th 100th}
}
\value{
A vector containing the mapped labels for specific percentiles and formatted strings for others.
}
\description{
This function takes a numeric vector of percentiles and maps specific values to their corresponding labels.
It also formats other numeric values based on whether they are integers or not.
}
\examples{
formatPercentiles(c(0, 0.5, 1))
formatPercentiles(c(0.05, 0.25, 0.75, 0.95), suffix = " CI")

}
