% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPkBoxplot.R
\name{generateBoxwhiskerPlotForPlotType}
\alias{generateBoxwhiskerPlotForPlotType}
\title{Generate Box-and-Whisker Plot for a Specific Plot Type}
\usage{
generateBoxwhiskerPlotForPlotType(
  onePlotConfig,
  pkParameterDT,
  percentiles,
  xAxisTextAngle,
  colorVector,
  facetAspectRatio,
  nBootstrap,
  asRatio,
  ...
)
}
\arguments{
\item{onePlotConfig}{Configuration for a single plot.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{percentiles}{A vector of percentiles to calculate.}

\item{xAxisTextAngle}{Angle for x-axis text rotation.}

\item{colorVector}{A named vector for colors.}

\item{facetAspectRatio}{Aspect ratio for facets.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{asRatio}{Logical indicating if the plot is for ratios.}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
A list of generated plots for the specified plot type.
}
\description{
Creates box-and-whisker plots for either absolute values or ratios of PK parameters.
}
\keyword{internal}
