% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{generateColorScaleVectors}
\alias{generateColorScaleVectors}
\title{generates named color vectors usable for sclae_color_manual}
\usage{
generateColorScaleVectors(
  dt,
  aesthetic = list(color = "dark", fill = "light"),
  index = "colorIndex"
)
}
\arguments{
\item{dt}{`data.table` with aesthetic an dindex column}

\item{aesthetic}{named `list`, names correspond to aesthic columns,
entries are either 'dark' or 'light'}

\item{index}{name of index column}
}
\value{
named list of color vectors
}
\description{
generates named color vectors usable for sclae_color_manual
}
