% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{generateHistograms}
\alias{generateHistograms}
\title{Generate Histograms}
\usage{
generateHistograms(
  onePlotConfig,
  plotData,
  colorVector,
  facetAspectRatio,
  nMaxFacetRows,
  ...
)
}
\arguments{
\item{onePlotConfig}{A configuration for the specific plot.}

\item{plotData}{A data.table containing the plot data.}

\item{colorVector}{A named vector for colors corresponding to scenarios.}

\item{facetAspectRatio}{A numeric value for the aspect ratio of the facets.}

\item{nMaxFacetRows}{Maximum number of facet rows.}

\item{...}{Additional arguments passed to the histogram plotting function.}
}
\value{
A list of histogram plot objects.
}
\description{
Generates histogram plots based on the provided configuration and data.
}
\keyword{internal}
