% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{generatePlotTag}
\alias{generatePlotTag}
\title{Generate a Plot Tag}
\usage{
generatePlotTag(index)
}
\arguments{
\item{index}{An integer representing the position in the alphabet}
}
\value{
A character string representing the uppercase letter
corresponding to the given index.
}
\description{
This function generates a plot tag based on the provided index.
The function takes an index and returns the corresponding letter
from the alphabet in uppercase.
}
\keyword{internal}
