% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{generateRangePlots}
\alias{generateRangePlots}
\title{Generate Range Plots}
\usage{
generateRangePlots(
  onePlotConfig,
  plotData,
  colorVector,
  facetAspectRatio,
  asStepPlot,
  aggregationFun,
  ...
)
}
\arguments{
\item{onePlotConfig}{A configuration for the specific plot.}

\item{plotData}{A data.table containing the plot data.}

\item{colorVector}{A named vector for colors corresponding to scenarios.}

\item{facetAspectRatio}{A numeric value for the aspect ratio of the facets.}

\item{asStepPlot}{Logical indicating if the plot should be a step plot.}

\item{aggregationFun}{A function for aggregating the data for plotting.}

\item{...}{Additional arguments passed to the range plotting function.}
}
\value{
A list of range plot objects.
}
\description{
Generates range plots based on the provided configuration and data.
}
\keyword{internal}
