% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{generateRmdContainer}
\alias{generateRmdContainer}
\title{Generate R Markdown Container for Plotting}
\usage{
generateRmdContainer(
  projectConfiguration,
  subfolder,
  configTable,
  plotFunction,
  ...
)
}
\arguments{
\item{projectConfiguration}{An object of class `ProjectConfiguration`, which contains project-specific settings.}

\item{subfolder}{A string specifying the subfolder within the output directory where the R Markdown files will be stored.}

\item{configTable}{A data frame or data table containing the configuration settings for the plots, including headers and levels.}

\item{plotFunction}{A function that takes a plot configuration and the R Markdown container as arguments and generates the corresponding plot.}

\item{...}{Additional arguments passed to the `plotFunction`.}
}
\value{
An instance of `RmdContainer` populated with the generated plots based on the configuration table.
}
\description{
This function initializes an R Markdown container and iterates through a configuration table to generate plots based on the provided plotting function.
}
\keyword{internal}
