% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{getApplicationTimes}
\alias{getApplicationTimes}
\title{Get Application Times}
\usage{
getApplicationTimes(outputPathsPerScenario, simulatedResults)
}
\arguments{
\item{outputPathsPerScenario}{A named list of output paths for each scenario.}

\item{simulatedResult}{The results of the simulation.}
}
\value{
A list containing the start and end times of applications for each scenario.
}
\description{
This helper function prepares the timerange filter for application times based on output paths and simulation results.
}
