% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPkBoxplot.R
\name{getCaptionForBoxwhiskerPlot}
\alias{getCaptionForBoxwhiskerPlot}
\title{Get Caption for Box-and-Whisker Plot}
\usage{
getCaptionForBoxwhiskerPlot(
  plotDataPk,
  percentiles = NULL,
  yScale = NULL,
  plotCaptionAddon,
  isPlotCaption = TRUE,
  ratioMode
)
}
\arguments{
\item{plotDataPk}{A data.table containing plot data.}

\item{percentiles}{A vector of percentiles to calculate.}

\item{yScale}{Scale type (linear or log).}

\item{plotCaptionAddon}{Additional text for the caption.}

\item{isPlotCaption}{Logical indicating if the caption is for the plot.}

\item{ratioMode}{Mode indicating if the plot is for ratios.}
}
\value{
A character string containing the caption.
}
\description{
Generates a caption for the box-and-whisker plot based on the data and configuration.
}
\keyword{interal}
