% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKRatioForestPlot.R
\name{getCaptionForDDIRatioPlot}
\alias{getCaptionForDDIRatioPlot}
\title{Get Caption for DDI Ratio Plot}
\usage{
getCaptionForDDIRatioPlot(outputPathIdLoop, plotData, pkParameterDT)
}
\arguments{
\item{outputPathIdLoop}{The output path ID for the current plot.}

\item{plotData}{A data.table containing the plot data.}

\item{pkParameterDT}{A data.table with PK parameters.}
}
\value{
A string containing the caption for the plot.
}
\description{
Generates a caption for the DDI ratio plot based on the output path ID and plot data.
}
\keyword{internal}
