% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{getCaptionForDemographicPlot}
\alias{getCaptionForDemographicPlot}
\title{Get Caption for Demographic Plot}
\usage{
getCaptionForDemographicPlot(
  idData,
  valueLabel,
  binLabel,
  valueScale,
  plotCaptionAddon
)
}
\arguments{
\item{idData}{A data.table containing the plot data.}

\item{valueLabel}{A label for the value axis.}

\item{binLabel}{A label for the bin axis (optional).}

\item{valueScale}{A scale type for the value axis (optional).}

\item{plotCaptionAddon}{Additional text to append to the caption (optional).}
}
\value{
A string containing the generated caption for the plot.
}
\description{
Generates a caption for demographic plots based on the provided data.
}
\keyword{internal}
