% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getCaptionForForestPlot}
\alias{getCaptionForForestPlot}
\title{Get Caption for Forest Plot}
\usage{
getCaptionForForestPlot(
  plotData,
  xScale,
  plotCaptionAddon,
  ratioMode,
  asPointeEstimate
)
}
\arguments{
\item{plotData}{Data table containing plot data.}

\item{xScale}{Scale of the x-axis.}

\item{plotCaptionAddon}{Additional caption text.}

\item{ratioMode}{Mode for ratio calculations.}

\item{asPointeEstimate}{Logical indicating if confidence intervals should be calculated.}
}
\value{
A string containing the caption for the plot.
}
\description{
Generates a caption for the forest plot based on the provided data.
}
\keyword{internal}
