% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getCaptionForPlot}
\alias{getCaptionForPlot}
\title{Generate a caption for the plot}
\usage{
getCaptionForPlot(plotData, yScale, timeRangeFilter, plotType, plotCounter)
}
\arguments{
\item{plotData}{Object containing the data for the plot.}

\item{yScale}{Y scale of the plot.}

\item{timeRangeFilter}{Name of the time range filter.}

\item{plotType}{Type of the plot.}

\item{plotCounter}{counter of plot}
}
\value{
Caption text for the plot.
}
\description{
Generate a caption for the plot
}
\keyword{internal}
