% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{getCaptionForSensitivityPlot}
\alias{getCaptionForSensitivityPlot}
\title{Get Caption for Sensitivity Plot}
\usage{
getCaptionForSensitivityPlot(plotData, projectConfiguration, plotCaptionAddon)
}
\arguments{
\item{plotData}{A data table containing the data for the sensitivity plot.}

\item{projectConfiguration}{A list containing project configuration settings.}

\item{plotCaptionAddon}{An optional string to be added to the caption.}
}
\value{
A string containing the generated caption for the sensitivity plot.
}
\description{
Generates a caption for the sensitivity plot based on the plot data and project configuration.
}
\keyword{internal}
