% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getCaptionTimeProfile}
\alias{getCaptionTimeProfile}
\title{creates a caption text}
\usage{
getCaptionTimeProfile(dtPlotId, yScale, timeRangeTag, PlotCaptionAddon)
}
\arguments{
\item{dtPlotId}{data.table with one row per plotID}

\item{yScale}{scale of Y axis}

\item{PlotCaptionAddon}{additional text for plot}
}
\value{
`character` with caption text
}
\description{
creates a caption text
}
