% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getCIbyBootstrapping}
\alias{getCIbyBootstrapping}
\title{Get Confidence Intervals by Bootstrapping}
\usage{
getCIbyBootstrapping(
  y,
  statFun = aggregationFun,
  scenarioName,
  pkParameter,
  outputPathId,
  nObservations,
  nBootstrap,
  confLevel
)
}
\arguments{
\item{y}{Numeric vector of values.}

\item{statFun}{Function used for aggregation.}

\item{scenarioName}{Name of the scenario.}

\item{pkParameter}{Name of the PK parameter.}

\item{outputPathId}{ID for the output path.}

\item{nObservations}{Number of observations.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}
}
\value{
A list containing the calculated confidence intervals.
}
\description{
Calculates confidence intervals using bootstrapping.
}
\keyword{internal}
