% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getColorVectorForLegend}
\alias{getColorVectorForLegend}
\title{Get Color Vector for Legend}
\usage{
getColorVectorForLegend(colorLegend, colorVector)
}
\arguments{
\item{colorLegend}{A character string containing color names
separated by a pipe (`|`).}

\item{colorVector}{A named character vector where names correspond
to color names in `colorLegend`. This vector
will be used to modify the default color values.}
}
\value{
A named character vector of colors corresponding to the
        provided color legend.
}
\description{
This function generates a color vector for a given color legend,
using a specified color vector to modify default colors. The
color legend is expected to be a character string containing
color names separated by a pipe (`|`). The function first validates
the provided color vector, then generates default colors, and
applies any colors from the specified color vector.
}
\keyword{internal}
