% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getColumnSelectionForPKForest}
\alias{getColumnSelectionForPKForest}
\title{Get Column Selection for PK Forest}
\usage{
getColumnSelectionForPKForest(plotData, ratioMode)
}
\arguments{
\item{plotData}{Data table containing plot data.}

\item{ratioMode}{Mode for ratio calculations.}
}
\value{
A list of selected columns for the plot.
}
\description{
Selects columns for the PK forest plot based on the provided data and ratio mode.
}
\keyword{internal}
