% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{getConfigDirectoryForWorkflow}
\alias{getConfigDirectoryForWorkflow}
\title{Get Configuration Directory for Workflow}
\usage{
getConfigDirectoryForWorkflow(wfIdentifier)
}
\arguments{
\item{wfIdentifier}{An integer identifier for the workflow.}
}
\value{
A character string representing the configuration directory path for the specified workflow.
}
\description{
This function generates the configuration directory path for a specific workflow based on its identifier.
}
\keyword{internal}
