% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getDefaultColorsForScaleVector}
\alias{getDefaultColorsForScaleVector}
\title{creates default color Vector}
\usage{
getDefaultColorsForScaleVector(shade = c("dark", "light"), n)
}
\arguments{
\item{shade}{for n < 10 differentiation between dark and light is possible}

\item{n}{}
}
\value{
named vector with default colors
}
\description{
creates default color Vector
}
