% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aggregation.R
\name{getErrorLabels}
\alias{getErrorLabels}
\title{Extracts error labels}
\usage{
getErrorLabels(yErrorType)
}
\arguments{
\item{yErrorType}{A string descriptor indicating the type of error.}
}
\value{
A character vector of length 2 containing descriptors for mean and range.
}
\description{
This function retrieves error labels based on the provided error type descriptor.
}
\keyword{internal}
