% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aggregation.R
\name{getErrorTypeForPercentiles}
\alias{getErrorTypeForPercentiles}
\title{Creates text for percentiles}
\usage{
getErrorTypeForPercentiles(percentiles)
}
\arguments{
\item{percentiles}{A numeric vector containing percentiles.}
}
\value{
A character string with the error type for the given percentiles.
}
\description{
This function generates a descriptive text based on the provided percentiles.
}
\keyword{internal}
