% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{getExportTableForRanges}
\alias{getExportTableForRanges}
\title{Get Export Table for Ranges}
\usage{
getExportTableForRanges(plotObject, aggregationFun, xLabel)
}
\arguments{
\item{plotObject}{A plot object generated from the plotting functions.}

\item{aggregationFun}{A function for aggregating the data for exporting.}

\item{xLabel}{A label for the x-axis.}
}
\value{
A data.table containing the export data for range plots.
}
\description{
Generates an export table for range plots.
}
\keyword{internal}
