% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getFoldDistanceForPvO}
\alias{getFoldDistanceForPvO}
\title{Get fold distance for plotType `PvO`}
\usage{
getFoldDistanceForPvO(plotData)
}
\arguments{
\item{plotData}{Object containing the data for the plot.}
}
\value{
List which can be used as input for comparisonLineVector.
}
\description{
Get fold distance for plotType `PvO`
}
\keyword{internal}
