% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getFootNoteLines}
\alias{getFootNoteLines}
\title{constructs footnote lines for aggregated data and data references}
\usage{
getFootNoteLines(observedData, projectConfiguration)
}
\arguments{
\item{observedData}{either `data.table` which was generated by \code{\link{readDataByDictionary}} or object of class `DataCombined`,
which will internally transfer to the data.table format by \code{\link{convertDataCombinedToDataTable}}}

\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
vector of characters, each entry is one footnote line
}
\description{
constructs footnote lines for aggregated data and data references
}
