% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getFootnoteLinesForForestPlots}
\alias{getFootnoteLinesForForestPlots}
\title{Get Footnote Lines for Forest Plots}
\usage{
getFootnoteLinesForForestPlots(
  plotData,
  ratioMode,
  asPointeEstimate,
  dtDataReference
)
}
\arguments{
\item{plotData}{Data table containing plot data.}

\item{ratioMode}{Mode for ratio calculations.}

\item{asPointeEstimate}{Logical indicating if confidence intervals should be calculated.}

\item{dtDataReference}{Optional data reference.}
}
\value{
A character vector containing footnote lines.
}
\description{
Generates footnote lines for the forest plot.
}
\keyword{internal}
