% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{getFootnoteLinesForRangePlots}
\alias{getFootnoteLinesForRangePlots}
\title{Get Footnote Lines for Range Plots}
\usage{
getFootnoteLinesForRangePlots(errorLabels)
}
\arguments{
\item{errorLabels}{A vector of error labels to include in the footnotes.}
}
\value{
A string containing the footnote lines for the plot.
}
\description{
Generates footnote lines for range plots based on error labels.
}
\keyword{internal}
