% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getFunctionKeys}
\alias{getFunctionKeys}
\title{Get Function Keys}
\usage{
getFunctionKeys()
}
\description{
This function returns a list of function keys associated with specific plotting functions and offset for subfolders
}
\keyword{internal}
