% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getGeomLineAttributesForTP}
\alias{getGeomLineAttributesForTP}
\title{get additional inputs for `TP` plotType to enable lines in shape legend}
\usage{
getGeomLineAttributesForTP(plotData)
}
\arguments{
\item{plotData}{`data.table` with plot data}
}
\value{
list with additional attributes for geom_line
}
\description{
get additional inputs for `TP` plotType to enable lines in shape legend
}
\keyword{internal}
