% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getGroupAesthetics}
\alias{getGroupAesthetics}
\title{Get group aesthetics from scale vectors}
\usage{
getGroupAesthetics(plotData)
}
\arguments{
\item{plotData}{Object containing the data for the plot.}
}
\value{
Vector with aesthetics to group by.
}
\description{
Get group aesthetics from scale vectors
}
\keyword{internal}
