% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{getIndividualMatchForScenario}
\alias{getIndividualMatchForScenario}
\title{Get Individual Match for a Scenario}
\usage{
getIndividualMatchForScenario(projectConfiguration, scenario, dtScenarios)
}
\arguments{
\item{projectConfiguration}{A list containing the project configuration,
including the path to the populations folder.}

\item{scenario}{A string specifying the name of the scenario for which
the individual match data is to be retrieved.}

\item{dtScenarios}{A data table containing scenario details.}
}
\value{
A data.table containing the 'IndividualId' and 'ObservedIndividualId'
if the population scenario is a virtual twin population; otherwise, returns NULL.
}
\description{
This function retrieves the individual match data for a specified scenario
from a project configuration. It checks if the scenario is a population
scenario with a static population file containing the column ObservedIndividualId
and reads the IndividualId of the simulated results and the ObservedIndividualId
of the observed data.
}
