% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getMappingForForestPlots}
\alias{getMappingForForestPlots}
\title{Get Mapping for Forest Plots}
\usage{
getMappingForForestPlots(plotData, columnList)
}
\arguments{
\item{plotData}{A data frame containing the data to be plotted. It should include
columns for x values, y values, and optionally error values.}

\item{columnList}{A list containing the names of the columns to be used in the
mapping. It should include at least `yColumn` for the y-axis.}
}
\value{
A mapping object created using `aes` from ggplot2, which can be used
  in a ggplot call to create forest plots.
}
\description{
This function generates a mapping for forest plots based on the provided plot data
and the specified column list. It uses the `aes` function from ggplot2 to create
the mapping, which includes x and y values, and optionally includes error bars
based on the plot data provided.
}
\keyword{internal}
