% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getMapSimulatedAndObserved}
\alias{getMapSimulatedAndObserved}
\title{Get mapping for simulated and observed data}
\usage{
getMapSimulatedAndObserved(plotData)
}
\arguments{
\item{plotData}{Object containing the data for the plot.}

\item{plotType}{Type of the plot.}
}
\value{
Data table mapping simulated and observed data.
}
\description{
Get mapping for simulated and observed data
}
\keyword{internal}
