% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configurationEnvironment.R
\name{getModelParameterDefinitions}
\alias{getModelParameterDefinitions}
\title{Load the Model Parameter Definitions}
\usage{
getModelParameterDefinitions(wbPlots)
}
\arguments{
\item{wbPlots}{The path to the workbook containing the model parameter definitions.}
}
\value{
A `data.table` with model parameter definitions.
}
\description{
This function loads the model parameter definitions from the specified workbook.
}
\seealso{
Other get identifier:
\code{\link[=getDataGroups]{getDataGroups()}},
\code{\link[=getOutputPathIds]{getOutputPathIds()}},
\code{\link[=getScenarioDefinitions]{getScenarioDefinitions()}},
\code{\link[=getTimeRangeTags]{getTimeRangeTags()}},
\code{\link[=loadConfigTableEnvironment]{loadConfigTableEnvironment()}}
}
\concept{get identifier}
