% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{getNFacetsForDemographics}
\alias{getNFacetsForDemographics}
\title{Get Number of Facets for Demographics}
\usage{
getNFacetsForDemographics(idData, isRangePlot, nMaxFacetRows = 2)
}
\arguments{
\item{idData}{A data.table containing the plot data.}

\item{isRangePlot}{Logical indicating if the plot is a range plot.}

\item{nMaxFacetRows}{Maximum number of facet rows (default is 2).}
}
\value{
The number of facet columns to use.
}
\description{
Determines the number of facets to use for demographic plots.
}
\keyword{internal}
