% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{getObservedUnitConversionDT}
\alias{getObservedUnitConversionDT}
\title{Get Observed Unit Conversion Data Table}
\usage{
getObservedUnitConversionDT(dataObserved, dtUnit)
}
\arguments{
\item{dataObserved}{A data frame containing observed data.}

\item{dtUnit}{A data table containing unit information.}
}
\value{
A data table with unit conversions for observed data.
}
\description{
This function retrieves the unit conversion data table for observed data.
}
\keyword{internal}
