% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configurationEnvironment.R
\name{getOutputPathIds}
\alias{getOutputPathIds}
\title{Load the Output Configurations}
\usage{
getOutputPathIds(wbPlots)
}
\arguments{
\item{wbPlots}{The path to the workbook containing the output configurations.}
}
\value{
A `data.table` with output configurations.
}
\description{
This function loads the output configurations from the specified workbook.
}
\seealso{
Other get identifier:
\code{\link[=getDataGroups]{getDataGroups()}},
\code{\link[=getModelParameterDefinitions]{getModelParameterDefinitions()}},
\code{\link[=getScenarioDefinitions]{getScenarioDefinitions()}},
\code{\link[=getTimeRangeTags]{getTimeRangeTags()}},
\code{\link[=loadConfigTableEnvironment]{loadConfigTableEnvironment()}}
}
\concept{get identifier}
