% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{getOutputPathsPerScenario}
\alias{getOutputPathsPerScenario}
\title{Get Output Paths per scenario}
\usage{
getOutputPathsPerScenario(configTable, dtOutputPaths)
}
\arguments{
\item{configTable}{A data frame containing the configuration data.}
}
\value{
A list of output paths per scenario.
}
\description{
This function retrieves output paths for each scenario based on the configuration table.
}
\keyword{internal}
