% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{getPKParameterOverview}
\alias{getPKParameterOverview}
\title{Get PK Parameter Overview}
\usage{
getPKParameterOverview(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class `ProjectConfiguration` containing paths to Excel files.}
}
\value{
A `data.table` containing the merged PK parameter data along with associated scenario names.
}
\description{
This function retrieves and compiles pharmacokinetic (PK) parameter data from specified scenario definitions
and a PK parameter Excel file. It processes the data by reading the relevant sheets and merging them with
scenario information.
}
\keyword{internal}
