% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{getPlotIdForColumns}
\alias{getPlotIdForColumns}
\title{Get Plot ID for Columns}
\usage{
getPlotIdForColumns(configTable, col)
}
\arguments{
\item{configTable}{A data frame containing the configuration data.}

\item{col}{The column name to be processed.}
}
\value{
A data table with updated Plot IDs.
}
\description{
This function retrieves the Plot ID for specified columns in the configuration table.
}
\keyword{internal}
