% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getPlotIdTable}
\alias{getPlotIdTable}
\title{Creates PlotId for each scenario outputs group}
\usage{
getPlotIdTable(panelConfig, projectConfiguration)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{configTable}{}
}
\value{
`data.table` with columns 'Scenario','OutputPathId','PlotId'
}
\description{
Creates PlotId for each scenario outputs group
}
