% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPkBoxplot.R
\name{getPopulationRatioByBootstrapping}
\alias{getPopulationRatioByBootstrapping}
\title{Calculate Population Ratio by Bootstrapping}
\usage{
getPopulationRatioByBootstrapping(
  plotData,
  pkParameterDT,
  nBootstrap,
  percentiles
)
}
\arguments{
\item{plotData}{A data.table containing plot data.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{percentiles}{A vector of percentiles to calculate.}
}
\value{
A data.table containing bootstrapped ratio statistics.
}
\description{
Calculates population ratios using bootstrapping techniques.
}
\keyword{internal}
