% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{getPredictionsForScenarios}
\alias{getPredictionsForScenarios}
\title{Get Predictions for Scenarios}
\usage{
getPredictionsForScenarios(
  scenarioResults,
  dataObservedForMatch,
  aggregationFun = NULL,
  identifier = c("outputPath", "individualId")
)
}
\arguments{
\item{scenarioResults}{A list of scenario results, each containing population and covariate information.}

\item{aggregationFun}{A function for aggregation (optional). If provided, it will be used to aggregate the simulated results.}

\item{identifier}{A character vector of identifiers for matching, defaults to c("outputPath", "individualId").}

\item{dataObserved}{A data.table containing observed data, which must include the required columns:
'scenario', 'outputPathId', 'yValues', and 'yUnit'.
The unit factors (`unitFactorX` and `unitFactorY`) are necessary for proper scaling of the predicted values
and can be generated using the `prepareDataForMatch` function prior to calling this function.}
}
\value{
A data.table containing the predicted values for each scenario, along with the scenario name and other relevant identifiers.
}
\description{
This function generates predictions for a set of scenarios based on observed data.
}
\details{
The function validates the input names, loops through each scenario to generate predictions, and combines all results into a single data.table.
It also handles individual matching if 'ObservedIndividualId' is present in the scenario results.
}
