% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{getProjectDirectory}
\alias{getProjectDirectory}
\title{Get Project Directory}
\usage{
getProjectDirectory(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object containing the project configuration details.}
}
\value{
A character string representing the common project directory.
}
\description{
This function retrieves the common project directory based on the configuration settings.
}
\keyword{internal}
