% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getRatioMode}
\alias{getRatioMode}
\title{Get Ratio Mode for Plot Configuration}
\usage{
getRatioMode(onePlotConfig, pkParameterDT, asRatio)
}
\arguments{
\item{onePlotConfig}{A data frame containing the plot configuration with columns 'plotName', 'scenario', and 'referenceScenario'.}

\item{pkParameterDT}{A data frame containing parameter details, including 'scenarioName' and 'populationId'.}

\item{asRatio}{A logical value indicating whether to calculate the ratio mode. If FALSE, the function returns 'none'.}
}
\value{
A character string indicating the ratio mode. Possible values are:
- 'none' if `asRatio` is FALSE.
- 'individualRatios' if all population IDs match between scenarios.
- 'ratioOfPopulation' if all population IDs are different between scenarios.
}
\description{
This function determines the ratio mode for a given plot configuration based on the provided parameters.
It checks if the scenarios in the plot configuration have the same or different base populations.
}
\keyword{internal}
