% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getScalevector}
\alias{getScalevector}
\title{returns a scalevector usable for manaul scaling in ggplot}
\usage{
getScalevector(namesOfScaleVector, listOfValues)
}
\arguments{
\item{namesOfScaleVector}{names of the vector}

\item{listOfValues}{list of possible entry, take the first where all values are not NA}
}
\value{
scaleVector
}
\description{
returns a scalevector usable for manaul scaling in ggplot
}
