% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getSimulatedTimeprofile}
\alias{getSimulatedTimeprofile}
\title{loads simulated result and convert for plotting}
\usage{
getSimulatedTimeprofile(
  simulatedResult,
  outputs,
  dtOutputPaths,
  targetTimeUnit,
  timeOffset
)
}
\arguments{
\item{simulatedResult}{simulation result for one scenario}

\item{outputs}{`outputPathsIds` selected for scenario}

\item{targetTimeUnit}{target unit for time display}

\item{timeOffset}{offset which is subtracted from time to adjust to Reference or data}
}
\value{
`data.table` with converted output
}
\description{
if population scenario, time profiles are aggregated
Units are converted to desired units
}
