% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getSubfolderByKey}
\alias{getSubfolderByKey}
\title{Get Subfolder by Key}
\usage{
getSubfolderByKey(key, subfolder, configTableSheet)
}
\arguments{
\item{key}{A character string representing the key that determines the suffix for the subfolder name.}

\item{subfolder}{An optional character string representing an existing subfolder name. If NULL, a new subfolder name will be generated.}

\item{configTableSheet}{A character string representing the configuration table sheet used to create the subfolder name.}
}
\value{
A character string representing the subfolder name, either provided or generated based on the function key.
}
\description{
This function generates a subfolder name based on a provided function key and configuration table sheet.
}
