% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getTableLabelsForPKForest}
\alias{getTableLabelsForPKForest}
\title{Get Table Labels for PK Forest}
\usage{
getTableLabelsForPKForest(plotData)
}
\arguments{
\item{plotData}{Data table containing plot data.}
}
\value{
A named vector of table labels.
}
\description{
Generates table labels for the PK forest plot.
}
\keyword{internal}
