% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{handleConfigTable}
\alias{handleConfigTable}
\title{Handle Configuration Table Processing}
\usage{
handleConfigTable(
  rmdPlotManager,
  configTable,
  projectConfiguration,
  inputs,
  suppressExport,
  theme
)
}
\arguments{
\item{rmdPlotManager}{A QmdPlotManager object responsible for managing Qmd file generation and plot exports.}

\item{configTable}{A data frame containing the configuration settings for the plots.}

\item{projectConfiguration}{A ProjectConfiguration object containing the project configuration settings.}

\item{inputs}{A list of additional inputs for the plot function.}

\item{suppressExport}{A logical value indicating whether to suppress the export of the Qmd file.}
}
\value{
A named list of generated plot objects.
}
\description{
This helper function processes the configuration table, generating plots according to the specified configurations.
}
\keyword{internal}
