% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{importProjectConfiguration}
\alias{importProjectConfiguration}
\title{Import Project Configuration}
\usage{
importProjectConfiguration(
  configurationDirectory,
  ePackageFolder,
  wfIdentifier
)
}
\arguments{
\item{configurationDirectory}{A character string representing the directory where the configuration files will be stored.}

\item{ePackageFolder}{A character string representing the path to the electronic package folder containing the configuration.}

\item{wfIdentifier}{An integer identifier for the workflow being imported.}
}
\value{
A character string indicating the direction of synchronization (e.g., "plotToScenario", "bothways").
}
\description{
This function imports the project configuration from an electronic package (ePackage) into a specified directory.
It reads the configuration data from a JSON file and creates the necessary Excel sheets.
}
\keyword{internal}
