% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{importWorkflow}
\alias{importWorkflow}
\title{Import Workflow}
\usage{
importWorkflow(
  projectDirectory,
  wfIdentifier,
  ePackageFolder,
  configurationDirectory
)
}
\arguments{
\item{projectDirectory}{A character string representing the path to the project directory where
the workflow will be imported.}

\item{wfIdentifier}{An integer identifier for the workflow being imported. This identifier is
used to distinguish between different workflows.}

\item{ePackageFolder}{A character string representing the path to the electronic package folder
that contains the workflow files and configurations to be imported.}

\item{configurationDirectory}{A character string representing the directory where the configuration
files will be stored.}
}
\value{
An object containing the updated project configuration after the workflow has been imported.
}
\description{
This function imports a workflow from an electronic package (ePackage) into a project directory.
It initializes the project configuration, retrieves relevant project files, and synchronizes
scenarios with plots. The function ensures that the necessary directories exist and that the
configuration is set up correctly for further processing.
}
\seealso{
Other electronic package:
\code{\link[=exportSimulationWorkflowToEPackage]{exportSimulationWorkflowToEPackage()}},
\code{\link[=exportTLFWorkflowToEPackage]{exportTLFWorkflowToEPackage()}}
}
\concept{electronic package}
