% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{importWorkflowFiles}
\alias{importWorkflowFiles}
\title{Import Workflow Files}
\usage{
importWorkflowFiles(projectConfigurationNew, ePackageFolder, wfIdentifier)
}
\arguments{
\item{projectConfigurationNew}{An object representing the new project configuration where files will be imported.}

\item{ePackageFolder}{A character string representing the path to the electronic package folder containing the workflow files.}

\item{wfIdentifier}{An integer identifier for the workflow being imported.}
}
\value{
Invisible NULL.
}
\description{
This function imports workflow files from an electronic package (ePackage) into a new project configuration.
It handles different file types and copies them to the appropriate directories.
}
\keyword{internal}
