% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotForrest.R
\name{iniForestPlot}
\alias{iniForestPlot}
\title{Initialize Forest Plot}
\usage{
iniForestPlot(plotData, xlabel)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted.}

\item{xlabel}{Character string for the x-axis label.}
}
\value{
A ggplot object representing the initialized forest plot.
}
\description{
This function initializes the basic structure of the forest plot.
}
\keyword{internal}
