% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{iniTestProject}
\alias{iniTestProject}
\title{Initialize Test Project}
\usage{
iniTestProject(
  rootDirectory,
  verbose,
  modelFiles,
  instDirectory = system.file(package = "ospsuite.reportingframework", "extdata",
    mustWork = TRUE)
)
}
\arguments{
\item{rootDirectory}{A character string specifying the root directory for the test project.}

\item{verbose}{A logical value indicating whether to print messages about the process.}

\item{modelFiles}{A named character vector containing paths to the model files used in the project.}

\item{instDirectory}{A character string specifying the installation directory for the package.}
}
\value{
A ProjectConfiguration object containing the details of the initialized project.
}
\description{
This function initializes the project structure and configuration for the test project.
}
\keyword{internal}
