% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{initializeParametersOfSheets}
\alias{initializeParametersOfSheets}
\title{Initialize PK Parameters in Specified Sheets}
\usage{
initializeParametersOfSheets(projectConfiguration, pkParameterSheets)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration settings,
including the path to the PK parameter configuration sheet}

\item{pkParameterSheets}{A vector of sheet names that contain the PK parameters
to be initialized.}
}
\value{
NULL. The function updates parameters in-place and does not return a value.
}
\description{
This function initializes pharmacokinetic (PK) parameters in the specified
sheets based on user-defined settings. It updates the parameters in-place
within the provided PK parameter configuration sheet.
}
\keyword{internal}
