% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{initLogfunction}
\alias{initLogfunction}
\title{Initializing a Log Function}
\usage{
initLogfunction(
  projectConfiguration,
  warningsNotDisplayed = c("introduced infinite values",
    "Each group consists of only one observation", "rows containing non-finite values",
    "rows containing missing values", "Ignoring unknown parameters",
    "was deprecated in ggplot2", "font family not found in Windows font database",
    "mbcsToSbcs"),
  messagesNotDisplayed = c("Each group consists of only one observation"),
  verbose = TRUE,
  loggingFolder = NULL
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and file names}

\item{warningsNotDisplayed}{A list of warnings that should not be logged.}

\item{messagesNotDisplayed}{A list of messages that should not be logged.}

\item{verbose}{boolean, if true log messages will be shown on the console}

\item{loggingFolder}{A character string specifying the folder where log files should be stored. If NULL, defaults to a "Logs" sub-folder within the output folder of the project configuration.}
}
\description{
This function initializes the logging during a workflow. It is called at the start of the workflow script.
It is used to configure options for the log file folder, warnings which should not be logged, and messages which should not be logged.
}
\examples{
\dontrun{
# Initialize the log function
logFunction <- initLogfunction(rootDirectory = "path/to/project")
}

}
\seealso{
Other log file management: 
\code{\link{addMessageToLog}()},
\code{\link{captureLog}()},
\code{\link{saveSessionInfo}()},
\code{\link{setShowLogMessages}()},
\code{\link{writeTableToLog}()},
\code{\link{writeToLog}()}
}
\concept{log file management}
