% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{isPlotTypeNeededAndPossible}
\alias{isPlotTypeNeededAndPossible}
\title{Check if a specific plot type is needed and possible}
\usage{
isPlotTypeNeededAndPossible(plotType, plotData)
}
\arguments{
\item{plotType}{Type of the plot to check.}

\item{plotData}{Object containing the data for the plot.}
}
\value{
Logical indicating if the plot type is needed and possible.
}
\description{
Check if a specific plot type is needed and possible
}
\keyword{internal}
