% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configurationEnvironment.R
\name{loadConfigTableEnvironment}
\alias{loadConfigTableEnvironment}
\title{Load Configuration Tables}
\usage{
loadConfigTableEnvironment(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class `ProjectConfiguration` containing paths to Excel files.}
}
\value{
An invisible value indicating that the loading was successful.
}
\description{
This function loads various configuration tables from specified Excel files into a global environment.
}
\seealso{
Other get identifier:
\code{\link[=getDataGroups]{getDataGroups()}},
\code{\link[=getModelParameterDefinitions]{getModelParameterDefinitions()}},
\code{\link[=getOutputPathIds]{getOutputPathIds()}},
\code{\link[=getScenarioDefinitions]{getScenarioDefinitions()}},
\code{\link[=getTimeRangeTags]{getTimeRangeTags()}}
}
\concept{get identifier}
