% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configurationEnvironment.R
\name{loadConfigTables}
\alias{loadConfigTables}
\title{Load Configuration Tables}
\usage{
loadConfigTables(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class `ProjectConfiguration` containing paths to Excel files.}
}
\value{
An invisible value indicating that the loading was successful.
}
\description{
This function loads various configuration tables from specified Excel files into a global environment.
}
