% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{loadDemographicParameters}
\alias{loadDemographicParameters}
\title{Load Demographic Parameters}
\usage{
loadDemographicParameters(onePlotConfig, scenarioList)
}
\arguments{
\item{onePlotConfig}{A configuration for the specific plot.}

\item{scenarioList}{A list of scenarios to consider.}
}
\value{
A data.table containing the loaded demographic parameters.
}
\description{
Loads demographic parameters from the configuration and scenario list.
}
\keyword{internal}
