% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{loadPKAnalysisPerScenario}
\alias{loadPKAnalysisPerScenario}
\title{Load or Calculate PK Analysis for a Given Scenario}
\usage{
loadPKAnalysisPerScenario(
  scenarioName,
  scenarioSimulation,
  pkParameterSheets,
  projectConfiguration
)
}
\arguments{
\item{scenarioName}{The name of the scenario to be processed.}

\item{scenarioSimulation}{A simulation object corresponding to the scenario.}

\item{pkParameterSheets}{A vector of sheet names from the PK parameter file to read.}

\item{projectConfiguration}{A list containing project configuration settings,
including the PK parameter file.}
}
\value{
A data.table containing the processed PK analyses for the specified scenario.
}
\description{
This function loads PK analysis results from a CSV file for a specific scenario.
}
\keyword{internal}
