% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{loadPkAnalysisRawData}
\alias{loadPkAnalysisRawData}
\title{Load PK Analysis Results from CSV}
\usage{
loadPkAnalysisRawData(projectConfiguration, scenarioName, scenarioSimulation)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration settings,
including the output folder path and the PK parameter file.}

\item{scenarioName}{The name of the scenario for which PK analysis is to be loaded.}

\item{scenarioSimulation}{A simulation object corresponding to the scenario.}
}
\value{
A data.table containing the PK analyses loaded from the CSV file.
}
\description{
This function loads PK analysis results from a CSV file or recalculates them
if the file is not found. It returns the loaded data as a data.table.
}
\keyword{internal}
