% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{loadPKParameter}
\alias{loadPKParameter}
\title{Load Pharmacokinetic (PK) Parameters for Specified Scenarios}
\usage{
loadPKParameter(projectConfiguration, scenarioListOrResult)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration settings,
including the PK parameter file.}

\item{scenarioListOrResult}{A named list of scenarios for which PK parameters are to be loaded.}
}
\value{
A data.table containing the processed PK analyses for all specified scenarios.

A data.table containing the processed PK analyses.
}
\description{
This function loads pharmacokinetic (PK) parameters for specified scenarios
based on project configuration and a list of scenarios. It processes each
scenario and returns the results as a data.table.
}
\seealso{
Other scenario management:
\code{\link[=calculatePKParameterForScenarios]{calculatePKParameterForScenarios()}},
\code{\link[=createScenariosWrapped]{createScenariosWrapped()}},
\code{\link[=loadScenarioResultsToFramework]{loadScenarioResultsToFramework()}},
\code{\link[=runAndSaveScenarios]{runAndSaveScenarios()}},
\code{\link[=runOrLoadScenarios]{runOrLoadScenarios()}}
}
\concept{scenario management}
