% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{loadPopulationParameterForScenario}
\alias{loadPopulationParameterForScenario}
\title{Load Population Parameter for Scenario}
\usage{
loadPopulationParameterForScenario(
  scenarioName,
  scenarioList,
  onePlotConfigIdentifier
)
}
\arguments{
\item{scenarioName}{The name of the scenario.}

\item{scenarioList}{A list of scenarios to consider.}

\item{onePlotConfigIdentifier}{A configuration identifier for the plot.}
}
\value{
A data.table containing the loaded population parameters.
}
\description{
Loads population parameters for a specific scenario.
}
\keyword{internal}
