% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{loadScenarioResultsToFramework}
\alias{loadScenarioResultsToFramework}
\title{Load existing scenario results}
\usage{
loadScenarioResultsToFramework(projectConfiguration, scenarioNames)
}
\arguments{
\item{projectConfiguration}{Configuration for the project, containing paths and settings necessary
to load the results.}

\item{scenarioNames}{Character vector of the names of the scenarios whose results are to be loaded.}
}
\value{
A list containing the loaded scenario results, including population data if available.
throws Error if the scenario results do not exist.
}
\description{
This function loads the results of specified scenarios. If the results do not exist,
it returns an error.
}
\seealso{
Other scenario management:
\code{\link[=calculatePKParameterForScenarios]{calculatePKParameterForScenarios()}},
\code{\link[=createScenariosWrapped]{createScenariosWrapped()}},
\code{\link[=loadPKParameter]{loadPKParameter()}},
\code{\link[=runAndSaveScenarios]{runAndSaveScenarios()}},
\code{\link[=runOrLoadScenarios]{runOrLoadScenarios()}}
}
\concept{scenario management}
