% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{loadScenarioTimeProfiles}
\alias{loadScenarioTimeProfiles}
\title{Load Scenario Time Profiles}
\usage{
loadScenarioTimeProfiles(
  projectConfiguration,
  simulatedResults,
  outputPathsPerScenario,
  aggregationFun
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{simulatedResults}{A list containing simulation results for each scenario.}

\item{outputPathsPerScenario}{A named list of output paths for each scenario.}

\item{aggregationFun}{A function to aggregate simulation data.}
}
\value{
A data.table containing the aggregated simulated time profiles for all scenarios.
}
\description{
This function aggregates simulated time profiles for different scenarios based on the provided results and output paths.
}
