% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{loadSimulatedResultsForTimeProfilePanelPlot}
\alias{loadSimulatedResultsForTimeProfilePanelPlot}
\title{load simulated results for all scenarios of this panel}
\usage{
loadSimulatedResultsForTimeProfilePanelPlot(projectConfiguration, panelConfig)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{panelConfig}{}
}
\value{
`data.table` with simulated results
}
\description{
load simulated results for all scenarios of this panel
}
