% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdBullet}
\alias{mdBullet}
\alias{mdBullet0}
\title{Insert bulleted item with indentation.}
\usage{
mdBullet(..., bullet = "-", level = 1)

mdBullet0(..., bullet = "-", level = 1)
}
\arguments{
\item{...}{Bullet content.}

\item{bullet}{Bullet character (default "-").}

\item{level}{Indentation level (default 1).}
}
\description{
Insert bulleted item with indentation.

Insert bulleted item using paste0.
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
