% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdCaption}
\alias{mdCaption}
\title{add Caption to .Rmd}
\usage{
mdCaption(subfolder, captionFile, captionPrefix, captionStyle = NULL)
}
\arguments{
\item{subfolder}{The folder where the file is located relative to Rmd}

\item{captionFile}{file containing caption}

\item{captionPrefix}{'a character which starts the caption like 'Figure 1:'}

\item{captionStyle}{'custom-style for captions'}
}
\description{
add Caption to .Rmd
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdBullet0]{mdBullet0()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
