% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdFootNote}
\alias{mdFootNote}
\title{add footnote lines}
\usage{
mdFootNote(subfolder, footNoteFile, footNoteCustomStyle = NULL)
}
\arguments{
\item{subfolder}{The folder where the file is located relative to Rmd}

\item{footNoteFile}{file containing footnotes}

\item{footNoteCustomStyle}{a character describing custom style footnotes}
}
\description{
add footnote lines
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdBullet0]{mdBullet0()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
