% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdPaste}
\alias{mdPaste}
\title{mdPaste}
\usage{
mdPaste(..., sep = " ", collapse = NULL, newlines = 1)
}
\arguments{
\item{...}{Elements to be `paste`d.}

\item{sep}{Passed to `paste`. Defaults to `" "`.}

\item{collapse}{Passed to `paste`. Defaults to `NULL`.}

\item{newlines}{Number of newlines to insert afterwards. Defaults to 1.}
}
\description{
`cat` text using `paste` and optionally end with newlines.
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdBullet0]{mdBullet0()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
