% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdPaste0}
\alias{mdPaste0}
\title{mdPaste0}
\usage{
mdPaste0(..., collapse = NULL, newlines = 1)
}
\arguments{
\item{...}{Elements to be `paste0`d.}

\item{collapse}{Collapse argument.}

\item{newlines}{Newlines afterwards (default 1).}
}
\description{
mdPaste0
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
