% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{mergeDataGroups}
\alias{mergeDataGroups}
\title{Merge Data Groups into Configuration}
\usage{
mergeDataGroups(dtNewConfig, dtDataGroups, dataObserved)
}
\arguments{
\item{dtNewConfig}{A data.table containing the new configuration for time profile plots.}

\item{dtDataGroups}{A data.table containing data group information.}

\item{dataObserved}{Optional. A data object containing observed data, if available.}
}
\value{
A data.table with merged data group information.
}
\description{
This internal function merges data group information into the configuration table based on the provided scenarios and observed data.
}
\keyword{internal}
