% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mergeRmds}
\alias{mergeRmds}
\title{Merge multiple Qmd files into one.}
\usage{
mergeRmds(
  newName = "appendix",
  title = "Appendix",
  sourceRmds = c("Demographics", "TimeProfile", "PKParameter", "DDIRatio", "myFigures"),
  projectConfiguration
)
}
\arguments{
\item{newName}{Output file name (default "appendix").}

\item{title}{Output title (default "Appendix").}

\item{sourceRmds}{Source Qmd files to include.}

\item{projectConfiguration}{ProjectConfiguration object.}
}
\description{
Merge multiple Qmd files into one.
}
\concept{functions called by workflow script}
