% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{messages}
\alias{messages}
\title{List of functions and strings used for log, warning, and error messages}
\usage{
messages
}
\description{
List of functions and strings used for log, warning, and error messages
}
\keyword{internal}
