% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{mockManualEditingsCleanup}
\alias{mockManualEditingsCleanup}
\title{Cleanup Mock Manual Editings}
\usage{
mockManualEditingsCleanup(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}
}
\description{
This function cleans upthe project configuration files
by clearing existing data from relevant Excel sheets.
}
\keyword{internal}
