% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{mockManualEditingsDataDictionary}
\alias{mockManualEditingsDataDictionary}
\title{Mock Manual editing for Data Dictionary}
\usage{
mockManualEditingsDataDictionary(projectConfiguration, withPK = TRUE)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}

\item{withPK}{boolean, if TRUE configuration for pkparameter  are added}
}
\description{
This function updates the data dictionary for the project configuration with new data files.
}
