% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{mockManualEditingsdisplayNames}
\alias{mockManualEditingsdisplayNames}
\title{Mock Manual editing for Display Names}
\usage{
mockManualEditingsdisplayNames(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}
}
\description{
This function updates the display names for scenarios and output paths in the project configuration.
}
