% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{mockManualEditingsPKParameter}
\alias{mockManualEditingsPKParameter}
\title{Mock Manual editing for PK Parameters}
\usage{
mockManualEditingsPKParameter(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}
}
\description{
This function adds PK parameters to the project configuration based on templates.
}
