% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{mockManualEditingsPopulation}
\alias{mockManualEditingsPopulation}
\title{Mock Manual editing for Population}
\usage{
mockManualEditingsPopulation(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}
}
\description{
This function adds biometric ranges for virtual pediatric and one adult populations to the configuration files
}
\keyword{internal}
