% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{mockManualEditingsScenario}
\alias{mockManualEditingsScenario}
\title{Mock Manual Editings for Scenarios}
\usage{
mockManualEditingsScenario(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}
}
\description{
This function sets up scenarios based on the defined populations and adds PK Parameters
and output paths to the project configuration.
}
