% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{mockManualEditingsSensitivity}
\alias{mockManualEditingsSensitivity}
\title{Mock Manual Editings for Sensitivity}
\usage{
mockManualEditingsSensitivity(projectConfiguration, sheetName)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}

\item{sheetName}{A character string specifying the name of the sensitivity sheet to update.}
}
\description{
This function updates the sensitivity analysis configuration in the project.
}
\keyword{internal}
