% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{open_figure_template}
\alias{open_figure_template}
\title{Opens the  template for figure creation as new document}
\usage{
open_figure_template()
}
\description{
Opens the  template for figure creation as new document
}
