% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{open_workflow_template}
\alias{open_workflow_template}
\title{Opens the workflow template as new document}
\usage{
open_workflow_template()
}
\description{
Per default the workflow template of the package is loaded.
To customize it save the file with name "template_workflow.R" in your template directory and
set the path to the template via option options(OSPSuite.REF.PathForWorkflowTemplate = 'myTemplateDirectory')
}
