% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{openAIDebugRunPrompt}
\alias{openAIDebugRunPrompt}
\title{Creates an AI-ready debug prompt in the console}
\usage{
openAIDebugRunPrompt()
}
\description{
Creates an AI-ready debug prompt in the console
}
\seealso{
Other Addins Support:
\code{\link[=createDocumentFromTemplate]{createDocumentFromTemplate()}},
\code{\link[=createWorkflowTemplate]{createWorkflowTemplate()}},
\code{\link[=openAIBuildPlotsPrompt]{openAIBuildPlotsPrompt()}},
\code{\link[=openAIImportDataPrompt]{openAIImportDataPrompt()}},
\code{\link[=openAIWorkflowPrompt]{openAIWorkflowPrompt()}},
\code{\link[=openEPackageTemplate]{openEPackageTemplate()}},
\code{\link[=openFigureTemplate]{openFigureTemplate()}},
\code{\link[=openWorkflowTemplate]{openWorkflowTemplate()}}
}
\concept{Addins Support}
