% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{openAIImportDataPrompt}
\alias{openAIImportDataPrompt}
\title{Creates an AI-ready import-data prompt in the console}
\usage{
openAIImportDataPrompt()
}
\description{
Creates an AI-ready import-data prompt in the console
}
\seealso{
Other Addins Support:
\code{\link[=createDocumentFromTemplate]{createDocumentFromTemplate()}},
\code{\link[=createWorkflowTemplate]{createWorkflowTemplate()}},
\code{\link[=openAIBuildPlotsPrompt]{openAIBuildPlotsPrompt()}},
\code{\link[=openAIDebugRunPrompt]{openAIDebugRunPrompt()}},
\code{\link[=openAIWorkflowPrompt]{openAIWorkflowPrompt()}},
\code{\link[=openEPackageTemplate]{openEPackageTemplate()}},
\code{\link[=openFigureTemplate]{openFigureTemplate()}},
\code{\link[=openWorkflowTemplate]{openWorkflowTemplate()}}
}
\concept{Addins Support}
