% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{openAIWorkflowPrompt}
\alias{openAIWorkflowPrompt}
\title{Creates an AI-ready workflow prompt in the console}
\usage{
openAIWorkflowPrompt()
}
\description{
Creates an AI-ready workflow prompt in the console
}
\seealso{
Other Addins Support:
\code{\link[=createDocumentFromTemplate]{createDocumentFromTemplate()}},
\code{\link[=createWorkflowTemplate]{createWorkflowTemplate()}},
\code{\link[=openAIBuildPlotsPrompt]{openAIBuildPlotsPrompt()}},
\code{\link[=openAIDebugRunPrompt]{openAIDebugRunPrompt()}},
\code{\link[=openAIImportDataPrompt]{openAIImportDataPrompt()}},
\code{\link[=openEPackageTemplate]{openEPackageTemplate()}},
\code{\link[=openFigureTemplate]{openFigureTemplate()}},
\code{\link[=openWorkflowTemplate]{openWorkflowTemplate()}}
}
\concept{Addins Support}
