% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{openWorkflowTemplate}
\alias{openWorkflowTemplate}
\title{Opens the workflow template as new document}
\usage{
openWorkflowTemplate()
}
\description{
Per default the workflow template of the package is loaded.
To customize it save the file with name "template_workflow.R" in your template directory and
set the path to the template via option `options(OSPSuite.RF.PathForWorkflowTemplate = 'myTemplateDirectory')`
}
\seealso{
Other Addins Support:
\code{\link[=createDocumentFromTemplate]{createDocumentFromTemplate()}},
\code{\link[=createWorkflowTemplate]{createWorkflowTemplate()}},
\code{\link[=openAIBuildPlotsPrompt]{openAIBuildPlotsPrompt()}},
\code{\link[=openAIDebugRunPrompt]{openAIDebugRunPrompt()}},
\code{\link[=openAIImportDataPrompt]{openAIImportDataPrompt()}},
\code{\link[=openAIWorkflowPrompt]{openAIWorkflowPrompt()}},
\code{\link[=openEPackageTemplate]{openEPackageTemplate()}},
\code{\link[=openFigureTemplate]{openFigureTemplate()}}
}
\concept{Addins Support}
