% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotForrest.R
\name{ospsuite_plotForest}
\alias{ospsuite_plotForest}
\title{Create a Forest Plot}
\usage{
ospsuite_plotForest(
  plotData,
  vlineIntercept = c(),
  xscale = "log",
  digitsToRound = 2,
  digitsToShow = 2,
  tableLabels = c("Ratio", "90\%\\nCI lower", "90\%\\nCI upper"),
  xlabel = "DDI Ratio",
  textwidthLeft = 0.22,
  textwidthRight = 0.3,
  withoutTable = FALSE
)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted. It must include the following columns:
\itemize{
  \item \code{displayGroup}: Grouping variable for displaying results.
  \item \code{displayName}: Name of the display for each row in the plot.
  \item \code{parameter}: The parameter for which the effect size is calculated.
  \item \code{type}: Type of effect size (e.g., treatment, control).
  \item \code{x}: The effect size (e.g., ratio).
  \item \code{xmin}: Lower bound of the confidence interval.
  \item \code{xmax}: Upper bound of the confidence interval.
}}

\item{vlineIntercept}{Numeric vector for vertical line intercept(s). This can be used to indicate reference lines (e.g., no effect line).
Default is an empty vector, which means no vertical lines will be drawn.}

\item{xscale}{Character string specifying the x-axis scale, either 'linear' or 'log'.
Default is 'log', which is often preferred for ratio data.}

\item{digitsToRound}{Integer indicating the number of digits to round the numerical values in the table. Default is 2.}

\item{digitsToShow}{Integer indicating the number of digits to display in the table. Default is 2.}

\item{tableLabels}{Character vector with labels for the table columns. Default includes:
\itemize{
  \item 'Ratio'
  \item '90%\nCI lower'
  \item '90%\nCI upper'
}}

\item{xlabel}{Character string for the x-axis label. Default is 'DDI Ratio'.}

\item{textwidthLeft}{Numeric value for the width of the left text area (0 to 1). Default is 0.22.}

\item{textwidthRight}{Numeric value for the width of the right text area (0 to 1). Default is 0.3.}

\item{withoutTable}{Logical indicating whether to display the table on the right side of the plot.
If TRUE, the table will not be displayed. Default is FALSE.}
}
\value{
A ggplot object representing the forest plot, which can be further customized or printed.
The plot includes error bars for the confidence intervals, points for the effect sizes,
and optional tables displaying the numerical values.
}
\description{
This function generates a forest plot based on the provided data. The plot visualizes the effect sizes (ratios)
along with their confidence intervals, allowing for easy comparison across different groups or parameters.
}
