% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotPredictedVsObserved}
\alias{ospsuite_plotPredictedVsObserved}
\title{plots predicted vs observed, data are grouped by "group"}
\usage{
ospsuite_plotPredictedVsObserved(plotData, metaData = NULL, ...)
}
\arguments{
\item{plotData}{object of class dataCombined or dataFrame generated by dataCombined object
units should be consistent, otherwise use convertUnits(myDataCombined)}

\item{metaData}{list with metaData, a default list is constructed from the data}

\item{...}{arguments passed on to opssuite.plots::plotTimeProfile}
}
\value{
`ggplot2`plotObject
}
\description{
plots predicted vs observed, data are grouped by "group"
}
\examples{
\dontrun{
plotPredictedVsObserved(convertUnits(
  myDataCombined,
  xUnit = ospUnits$Time$h,
  yUnit = ospUnits$`Concentration [mass]`$`µg/l`
))
}
}
