% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotQuantileQuantilePlot}
\alias{ospsuite_plotQuantileQuantilePlot}
\title{Plots a Quantile-Quantile plot, grouped by "group".}
\usage{
ospsuite_plotQuantileQuantilePlot(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  ...
)
}
\arguments{
\item{plotData}{An object of class DataCombined or a data.frame generated by a DataCombined object.
Units should be consistent; if not, use `convertUnits(myDataCombined)` to standardize.}

\item{metaData}{A list containing metadata for the plot. If NULL, a default list is constructed from the data.}

\item{mapping}{additional mappings specified by user}

\item{...}{Additional arguments passed to `ospsuite.plots::plotQQ`.}
}
\value{
A `ggplot2` plot object representing the Q-Q plot.
}
\description{
This function visualizes the distribution of predicted vs observed values using a Q-Q plot.
}
\examples{
\dontrun{
# Generate a Q-Q plot for the provided data
plotQuantileQuantilePlot(convertUnits(
  myDataCombined,
  xUnit = ospUnits$Time$h,
  yUnit = ospUnits$`Concentration [mass]`$`µg/l`
))
}
}
