% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotResidualsAsHistogram}
\alias{ospsuite_plotResidualsAsHistogram}
\title{Plots residuals as a histogram, grouped by "group".}
\usage{
ospsuite_plotResidualsAsHistogram(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  ...
)
}
\arguments{
\item{plotData}{An object of class DataCombined or a data.frame generated by a DataCombined object.
Units should be consistent; if not, use `convertUnits(myDataCombined)` to standardize.}

\item{metaData}{A list containing metadata for the plot. If NULL, a default list is constructed from the data.}

\item{mapping}{additional mappings specified by user}

\item{...}{Additional arguments passed to `ospsuite.plots::plotHistogram`.}
}
\value{
A `ggplot2` plot object representing the histogram of residuals.
}
\description{
This function generates a histogram of the residuals, providing a visual representation of their distribution.
}
\examples{
\dontrun{
# Generate a histogram of residuals for the provided data
plotResidualsAsHistogram(convertUnits(
  myDataCombined,
  xUnit = ospUnits$Time$h,
  yUnit = ospUnits$`Concentration [mass]`$`µg/l`
))
}
}
