% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotResidualsVsObserved}
\alias{ospsuite_plotResidualsVsObserved}
\title{Plots residuals vs observed values, grouped by "group".}
\usage{
ospsuite_plotResidualsVsObserved(
  plotData,
  metaData = NULL,
  mapping = ggplot2::aes(),
  ...
)
}
\arguments{
\item{plotData}{An object of class DataCombined or a data.frame generated by a DataCombined object.
Units should be consistent; if not, use `convertUnits(myDataCombined)` to standardize.}

\item{metaData}{A list containing metadata for the plot. If NULL, a default list is constructed from the data.}

\item{mapping}{additional mappings specified by user}

\item{...}{Additional arguments passed to `ospsuite.plots::plotResVsCov`.}
}
\value{
A `ggplot2` plot object representing residuals vs observed values.
}
\description{
This function visualizes the residuals against observed values, helping to assess model performance.
}
\examples{
\dontrun{
# Generate a residuals vs observed plot for the provided data
plotResidualsVsObserved(convertUnits(
  myDataCombined,
  xUnit = ospUnits$Time$h,
  yUnit = ospUnits$`Concentration [mass]`$`µg/l`
))
}
}
