% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotTimeProfile}
\alias{ospsuite_plotTimeProfile}
\title{creates a time profile for plotData, data are grouped ny "group"}
\usage{
ospsuite_plotTimeProfile(plotData, metaData = NULL, ...)
}
\arguments{
\item{plotData}{object of class dataCombined or dataFrame generated by dataCombined object
units should be consistent, otherwise use convertUnits(myDataCombined)}

\item{metaData}{list with metaData, a defualt list is constructed from the data}

\item{...}{arguments passed on to opssuite.plots::plotTimeProfile}
}
\value{
`ggplot2`plotObject
}
\description{
creates a time profile for plotData, data are grouped ny "group"
}
\examples{
\dontrun{
plotTimeProfile(convertUnits(
     myDataCombined,
     xUnit = ospUnits$Time$s,
     yUnit = ospUnits$`Concentration [mass]`$`µg/l`
 ))
 }
}
