% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{pasteFigureTags}
\alias{pasteFigureTags}
\title{Title}
\usage{
pasteFigureTags(dtCaption, captionColumn)
}
\arguments{
\item{dtCaption}{`data.table` with caption information must have column PlotTag}

\item{captionColumn}{`character`column names which should be sorted to Tags}
}
\value{
`character` text for cpation
}
\description{
Title
}
