% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\docType{class}
\name{PlotDataTimeProfile}
\alias{PlotDataTimeProfile}
\title{plotData for Time profile plot Data}
\description{
An object to collect plot Data and other information to create a time profile plot
}
\keyword{internal}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{PlotDataTimeProfile}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{data}}{Data for Plot}

    \item{\code{dataReference}}{Data with references for data}

    \item{\code{configTable}}{Configuration table for one plot}

    \item{\code{plotName}}{plotName of configTable}

    \item{\code{dtCaption}}{Data.table with caption information}

    \item{\code{scaleVectors}}{List with scaling vectors to manually scale aesthetics}

    \item{\code{tpLabelSimulatedMean}}{Label for simulated mean}

    \item{\code{tpLabelSimulatedRange}}{Label for simulated range}

    \item{\code{tpLabelObserved}}{Label for observed data}

    \item{\code{tpLabels}}{List of time profile labels to add on ggplot}

    \item{\code{nFacetColumns}}{Number of facet columns}

    \item{\code{timeRangeTagFilter}}{List with filters for time range tags}

    \item{\code{reverseLegend}}{if true color legend will be reversed}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PlotDataTimeProfile-initialize}{\code{PlotDataTimeProfile$new()}}
    \item \href{#method-PlotDataTimeProfile-loadSimulatedResults}{\code{PlotDataTimeProfile$loadSimulatedResults()}}
    \item \href{#method-PlotDataTimeProfile-filterObservedDataForPlot}{\code{PlotDataTimeProfile$filterObservedDataForPlot()}}
    \item \href{#method-PlotDataTimeProfile-addTimeRangeTags}{\code{PlotDataTimeProfile$addTimeRangeTags()}}
    \item \href{#method-PlotDataTimeProfile-splitDataToPanels}{\code{PlotDataTimeProfile$splitDataToPanels()}}
    \item \href{#method-PlotDataTimeProfile-setOrderAndFactors}{\code{PlotDataTimeProfile$setOrderAndFactors()}}
    \item \href{#method-PlotDataTimeProfile-setIndexColumns}{\code{PlotDataTimeProfile$setIndexColumns()}}
    \item \href{#method-PlotDataTimeProfile-getTimeLabelForTimeRange}{\code{PlotDataTimeProfile$getTimeLabelForTimeRange()}}
    \item \href{#method-PlotDataTimeProfile-addPredictedForObserved}{\code{PlotDataTimeProfile$addPredictedForObserved()}}
    \item \href{#method-PlotDataTimeProfile-getDataForTimeRange}{\code{PlotDataTimeProfile$getDataForTimeRange()}}
    \item \href{#method-PlotDataTimeProfile-useColorIndex}{\code{PlotDataTimeProfile$useColorIndex()}}
    \item \href{#method-PlotDataTimeProfile-useShapeIndex}{\code{PlotDataTimeProfile$useShapeIndex()}}
    \item \href{#method-PlotDataTimeProfile-hasSimulatedPop}{\code{PlotDataTimeProfile$hasSimulatedPop()}}
    \item \href{#method-PlotDataTimeProfile-hasObservedData}{\code{PlotDataTimeProfile$hasObservedData()}}
    \item \href{#method-PlotDataTimeProfile-hasObservedDataRange}{\code{PlotDataTimeProfile$hasObservedDataRange()}}
    \item \href{#method-PlotDataTimeProfile-hasReferenceComparison}{\code{PlotDataTimeProfile$hasReferenceComparison()}}
    \item \href{#method-PlotDataTimeProfile-clone}{\code{PlotDataTimeProfile$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-initialize}{}}}
\subsection{\code{PlotDataTimeProfile$new()}}{
  Initialize a new instance of the class
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$new(
  projectConfiguration = projectConfiguration,
  onePlotConfig = onePlotConfig
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{projectConfiguration}}{Object of class `ProjectConfiguration` containing information on paths and file names}
      \item{\code{onePlotConfig}}{plot configuration for one plot}
      \item{\code{dataObserved}}{`data.table` with observed data}
      \item{\code{aggregationFun}}{function to aggregate simulated data}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    QmdPlotManager object
Load simulated results for the specified project configuration
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-loadSimulatedResults"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-loadSimulatedResults}{}}}
\subsection{\code{PlotDataTimeProfile$loadSimulatedResults()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$loadSimulatedResults(
  projectConfiguration,
  aggregationFun,
  scenarioResults
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{projectConfiguration}}{Configuration for the project}
      \item{\code{aggregationFun}}{Function to aggregate simulated data}
      \item{\code{scenarioResults}}{list with simulated scenario results
Filter observed data for plotting}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-filterObservedDataForPlot"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-filterObservedDataForPlot}{}}}
\subsection{\code{PlotDataTimeProfile$filterObservedDataForPlot()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$filterObservedDataForPlot(dataObserved)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{dataObserved}}{`data.table` with observed data}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    `data.table` with filtered observed data for plotting
Replicates data for each time tag
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-addTimeRangeTags"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-addTimeRangeTags}{}}}
\subsection{\code{PlotDataTimeProfile$addTimeRangeTags()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$addTimeRangeTags()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    invisible(NULL)
Creates plotId for each scenario outputs group
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-splitDataToPanels"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-splitDataToPanels}{}}}
\subsection{\code{PlotDataTimeProfile$splitDataToPanels()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$splitDataToPanels(nMaxFacetRows)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{nMaxFacetRows}}{maximal number of facet rows
@title Set factors for plotting}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-setOrderAndFactors"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-setOrderAndFactors}{}}}
\subsection{\code{PlotDataTimeProfile$setOrderAndFactors()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$setOrderAndFactors()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    invisible(NULL)
set columns for aesthetics
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-setIndexColumns"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-setIndexColumns}{}}}
\subsection{\code{PlotDataTimeProfile$setIndexColumns()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$setIndexColumns(referenceScaleVector)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{referenceScaleVector}}{scale vector to scale aesthetic color and fill for scenarios with reference scenario
Get the time label for the filtered time range}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-getTimeLabelForTimeRange"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-getTimeLabelForTimeRange}{}}}
\subsection{\code{PlotDataTimeProfile$getTimeLabelForTimeRange()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$getTimeLabelForTimeRange(filterName)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filterName}}{character with name of time range filter
@title Add predicted values to observed data}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-addPredictedForObserved"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-addPredictedForObserved}{}}}
\subsection{\code{PlotDataTimeProfile$addPredictedForObserved()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$addPredictedForObserved()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    invisible(NULL)
Get the data for the filtered time range
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-getDataForTimeRange"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-getDataForTimeRange}{}}}
\subsection{\code{PlotDataTimeProfile$getDataForTimeRange()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$getDataForTimeRange(
  filterName,
  plotCounter,
  yScale,
  typeFilter = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{filterName}}{name of time range filter}
      \item{\code{plotCounter}}{counter for different plots}
      \item{\code{yScale}}{scale of y-axis}
      \item{\code{typeFilter}}{filter for data type}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    `data.table` with filtered plot data
Function to determine if color legend is needed
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-useColorIndex"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-useColorIndex}{}}}
\subsection{\code{PlotDataTimeProfile$useColorIndex()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$useColorIndex()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Logical
Function to determine if shape legend is needed
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-useShapeIndex"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-useShapeIndex}{}}}
\subsection{\code{PlotDataTimeProfile$useShapeIndex()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$useShapeIndex()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Logical
Function to determine if simulation is aggregated
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-hasSimulatedPop"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-hasSimulatedPop}{}}}
\subsection{\code{PlotDataTimeProfile$hasSimulatedPop()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$hasSimulatedPop()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Logical
Function to determine if data contains observed data
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-hasObservedData"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-hasObservedData}{}}}
\subsection{\code{PlotDataTimeProfile$hasObservedData()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$hasObservedData()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Logical
Function to determine if data contains observed data range
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-hasObservedDataRange"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-hasObservedDataRange}{}}}
\subsection{\code{PlotDataTimeProfile$hasObservedDataRange()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$hasObservedDataRange()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Logical
Function to determine if a reference population should be plotted
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-hasReferenceComparison"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-hasReferenceComparison}{}}}
\subsection{\code{PlotDataTimeProfile$hasReferenceComparison()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$hasReferenceComparison()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Logical
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotDataTimeProfile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotDataTimeProfile-clone}{}}}
\subsection{\code{PlotDataTimeProfile$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotDataTimeProfile$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
