% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{plotDemographics}
\alias{plotDemographics}
\title{Plot Demographics}
\usage{
plotDemographics(
  projectConfiguration,
  onePlotConfig,
  pkParameterDT = NULL,
  scenarioList = NULL,
  asStepPlot = TRUE,
  aggregationFlag = c("Percentiles", "GeometricStdDev", "ArithmeticStdDev", "Custom"),
  customFunction = NULL,
  percentiles = opsuite.plots::getOspsuite.plots.option(optionKey =
    OptionKeys$Percentiles)[c(1, 3, 5)],
  facetAspectRatio = 0.5,
  colorVector = c(scenario = NA, referenceScenario = NA),
  nMaxFacetRows = 2,
  ...
)
}
\arguments{
\item{projectConfiguration}{A configuration object for the project.}

\item{onePlotConfig}{A configuration for the specific plot.}

\item{pkParameterDT}{A data.table containing PK parameter data (optional).}

\item{scenarioList}{A list of scenarios to consider (optional).}

\item{asStepPlot}{Logical indicating if the plot should be a step plot (default is TRUE).}

\item{aggregationFlag}{A character vector indicating the type of aggregation function to use
for simulated data. Options include "GeometricStdDev" (Default), "ArithmeticStdDev", "Percentiles",
and "Custom".}

\item{customFunction}{An optional custom function for aggregation.
A custom function should take a numeric vector `y` as input and return a list containing:

- `yValues`: The aggregated value (e.g., mean).

- `yMin`: The lower value of the aggregated data, (e.g. mean - sd).

- `yMax`: The upper value of the aggregated data, (e.g. mean + sd).

- `yErrorType`: A string indicating the type of error associated with the aggregation,
it is used in plot legends and captions. It must be a concatenation of the descriptor of yValues and the descriptor of yMin - yMax range
separated by "|" (e.g., "mean | standard deviation" or "median | 5th - 95th percentile").}

\item{percentiles}{A numeric vector of percentiles to consider if the aggregation method is set
to "Percentiles" (default is c(5, 50, 95)).}

\item{facetAspectRatio}{A numeric value for the aspect ratio of the facets (default is 0.5).}

\item{colorVector}{A named vector for colors corresponding to scenarios (default is c(scenario = NA, referenceScenario = NA)).}

\item{nMaxFacetRows}{Maximum number of facet rows (default is 2).}

\item{...}{Additional arguments passed to plotting function ospsuite.plots::plotHistogram or for rangePlots plotRangeDistribution.}
}
\value{
A list of plot objects generated based on the input configuration.
}
\description{
This function generates demographic plots based on the provided configuration and data.
}
