% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{plotForest}
\alias{plotForest}
\title{Create a Forest Plot}
\usage{
plotForest(
  plotData,
  mapping = aes(y = y, x = x, groupby = dataType),
  xLabel,
  yFacetColumns = NULL,
  xFacetColumn = NULL,
  xscale = c("linear", "log"),
  xscale.args = list(),
  groupAesthetics = c("color", "fill", "shape"),
  tableColumns = c("yValues", "yErrorValues"),
  tableLabels = c("M", "Variance"),
  labelWrapWidth = 10,
  digitsToRound = 2,
  digitsToShow = 2,
  withTable = is.null(xFacetColumn),
  geomPointAttributes = getDefaultGeomAttributes("Point"),
  geomErrorbarAttributes = getDefaultGeomAttributes("Errorbar"),
  facetScales = c("free_y", "free")
)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted. Must include columns specified in `yFacetColumns`, `xFacetColumn`, `tableColumns`, and others.}

\item{xLabel}{A string representing the label for the x-axis.}

\item{yFacetColumns}{A character vector of column names used for faceting on the y-axis. Can be NULL or of length up to 2.}

\item{xFacetColumn}{A character string specifying the column name for the x-axis facet. Must be of length 1 or NULL.}

\item{xscale}{A character string indicating the scale type for the x-axis. Options are "linear" or "log".}

\item{xscale.args}{A list of additional arguments for customizing the x-axis scale.}

\item{tableColumns}{A character vector of column names to be included in the table.}

\item{tableLabels}{A character vector of labels corresponding to `tableColumns`.}

\item{labelWrapWidth}{A numeric value specifying the width for label wrapping in facets.}

\item{digitsToRound}{An integer specifying the number of digits to round in the table.}

\item{digitsToShow}{An integer specifying the number of digits to display in the table.}

\item{withTable}{A logical flag indicating whether to include the table in the output. Defaults to TRUE if `xFacetColumn` is not NULL.}

\item{yColumn}{A character string specifying the name of the column to be used for the y-axis.}
}
\value{
A combined plot object containing the forest plot and the table (if applicable).
}
\description{
This function generates a forest plot with optional faceting and a corresponding table.
}
