% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotOfOnePanel}
\alias{plotOfOnePanel}
\title{function to generate one plot as facet panels}
\usage{
plotOfOnePanel(
  projectConfiguration,
  onePlotConfig,
  dataObserved,
  nFacetColumns,
  rmdContainer,
  facetAspectRatio
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{rmdContainer}{object of class `rmdContainer`}

\item{observedData}{`data.table` format corresponds to tables  generated by \code{\link{readDataByDictionary}}}
}
\value{
object of class `rmdContainer` with added figures
}
\description{
function to generate one plot as facet panels
}
