% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotOfOneTimeProfilePanel}
\alias{plotOfOneTimeProfilePanel}
\title{function to generate one time profile plot as facet panels}
\usage{
plotOfOneTimeProfilePanel(
  projectConfiguration,
  panelConfig,
  observedData,
  nFacetColumns,
  rmdContainer,
  facetAspectRatio
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{panelConfig}{part of configuration table which defines the plot}

\item{observedData}{`data.table` format corresponds to tables  generated by \code{\link{readDataByDictionary}}}

\item{rmdContainer}{object of class `rmdContainer`}
}
\value{
object of class `rmdContainer` with added figures
}
\description{
function to generate one time profile plot as facet panels
}
