% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{plotPKForestAbsoluteValuesWithVariance}
\alias{plotPKForestAbsoluteValuesWithVariance}
\title{PK Forest Plots}
\usage{
plotPKForestAbsoluteValuesWithVariance(
  projectConfiguration,
  onePlotConfig,
  pkParameterDT,
  pkParameterObserved = NULL,
  aggregationFlag = c("GeometricStdDev", "ArithmeticStdDev", "Percentiles"),
  percentiles = getOspsuite.plots.option(optionKey = OptionKeys$Percentiles)[c(1, 3, 5)],
  scaleVectors = list(),
  labelWrapWidth = 10,
  vlineIntercept = NULL,
  withTable = TRUE,
  relWidths = NULL,
  digitsToRound = 2,
  digitsToShow = 2
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing the
necessary settings and file paths for the project.}

\item{onePlotConfig}{A data.table containing configuration settings for a
single plot, including plot name, x-axis scale, and other aesthetic settings.}

\item{pkParameterDT}{A data.table containing simulated PK parameter data,
including columns for scenario names, parameters, values, and output paths.}

\item{pkParameterObserved}{Optional data.table containing observed PK parameter
data for comparison, which can include columns for observed values and associated
metadata.}

\item{aggregationFlag}{A character string indicating the method of aggregation
 for variance calculation. Options include "GeometricStdDev", "ArithmeticStdDev",
 and "Percentiles".
(for `plotPKForestAbsoluteValuesWithVariance` and `plotPKForestRatiosWithVariance`):}

\item{percentiles}{A numeric vector specifying which percentiles to calculate
and display in the plot, only relevant if `aggregationFlag = "Percentiles"`
(for `plotPKForestAbsoluteValuesWithVariance` and `plotPKForestRatiosWithVariance`):}

\item{scaleVectors}{A list containing user-defined scale vectors for the plot aesthetics
for simulated and observed data. The list can have up to two components: `simulated` and
`observed`, each containing a list of properties (`color`, `fill`, `shape`) to modify the
default scale vector settings.}

\item{labelWrapWidth}{Numeric value specifying the maximum width for wrapping
labels in the plot to enhance readability.}

\item{vlineIntercept}{Optional numeric value indicating where to draw a vertical
line on the plot, often used to denote a reference value.}

\item{withTable}{logical, if TRUE (default) values are displaye as table beside the plot}

\item{relWidths}{Optional numeric vector specifying relative widths for the plot and table.}

\item{digitsToRound}{An integer specifying the number of digits to round in the
displayed values.}

\item{digitsToShow}{An integer specifying the number of digits to show in the
displayed values.}
}
\value{
A list containing the generated plot objects.
}
\description{
This is one of a group of functions which generates PK forest plots for pharmacokinetic (PK) parameters, providing visualizations
that can include absolute values or ratios, along with options for displaying variance
or confidence intervals. These functions allow for the comparison of simulated and observed
PK parameter data across different scenarios, aiding in the interpretation of variability
and uncertainty in the estimates.

Available functions are:

\describe{
  \item{\code{plotPKForestAbsoluteValuesWithVariance}}{Generates a PK forest plot displaying absolute values along with variance.}
  \item{\code{plotPKForestAbsoluteValuesWithCI}}{Generates a PK forest plot displaying absolute values along with confidence intervals using bootstrapping.}
  \item{\code{plotPKForestRatiosWithVariance}}{Generates a PK forest plot displaying ratios of PK parameters along with variance.}
  \item{\code{plotPKForestRatiosWithCI}}{Generates a PK forest plot displaying ratios of PK parameters along with confidence intervals using bootstrapping.}
}
}
\details{
For the ratio calculation, the function distinguishes between two cases:


\describe{
  \item{Case 1:}{If the scenarios being compared are based on the \strong{same populations}
  (same `PopulationId` in scenario configuration), the plots will show summary statistics of
  individual ratios.}
  \item{Case 2:}{If the scenarios are based on \strong{different populations}, the plots
  will display the ratio of summary statistics, using bootstrapping methods. During each bootstrapping
  step, `nSamples` are drawn and aggregated. `nSamples` is the minimum number of individuals from both populations.
  Displayed are the median values of the aggregated results.}
}

To support the creation of the configuration table a support function \code{addDefaultConfigForPKForestPlots} is available

For the bootstrapping process, a unique seed is set for each combination of scenario, referenceScenario (if available),
outputPathId, and PKParameter identifier. This ensures that results are consistent when a specific
combination is used in different plots and that they are reproducible.
}
\seealso{
\code{\link{plotPKForestAbsoluteValuesWithVariance}}, \code{\link{plotPKForestAbsoluteValuesWithCI}},
\code{\link{plotPKForestRatiosWithVariance}}, \code{\link{plotPKForestRatiosWithCI}},
\code{\link{addDefaultConfigForPKForestPlots}}
}
