% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKRatioForestPlot.R
\name{plotPKRatioForestPlot}
\alias{plotPKRatioForestPlot}
\title{Plot PK Ratio}
\usage{
plotPKRatioForestPlot(
  projectConfiguration,
  subfolder,
  configTableSheet,
  pkParameterDT,
  coefficientOfVariation,
  nObservationDefault,
  nBootstrap = 1000,
  statFun = function(x) exp(mean(log(x))),
  confLevel = 0.9,
  seed = 123,
  vlineIntercept = 1,
  scaleVactors = list(simulated = list(color = "darkgrey", fill = "darkgrey", shape =
    "circle"), observed = list(color = "darkgrey", fill = "lightgrey", shape =
    "triangle filled")),
  digitsToRound = 2,
  digitsToShow = 2,
  xlabel = "DDI Ratio"
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{subfolder}{A string specifying the subfolder for output.}

\item{configTableSheet}{Name of the sheet containing configuration tables.}

\item{pkParameterDT}{A data.table with PK parameters.}

\item{coefficientOfVariation}{Coefficient of variation for simulations.}

\item{nObservationDefault}{Default number of observations.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{statFun}{Function to compute statistics.}

\item{confLevel}{Confidence level for intervals.}

\item{seed}{Random seed for reproducibility.}

\item{vlineIntercept}{Intercept for vertical line on plots.}

\item{scaleVactors}{A list of aesthetic parameters for plots.}

\item{digitsToRound}{Number of digits to round in output.}

\item{digitsToShow}{Number of digits to display.}

\item{xlabel}{Label for x-axis.}
}
\value{
An RmdContainer object with generated plots.
}
\description{
This function generates plots for PK ratios based on provided configurations.
}
\keyword{internal}
