% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{plotPKRatioForestPlotByBoostrapping}
\alias{plotPKRatioForestPlotByBoostrapping}
\title{Plot PK Ratio by Bootstrapping}
\usage{
plotPKRatioForestPlotByBoostrapping(
  projectConfiguration,
  subfolder,
  configTableSheet,
  pkParameterDT,
  pkParameterObserved = NULL,
  coefficientOfVariation,
  nObservationDefault,
  nBootstrap = 1000,
  statFun = function(x) exp(mean(log(x))),
  confLevel = 0.9,
  seed = 123,
  vlineIntercept = 1,
  scaleVectors = list(simulated = list(color = "darkgrey", fill = "darkgrey", shape =
    "circle"), observed = list(color = "darkgrey", fill = "lightgrey", shape =
    "triangle filled")),
  digitsToRound = 2,
  digitsToShow = 2,
  xlabel = "DDI Ratio",
  tableLabels = c("Ratio", paste0(confLevel * 100, "\%\\nCI lower"), paste0(confLevel *
    100, "\%\\nCI upper"))
)
}
\arguments{
\item{subfolder}{A string specifying the output directory for the generated plots.}

\item{configTableSheet}{A string representing the name of the sheet that contains configuration tables.}

\item{pkParameterDT}{A data.table with PK parameters, including required columns for analysis.}

\item{pkParameterObserved}{A data.table with observed PK parameters. This parameter is optional.}

\item{coefficientOfVariation}{A numeric value representing the coefficient of variation for simulations.}

\item{nObservationDefault}{An integer representing the default number of observations for simulations.}

\item{nBootstrap}{An integer specifying the number of bootstrap samples to generate.}

\item{statFun}{A function to compute statistics from the bootstrap samples. Default is the geometric mean.}

\item{confLevel}{A numeric value representing the confidence level for intervals (e.g., 0.9 for 90%).}

\item{seed}{An integer for setting the random seed to ensure reproducibility of results.}

\item{vlineIntercept}{A numeric value for the intercept of the vertical line on plots.}

\item{scaleVectors}{A list of aesthetic parameters for the plots, including colors and shapes.}

\item{digitsToRound}{An integer specifying the number of digits to round in output.}

\item{digitsToShow}{An integer specifying the number of digits to display in plot annotations.}

\item{xlabel}{A string representing the label for the x-axis in the plots.}

\item{tableLabels}{A character vector of labels for tables generated in the plot.}

\item{description}{projectConfiguration A ProjectConfiguration object containing project settings and parameters.}
}
\value{
An RmdPlotManager object containing the generated plots and related information for further
        processing or exporting.
}
\description{
This function generates plots for PK ratios using bootstrapping methods. It allows users to visualize
the variability and uncertainty in PK ratios based on simulated data.
}
