% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{plotPKRatioForestPlotByRatioAggregation}
\alias{plotPKRatioForestPlotByRatioAggregation}
\title{Plot PK Ratio by Ratio Aggregation}
\usage{
plotPKRatioForestPlotByRatioAggregation(
  projectConfiguration,
  subfolder,
  configTableSheet,
  pkParameterDT,
  pkParameterObserved = NULL,
  aggregationFlag = c("GeometricStdDev", "ArithmeticStdDev", "Percentiles", "Custom"),
  percentiles = c(0.05, 0.5, 0.95),
  customFunction = NULL,
  vlineIntercept = 1,
  scaleVectors = list(simulated = list(color = "darkgrey", fill = "darkgrey", shape =
    "circle"), observed = list(color = "darkgrey", fill = "lightgrey", shape =
    "triangle filled")),
  digitsToRound = 2,
  digitsToShow = 2,
  xlabel = "Ratio",
  tableLabels = NULL
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project settings and parameters.}

\item{subfolder}{A string specifying the output directory where the plots will be saved.}

\item{configTableSheet}{A string representing the name of the sheet that contains configuration tables.}

\item{pkParameterDT}{A data.table with PK parameters, including required columns for analysis.}

\item{pkParameterObserved}{A data.table with observed PK parameters. This parameter is optional.}

\item{aggregationFlag}{A character string indicating the aggregation method. Options include:
- "GeometricStdDev"
- "ArithmeticStdDev"
- "Percentiles"
- "Custom"}

\item{percentiles}{A numeric vector of percentiles to calculate if aggregationFlag is "Percentiles".
Default is c(5, 50, 95).}

\item{customFunction}{A custom function for aggregation if aggregationFlag is "Custom". Default is NULL.
       The function should take a numeric vector `y` as input and return a list containing:
- `yValues`: The aggregated value (e.g., mean).
- `yMin`: The lower value of the aggregated data (e.g., mean - sd).
- `yMax`: The upper value of the aggregated data (e.g., mean + sd).}

\item{vlineIntercept}{A numeric value for the intercept of the vertical line on plots.}

\item{scaleVectors}{A list of aesthetic parameters for the plots, including colors and shapes.}

\item{digitsToRound}{An integer specifying the number of digits to round in output.}

\item{digitsToShow}{An integer specifying the number of digits to display in plot annotations.}

\item{xlabel}{A string representing the label for the x-axis in the plots.}

\item{tableLabels}{A list of labels for tables generated in the plot.}
}
\value{
An RmdPlotManager object containing the generated plots and related information for further
        processing or exporting.
}
\description{
This function generates plots for pharmacokinetic (PK) ratios using various aggregation methods.
It allows users to visualize the ratios based on different statistical techniques, making it easier
to interpret the pharmacokinetic data.
}
\details{
The function supports multiple aggregation methods, including:
- Geometric Standard Deviations
- Arithmetic Standard Deviations
- Percentiles
- A user-defined custom function for aggregation
}
