% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotPredictedVsObservedPanel}
\alias{plotPredictedVsObservedPanel}
\title{plot predicted vs observed}
\usage{
plotPredictedVsObservedPanel(plotData, rmdContainer)
}
\value{
RmdContainer object with added plot
}
\description{
plot predicted vs observed
}
