% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotQQPanel}
\alias{plotQQPanel}
\title{plot residuals as QQ-plot}
\usage{
plotQQPanel(plotData, rmdContainer)
}
\value{
RmdContainer object with added plot
}
\description{
plot residuals as QQ-plot
}
