% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotResidualsAsHistogramPanel}
\alias{plotResidualsAsHistogramPanel}
\title{plot residuals as histogram}
\usage{
plotResidualsAsHistogramPanel(plotData, rmdContainer, facetAspectRatio)
}
\value{
RmdContainer object with added plot
}
\description{
plot residuals as histogram
}
