% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotResidualsVsObservedPanel}
\alias{plotResidualsVsObservedPanel}
\title{plot residuals vs Observed}
\usage{
plotResidualsVsObservedPanel(plotData, rmdContainer, facetAspectRatio)
}
\value{
RmdContainer object with added plot
}
\description{
plot residuals vs Observed
}
