% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotResidualsVsTimePanel}
\alias{plotResidualsVsTimePanel}
\title{plot residuals ve Time}
\usage{
plotResidualsVsTimePanel(plotData, rmdContainer, facetAspectRatio)
}
\description{
plot residuals ve Time
}
