% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotTimeProfilePanels}
\alias{plotTimeProfilePanels}
\title{function to generate list of time profiles as facet panels}
\usage{
plotTimeProfilePanels(
  projectConfiguration,
  subfolder,
  configTableSheet,
  observedData,
  nFacetColumns = 2,
  facetAspectRatio = 0.5
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{subfolder}{is name of subfolder in the project/to/path/report directory}

\item{configTableSheet}{name of sheet in plot configuration xlsx, which defines the plot}

\item{observedData}{either `data.table` which was generated by \code{\link{readDataByDictionary}} or object of class `DataCombined`,
which will internally transfer to the data.table format by \code{\link{convertDataCombinedToDataTable}}}

\item{nFacetColumns}{maximal number of facet columns (default 2) used in for facet type "by Order"}
}
\value{
object of class `rmdContainer`
}
\description{
function to generate list of time profiles as facet panels
}
