% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotTPPanel}
\alias{plotTPPanel}
\title{create the TimeProfile panels}
\usage{
plotTPPanel(
  projectConfiguration,
  onePlotConfig,
  plotData,
  dtPlotId,
  rmdContainer,
  nFacetColumns,
  facetAspectRatio
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{facetAspectRatio}{}
}
\description{
create the TimeProfile panels
}
