% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{prepareConfigTableExport}
\alias{prepareConfigTableExport}
\title{Prepare Configuration Table}
\usage{
prepareConfigTableExport(configTable)
}
\arguments{
\item{configTable}{A data.table containing configuration details for scenarios.}
}
\value{
A prepared configuration table for export.
}
\description{
This function prepares the configuration table for export by merging it with
data group and scenario definitions.
}
\keyword{internal}
