% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{prepareDataForAggregation}
\alias{prepareDataForAggregation}
\title{Prepares data for aggregation}
\usage{
prepareDataForAggregation(dataObserved, groups, groupSuffix)
}
\arguments{
\item{dataObserved}{A `data.table` containing observed data.}

\item{groups}{A character vector specifying the groups to aggregate.
If NULL, all available groups are used.}

\item{groupSuffix}{A character string to append to group names in the aggregated output.
Default is 'aggregated'.}
}
\value{
Prepared data for aggregation as a `data.table`.
}
\description{
Prepares data for aggregation
}
\keyword{internal}
