% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{prepareDataForMatch}
\alias{prepareDataForMatch}
\title{Prepare Data for Matching}
\usage{
prepareDataForMatch(projectConfiguration, dataObserved, scenarioList)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details, including data groups and output paths.}

\item{dataObserved}{A data.table containing observed data. It must include columns for 'group', 'yValues', and 'yUnit'.}

\item{scenarioList}{A list of scenarios, each containing simulation details relevant to the observed data.}
}
\value{
A data.table containing the prepared observed data, with adjusted values and units for matching with simulation results.
}
\description{
This function prepares observed data for matching with simulation results based on project configuration and scenario definitions.
}
\details{
The function checks for valid scenarios, merges data tables, calculates conversion factors, and ensures that the data class matches the expected format.
}
