% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKBoxwhisker.R
\name{prepareDataForPKBoxplot}
\alias{prepareDataForPKBoxplot}
\title{Prepare Data for PK Boxplot}
\usage{
prepareDataForPKBoxplot(onePlotConfig, pkParameterDT, colorVector, asRatio)
}
\arguments{
\item{onePlotConfig}{Configuration for a single plot.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{colorVector}{A named vector for colors.}

\item{asRatio}{boolean indicating if the plot is for ratios.}
}
\value{
A data.table prepared for plotting.
}
\description{
Prepares and cleans data for box-and-whisker plotting.
}
\keyword{internal}
