% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{prepareDataForPKRatioPlot}
\alias{prepareDataForPKRatioPlot}
\title{Prepare Data for DDI Ratio Plot}
\usage{
prepareDataForPKRatioPlot(
  onePlotConfig,
  pkParameterDT,
  pkParameterObserved,
  ratioCalculationMethod,
  ratioCalculationInputs
)
}
\arguments{
\item{onePlotConfig}{A data.table containing the plot configuration for the current plot.}

\item{pkParameterDT}{A data.table with PK parameters that will be used for ratio calculations.}

\item{pkParameterObserved}{A data.table containing observed PK parameters. This parameter is optional.}

\item{ratioCalculationMethod}{A string specifying the method used for calculating ratios (e.g., 'byBootsTrapping').}

\item{ratioCalculationInputs}{A list containing inputs specific to the method used for calculating ratios.}
}
\value{
A data.table containing the prepared plot data, including calculated ratios and relevant
        statistics for plotting.
}
\description{
Prepares the necessary data for generating a DDI ratio plot. This function handles the merging
of configuration data with observed PK parameters and calculates the ratios based on the specified
method.
}
\keyword{internal}
