% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markdown.R
\name{prepareDTForPrinting}
\alias{prepareDTForPrinting}
\title{Prepare a data.table for prettier printing by converting line breaks, large
number and percentages}
\usage{
prepareDTForPrinting(
  dt,
  replaceLinebreaksIn = "!guess",
  replaceLinebreaksBy = " ",
  convertToComma = "!guess",
  commaAccuracy = 1,
  convertToPercent = "!guess",
  percentAccuracy = 0.1,
  convertToYesNo = "!guess",
  skip = NULL
)
}
\arguments{
\item{dt}{`data.frame`, will be converted to data.table}

\item{replaceLinebreaksIn}{`character` vector naming the columns of `dt` in
which line breaks will be replaced. `!guess` selects all character columns
of `dt`}

\item{replaceLinebreaksBy}{`string` specifying how to replace line breaks}

\item{convertToComma}{`character` vector naming the columns of `dt` to which
`scales::comma` will be applied. `!guess` selects all integer columns of
`dt`}

\item{commaAccuracy}{`number` specifying the rounding by `scales::comma`}

\item{convertToPercent}{`character` vector naming the columns of `dt` to
which `scales::percent` will be applied. `!guess` selects all numeric,
non-integer columns of `dt` with values between 0 and 1}

\item{percentAccuracy}{`number` specifying the rounding of `scales::percent`}

\item{convertToYesNo}{`character` vector naming the columns of `dt` which
will be converted to "yes"/"no" depending on their logical status. `!guess`
selects all logical columns of `dt`}

\item{skip}{`character` vector naming columns of `dt` that won't be modified}
}
\value{
`data.table` prepared for display
}
\description{
Prepare a data.table for prettier printing by converting line breaks, large
number and percentages
}
