% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{prepareSensitivityPlotData}
\alias{prepareSensitivityPlotData}
\title{Prepare Sensitivity Plot Data}
\usage{
prepareSensitivityPlotData(onePlotConfig, projectConfiguration, scenarioList)
}
\arguments{
\item{onePlotConfig}{A data frame containing the configuration for a single plot.}

\item{projectConfiguration}{A list containing project configuration settings.}

\item{scenarioList}{A list of scenarios for which the sensitivity analysis will be performed.}
}
\value{
A data table containing the prepared data for the sensitivity plot.
}
\description{
Prepares the data required for creating a sensitivity plot based on the provided configuration.
}
\keyword{internal}
