% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{prepareTestProject}
\alias{prepareTestProject}
\title{Run Example 1 of the Reporting Framework}
\usage{
prepareTestProject(rootDirectory)
}
\arguments{
\item{rootDirectory}{A character string specifying the root directory for the test project.}
}
\value{
A list containing:
  - `projectConfiguration`: The configuration of the initialized project.
  - `scenarioList`: A list of scenarios initialized for the project.
  - `scenarioResults`: The results from running the scenarios.
}
\description{
This function sets up the environment for running Example 1 of the
Reporting Framework, following the workflow outlined in the tutorial 'Tutorial-timeprofiles'.
It creates a directory structure if it does not exist, and sources
the necessary R scripts to execute the workflow.
}
\details{
Build Test Project Directory

This function creates a test project directory for users to experiment with examples
in the `ospsuite.reportingframework` package.
}
\examples{
\dontrun{
prepareTestProject("path/to/my/testDirectory")
}
}
