% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{processPKAnalyses}
\alias{processPKAnalyses}
\title{Process PK Analyses with Parameter Definitions}
\usage{
processPKAnalyses(pkAnalyses, dtPkParameterDefinition, dtOutputPaths)
}
\arguments{
\item{pkAnalyses}{A data.table containing PK analyses.}

\item{dtPkParameterDefinition}{A data.table containing PK parameter definitions.}

\item{dtOutputPaths}{A data.table containing output paths.}
}
\value{
A processed data.table containing PK analyses with parameters.
}
\description{
This function processes the PK analyses using the provided parameter definitions.
}
\keyword{internal}
