% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{processPKDefinition}
\alias{processPKDefinition}
\title{Process PK Parameter Definitions}
\usage{
processPKDefinition(
  scenarioResult,
  dtOutputPaths,
  dtPkAnalyses,
  dtPkParameterDefinition
)
}
\arguments{
\item{scenarioResult}{A list containing the results of the scenario.}

\item{dtOutputPaths}{A data.table containing output paths.}

\item{dtPkAnalyses}{A data.table containing PK analyses.}

\item{dtPkParameterDefinition}{A data.table containing PK parameter definitions.}
}
\value{
A data.table containing processed PK parameter definitions for the scenario.
}
\description{
This function processes the PK parameter definitions for a specific scenario, including unit conversions and output filters.
}
