% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{processSheetData}
\alias{processSheetData}
\title{Process the data read from the sheet}
\usage{
processSheetData(
  dt,
  skipDescriptionRow,
  alwaysCharacter,
  emptyAsNA,
  convertHeaders
)
}
\arguments{
\item{dt}{A `data.table` containing the raw data.}

\item{skipDescriptionRow}{A logical value or an integer indicating whether to skip description rows.}

\item{alwaysCharacter}{A character vector with column names or regex patterns that should be returned as character.}

\item{emptyAsNA}{A logical value. If TRUE, empty strings in character columns are converted to NA.
If FALSE NA is returned as empty string. Numeric columns Return always NA}

\item{convertHeaders}{A logical value. If TRUE, column names are converted to start with a lowercase letter.}
}
\value{
A processed `data.table`.
}
\description{
This helper function processes the data by skipping description rows, converting specified columns to character,
converting empty strings to NA, and adjusting column headers.
}
\keyword{internal}
