% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QmdPlotManager.R
\docType{class}
\name{QmdPlotManager}
\alias{QmdPlotManager}
\title{QmdPlotManager}
\description{
Manages the creation and writing of .qmd files for plots.
}
\keyword{internal}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{QmdPlotManager}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{digitsOfSignificance}}{Digits for significance in table display.}

    \item{\code{configTable}}{Configuration table.}

    \item{\code{plotFunction}}{Function to create plot list.}

    \item{\code{validateConfigTableFunction}}{Function to read config table.}

    \item{\code{suppressExport}}{A logical value indicating whether to suppress export.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-QmdPlotManager-initialize}{\code{QmdPlotManager$new()}}
    \item \href{#method-QmdPlotManager-writeQmd}{\code{QmdPlotManager$writeQmd()}}
    \item \href{#method-QmdPlotManager-addHeader}{\code{QmdPlotManager$addHeader()}}
    \item \href{#method-QmdPlotManager-addNewline}{\code{QmdPlotManager$addNewline()}}
    \item \href{#method-QmdPlotManager-addNewpage}{\code{QmdPlotManager$addNewpage()}}
    \item \href{#method-QmdPlotManager-exportPlotList}{\code{QmdPlotManager$exportPlotList()}}
    \item \href{#method-QmdPlotManager-addAndExportFigure}{\code{QmdPlotManager$addAndExportFigure()}}
    \item \href{#method-QmdPlotManager-addAndExportTable}{\code{QmdPlotManager$addAndExportTable()}}
    \item \href{#method-QmdPlotManager-clone}{\code{QmdPlotManager$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-initialize}{}}}
\subsection{\code{QmdPlotManager$new()}}{
  Initialize a new instance of the class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$new(
  qmdfolder,
  qmdName,
  nameOfplotFunction,
  suppressExport = FALSE,
  digitsOfSignificance = 3
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{qmdfolder}}{Folder where the .qmd file should be saved.}
      \item{\code{qmdName}}{A character string for the name of the .qmd file (without extension).}
      \item{\code{nameOfplotFunction}}{The name of the plot-function as character.}
      \item{\code{suppressExport}}{A logical value indicating whether to suppress export. Default is FALSE.}
      \item{\code{digitsOfSignificance}}{Number of significant digits to display in tables.
#'}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    An instance of the QmdPlotManager object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-writeQmd"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-writeQmd}{}}}
\subsection{\code{QmdPlotManager$writeQmd()}}{
  Write the .qmd file.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$writeQmd(fileName = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fileName}}{Name of the .qmd file. If NULL, the file name will default to the subfolder name.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. This function writes the .qmd file to the specified location and returns nothing.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-addHeader"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-addHeader}{}}}
\subsection{\code{QmdPlotManager$addHeader()}}{
  Insert a heading with a specified level.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$addHeader(..., level = 1, newlines = 2)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Arguments passed to `mdPaste`.}
      \item{\code{level}}{The header level, i.e., the number of `#`s. Defaults to 1.}
      \item{\code{newlines}}{The number of newlines inserted after the heading. Defaults to 2.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function modifies the internal qmd lines.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-addNewline"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-addNewline}{}}}
\subsection{\code{QmdPlotManager$addNewline()}}{
  Insert line endings and start a new line.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$addNewline(n = 1)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{Number of new lines. Defaults to 1.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function modifies the internal qmd lines.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-addNewpage"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-addNewpage}{}}}
\subsection{\code{QmdPlotManager$addNewpage()}}{
  Insert a page break and a newline.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$addNewpage()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function modifies the internal qmd lines.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-exportPlotList"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-exportPlotList}{}}}
\subsection{\code{QmdPlotManager$exportPlotList()}}{
  Export a list of plots.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$exportPlotList(plotList)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plotList}}{A list of plots to export.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function exports the plots to the specified location.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-addAndExportFigure"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-addAndExportFigure}{}}}
\subsection{\code{QmdPlotManager$addAndExportFigure()}}{
  Add and export a figure with caption and footnote.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$addAndExportFigure(
  plotObject,
  caption,
  figureKey,
  footNoteLines = NULL,
  exportArguments = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plotObject}}{A ggplot object to export.}
      \item{\code{caption}}{A character string for the caption text.}
      \item{\code{figureKey}}{A key used to generate file names; it should be unique for the folder.}
      \item{\code{footNoteLines}}{A character string for figure footnotes.}
      \item{\code{exportArguments}}{additional arguments passed on to ospsuite.plots::export}
      \item{\code{...}}{Additional parameters passed to `ospsuite.plots::exportPlot()`.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function exports the figure and its metadata.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-addAndExportTable"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-addAndExportTable}{}}}
\subsection{\code{QmdPlotManager$addAndExportTable()}}{
  Add and export tables with caption and footnote.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$addAndExportTable(
  table,
  caption,
  tableKey,
  footNoteLines = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{table}}{A data.table object to export.}
      \item{\code{caption}}{A character string for the caption text.}
      \item{\code{tableKey}}{A key used to generate filenames; it should be unique for the folder.}
      \item{\code{footNoteLines}}{A character string for table footnotes.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function exports the table and its metadata.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QmdPlotManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QmdPlotManager-clone}{}}}
\subsection{\code{QmdPlotManager$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{QmdPlotManager$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
