% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{readDataDictionary}
\alias{readDataDictionary}
\title{Read data dictionary}
\usage{
readDataDictionary(dictionaryFile, sheet, data)
}
\arguments{
\item{dictionaryFile}{The file containing the data dictionary}

\item{sheet}{The sheet within the data dictionary file}

\item{data}{The data to be used with the dictionary}
}
\value{
The data dictionary
}
\description{
This function reads the data dictionary based on the provided file and sheet.
}
