% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{readObservedDataByDictionary}
\alias{readObservedDataByDictionary}
\title{Read data by dictionary}
\usage{
readObservedDataByDictionary(
  projectConfiguration,
  spreadData = TRUE,
  dataClassType = c("timeprofile", "pkParameter", "numericValues"),
  fileIds = NULL
)
}
\arguments{
\item{projectConfiguration}{An object containing project configuration details, including the path to the data importer configuration file.}

\item{spreadData}{If TRUE, information derived from observed data, such as identifiers and biometrics, is spread to other tables.}

\item{dataClassType}{A character string indicating the type of data class to process. Options are "timeprofile", "pkParameter", or "numericValues".}

\item{fileIds}{A character vector with file identifiers to be selected, if NULL (default) all are selected.}
}
\value{
A `data.table` containing the processed data based on the dictionary. The structure includes relevant columns defined in the data dictionary.
}
\description{
This function reads and processes data based on the provided project configuration.
}
\seealso{
Other observed data processing:
\code{\link[=addBiometricsToConfig]{addBiometricsToConfig()}},
\code{\link[=aggregateObservedDataGroups]{aggregateObservedDataGroups()}},
\code{\link[=convertDataCombinedToDataTable]{convertDataCombinedToDataTable()}},
\code{\link[=convertDataTableToDataCombined]{convertDataTableToDataCombined()}},
\code{\link[=getNumericValue]{getNumericValue()}},
\code{\link[=updateDataGroupId]{updateDataGroupId()}},
\code{\link[=updateOutputPathId]{updateOutputPathId()}},
\code{\link[=validateObservedData]{validateObservedData()}}
}
\concept{observed data processing}
