% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{readOntongenies}
\alias{readOntongenies}
\title{Read Ontogenies from Data}
\usage{
readOntongenies(data)
}
\arguments{
\item{data}{A data frame containing a column named "Protein Ontogenies".}
}
\value{
A list of `MoleculeOntogeny` objects, each representing a protein
and its corresponding ontogeny. Returns NULL if the "Protein Ontogenies"
field is NA.
}
\description{
based on esqlabsR:::.readOntongeniesFromXLS
}
\details{
This function extracts protein ontogeny mappings from the provided data.
It splits the mappings into individual protein-ontogeny pairs and validates
the structure of each pair. Each valid pair is then converted into a
`MoleculeOntogeny` object.
}
\keyword{internal}
