% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{readOrInitializeFileNameDictionary}
\alias{readOrInitializeFileNameDictionary}
\title{Read or Initialize File Name Dictionary}
\usage{
readOrInitializeFileNameDictionary(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{A list containing project-specific configurations.}
}
\value{
A data.table containing the filename dictionary.
}
\description{
This function reads the filename dictionary from the specified path or initializes
an empty one if it does not exist.
}
\keyword{internal}
