% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{readSheetData}
\alias{readSheetData}
\title{Read data from the specified sheet using openxlsx}
\usage{
readSheetData(wb, sheetName)
}
\arguments{
\item{wb}{A workbook object created by `openxlsx::loadWorkbook()`.}

\item{sheetName}{A character string specifying the name of the sheet to read.}
}
\value{
A `data.table` containing the raw data from the sheet.
}
\description{
This helper function reads data from a specified sheet in a workbook.
}
\keyword{internal}
