% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{readTimeprofileConfigTable}
\alias{readTimeprofileConfigTable}
\title{Read the configuration table for time profiles}
\usage{
readTimeprofileConfigTable(projectConfiguration, sheetName, dataObserved)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing the project settings,
including file paths and scenario definitions. This is essential for the function to access the necessary data.}

\item{dataObserved}{A data.table or DataCombined object containing the observed data to be plotted.}
}
\value{
`data.table` with plot configurations.
}
\description{
Read the configuration table for time profiles
}
\keyword{internal}
