% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{readUserDefinedPKParameters}
\alias{readUserDefinedPKParameters}
\title{Read User Defined PK Parameters}
\usage{
readUserDefinedPKParameters(file)
}
\arguments{
\item{file}{The path to the Excel file containing the user-defined PK parameters.}
}
\value{
A data frame of user-defined PK parameters.
throws Error if validation fails.
}
\description{
Reads user-defined pharmacokinetic (PK) parameters from an Excel file.
}
\keyword{internal}
