% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reporting.R
\name{renderWord}
\alias{renderWord}
\title{converts .Rmd file to word}
\usage{
renderWord(
  fileName,
  wordConversionTemplate = NULL,
  doAppendDocToTemplate = TRUE,
  customStyles = list(FigureCaption = NULL, FigureFootnote = NULL, TableCaption = NULL,
    TableFootnote = NULL),
  ...
)
}
\arguments{
\item{fileName}{name of .Rmd file to convert to word format (".docx")}

\item{wordConversionTemplate}{template used for conversion}

\item{doAppendDocToTemplate}{`boolean`, if TRUE and `wordConversionTemplate` is given,
new generated document is added to the template after bookmark REPORTBLOCK}

\item{customStyles}{list of custom styles usable for figure and table captions and footnotes
available list elements for styles are: "FigureCaption", "FigureFootnote", "TableCaption" and "TableFootnote"
The selected styles should be defined in the `wordConversionTemplate`}

\item{...}{passed to render}
}
\description{
converts .Rmd file to word
}
\examples{
\dontrun{
# Example with customstyles for Footnotes
renderWord(
  fileName = "myReport.Rmd",
  wordConversionTemplate = "path/to/template/myTemplate.docx",
  customStyles = list(FigureFootnote = "myFootnoteFormat", TableFootnote = "myFootnoteFormat")
)
}
}
