% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{REPORTING_FRAMEWORK_VERSION_PROPERTY}
\alias{REPORTING_FRAMEWORK_VERSION_PROPERTY}
\title{Property name for the reporting framework version in ProjectConfiguration.xlsx}
\usage{
REPORTING_FRAMEWORK_VERSION_PROPERTY
}
\description{
Property name for the reporting framework version in ProjectConfiguration.xlsx
}
\keyword{internal}
