% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{restructureApplicationTimeByScenarioIndex}
\alias{restructureApplicationTimeByScenarioIndex}
\title{Restructure Application Time by Scenario Index}
\usage{
restructureApplicationTimeByScenarioIndex(applicationTimes, configTable)
}
\arguments{
\item{applicationTimes}{A list of application times.}

\item{configTable}{A data frame containing the configuration data.}
}
\value{
A list of application times by scenario index.
}
\description{
This function restructures application time based on the scenario index.
}
\keyword{internal}
