% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmdContainer.R
\docType{class}
\name{RmdContainer}
\alias{RmdContainer}
\title{RmdContainer}
\value{
RmdContainer object
writes the .Rmd File
}
\description{
An object to create a .Rmd file during a plot function
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{RmdContainer}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{digitsOfSignificance}}{digits for significance in table display}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RmdContainer-new}{\code{RmdContainer$new()}}
\item \href{#method-RmdContainer-writeRmd}{\code{RmdContainer$writeRmd()}}
\item \href{#method-RmdContainer-addHeader}{\code{RmdContainer$addHeader()}}
\item \href{#method-RmdContainer-addNewline}{\code{RmdContainer$addNewline()}}
\item \href{#method-RmdContainer-addNewpage}{\code{RmdContainer$addNewpage()}}
\item \href{#method-RmdContainer-addAndExportFigure}{\code{RmdContainer$addAndExportFigure()}}
\item \href{#method-RmdContainer-addAndExportTable}{\code{RmdContainer$addAndExportTable()}}
\item \href{#method-RmdContainer-clone}{\code{RmdContainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$new(rmdfolder, subfolder)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rmdfolder}}{folder where rmd file should be saved}

\item{\code{subfolder}}{folder name relative to rmd folder where figures and tables are exported}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-writeRmd"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-writeRmd}{}}}
\subsection{Method \code{writeRmd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$writeRmd(fileName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fileName}}{name of .Rmd file
Insert a heading with a specified level.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-addHeader"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-addHeader}{}}}
\subsection{Method \code{addHeader()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$addHeader(..., level = 1, newlines = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to mdPaste}

\item{\code{level}}{The header level, i.e. the number of `#`s. Defaults to 1.}

\item{\code{newlines}}{The number of newlines inserted after the heading. Defaults to 2.
line end and start a new line.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-addNewline"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-addNewline}{}}}
\subsection{Method \code{addNewline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$addNewline(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of new lines. Defaults to 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-addNewpage"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-addNewpage}{}}}
\subsection{Method \code{addNewpage()}}{
Insert a page break and a newline.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$addNewpage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-addAndExportFigure"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-addAndExportFigure}{}}}
\subsection{Method \code{addAndExportFigure()}}{
adds and exports figure with caption and footnote
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$addAndExportFigure(
  plotObject,
  caption,
  figureKey,
  footNoteLines = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plotObject}}{ggplot object to export}

\item{\code{caption}}{character caption text}

\item{\code{figureKey}}{key which is used to generate file names should be unique for folder}

\item{\code{footNoteLines}}{character text of figure footnotes}

\item{\code{...}}{parameters passed to `ospsuite.plots::exportPlot()`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-addAndExportTable"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-addAndExportTable}{}}}
\subsection{Method \code{addAndExportTable()}}{
adds and exports tables with caption and footnote
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$addAndExportTable(
  table,
  caption,
  tableKey,
  footNoteLines = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{data.table object to export}

\item{\code{caption}}{character caption text}

\item{\code{tableKey}}{key which is used to generate filenames should be unique for folder}

\item{\code{footNoteLines}}{character text of figure footnotes}

\item{\code{...}}{parameters passed to `utils::write.csv`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RmdContainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
