% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmdPlotManager.R
\docType{class}
\name{RmdPlotManager}
\alias{RmdPlotManager}
\title{RmdPlotManager}
\description{
Manages the creation and writing of .Rmd files for plots.
}
\keyword{internal}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{RmdPlotManager}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{digitsOfSignificance}}{Digits for significance in table display.}

    \item{\code{configTable}}{Configuration table.}

    \item{\code{plotFunction}}{Function to create plot list.}

    \item{\code{validateConfigTableFunction}}{Function to read config table.}

    \item{\code{suppressExport}}{A logical value indicating whether to suppress export.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-RmdPlotManager-initialize}{\code{RmdPlotManager$new()}}
    \item \href{#method-RmdPlotManager-writeRmd}{\code{RmdPlotManager$writeRmd()}}
    \item \href{#method-RmdPlotManager-addHeader}{\code{RmdPlotManager$addHeader()}}
    \item \href{#method-RmdPlotManager-addNewline}{\code{RmdPlotManager$addNewline()}}
    \item \href{#method-RmdPlotManager-addNewpage}{\code{RmdPlotManager$addNewpage()}}
    \item \href{#method-RmdPlotManager-exportPlotList}{\code{RmdPlotManager$exportPlotList()}}
    \item \href{#method-RmdPlotManager-addAndExportFigure}{\code{RmdPlotManager$addAndExportFigure()}}
    \item \href{#method-RmdPlotManager-addAndExportTable}{\code{RmdPlotManager$addAndExportTable()}}
    \item \href{#method-RmdPlotManager-clone}{\code{RmdPlotManager$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-initialize}{}}}
\subsection{\code{RmdPlotManager$new()}}{
  Initialize a new instance of the class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$new(
  rmdfolder,
  rmdName,
  nameOfplotFunction,
  suppressExport = FALSE,
  digitsOfSignificance = 3
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{rmdfolder}}{Folder where the .Rmd file should be saved.}
      \item{\code{rmdName}}{A character string for the name of the .Rmd file (without extension).}
      \item{\code{nameOfplotFunction}}{The name of the plot-function as character.}
      \item{\code{suppressExport}}{A logical value indicating whether to suppress export. Default is FALSE.}
      \item{\code{digitsOfSignificance}}{Number of significant digits to display in tables.
#'}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    An instance of the RmdPlotManager object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-writeRmd"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-writeRmd}{}}}
\subsection{\code{RmdPlotManager$writeRmd()}}{
  Write the .Rmd file.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$writeRmd(fileName = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fileName}}{Name of the .Rmd file. If NULL, the file name will default to the subfolder name.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. This function writes the .Rmd file to the specified location and returns nothing.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-addHeader"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-addHeader}{}}}
\subsection{\code{RmdPlotManager$addHeader()}}{
  Insert a heading with a specified level.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$addHeader(..., level = 1, newlines = 2)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{Arguments passed to `mdPaste`.}
      \item{\code{level}}{The header level, i.e., the number of `#`s. Defaults to 1.}
      \item{\code{newlines}}{The number of newlines inserted after the heading. Defaults to 2.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function modifies the internal Rmd lines.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-addNewline"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-addNewline}{}}}
\subsection{\code{RmdPlotManager$addNewline()}}{
  Insert line endings and start a new line.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$addNewline(n = 1)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{Number of new lines. Defaults to 1.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function modifies the internal Rmd lines.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-addNewpage"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-addNewpage}{}}}
\subsection{\code{RmdPlotManager$addNewpage()}}{
  Insert a page break and a newline.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$addNewpage()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function modifies the internal Rmd lines.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-exportPlotList"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-exportPlotList}{}}}
\subsection{\code{RmdPlotManager$exportPlotList()}}{
  Export a list of plots.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$exportPlotList(plotList)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plotList}}{A list of plots to export.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function exports the plots to the specified location.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-addAndExportFigure"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-addAndExportFigure}{}}}
\subsection{\code{RmdPlotManager$addAndExportFigure()}}{
  Add and export a figure with caption and footnote.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$addAndExportFigure(
  plotObject,
  caption,
  figureKey,
  footNoteLines = NULL,
  exportArguments = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plotObject}}{A ggplot object to export.}
      \item{\code{caption}}{A character string for the caption text.}
      \item{\code{figureKey}}{A key used to generate file names; it should be unique for the folder.}
      \item{\code{footNoteLines}}{A character string for figure footnotes.}
      \item{\code{exportArguments}}{additional arguments passed on to ospsuite.plots::export}
      \item{\code{...}}{Additional parameters passed to `ospsuite.plots::exportPlot()`.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function exports the figure and its metadata.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-addAndExportTable"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-addAndExportTable}{}}}
\subsection{\code{RmdPlotManager$addAndExportTable()}}{
  Add and export tables with caption and footnote.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$addAndExportTable(
  table,
  caption,
  tableKey,
  footNoteLines = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{table}}{A data.table object to export.}
      \item{\code{caption}}{A character string for the caption text.}
      \item{\code{tableKey}}{A key used to generate filenames; it should be unique for the folder.}
      \item{\code{footNoteLines}}{A character string for table footnotes.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    NULL. The function exports the table and its metadata.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdPlotManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RmdPlotManager-clone}{}}}
\subsection{\code{RmdPlotManager$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RmdPlotManager$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
