% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario-management.R
\name{runAndSaveScenarios}
\alias{runAndSaveScenarios}
\title{Run and save scenarios}
\usage{
runAndSaveScenarios(
  projectConfiguration,
  scenarioList,
  simulationRunOptions = NULL,
  ...
)
}
\arguments{
\item{projectConfiguration}{Configuration for the project, containing paths and settings necessary
to run the simulations and save the results.}

\item{scenarioList}{Named list of Scenario objects to be simulated.}

\item{simulationRunOptions}{Object of type `SimulationRunOptions` that will be passed to simulation runs.
If `NULL`, default options are used.}

\item{...}{Additional arguments passed to `esqlabsR::saveScenarioResults`.}
}
\value{
A list containing the simulation results for each scenario that was run.
}
\description{
This function simulates a list of scenarios and saves the results.
If results already exist for a scenario, it will overwrite them based on the provided options.
}
\examples{
\dontrun{
runAndSaveScenarios(
  projectConfiguration = myProjectConfig,
  scenarioList = myScenarioList,
  simulationRunOptions = myRunOptions
)
}

}
\seealso{
Other scenario management:
\code{\link[=calculatePKParameterForScenarios]{calculatePKParameterForScenarios()}},
\code{\link[=createScenariosWrapped]{createScenariosWrapped()}},
\code{\link[=loadPKParameter]{loadPKParameter()}},
\code{\link[=loadScenarioResultsToFramework]{loadScenarioResultsToFramework()}},
\code{\link[=runOrLoadScenarios]{runOrLoadScenarios()}}
}
\concept{scenario management}
