% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-examples.R
\name{runExample1}
\alias{runExample1}
\title{Run Example 1 of the Reporting Framework}
\usage{
runExample1(rootDirectory = NULL)
}
\arguments{
\item{rootDirectory}{A character string specifying the root directory
where the example files will be created. If NULL,
a temporary directory will be used.}
}
\value{
A character string indicating the path to the root directory
        where the example files were created.
}
\description{
This function sets up the environment for running Example 1 of the
Reporting Framework, following the workflow outlined in the tutorial 'Tutorial-timeprofiles'.
It creates a directory structure if it does not exist, and sources
the necessary R scripts to execute the workflow.
}
