% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{runExample2}
\alias{runExample2}
\title{Run Example 2 of the Reporting Framework}
\usage{
runExample2(rootDirectory = NULL)
}
\arguments{
\item{rootDirectory}{A character string specifying the root directory
where the example files will be created. If NULL,
a temporary directory will be used.}
}
\value{
A character string indicating the path to the root directory
        where the example files were created.
}
\description{
This function sets up the environment for running Example 2 of the
Reporting Framework,
}
