% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{runScenarios.wrapped}
\alias{runScenarios.wrapped}
\title{Run a set of scenarios.}
\usage{
runScenarios.wrapped(scenarioList, ...)
}
\arguments{
\item{scenarioList}{Named list of Scenario objects.}

\item{...}{passed to esqlabsR::runScenarios}
}
\value{
Named list of simulation results
}
\description{
wrap of `esqlabsR::runScenarios`
}
