% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity.R
\name{runSensitivityAnalysisForScenarios}
\alias{runSensitivityAnalysisForScenarios}
\title{Run Sensitivity Analysis for Specified Scenarios}
\usage{
runSensitivityAnalysisForScenarios(
  projectConfiguration,
  scenarioList,
  scenarioNames = NULL,
  sensitivitysheet,
  sensitivityAnalysisRunOptions = SensitivityAnalysisRunOptions$new(showProgress = TRUE),
  overwrite = TRUE
)
}
\arguments{
\item{projectConfiguration}{An object representing the project configuration.}

\item{scenarioList}{A list of scenarios to analyze.}

\item{scenarioNames}{A vector of scenario names to run the analysis on. Defaults to NULL.}

\item{sensitivitysheet}{The name of the sheet in the sensitivity Excel file to use for analysis.}

\item{sensitivityAnalysisRunOptions}{Options for running the sensitivity analysis.
Defaults to an instance of SensitivityAnalysisRunOptions with showProgress set to TRUE.}

\item{overwrite}{A logical indicating whether to overwrite existing results. Defaults to TRUE.}
}
\description{
This function executes sensitivity analysis for the specified scenarios in the project configuration.
It checks for the presence of the sensitivity file and runs the analysis, saving the results to CSV files.
}
