% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{saveSessionInfo}
\alias{saveSessionInfo}
\title{Save session information to the log}
\usage{
saveSessionInfo()
}
\description{
Save session information to the log
}
\examples{
\dontrun{
saveSessionInfo()
}

}
\seealso{
Other log file management:
\code{\link[=addMessageToLog]{addMessageToLog()}},
\code{\link[=captureLog]{captureLog()}},
\code{\link[=initLogfunction]{initLogfunction()}},
\code{\link[=setShowLogMessages]{setShowLogMessages()}},
\code{\link[=writeTableToLog]{writeTableToLog()}}
}
\concept{log file management}
