% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotForrest.R
\name{scaleForestPlot}
\alias{scaleForestPlot}
\title{Scale Forest Plot}
\usage{
scaleForestPlot(
  plotObject,
  plotData,
  xscale,
  xgridLabel,
  xgridTable,
  xBreaks,
  tableLabels,
  withoutTable
)
}
\arguments{
\item{plotObject}{A ggplot object representing the forest plot.}

\item{plotData}{A data.table containing the data to be plotted.}

\item{xscale}{Character string specifying the x-axis scale, either 'linear' or 'log'.}

\item{xgridLabel}{Numeric vector for x-axis grid labels.}

\item{xgridTable}{Numeric vector for x-axis grid table positions.}

\item{xBreaks}{Numeric vector for x-axis breaks.}

\item{tableLabels}{Character vector with labels for the table.}

\item{withoutTable}{Logical indicating whether to display the table.}
}
\value{
A ggplot object with applied scaling.
}
\description{
This function applies scaling and formatting to the forest plot.
}
\keyword{internal}
