% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sensitivity.R
\name{sensitivityAnalyisName}
\alias{sensitivityAnalyisName}
\title{Generate Sensitivity Analysis Output File Name}
\usage{
sensitivityAnalyisName(scenarioName, sensitivitysheet)
}
\arguments{
\item{scenarioName}{The name of the scenario.}

\item{sensitivitysheet}{The name of the sensitivity sheet.}
}
\value{
A string representing the generated file name for the sensitivity analysis results.
}
\description{
This function generates a name for the sensitivity analysis output file based on the scenario name and sheet name.
}
