% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{separateAndTrimColumn}
\alias{separateAndTrimColumn}
\title{Separate and Trim Helper Function}
\usage{
separateAndTrimColumn(data, columnName)
}
\arguments{
\item{data}{A data.table containing the column to be processed.}

\item{columnName}{A string representing the name of the column to separate.}
}
\value{
A data.table with the specified column separated into multiple rows,
        trimmed of whitespace, and renamed to remove the plural 's'.
}
\description{
This function separates the rows of a specified column in a data.table,
trims whitespace from the resulting values, and renames the column to remove
the plural 's' at the end of the column name.
}
\seealso{
Other function to read from and write to xlsx: 
\code{\link{copyConfigSheet}()},
\code{\link{setHeadersToLowerCase}()},
\code{\link{splitInputs}()},
\code{\link{synchronizeScenariosOutputsWithPlots}()},
\code{\link{synchronizeScenariosWithPlots}()},
\code{\link{xlsxAddDataUsingTemplate}()},
\code{\link{xlsxAddSheet}()},
\code{\link{xlsxCloneAndSet}()},
\code{\link{xlsxReadData}()},
\code{\link{xlsxWriteData}()}
}
\concept{function to read from and write to xlsx}
