% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{setColorIndex}
\alias{setColorIndex}
\title{add colorIndex column}
\usage{
setColorIndex(projectConfiguration, plotDataList)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{plotDataList}{list which serves as bracket for data related objects needed for plots
plotData `data.table` with data to plot
dtPlotId `data.table`  with caption information and tags for each plot
scaleVectors, scaling vectors to adjust aesthetics like color}
}
\value{
list with entries
}
\description{
add colorIndex column
}
