% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{setDataTypeAttributes}
\alias{setDataTypeAttributes}
\title{Sets columnType attribute according to dictionary}
\usage{
setDataTypeAttributes(dataDT, dict = NULL)
}
\arguments{
\item{dataDT}{`data.table` with observed data.}

\item{dict}{Named list with columnNames and columnTypes; if NULL,
list is produced by template saved in package installation.}
}
\value{
`data.table` with attributes.
}
\description{
Sets columnType attribute according to dictionary
}
\keyword{internal}
