% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{setExportAttributes}
\alias{setExportAttributes}
\title{Set Export Attributes for Plot and Table Objects}
\usage{
setExportAttributes(
  object,
  caption = NULL,
  footNoteLines = NULL,
  exportArguments = NULL
)
}
\arguments{
\item{object}{An object (e.g., a plot or table) for which export attributes will be set.}

\item{caption}{A character string specifying the caption for the object. If NULL, no caption will be set.}

\item{footNoteLines}{A character vector specifying footnote lines for the object. If, no footnote lines will be set.}

\item{exportArguments}{A list of additional arguments such as width or height passed on to `ospsuite.plots::exportPlot`.}
}
\value{
The modified object with the specified export attributes set.
}
\description{
This function prepares plot and table objects by setting their export attributes,
such as caption and footnote lines, which will be used when exporting through
the `PlotManager` class.
}
