% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotForrest.R
\name{setLabelsAndPosition}
\alias{setLabelsAndPosition}
\title{Set Labels and Position for Plot Data}
\usage{
setLabelsAndPosition(plotData)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted.}
}
\value{
A data.table with added columns for y-positions and labels.
}
\description{
This function calculates y-positions and labels for the plot data.
}
\keyword{internal}
