% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{setManualScaleVectors}
\alias{setManualScaleVectors}
\title{Set manual scale vectors for plots}
\usage{
setManualScaleVectors(plotObject, plotData, plotType)
}
\arguments{
\item{plotObject}{The ggplot object to update.}

\item{plotData}{Object containing the data for the plot.}

\item{plotType}{Type of the plot.}
}
\value{
Updated ggplot object with manual scale vectors.
}
\description{
Set manual scale vectors for plots
}
\keyword{internal}
