% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{setOrderAndFactors}
\alias{setOrderAndFactors}
\title{setfactors to plot}
\usage{
setOrderAndFactors(projectConfiguration, plotData, identifierColumns)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{plotData}{`data.table`  with all data}

\item{identifierColumns}{identifier columns which define sort order}
}
\description{
setfactors to plot
}
