% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{setPlotTag}
\alias{setPlotTag}
\title{Set Plot Tag}
\usage{
setPlotTag(plotData, asRangePlot, usePKParameter)
}
\arguments{
\item{plotData}{A data.table containing the plot data.}

\item{asRangePlot}{Logical indicating if the plot is a range plot.}

\item{usePKParameter}{Logical indicating if PK parameters are to be used.}
}
\value{
A data.table containing the updated plot data with assigned plot tags.
}
\description{
Assigns plot tags to the data for plotting.
}
\keyword{internal}
