% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{setRunModeIsValid}
\alias{setRunModeIsValid}
\title{set a group of options, to generate a valid run of a workflow}
\usage{
setRunModeIsValid(isValid)
}
\arguments{
\item{isValid}{`boolean`if TRUE run of workflow is executed in valid mode}
}
\description{
if TRUE
 Ospsuite.plots.option watermark_enabled is set to FALSE
}
