% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{setSeedForPKParameter}
\alias{setSeedForPKParameter}
\title{Set Seed for PK Parameter}
\usage{
setSeedForPKParameter(
  scenarioName,
  pkParameter,
  outputPathId,
  referenceScenarion = NULL
)
}
\arguments{
\item{scenarioName}{Name of the scenario.}

\item{pkParameter}{Name of the PK parameter.}

\item{outputPathId}{ID for the output path.}

\item{referenceScenarion}{Optional reference scenario name.}
}
\value{
NULL (invisible).
}
\description{
Sets the seed for random number generation based on the provided parameters.
}
\keyword{internal}
