% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{setShowLogMessages}
\alias{setShowLogMessages}
\title{Toggle display of log messages on the console}
\usage{
setShowLogMessages(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean, if true log messages will be shown}
}
\description{
Toggle display of log messages on the console
}
\seealso{
Other log file management:
\code{\link[=addMessageToLog]{addMessageToLog()}},
\code{\link[=captureLog]{captureLog()}},
\code{\link[=initLogfunction]{initLogfunction()}},
\code{\link[=saveSessionInfo]{saveSessionInfo()}},
\code{\link[=writeTableToLog]{writeTableToLog()}}
}
\concept{log file management}
