% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{setupPopulationsForTest}
\alias{setupPopulationsForTest}
\title{Setup Populations for Test}
\usage{
setupPopulationsForTest(projectConfiguration, instDirectory, writeTestData)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing the project configuration details.}

\item{instDirectory}{A character string specifying the installation directory for the package.}

\item{writeTestData}{A logical value indicating whether newly created test data should be filed to the inst directory of the package.}
}
\description{
This function sets up populations for the test project by copying necessary files
and exporting defined populations.
}
\keyword{internal}
