% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{setupRandomDataForTest}
\alias{setupRandomDataForTest}
\title{Setup Random Data for Test}
\usage{
setupRandomDataForTest(
  pBuilder,
  projectConfiguration,
  scenarioList,
  scenarioResults
)
}
\arguments{
\item{pBuilder}{An TestProjectBuilder object, which provides function to build the test project}

\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}

\item{scenarioList}{A list of scenarios initialized for the project.}

\item{scenarioResults}{A list containing results from running the scenarios.}
}
\value{
Invisible NULL.
}
\description{
This function adds random pharmacokinetic (PK and time profile) data to the test project,
including shifting values for selected individuals.
}
\keyword{internal}
