% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{setUpSensitivity}
\alias{setUpSensitivity}
\title{Set Up Sensitivity Analysis}
\usage{
setUpSensitivity(
  projectConfiguration,
  scenarioListInd,
  instDirectory,
  writeTestData
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}

\item{scenarioListInd}{A list of individual scenarios initialized for the project.}

\item{instDirectory}{A character string specifying the installation directory for the package.}

\item{writeTestData}{A logical value indicating whether newly created test data should be filed to the inst directory of the package.}
}
\value{
A ProjectConfiguration object containing the updated project configuration.
}
\description{
This function sets up sensitivity analysis for the test project based on the defined scenarios.
}
\keyword{internal}
