% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{setupSimulationsForTest}
\alias{setupSimulationsForTest}
\title{Setup Simulations for Test}
\usage{
setupSimulationsForTest(
  projectConfiguration,
  scenarioList,
  instDirectory,
  writeTestData
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing the project configuration details.}

\item{scenarioList}{A list of scenarios initialized for the project.}

\item{instDirectory}{A character string specifying the installation directory for the package.}

\item{writeTestData}{A logical value indicating whether newly created test data should be filed to the inst directory of the package.}
}
\value{
A list containing the results from running the scenarios.
}
\description{
This function sets up simulations for the test project by synchronizing necessary directories
and running initialized scenarios to calculate pharmacokinetic (PK) parameters.
}
\keyword{internal}
