% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{setupVirtualTwinScenariosForTest}
\alias{setupVirtualTwinScenariosForTest}
\title{Setup Virtual Twin Scenarios for Test}
\usage{
setupVirtualTwinScenariosForTest(projectConfiguration, modelFiles)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project configuration details.}

\item{modelFiles}{A named character vector containing paths to the model files used in the project.}
}
\value{
A list of scenario names initialized for the virtual twin scenarios.
}
\description{
This function sets up virtual twin scenarios for the test project based on the defined populations
and model files.
}
\keyword{internal}
