% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{shouldSuppressExport}
\alias{shouldSuppressExport}
\title{Determine if Export Should be Suppressed}
\usage{
shouldSuppressExport(suppressExport, plotNames, inputs)
}
\arguments{
\item{suppressExport}{A logical value indicating the current state of export suppression.}

\item{plotNames}{A character vector of plot names to filter which plots should be generated. Default is NULL.}

\item{inputs}{A list of additional inputs that may contain parameters influencing the export behavior.}
}
\value{
A logical value indicating whether the export should be suppressed (TRUE) or not (FALSE).
}
\description{
This helper function checks whether the export of the Rmd file should be suppressed based on user-defined conditions.
}
\keyword{internal}
