% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{splitCaptionByIndividuals}
\alias{splitCaptionByIndividuals}
\title{Split Caption by Individuals}
\usage{
splitCaptionByIndividuals(configTable, individualIdVector, dtCaption)
}
\arguments{
\item{configTable}{A data frame containing the configuration data.}

\item{individualIdVector}{A vector of individual IDs.}

\item{dtCaption}{A data table containing the caption data.}
}
\value{
A data table with merged caption data by individuals.
}
\description{
This function splits the caption data by individual IDs.
}
\keyword{internal}
