% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotForrest.R
\name{splitWindowsForTable}
\alias{splitWindowsForTable}
\title{Split Windows for Table}
\usage{
splitWindowsForTable(
  plotData,
  textwidthLeft,
  textwidthRight,
  xscale,
  vlineIntercept,
  withoutTable
)
}
\arguments{
\item{plotData}{A data.table containing the data to be plotted.}

\item{textwidthLeft}{Numeric width for the left text area (0 to 1).}

\item{textwidthRight}{Numeric width for the right text area (0 to 1).}

\item{xscale}{Character string specifying the x-axis scale, either 'linear' or 'log'.}

\item{vlineIntercept}{Numeric vector for vertical line intercept(s).}

\item{withoutTable}{Logical indicating whether to display the table.}
}
\value{
A list containing x-limits, breaks, and grid positions for labels and tables.
}
\description{
This function splits the plotting area into labels, graphics, and tables.
}
\keyword{internal}
