% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{STANDARD_ESQLABSR_PROJECT_CONFIGURATION_PROPERTIES}
\alias{STANDARD_ESQLABSR_PROJECT_CONFIGURATION_PROPERTIES}
\title{Standard property names stored by esqlabsR::ProjectConfiguration}
\usage{
STANDARD_ESQLABSR_PROJECT_CONFIGURATION_PROPERTIES
}
\description{
Standard property names stored by esqlabsR::ProjectConfiguration
}
\keyword{internal}
