% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{startQmd}
\alias{startQmd}
\title{Create standard Qmd header with params.}
\usage{
startQmd(title = "Report")
}
\arguments{
\item{title}{Report title (default "Report").}
}
\value{
Character vector of header lines.
}
\description{
Create standard Qmd header with params.
}
\keyword{internal}
