% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markdown.R
\name{startRmd}
\alias{startRmd}
\title{creates startlines of Rmd file}
\usage{
startRmd(title = "Report")
}
\arguments{
\item{title}{}
}
\value{
character with startlines
}
\description{
creates startlines of Rmd file
}
