% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{synchronizeDirectories}
\alias{synchronizeDirectories}
\title{Synchronize Directories}
\usage{
synchronizeDirectories(fromDir, toDir, pattern = NULL)
}
\arguments{
\item{fromDir}{A character string specifying the source directory.}

\item{toDir}{A character string specifying the target directory.}

\item{pattern}{A regular expression pattern to filter the files to be copied (default is NULL).}
}
\value{
Invisible NULL.
}
\description{
This function synchronizes files from one directory to another, copying only the files that are not already present.
}
\keyword{internal}
