% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{synchronizeScenariosOutputsWithPlots}
\alias{synchronizeScenariosOutputsWithPlots}
\title{Synchronize Scenario Outputs with Plot Outputs}
\usage{
synchronizeScenariosOutputsWithPlots(
  projectConfiguration,
  direction = c("bothways", "scenarioToPlot", "plotToScenario")
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing the file paths for scenariosFile and plotsFile.}
}
\value{
Returns invisibly.
}
\description{
This function synchronizes output paths between two Excel files: one containing
scenario outputs and the other containing plot outputs. It only writes back to the
Excel files if changes are detected in the output paths.
}
\seealso{
Other function to read from and write to xlsx: 
\code{\link{copyConfigSheet}()},
\code{\link{separateAndTrimColumn}()},
\code{\link{setHeadersToLowerCase}()},
\code{\link{splitInputs}()},
\code{\link{synchronizeScenariosWithPlots}()},
\code{\link{xlsxAddDataUsingTemplate}()},
\code{\link{xlsxAddSheet}()},
\code{\link{xlsxCloneAndSet}()},
\code{\link{xlsxReadData}()},
\code{\link{xlsxWriteData}()}
}
\concept{function to read from and write to xlsx}
