% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{synchronizeScenariosWithPlots}
\alias{synchronizeScenariosWithPlots}
\title{Synchronize Scenarios Between Scenario and Plot Files}
\usage{
synchronizeScenariosWithPlots(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing the file paths for scenariosFile and plotsFile.}
}
\value{
Returns invisibly.
}
\description{
This function synchronizes scenarios between two Excel files: one containing
scenarios for a project and the other containing scenarios related to plots.
It adds any missing scenarios from the scenarios file to the plots file.
}
\seealso{
Other function to read from and write to xlsx: 
\code{\link{copyConfigSheet}()},
\code{\link{separateAndTrimColumn}()},
\code{\link{setHeadersToLowerCase}()},
\code{\link{splitInputs}()},
\code{\link{synchronizeScenariosOutputsWithPlots}()},
\code{\link{xlsxAddDataUsingTemplate}()},
\code{\link{xlsxAddSheet}()},
\code{\link{xlsxCloneAndSet}()},
\code{\link{xlsxReadData}()},
\code{\link{xlsxWriteData}()}
}
\concept{function to read from and write to xlsx}
