% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestProjectBuilder.R
\docType{class}
\name{TestProjectBuilder}
\alias{TestProjectBuilder}
\title{TestProjectBuilder}
\description{
Manages the creation and writing test project, which can be used for this package or
packages which use the reporting framework as base to build a test project
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{TestProjectBuilder}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rootDirectory}}{root directory of test project.
This function synchronizes files from one directory to another, copying only the files that are not already present.
This function updates the display names for scenarios and output paths in the project configuration.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TestProjectBuilder-new}{\code{TestProjectBuilder$new()}}
\item \href{#method-TestProjectBuilder-iniTestProject}{\code{TestProjectBuilder$iniTestProject()}}
\item \href{#method-TestProjectBuilder-setupRandomPopulations}{\code{TestProjectBuilder$setupRandomPopulations()}}
\item \href{#method-TestProjectBuilder-setupSimulations}{\code{TestProjectBuilder$setupSimulations()}}
\item \href{#method-TestProjectBuilder-setUpSensitivity}{\code{TestProjectBuilder$setUpSensitivity()}}
\item \href{#method-TestProjectBuilder-addRandomTPData}{\code{TestProjectBuilder$addRandomTPData()}}
\item \href{#method-TestProjectBuilder-addRandomPKData}{\code{TestProjectBuilder$addRandomPKData()}}
\item \href{#method-TestProjectBuilder-mockManualEditingsCleanup}{\code{TestProjectBuilder$mockManualEditingsCleanup()}}
\item \href{#method-TestProjectBuilder-mockManualEditingsPopulation}{\code{TestProjectBuilder$mockManualEditingsPopulation()}}
\item \href{#method-TestProjectBuilder-mockManualEditingsScenario}{\code{TestProjectBuilder$mockManualEditingsScenario()}}
\item \href{#method-TestProjectBuilder-mockManualEditingsPKParameter}{\code{TestProjectBuilder$mockManualEditingsPKParameter()}}
\item \href{#method-TestProjectBuilder-mockManualEditingsIndividuals}{\code{TestProjectBuilder$mockManualEditingsIndividuals()}}
\item \href{#method-TestProjectBuilder-mockManualEditingsSensitivity}{\code{TestProjectBuilder$mockManualEditingsSensitivity()}}
\item \href{#method-TestProjectBuilder-clone}{\code{TestProjectBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-iniTestProject"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-iniTestProject}{}}}
\subsection{Method \code{iniTestProject()}}{
A short description...

This function initializes the project structure and configuration for the test project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$iniTestProject(rootDirectory)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rootDirectory}}{A character string specifying the root directory for the test project.}

\item{\code{verbose}}{A logical value indicating whether to print messages about the process.}

\item{\code{modelFiles}}{A named character vector containing paths to the model files used in the project.}

\item{\code{instDirectory}}{A character string specifying the installation directory for the package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ProjectConfiguration object containing the details of the initialized project.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-setupRandomPopulations"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-setupRandomPopulations}{}}}
\subsection{Method \code{setupRandomPopulations()}}{
This function sets up populations for the test project by copying necessary files
and exporting defined populations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$setupRandomPopulations(
  projectConfiguration,
  populationNames,
  writeTestData = FALSE,
  templateDir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing the project configuration details.}

\item{\code{populationNames}}{A character vector of population names to be copied.}

\item{\code{writeTestData}}{A logical value indicating whether newly created test data should be filed to the inst directory of the package.}

\item{\code{templateDir}}{character string specifying the installation directory for the package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-setupSimulations"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-setupSimulations}{}}}
\subsection{Method \code{setupSimulations()}}{
This function sets up simulations for the test project by synchronizing necessary directories
and running initialized scenarios to calculate pharmacokinetic (PK) parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$setupSimulations(
  projectConfiguration,
  scenarioList,
  writeTestData,
  templateDir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing the project configuration details.}

\item{\code{scenarioList}}{A list of scenarios initialized for the project.}

\item{\code{writeTestData}}{A logical value indicating whether newly created test data should be filed to the inst directory of the package.}

\item{\code{templateDir}}{A character string specifying the installation directory for the package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the results from running the scenarios.
Set Up Sensitivity Analysis

This function sets up sensitivity analysis for the test project based on the defined scenarios.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-setUpSensitivity"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-setUpSensitivity}{}}}
\subsection{Method \code{setUpSensitivity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$setUpSensitivity(
  projectConfiguration,
  scenarioList,
  scenarioName,
  writeTestData,
  parameterList,
  sensitivitySheet = "smallSelection",
  templateDir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing project configuration details.}

\item{\code{scenarioList}}{A list of scenarios initialized for the project.}

\item{\code{scenarioName}}{The name of the scenario used for sensitivity}

\item{\code{writeTestData}}{A logical value indicating whether newly created test data should be filed to the inst directory of the package.}

\item{\code{parameterList}}{A named list where each name corresponds to a parameter in the
Excel sheet, and each value is the new value to set for that
parameter.}

\item{\code{sensitivitySheet}}{Name of sheet with parameter selection}

\item{\code{templateDir}}{A character string specifying the installation directory for the package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ProjectConfiguration object containing the updated project configuration.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-addRandomTPData"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-addRandomTPData}{}}}
\subsection{Method \code{addRandomTPData()}}{
This function adds random time profile data to the test project for selected individuals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$addRandomTPData(
  projectConfiguration,
  scenarioResults,
  ids,
  timePoints = c(15, 30, 45, 60, 90, c(3, 6, 9, 12, 18, 24) * 60),
  outputPathIds = c("Plasma", "CYP3A4total", "CYP3A4Liver"),
  sdShift = 0.05,
  lloqValue = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing project configuration details.}

\item{\code{scenarioResults}}{List of scenario results.}

\item{\code{ids}}{A vector of individual IDs to include in the random data.}

\item{\code{timePoints}}{A numeric vector specifying the time points at which to generate data (default is c(15, 30, 45, 60, 90, c(3, 6, 9, 12, 18, 24) * 60)).}

\item{\code{outputPathIds}}{A character vector specifying the output path IDs to include in the random data (default is c('Plasma', 'CYP3A4total', 'CYP3A4Liver')).}

\item{\code{sdShift}}{A numeric value indicating the standard deviation shift applied to the random values (default is 0.05).}

\item{\code{lloqValue}}{A numeric value representing the lower limit of quantification (default is 0.1).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible NULL.
This function adds random pharmacokinetic (PK) data to the test project,
including shifting values and calculating statistics for selected individuals.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-addRandomPKData"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-addRandomPKData}{}}}
\subsection{Method \code{addRandomPKData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$addRandomPKData(
  projectConfiguration,
  pkParameterDT,
  ids,
  outputPathIds = c("Plasma", "CYP3A4total", "CYP3A4Liver"),
  sdShift = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing project configuration details.}

\item{\code{pkParameterDT}}{A data.table containing PK parameters and values.}

\item{\code{ids}}{A vector of individual IDs to include in the random data.}

\item{\code{outputPathIds}}{A character vector specifying the output path IDs to include in the random data (default is c('Plasma', 'CYP3A4total', 'CYP3A4Liver')).}

\item{\code{sdShift}}{A numeric value indicating the standard deviation shift applied to the random values (default is 0.05).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-mockManualEditingsCleanup"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-mockManualEditingsCleanup}{}}}
\subsection{Method \code{mockManualEditingsCleanup()}}{
This function cleans up the project configuration files
by clearing existing data from relevant Excel sheets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$mockManualEditingsCleanup(projectConfiguration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing project configuration details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-mockManualEditingsPopulation"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-mockManualEditingsPopulation}{}}}
\subsection{Method \code{mockManualEditingsPopulation()}}{
This function adds biometric ranges for virtual pediatric and one adult populations to the configuration files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$mockManualEditingsPopulation(
  projectConfiguration,
  randomPops = data.table(populationName = c("adults", "toddler", "children",
    "school-children", "adolescents"), species = "Human", population =
    "European_ICRP_2002", numberOfIndividuals = 100, proportionOfFemales = 50, ageMin =
    c(20, 0.5, 2, 6, 12), ageMax = c(40, 2, 6, 12, 18), weightUnit = "kg", heightUnit =
    "cm", bMIUnit = "kg/m²", protein = "CYP3A4,UGT1A4", ontogeny = "CYP3A4,UGT1A4")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing project configuration details.}

\item{\code{randomPops}}{data.table with content for sheet Demographics in Population.xlsx}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-mockManualEditingsScenario"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-mockManualEditingsScenario}{}}}
\subsection{Method \code{mockManualEditingsScenario()}}{
This function sets up scenarios based on the defined populations and adds PK Parameters
and output paths to the project configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$mockManualEditingsScenario(
  projectConfiguration,
  dtTestScenarios,
  pKParameter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing project configuration details.}

\item{\code{dtTestScenarios}}{`data.table` with content for scenarios}

\item{\code{pKParameter}}{vector with pkParameter sheets}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-mockManualEditingsPKParameter"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-mockManualEditingsPKParameter}{}}}
\subsection{Method \code{mockManualEditingsPKParameter()}}{
This function adds PK parameters to the project configuration based on templates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$mockManualEditingsPKParameter(projectConfiguration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A ProjectConfiguration object containing project configuration details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-mockManualEditingsIndividuals"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-mockManualEditingsIndividuals}{}}}
\subsection{Method \code{mockManualEditingsIndividuals()}}{
This function adjust the bio-metrics with in individuals and exports virtual twin populations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$mockManualEditingsIndividuals(projectConfiguration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A `ProjectConfiguration` object containing project configuration details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`data.table` with content of sheet `VirtualTwinPopulation`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-mockManualEditingsSensitivity"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-mockManualEditingsSensitivity}{}}}
\subsection{Method \code{mockManualEditingsSensitivity()}}{
This function modifies specific parameters in a given Excel sheet based on the provided
parameter list and saves the changes back to the workbook.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$mockManualEditingsSensitivity(
  projectConfiguration,
  sheetName,
  parameterList
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{A list containing project configuration details, including
the path to the sensitivity file.}

\item{\code{sheetName}}{A string representing the name of the sheet in the Excel workbook
that will be edited.}

\item{\code{parameterList}}{A named list where each name corresponds to a parameter in the
Excel sheet, and each value is the new value to set for that
parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL This function is called for its side effects (i.e., modifying the Excel file).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TestProjectBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TestProjectBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TestProjectBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
