% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\name{TIMERANGE}
\alias{TIMERANGE}
\title{enumeration keys for Time range shortcut}
\usage{
TIMERANGE
}
\description{
enumeration keys for Time range shortcut
}
\seealso{
Other enumerations:
\code{\link{BIOMETRICUNITS}},
\code{\link{DATACLASS}},
\code{\link{EXPORTDIR}}
}
\concept{enumerations}
