% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{updateConfigTablesForTestProject}
\alias{updateConfigTablesForTestProject}
\title{Update Tables for Test Projects}
\usage{
updateConfigTablesForTestProject(projectConfiguration, modelFiles)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing the
details of the project configuration that needs to be updated.}

\item{modelFiles}{A named character vector containing paths to the model
files used in the project.}
}
\value{
invisible(NULL)
}
\description{
This function updates the configuration tables for test projects in the
Reporting Framework. It performs necessary modifications to the project
configuration, including clearing existing data, setting up scenarios,
and adding pharmacokinetic (PK) parameters.
}
\details{
The function is designed to ensure that the project configuration is
correctly set up for running simulations and analyses within the
`ospsuite.reportingframework` package.
}
\keyword{internal}
