% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{updateDtCaption}
\alias{updateDtCaption}
\title{Update Data Table Caption}
\usage{
updateDtCaption(dtCaption, configTable)
}
\arguments{
\item{dtCaption}{A data.table containing the captions to be updated.}

\item{configTable}{A data.table containing configuration information,
including individual IDs.}
}
\value{
A data.table with updated captions, where individual IDs marked
        with '(*)' are removed.
}
\description{
This function updates the provided data table caption by modifying the
individual IDs based on a given configuration table. If the configuration
table contains any individual IDs marked with '(*)', those IDs will be
removed from the data table caption.
}
\keyword{internal}
