% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{updateExportedPopulation}
\alias{updateExportedPopulation}
\title{Update Exported Population}
\usage{
updateExportedPopulation(
  projectConfiguration,
  sourcePopulation,
  targetPopulation,
  sheetName,
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A list containing configuration settings, including
the path to the populations folder and the populations file.}

\item{sourcePopulation}{A character string representing the name of the source
population to load.}

\item{targetPopulation}{A character string representing the name of the target
population to be updated.}

\item{sheetName}{A character string specifying the name of the sheet in the Excel
file from which to extend the population data.}

\item{overwrite}{A logical value indicating whether to overwrite the existing
target population file. Default is FALSE.}
}
\value{
NULL (invisible), or a warning if the target population already exists
        and overwrite is FALSE.
}
\description{
This function updates an existing exported population by loading data from a
specified source population file and extending it with additional data from
an Excel sheet. It checks for the existence of the target population file and
can optionally overwrite existing files.
}
