% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{updateFileNameDictionary}
\alias{updateFileNameDictionary}
\title{Update File Name Dictionary}
\usage{
updateFileNameDictionary(fileNameDict, invalidFilenames, projectConfiguration)
}
\arguments{
\item{fileNameDict}{A data.table containing the filename dictionary.}

\item{invalidFilenames}{A vector of invalid filenames encountered during validation.}

\item{projectConfiguration}{A list containing project-specific configurations.}
}
\value{
The updated filename dictionary.
}
\description{
This function updates the filename dictionary with invalid filenames encountered.
}
\keyword{internal}
