% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{updateGuides}
\alias{updateGuides}
\title{Update guides and scales}
\usage{
updateGuides(plotData, plotObject, plotType)
}
\arguments{
\item{plotData}{Object containing the data for the plot.}

\item{plotObject}{The ggplot object to update.}

\item{plotType}{Type of the plot.}
}
\value{
Updated ggplot object with guides.
}
\description{
Update guides and scales
}
\keyword{internal}
