% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{updateParameterOfSheets}
\alias{updateParameterOfSheets}
\title{Update Parameters of Sheets}
\usage{
updateParameterOfSheets(projectConfiguration, pkParameterSheets)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration, including the PK parameter file.}

\item{pkParameterSheets}{A vector of sheet names to update.}
}
\value{
NULL (updates are made in-place).
}
\description{
This function updates pharmacokinetic (PK) parameters in specified sheets based on user-defined parameters.
}
