% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{updateScalevector}
\alias{updateScalevector}
\title{Update Scale Vectors}
\usage{
updateScalevector(scaleVectorsInput)
}
\value{
A list containing updated scale vectors for both simulated and observed data. Each component of the
        list includes properties such as `color`, `fill`, and `shape`.
}
\description{
This function updates the scale vectors for simulated and observed data based on the provided input.
}
\keyword{internal}
