% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateAtleastOneEntry}
\alias{validateAtleastOneEntry}
\title{check if at least one of the following columns is selected}
\usage{
validateAtleastOneEntry(configTablePlots, columnVector)
}
\arguments{
\item{configTablePlots}{`data.table` ConfigurationTable without header lines}

\item{columnVector}{vector of columns to check}
}
\description{
check if at least one of the following columns is selected
}
