% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{validateBootstrappingInputs}
\alias{validateBootstrappingInputs}
\title{Validate Bootstrapping Inputs}
\usage{
validateBootstrappingInputs(
  nObservationDefault,
  nBootstrap,
  confLevel,
  coefficientOfVariation,
  statFun,
  ratioMode
)
}
\arguments{
\item{nObservationDefault}{Default number of observations.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}

\item{coefficientOfVariation}{Coefficient of variation for the data.}

\item{statFun}{Statistical function for bootstrapping.}

\item{ratioMode}{Mode for ratio calculations.}
}
\value{
NULL (invisible).
}
\description{
Validates the inputs for bootstrapping in PK forest plots.
}
\keyword{internal}
