% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateColorVector}
\alias{validateColorVector}
\title{Validate Color Vector}
\usage{
validateColorVector(colorVector)
}
\arguments{
\item{colorVector}{A character vector of colors, which should have a length of 2.}
}
\value{
Returns `invisible()` if the checks pass.
}
\description{
This function checks if the provided color vector meets the following criteria:
- It is named.
- All values are valid color names.
}
\keyword{internal}
