% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateColumn}
\alias{validateColumn}
\title{validate a Column}
\usage{
validateColumn(col, data, type, anyMissing = FALSE)
}
\arguments{
\item{col}{A string representing the name of the column to validate.}

\item{data}{A data frame containing the column to be validated.}

\item{type}{A string specifying the type of validation to perform.
Options are "character", "numeric", or "logical".}

\item{anyMissing}{A logical value indicating whether missing values are allowed.
Default is FALSE.}
}
\value{
invisible(NULL). Throws an error if validation fails.
}
\description{
This function validates a specified column in a data frame based on the
provided type. It checks for character, numeric, or logical types and
can enforce the presence or absence of missing values.
}
\keyword{internal}
