% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateConfigTableForTimeProfiles}
\alias{validateConfigTableForTimeProfiles}
\title{Validation of config table fo time profiles plots}
\usage{
validateConfigTableForTimeProfiles(
  configTable,
  dtScenarios,
  dtOutputPaths,
  observedData
)
}
\arguments{
\item{configTable}{plot Configuration table}

\item{dtScenarios}{data.table with scenario definitions}

\item{dtOutputPaths}{data.table with output definitions}

\item{observedData}{observed data as data.table}
}
\description{
Validation of config table fo time profiles plots
}
