% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateConfigTablePlots}
\alias{validateConfigTablePlots}
\title{validate types of plot configuration tables}
\usage{
validateConfigTablePlots(
  configTablePlots,
  charactersWithoutMissing = NULL,
  charactersWithMissing = NULL,
  numericColumns = NULL,
  numericWithoutMissing = NULL,
  logicalColumns = NULL,
  numericRangeColumns = NULL,
  subsetList = list()
)
}
\arguments{
\item{configTablePlots}{`data.table` configuration table without header lines}

\item{charactersWithoutMissing}{vector with character columns, where no missing value is allowed}

\item{charactersWithMissing}{vector with character column, where values may missing}

\item{numericColumns}{vector with numeric columns}

\item{logicalColumns}{vector with booleans}

\item{numericRangeColumns}{vector with columns where entries must be evaluate to a numeric of length 2}

\item{subsetList}{list where each entry is a list:
list(cols = 'vector with columns',
allowedValues = vector with allowed values)}
}
\description{
validate types of plot configuration tables
}
