% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateDataGroupIdsForPlot}
\alias{validateDataGroupIdsForPlot}
\title{Validation of data.table with outputPath Ids}
\usage{
validateDataGroupIdsForPlot()
}
\description{
Validation of data.table with outputPath Ids
}
\keyword{internal}
