% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateDemographicsConfig}
\alias{validateDemographicsConfig}
\title{Validate Demographics Configuration Table}
\usage{
validateDemographicsConfig(configTable, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{...}{Additional arguments for validation.}

\item{pkParameterDT}{A data.table containing PK parameter data.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration table for PK box-and-whisker plots.
}
