% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateGroupConsistency}
\alias{validateGroupConsistency}
\title{Validate Consistency of Values Within Groups}
\usage{
validateGroupConsistency(dt, valueColumns, groupingColumns = "plotName")
}
\arguments{
\item{dt}{A data.table or data.frame containing the data to be validated.}

\item{valueColumns}{A character vector of column names to check for consistency.}

\item{groupingColumns}{A character vector of column names that define the groups.
Default is 'plotName'.}
}
\value{
invisible(NULL). Raises an error if any value column contains inconsistent values within a group.
}
\description{
This function checks whether specified columns in a data table have consistent values
within groups defined by one or more grouping columns. If any column contains more than
one unique value within a group, an error is raised.
}
