% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateHeaders}
\alias{validateHeaders}
\title{checks if config table header and plot rows are strict separated}
\usage{
validateHeaders(configTable)
}
\arguments{
\item{configTable}{`data.table` configuration table to check}
}
\value{
configuration table without header lines
}
\description{
checks if config table header and plot rows are strict separated
}
