% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateHistogramsConfig}
\alias{validateHistogramsConfig}
\title{Validate Histograms Configuration}
\usage{
validateHistogramsConfig(configTable, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration table for histogram plots.
}
\seealso{
Other plot configuration validation function:
\code{\link[=validateDistributionVsDemographicsConfig]{validateDistributionVsDemographicsConfig()}},
\code{\link[=validatePKBoxwhiskerConfig]{validatePKBoxwhiskerConfig()}},
\code{\link[=validatePKForestAggregatedAbsoluteValuesConfig]{validatePKForestAggregatedAbsoluteValuesConfig()}},
\code{\link[=validatePKForestAggregatedRatiosConfig]{validatePKForestAggregatedRatiosConfig()}},
\code{\link[=validatePKForestPointEstimateOfAbsoluteValuesConfig]{validatePKForestPointEstimateOfAbsoluteValuesConfig()}},
\code{\link[=validatePKForestPointEstimateOfRatiosConfig]{validatePKForestPointEstimateOfRatiosConfig()}},
\code{\link[=validateSensitivityConfig]{validateSensitivityConfig()}},
\code{\link[=validateTimeProfilesConfig]{validateTimeProfilesConfig()}}

Other functions to generate histograms:
\code{\link[=addDefaultConfigForDistributionsVsDemographics]{addDefaultConfigForDistributionsVsDemographics()}},
\code{\link[=plotHistograms]{plotHistograms()}}
}
\concept{functions to generate histograms}
\concept{plot configuration validation function}
