% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPkBoxplot.R
\name{validateIsCrossOverStudy}
\alias{validateIsCrossOverStudy}
\title{Validate Cross-Over Study}
\usage{
validateIsCrossOverStudy(configTablePlots, pkParameterDT)
}
\arguments{
\item{configTablePlots}{A data.table containing configuration for plots,
including scenario and referenceScenario columns.}

\item{pkParameterDT}{A data.table containing pharmacokinetic parameters
including scenario names and population IDs.}
}
\value{
None. The function will print any invalid configurations and stop
        execution if the validation fails.
}
\description{
This function checks if the provided `configTablePlots` and `pkParameterDT`
data frames represent a valid cross-over study by ensuring that the scenario
and reference scenario are based on the same population.
}
\keyword{internals}
