% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateNumericRangeColumns}
\alias{validateNumericRangeColumns}
\title{validate Numeric Range Columns}
\usage{
validateNumericRangeColumns(columns, data)
}
\arguments{
\item{columns}{A vector of column names to validate.}

\item{data}{A data frame containing the columns to validate.}
}
\description{
Validates numeric range columns in the provided data frame.
}
