% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateNumericVectorColumns}
\alias{validateNumericVectorColumns}
\title{validate Numeric Range Columns}
\usage{
validateNumericVectorColumns(columns, data, ...)
}
\arguments{
\item{columns}{A vector of column names to validate.}

\item{data}{A data frame containing the columns to validate.}

\item{...}{additionally parameters parsed to checkmate::assertNumeric}
}
\description{
Validates numeric range columns in the provided data frame.
}
