% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateOutputPathIdFormat}
\alias{validateOutputPathIdFormat}
\title{Validates output path ID format}
\usage{
validateOutputPathIdFormat(configTablePlots, column = "outputPathIds")
}
\arguments{
\item{configTablePlots}{`data.table` ConfigurationTable without header lines}
}
\description{
Validates output path ID format
}
\keyword{internal}
