% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validatePanelConsistency}
\alias{validatePanelConsistency}
\title{Check if panel columns are filled consistently}
\usage{
validatePanelConsistency(configTablePlots, panelColumns)
}
\arguments{
\item{configTablePlots}{`data.table` ConfigurationTable without header lines}

\item{panelColumns}{Vector of columns which should be consistent.}
}
\description{
Check if panel columns are filled consistently
}
\keyword{internal}
