% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateParameterID}
\alias{validateParameterID}
\title{Validate Parameter ID}
\usage{
validateParameterID(configTablePlots, ...)
}
\arguments{
\item{configTablePlots}{A data.table containing the configuration table for plots.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the parameter IDs in the configuration table for plots.
}
\keyword{internal}
