% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{validatePKForestPointEstimateOfAbsoluteValuesConfig}
\alias{validatePKForestPointEstimateOfAbsoluteValuesConfig}
\title{Validate PK Forest Absolute Values with Confidence Intervals Configuration}
\usage{
validatePKForestPointEstimateOfAbsoluteValuesConfig(
  configTable,
  pkParameterDT,
  ...
)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration for PK forest plots with absolute values and confidence intervals.
}
\seealso{
Other plot configuration validation function:
\code{\link[=validateDistributionVsDemographicsConfig]{validateDistributionVsDemographicsConfig()}},
\code{\link[=validateHistogramsConfig]{validateHistogramsConfig()}},
\code{\link[=validatePKBoxwhiskerConfig]{validatePKBoxwhiskerConfig()}},
\code{\link[=validatePKForestAggregatedAbsoluteValuesConfig]{validatePKForestAggregatedAbsoluteValuesConfig()}},
\code{\link[=validatePKForestAggregatedRatiosConfig]{validatePKForestAggregatedRatiosConfig()}},
\code{\link[=validatePKForestPointEstimateOfRatiosConfig]{validatePKForestPointEstimateOfRatiosConfig()}},
\code{\link[=validateSensitivityConfig]{validateSensitivityConfig()}},
\code{\link[=validateTimeProfilesConfig]{validateTimeProfilesConfig()}}

Other functions to generate forest plots:
\code{\link[=addDefaultConfigForPKForestPlots]{addDefaultConfigForPKForestPlots()}},
\code{\link[=plotPKForestAggregatedAbsoluteValues]{plotPKForestAggregatedAbsoluteValues()}},
\code{\link[=plotPKForestAggregatedRatios]{plotPKForestAggregatedRatios()}},
\code{\link[=plotPKForestPointEstimateOfAbsoluteValues]{plotPKForestPointEstimateOfAbsoluteValues()}},
\code{\link[=plotPKForestPointEstimateOfRatios]{plotPKForestPointEstimateOfRatios()}},
\code{\link[=validatePKForestAggregatedAbsoluteValuesConfig]{validatePKForestAggregatedAbsoluteValuesConfig()}},
\code{\link[=validatePKForestAggregatedRatiosConfig]{validatePKForestAggregatedRatiosConfig()}},
\code{\link[=validatePKForestPointEstimateOfRatiosConfig]{validatePKForestPointEstimateOfRatiosConfig()}}
}
\concept{functions to generate forest plots}
\concept{plot configuration validation function}
