% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{validatePKParameterDT}
\alias{validatePKParameterDT}
\title{Validate Structure and Content of PK Parameter Data Table}
\usage{
validatePKParameterDT(pkParameterDT)
}
\arguments{
\item{pkParameterDT}{A data.table containing pharmacokinetic parameters
with the required columns: scenario, pkParameter,
individualId, value, outputPathId, displayNamePKParameter,
and displayUnitPKParameter.}
}
\value{
NULL. The function will stop execution if validation fails,
        otherwise returns invisibly.
}
\description{
This function validates the structure and content of the provided
pharmacokinetic parameter data table (`pkParameterDT`). It checks
for the presence of required columns and ensures that the combination
of `outputPathId` and `parameter` is unique with consistent
`displayUnitPKParameter`.
}
\keyword{internal}
