% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validatePlotInput}
\alias{validatePlotInput}
\title{Validates additional Input columns}
\usage{
validatePlotInput(configTablePlots, plotInputColumns)
}
\arguments{
\item{configTablePlots}{`data.table` ConfigurationTable without header lines}

\item{plotInputColumns}{vector with names of additional inputs}
}
\description{
Validates additional Input columns
}
