% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{validatePointEstimateInputs}
\alias{validatePointEstimateInputs}
\title{Validate Bootstrapping Inputs}
\usage{
validatePointEstimateInputs(nBootstrap, confLevel, statFun)
}
\arguments{
\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}

\item{statFun}{Statistical function for bootstrapping.}

\item{nObservationDefault}{Default number of observations.}

\item{coefficientOfVariation}{Coefficient of variation for the data.}

\item{ratioMode}{Mode for ratio calculations.}
}
\value{
NULL (invisible).
}
\description{
Validates the inputs for bootstrapping in PK forest plots.
}
\keyword{internal}
