% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateRangePlotsColumns}
\alias{validateRangePlotsColumns}
\title{Validate Range Plots Columns}
\usage{
validateRangePlotsColumns(configTablePlots)
}
\arguments{
\item{configTablePlots}{A data.table containing the configuration table for plots.}
}
\value{
NULL (invisible).
}
\description{
Validates the columns used for range plots in the configuration table.
}
\keyword{internal}
