% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{validateSensitivityConfig}
\alias{validateSensitivityConfig}
\title{Validate Sensitivity Configuration Table}
\usage{
validateSensitivityConfig(configTable, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration for sensitivity analysis.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A validated data frame containing the configuration table for sensitivity plots.
}
\description{
This function checks the headers of the configuration table, validates output IDs and data group IDs for plotting,
and ensures the configuration adheres to specified criteria. It also checks for file existence in the output folder.
}
\seealso{
Other plot configuration validation function:
\code{\link[=validateDistributionVsDemographicsConfig]{validateDistributionVsDemographicsConfig()}},
\code{\link[=validateHistogramsConfig]{validateHistogramsConfig()}},
\code{\link[=validatePKBoxwhiskerConfig]{validatePKBoxwhiskerConfig()}},
\code{\link[=validatePKForestAggregatedAbsoluteValuesConfig]{validatePKForestAggregatedAbsoluteValuesConfig()}},
\code{\link[=validatePKForestAggregatedRatiosConfig]{validatePKForestAggregatedRatiosConfig()}},
\code{\link[=validatePKForestPointEstimateOfAbsoluteValuesConfig]{validatePKForestPointEstimateOfAbsoluteValuesConfig()}},
\code{\link[=validatePKForestPointEstimateOfRatiosConfig]{validatePKForestPointEstimateOfRatiosConfig()}},
\code{\link[=validateTimeProfilesConfig]{validateTimeProfilesConfig()}}
}
\concept{functions to generate sensitivity plots}
\concept{plot configuration validation function}
