% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateSubsetList}
\alias{validateSubsetList}
\title{validate Subset List}
\usage{
validateSubsetList(subsetList, data)
}
\arguments{
\item{subsetList}{A list containing subsets to validate.}

\item{data}{A data frame containing the columns to validate.}
}
\description{
Validates the subset list against the provided data frame.
}
\keyword{internal}
