% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateTimeProfilesConfig}
\alias{validateTimeProfilesConfig}
\title{Validation of config table for time profiles plots}
\usage{
validateTimeProfilesConfig(
  configTable,
  dataObserved = NULL,
  scenarioResults,
  ...
)
}
\arguments{
\item{configTable}{Plot configuration table.}

\item{dataObserved}{A `data.table` (formatted as produced by `readObservedDataByDictionary`)
or `DataCombined` object containing the observed data to be plotted.}

\item{scenarioResults}{A list containing simulated scenario results.}

\item{...}{Additional arguments passed to `ospsuite.plots::plotTimeprofile`.}
}
\description{
Validation of config table for time profiles plots
}
\seealso{
Other plot configuration validation function:
\code{\link[=validateDistributionVsDemographicsConfig]{validateDistributionVsDemographicsConfig()}},
\code{\link[=validateHistogramsConfig]{validateHistogramsConfig()}},
\code{\link[=validatePKBoxwhiskerConfig]{validatePKBoxwhiskerConfig()}},
\code{\link[=validatePKForestAggregatedAbsoluteValuesConfig]{validatePKForestAggregatedAbsoluteValuesConfig()}},
\code{\link[=validatePKForestAggregatedRatiosConfig]{validatePKForestAggregatedRatiosConfig()}},
\code{\link[=validatePKForestPointEstimateOfAbsoluteValuesConfig]{validatePKForestPointEstimateOfAbsoluteValuesConfig()}},
\code{\link[=validatePKForestPointEstimateOfRatiosConfig]{validatePKForestPointEstimateOfRatiosConfig()}},
\code{\link[=validateSensitivityConfig]{validateSensitivityConfig()}}
}
\concept{plot configuration validation function}
