% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateTimeRangeColumns}
\alias{validateTimeRangeColumns}
\title{Validates time range columns}
\usage{
validateTimeRangeColumns(configTablePlots)
}
\arguments{
\item{configTablePlots}{`data.table` configuration table without header lines}
}
\description{
Time range columns must be character and must contain NA
'total','firstApplication','lastApplication'
or a string which evaluates in R to a numeric vector length 2  (e.g. 'c(2,3)' or 'c(2,NA)'
}
\keyword{internal}
