% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{writeTableToLog}
\alias{writeTableToLog}
\title{Write a data frame to the log}
\usage{
writeTableToLog(dt, filename = NULL)
}
\arguments{
\item{dt}{table to log}

\item{filename}{Ignored; kept for backwards compatibility.}
}
\description{
Write a data frame to the log
}
\seealso{
Other log file management:
\code{\link[=addMessageToLog]{addMessageToLog()}},
\code{\link[=captureLog]{captureLog()}},
\code{\link[=initLogfunction]{initLogfunction()}},
\code{\link[=saveSessionInfo]{saveSessionInfo()}},
\code{\link[=setShowLogMessages]{setShowLogMessages()}}
}
\concept{log file management}
