% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{writeToLog}
\alias{writeToLog}
\title{Writing to Log}
\usage{
writeToLog(type, msg, filename = NULL)
}
\arguments{
\item{type}{The type of message (e.g., Error, Info).}

\item{msg}{The message to be logged.}

\item{filename}{The name of the log file.}
}
\description{
The `writeToLog` function is used to append log messages to a log file.
The path for the logfile has to be initialized by the function initLogfile
}
\examples{
\dontrun{
# Write a log message
writeToLog(type = "Info", msg = "This is an information message", filename = "run.log")
}

}
