% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addBiometricsToConfig}
\alias{addBiometricsToConfig}
\title{add Biometric information to config}
\usage{
addBiometricsToConfig(projectConfiguration, observedData, overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\description{
add Biometric information to config
}
