% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addBiometricsToConfig}
\alias{addBiometricsToConfig}
\title{add Biometric information to config}
\usage{
addBiometricsToConfig(projectConfiguration, observedData, overwrite = FALSE)
}
\arguments{
\item{projectConfiguration}{The project configuration}

\item{observedData}{`data.table` with observed data}

\item{overwrite}{if TRUE existing rows will be overwritten}
}
\description{
add Biometric information to config
}
