% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addBiometricsToConfig}
\alias{addBiometricsToConfig}
\title{Add biometrics information to config}
\usage{
addBiometricsToConfig(projectConfiguration, dataDT, overwrite = FALSE)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{dataDT}{A `data.table` with observed data.}

\item{overwrite}{If TRUE, existing rows will be overwritten.}
}
\description{
Add biometrics information to config
}
