% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{addDefaultConfigForDistributionsVsDemographics}
\alias{addDefaultConfigForDistributionsVsDemographics}
\title{Add Default Configuration for Distribution vs Demographics}
\usage{
addDefaultConfigForDistributionsVsDemographics(
  projectConfiguration,
  pkParameterDT = NULL,
  sheetName = "RangePlots",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{pkParameterDT}{A data.table containing PK parameter data (optional).}

\item{sheetName}{Name of the sheet to create (default is "RangePlots").}

\item{overwrite}{Logical indicating if existing data should be overwritten (default is FALSE).}
}
\value{
NULL (invisible).
}
\description{
Adds default configurations for distribution vs demographics plots to the `Plots.xlsx` configuration file.
}
