% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{addDefaultConfigForHistograms}
\alias{addDefaultConfigForHistograms}
\title{Add Default Configuration for Histograms}
\usage{
addDefaultConfigForHistograms(
  projectConfiguration,
  pkParameterDT = NULL,
  sheetName = "Histograms",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{pkParameterDT}{A data.table containing PK parameter data (optional).}

\item{sheetName}{Name of the sheet to create (default is "Histograms").}

\item{overwrite}{Logical indicating if existing data should be overwritten (default is FALSE).}
}
\value{
NULL (invisible).
}
\description{
Adds default configurations for histogram plots to the `Plots.xlsx` configuration file.
}
