% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPkBoxplot.R
\name{addDefaultConfigForPKBoxwhsikerPlots}
\alias{addDefaultConfigForPKBoxwhsikerPlots}
\title{Add Default Configuration for PK Box-and-Whisker Plots}
\usage{
addDefaultConfigForPKBoxwhsikerPlots(
  projectConfiguration,
  pkParameterDT,
  sheetName = "PKParameter_Boxplot",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{sheetName}{Name of the sheet to create.}

\item{overwrite}{Logical indicating if existing data should be overwritten.}
}
\value{
NULL (invisible).
}
\description{
Adds default configurations for box-and-whisker plots to the `Plots.xlsx` configuration file.
}
