% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addDefaultConfigForPKForestPlots}
\alias{addDefaultConfigForPKForestPlots}
\title{Add Default Configuration for PK Forest Plots}
\usage{
addDefaultConfigForPKForestPlots(
  projectConfiguration,
  pkParameterDT,
  sheetName = "PKParameter_Forest",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{sheetName}{Name of the sheet to create.}

\item{overwrite}{Logical indicating if existing data should be overwritten.}
}
\value{
NULL (invisible).
}
\description{
Adds default configurations for forest plots to the `Plots.xlsx` configuration file.
}
\seealso{
\code{\link{plotPKForestAggregatedAbsoluteValues}}, \code{\link{plotPKForestPointEstimateOfAbsoluteValues}},
\code{\link{plotPKForestAggregatedRatios}}, \code{\link{plotPKForestPointEstimateOfRatios}},

Other plot functions: 
\code{\link{addDefaultConfigForTimeProfilePlots}()},
\code{\link{plotDistributionVsDemographics}()},
\code{\link{plotHistograms}()},
\code{\link{plotPKBoxwhisker}()},
\code{\link{plotPKForestAggregatedAbsoluteValues}()},
\code{\link{plotPKForestAggregatedRatios}()},
\code{\link{plotPKForestPointEstimateOfAbsoluteValues}()},
\code{\link{plotPKForestPointEstimateOfRatios}()},
\code{\link{plotSensitivity}()},
\code{\link{plotTimeProfiles}()}

Other plot configuration helper function: 
\code{\link{addDefaultConfigForDistributionsVsDemographics}()},
\code{\link{addDefaultConfigForHistograms}()},
\code{\link{addDefaultConfigForPKBoxwhsikerPlots}()},
\code{\link{addDefaultConfigForTimeProfilePlots}()}

Other functions to generate forest plots: 
\code{\link{plotPKForestAggregatedAbsoluteValues}()},
\code{\link{plotPKForestAggregatedRatios}()},
\code{\link{plotPKForestPointEstimateOfAbsoluteValues}()},
\code{\link{plotPKForestPointEstimateOfRatios}()},
\code{\link{validatePKForestAggregatedAbsoluteValuesConfig}()},
\code{\link{validatePKForestAggregatedRatiosConfig}()},
\code{\link{validatePKForestPointEstimateOfAbsoluteValuesConfig}()},
\code{\link{validatePKForestPointEstimateOfRatiosConfig}()}
}
\concept{functions to generate forest plots}
\concept{plot configuration helper function}
\concept{plot functions}
