% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addDefaultConfigForPKForestPlots}
\alias{addDefaultConfigForPKForestPlots}
\title{Add Default Configuration for PK Forest Plots}
\usage{
addDefaultConfigForPKForestPlots(
  projectConfiguration,
  pkParameterDT,
  sheetName = "PKParameter_Forest",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{sheetName}{Name of the sheet to create.}

\item{overwrite}{Logical indicating if existing data should be overwritten.}
}
\value{
NULL (invisible).

#' @seealso
\code{\link{plotPKForestAbsoluteValuesWithVariance}}, \code{\link{plotPKForestAbsoluteValuesWithCI}},
\code{\link{plotPKForestRatiosWithVariance}}, \code{\link{plotPKForestRatiosWithCI}},
}
\description{
Adds default configurations for forest plots to the `Plots.xlsx` configuration file.
}
