% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{addDefaultConfigForTimeProfilePlots}
\alias{addDefaultConfigForTimeProfilePlots}
\title{Adds a default sheet to the plot configuration table}
\usage{
addDefaultConfigForTimeProfilePlots(
  projectConfiguration,
  sheetName = "TimeProfile_Panel",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{sheetName}{Name of the sheet in the plot configuration table.}

\item{overwrite}{Boolean indicating if existing configurations will be overwritten.}
}
\description{
Adds a default sheet to the plot configuration table
}
