% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{addDefaultConfigForTimeProfilePlots}
\alias{addDefaultConfigForTimeProfilePlots}
\title{Adds a default sheet to the plot configuration table}
\usage{
addDefaultConfigForTimeProfilePlots(
  projectConfiguration,
  sheetName = "TimeProfiles",
  overwrite = FALSE
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{sheetName}{name of sheet in plot configuration table}

\item{overwrite}{`boolean` if TRUE existing configurations will be overwrtitten}
}
\description{
Adds a default sheet to the plot configuration table
}
