% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{addDefaultDemographicPlots}
\alias{addDefaultDemographicPlots}
\title{Add Default Configuration for Pediatric Demography Plots}
\usage{
addDefaultDemographicPlots(
  projectConfiguration,
  pkParameterDT,
  sheetName,
  overwrite,
  asRangePlot,
  templateSheet
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{sheetName}{Name of the sheet to create.}

\item{overwrite}{Logical indicating if existing data should be overwritten.}

\item{templateSheet}{Name of templateSheet in Configuration table}
}
\value{
NULL (invisible).
}
\description{
Adds default configurations for Pediatric demography plots to the `Plots.xlsx` configuration file.
}
\keyword{internal}
