% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{addDescriptions}
\alias{addDescriptions}
\title{Add Descriptions to Plot Data}
\usage{
addDescriptions(plotData, onePlotConfig, pkParameterDT)
}
\arguments{
\item{plotData}{Data table with calculated plot data.}

\item{onePlotConfig}{Configuration for the plot.}

\item{pkParameterDT}{Data table containing PK parameter data.}
}
\value{
Updated plot data with descriptions added.
}
\description{
This function adds descriptions to the plot data.
}
\keyword{Add}
\keyword{Data}
\keyword{Details}
\keyword{Output}
\keyword{PK}
\keyword{Parameter}
\keyword{Path}
\keyword{Plot}
\keyword{internal}
\keyword{to}
