% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{addFacets}
\alias{addFacets}
\title{Add facets to a ggplot object}
\usage{
addFacets(plotObject, plotData, facetAspectRatio = 0.5)
}
\arguments{
\item{plotObject}{ggplot object to which the facets should be added.}

\item{nFacetColumns}{Maximal number of facets used for facet type by Order.}
}
\value{
Updated ggplot object with facets.
}
\description{
Add facets to a ggplot object
}
\keyword{internal}
