% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{addFacets}
\alias{addFacets}
\title{addFacetColum}
\usage{
addFacets(plotObject, plotData, facetAspectRatio = 0.5)
}
\arguments{
\item{plotObject}{ggplot object to which the facet should be added}

\item{nFacetColumns}{maximal number of facet used for facettype byOrder}
}
\value{
ggplot Object with facets
}
\description{
addFacetColum
}
