% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{addFacets}
\alias{addFacets}
\title{addFacetColum}
\usage{
addFacets(plotObject, panelConfig, nFacetColumns, facetAspectRatio = 0.5)
}
\arguments{
\item{plotObject}{ggplot object}

\item{panelConfig}{configuration table}

\item{nFacetColumns}{maxinaml number of facet used for facettype byOrder}
}
\value{
ggplot Object with facets
}
\description{
addFacetColum
}
