% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{addFiguresAndTables}
\alias{addFiguresAndTables}
\title{Add figures and tables from keyTypes or keyList.}
\usage{
addFiguresAndTables(
  keyTypes = NULL,
  subfolder,
  numbersOf,
  customStyles = list(),
  digitsOfSignificance = 3,
  keyList = NULL
)
}
\arguments{
\item{keyTypes}{Named list of figure/table types.}

\item{subfolder}{Subfolder relative to Rmd.}

\item{numbersOf}{List with figure/table counts.}

\item{customStyles}{Custom styles list.}

\item{digitsOfSignificance}{Digits for tables (default 3).}

\item{keyList}{Fallback list of keys.}
}
\value{
Updated numbersOf list.
}
\description{
Add figures and tables from keyTypes or keyList.
}
\seealso{
Other markdown helper function:
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
