% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{addFiguresAndTables}
\alias{addFiguresAndTables}
\title{Loop on figure keys, add figure or table}
\usage{
addFiguresAndTables(
  keyList,
  subfolder,
  numbersOf,
  customStyles = list(),
  digitsOfSignificance = 3
)
}
\arguments{
\item{keyList}{list of keys}

\item{subfolder}{sub-folder of relative to .Rmd file where figures are saved}

\item{numbersOf}{list with numbers of tables and figures in rmd}

\item{customStyles}{custom-styles to render word document}

\item{digitsOfSignificance}{significance digits for tables}
}
\value{
list of numbers and figures after loop
}
\description{
Loop on figure keys, add figure or table
}
\seealso{
Other markdown helper function: 
\code{\link{mdBullet}()},
\code{\link{mdBullet0}()},
\code{\link{mdCaption}()},
\code{\link{mdFigure}()},
\code{\link{mdFootNote}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdNewpage}()},
\code{\link{mdPaste}()},
\code{\link{mdPaste0}()},
\code{\link{mdTable}()}
}
\concept{markdown helper function}
