% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{addFiguresAndTables}
\alias{addFiguresAndTables}
\title{Loop on figure keys, add figure or table}
\usage{
addFiguresAndTables(
  keyList,
  subfolder,
  numbersOf,
  customStyles = list(),
  digitsOfSignificance = 3
)
}
\arguments{
\item{keyList}{list of keys}

\item{subfolder}{sub-folder of relative to .Rmd file where figures are saved}

\item{numbersOf}{list with numbers of tables and figures in rmd}

\item{customStyles}{custom-styles to render word document}

\item{digitsOfSignificance}{significance digits for tables}
}
\value{
list of numbers and figures after loop
}
\description{
Loop on figure keys, add figure or table
}
