% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{addMessageToLog}
\alias{addMessageToLog}
\title{Used to add message to log file}
\usage{
addMessageToLog(messageText)
}
\arguments{
\item{messageText}{character with message text}
}
\description{
This function is for the usage outside a `captureLog` bracket.
Inside the bracket `message("my message Text")` can be used
}
\seealso{
Other log file management: 
\code{\link{captureLog}()},
\code{\link{initLogfunction}()},
\code{\link{saveSessionInfo}()},
\code{\link{setShowLogMessages}()},
\code{\link{writeTableToLog}()},
\code{\link{writeToLog}()}
}
\concept{log file management}
