% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_logging.R
\name{addMessageToLog}
\alias{addMessageToLog}
\title{Used to add message to log file}
\usage{
addMessageToLog(messageText)
}
\arguments{
\item{messageText}{character whit message text}
}
\description{
This function is for the usage outside a logCatch bracket.
Inside message(messageText) can be used
}
