% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addMetaDataToDataSet}
\alias{addMetaDataToDataSet}
\title{Add meta data to a data set}
\usage{
addMetaDataToDataSet(dataSet, groupData)
}
\arguments{
\item{dataSet}{`DataSet` object with observed data.}

\item{groupData}{Corresponding `data.table` with meta data.}
}
\value{
`DataSet` with observed data with added metadata.
}
\description{
Add meta data to a data set
}
\keyword{internal}
