% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{addUniqueColumns}
\alias{addUniqueColumns}
\title{Add all columns of observedData which are unique for a group}
\usage{
addUniqueColumns(dataObserved, aggregatedData)
}
\arguments{
\item{dataObserved}{The observed data.}

\item{aggregatedData}{The aggregated data.}
}
\value{
The updated aggregated data.
}
\description{
Add all columns of observedData which are unique for a group
}
\keyword{internal}
