% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{BIOMETRICUNITS}
\alias{BIOMETRICUNITS}
\title{enumeration biometric units}
\format{
An object of class \code{list} of length 3.
}
\usage{
BIOMETRICUNITS
}
\description{
enumeration biometric units
}
\seealso{
Other enumerations: 
\code{\link{DATACLASS}},
\code{\link{EXPORTDIR}},
\code{\link{TIMERANGE}}
}
\concept{enumerations}
\keyword{datasets}
