% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{BIOMETRICUNITS}
\alias{BIOMETRICUNITS}
\title{enumeration biomrtric units}
\format{
An object of class \code{list} of length 3.
}
\usage{
BIOMETRICUNITS
}
\description{
enumeration biomrtric units
}
\keyword{datasets}
