% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{calculatePKParameterForScenarios}
\alias{calculatePKParameterForScenarios}
\title{Calculate and load PK Parameters}
\usage{
calculatePKParameterForScenarios(projectConfiguration, scenarioResults)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration, including output folder and PK parameter file.}

\item{scenarioResults}{A list of scenario results containing simulation data.}
}
\value{
A data.table containing combined PK analyses for all scenarios.
}
\description{
This function calculates pharmacokinetic (PK) parameters for given scenarios based on project configuration and results.
}
