% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{calculateRatiosByAggregation}
\alias{calculateRatiosByAggregation}
\title{Calculate Ratios for Workflow}
\usage{
calculateRatiosByAggregation(configList, pkParameterDT, aggregationFun)
}
\arguments{
\item{configList}{A data.table containing the configuration for the current workflow, including
reference and scenario details.}

\item{pkParameterDT}{A data.table with PK parameters that will be used for ratio calculations.}

\item{aggregationFun}{A function used for aggregating the ratios calculated from the data.}
}
\value{
A data.table containing the aggregated statistics, including the calculated ratio and relevant
        confidence intervals.
}
\description{
Calculates ratios for the specified workflow based on the provided configuration, using an aggregation
function. This function merges reference and scenario data and computes the ratios accordingly.
}
\keyword{internal}
