% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{captureLog}
\alias{captureLog}
\title{function that catches messages, warnings, and errors.
This function has to be initialized by  `initLogfunction` function}
\usage{
captureLog(expr, finallyExpression = invisible())
}
\arguments{
\item{expr}{The expression to evaluate.}

\item{finallyExpression}{The expression to evaluate finally}
}
\description{
function that catches messages, warnings, and errors.
This function has to be initialized by  `initLogfunction` function
}
\seealso{
Other log file management: 
\code{\link{addMessageToLog}()},
\code{\link{initLogfunction}()},
\code{\link{saveSessionInfo}()},
\code{\link{setShowLogMessages}()},
\code{\link{writeTableToLog}()},
\code{\link{writeToLog}()}
}
\concept{log file management}
