% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{checkAndAdjustYlimits}
\alias{checkAndAdjustYlimits}
\title{check if Observed data are outside limits and set yLimits for log scale}
\usage{
checkAndAdjustYlimits(plotData, yScale, timeRangeFilter, plotType)
}
\arguments{
\item{plotData}{object of class PlotDataTimeProfile}

\item{yScale}{y scale of Plot}

\item{timeRangeFilter}{Name of time range filter}
}
\value{
yLimits as numeric vector
}
\description{
check if Observed data are outside limits and set yLimits for log scale
}
