% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{checkFileNameValidity}
\alias{checkFileNameValidity}
\title{Checks if files given as a character vector fullfills conditions for filen ames for electronic package}
\usage{
checkFileNameValidity(files, dataSets = NULL)
}
\arguments{
\item{files}{vector wtih file names}

\item{dataSets}{ector with data set names (may be NULL)}
}
\value{
list with inValid file names
}
\description{
Checks if files given as a character vector fullfills conditions for filen ames for electronic package
}
