% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-electronicPackage.R
\name{checkFileNameValidityForDirectory}
\alias{checkFileNameValidityForDirectory}
\title{Check on directory for valid filenams}
\usage{
checkFileNameValidityForDirectory(directory_path)
}
\arguments{
\item{directory_path}{path to the driectory where the electronic package is filed}
}
\description{
Checks if all files and all files of datasets of a directory fullfill conditions for file names for electronic package
}
