% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{collectPlotData}
\alias{collectPlotData}
\title{Collects all observed and simulated data needed for this plots}
\usage{
collectPlotData(projectConfiguration, onePlotConfig, dataObserved, facetType)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{observedData}{`data.table` format corresponds to tables  generated by \code{\link{readDataByDictionary}}}
}
\value{
`data.table` with plotData and attribute metaData
}
\description{
Collects all observed and simulated data needed for this plots
}
