% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{commonPKRatioPlot}
\alias{commonPKRatioPlot}
\title{This internal function performs common validation and configuration setup for PK ratio plots.}
\usage{
commonPKRatioPlot(
  projectConfiguration,
  subfolder,
  configTableSheet,
  pkParameterDT,
  pkParameterObserved,
  ratioCalculationInputs,
  ratioCalculationMethod,
  vlineIntercept,
  scaleVectors,
  digitsToRound,
  digitsToShow,
  xlabel,
  tableLabels
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object.}

\item{subfolder}{A string specifying the subfolder for output.}

\item{configTableSheet}{Name of the sheet containing configuration tables.}

\item{pkParameterDT}{A data.table with PK parameters.}

\item{pkParameterObserved}{A data.table with observed PK parameters (optional).}

\item{ratioCalculationInputs}{A list of inputs specific to the ratio calculation method.}

\item{ratioCalculationMethod}{A string specifying the method for ratio calculation.}

\item{vlineIntercept}{Intercept for vertical line on plots.}

\item{scaleVectors}{A list of aesthetic parameters for plots.}

\item{digitsToRound}{Number of digits to round in output.}

\item{digitsToShow}{Number of digits to display.}

\item{xlabel}{Label for x-axis.}
}
\value{
An RmdContainer object with generated plots.
}
\description{
This internal function performs common validation and configuration setup for PK ratio plots.
}
\keyword{internal}
