% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertBiometrics}
\alias{convertBiometrics}
\title{Converts biometric columns to default unit}
\usage{
convertBiometrics(data, dict, dictionaryName)
}
\arguments{
\item{data}{The data to be converted.}

\item{dict}{The dictionary to be used for conversion.}

\item{dictionaryName}{The name of the dictionary.}
}
\value{
`data.table` with converted columns.
}
\description{
Converts biometric columns to default unit
}
\keyword{internal}
