% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertDataByDictionary}
\alias{convertDataByDictionary}
\title{Convert data by dictionary}
\usage{
convertDataByDictionary(data, dataFilter, dict, dictionaryName)
}
\arguments{
\item{data}{The data to be converted.}

\item{dataFilter}{The filter to be applied to the data.}

\item{dict}{The dictionary to be used for conversion.}

\item{dictionaryName}{The name of the dictionary.}
}
\value{
A `data.table` containing the converted data.
}
\description{
This function converts the data based on the provided dictionary and filters.
}
\keyword{internal}
