% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertDataCombinedToDataTable}
\alias{convertDataCombinedToDataTable}
\title{Converts object of class `DataCombined` to data.table with attributes.}
\usage{
convertDataCombinedToDataTable(datacombined)
}
\arguments{
\item{datacombined}{The input `DataCombined` object.}
}
\value{
A `data.table` containing the converted data.
}
\description{
Format corresponds to `data.table` produced by `readObservedDataByDictionary`.
}
\seealso{
Other observed data processing:
\code{\link[=addBiometricsToConfig]{addBiometricsToConfig()}},
\code{\link[=aggregateObservedDataGroups]{aggregateObservedDataGroups()}},
\code{\link[=convertDataTableToDataCombined]{convertDataTableToDataCombined()}},
\code{\link[=readObservedDataByDictionary]{readObservedDataByDictionary()}},
\code{\link[=updateDataGroupId]{updateDataGroupId()}},
\code{\link[=updateOutputPathId]{updateOutputPathId()}},
\code{\link[=validateObservedData]{validateObservedData()}}
}
\concept{observed data processing}
