% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertDataCombinedToDataTable}
\alias{convertDataCombinedToDataTable}
\title{converst object of class dataCombined to data.table with atttributes}
\usage{
convertDataCombinedToDataTable(datacombined)
}
\arguments{
\item{datacombined}{}
}
\value{
`data.table`
}
\description{
format correspondends to data.table produced by `readObservedDataByDictionary`
}
