% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertDataCombinedToDataTable}
\alias{convertDataCombinedToDataTable}
\title{Converts object of class `DataCombined` to data.table with attributes.}
\usage{
convertDataCombinedToDataTable(datacombined)
}
\arguments{
\item{datacombined}{The input `DataCombined` object.}
}
\value{
A `data.table` containing the converted data.
}
\description{
Format corresponds to `data.table` produced by `readObservedDataByDictionary`.
}
