% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertDataCombinedToDataTable}
\alias{convertDataCombinedToDataTable}
\title{converts object of class `DataCombined` to data.table with attributes}
\usage{
convertDataCombinedToDataTable(datacombined)
}
\arguments{
\item{datacombined}{}
}
\value{
`data.table`
}
\description{
format correspondents to data.table produced by `readObservedDataByDictionary`
}
