% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{convertDataTableToDataCombined}
\alias{convertDataTableToDataCombined}
\title{Converts data.table with observed data to `ospsuite::DataCombined` object}
\usage{
convertDataTableToDataCombined(dataDT)
}
\arguments{
\item{dataDT}{A `data.table` to convert.}
}
\value{
An object of class `DataCombined`.
}
\description{
The `data.table` must be formatted like a table produced by `readObservedDataByDictionary`.
}
