% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{convertYunit}
\alias{convertYunit}
\title{Convert Y Unit}
\usage{
convertYunit(timeprofile, dtUnit)
}
\arguments{
\item{timeprofile}{A data.table containing time profile data.}

\item{dtUnit}{A data.table with unit conversion factors.}
}
\value{
A data.table with the y-values converted to the target unit.
}
\description{
This function converts the y-values of a time profile based on unit conversion factors.
}
\keyword{internal}
