% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{createDataSets}
\alias{createDataSets}
\title{Function to create data sets from grouped data}
\usage{
createDataSets(groupData)
}
\arguments{
\item{groupData}{A `data.table` unique for identifier.}
}
\value{
An object of class 'DataSet'.
}
\description{
Function to create data sets from grouped data
}
\keyword{internal}
