% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{createDocumentFromTemplate}
\alias{createDocumentFromTemplate}
\title{Opens a template file a new document}
\usage{
createDocumentFromTemplate(
  template = "template_workflow",
  templatePath = system.file("templates", package = "OSPSuite.ReportingFramework")
)
}
\arguments{
\item{template}{name of template script}

\item{templatePath}{path of template script}
}
\description{
Opens a template file a new document
}
