% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{createPanelPlotsForPlotName}
\alias{createPanelPlotsForPlotName}
\title{Generate one plot as facet panels}
\usage{
createPanelPlotsForPlotName(
  onePlotConfig,
  rmdContainer,
  projectConfiguration,
  dataObserved,
  scenarioResults,
  nFacetColumns,
  nMaxFacetRows,
  facetAspectRatio,
  aggregationFun,
  referenceScaleVector
)
}
\arguments{
\item{onePlotConfig}{configuration table for this plot}

\item{rmdContainer}{Object of class `RmdContainer`.}

\item{projectConfiguration}{An object of class ProjectConfiguration containing the project settings,
including file paths and scenario definitions. This is essential for the function to access the necessary data.}

\item{dataObserved}{A data.table or DataCombined object containing the observed data to be plotted.}

\item{scenarioResults}{A list with simulated scenario results. This parameter may be an empty list or
contain only some of all scenarios defined in the project configuration, in which
case the function will attempt to load missing scenario results from the specified folder.}

\item{nFacetColumns}{An integer specifying the maximum number of facet columns (default is 2)
used for the facet type "by Order".}

\item{nMaxFacetRows}{An integer specifying the maximum number of facet rows (default is 4)
used for all facet types.}

\item{facetAspectRatio}{A numeric value specifying the aspect ratio for the facets (default is 0.5).}

\item{aggregationFun}{Function to aggregate simulated data.}

\item{referenceScaleVector}{A named list that configures labels and colors for the display of
reference scenarios. The names of the list are displayed as labels in the legend.}
}
\value{
Object of class `rmdContainer` with added figures.
}
\description{
Generate one plot as facet panels
}
