% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{createPanelPlotsForPlotName}
\alias{createPanelPlotsForPlotName}
\title{Generate one plot as facet panels}
\usage{
createPanelPlotsForPlotName(
  projectConfiguration,
  onePlotConfig,
  dataObserved,
  nFacetColumns,
  nMaxFacetRows,
  rmdContainer,
  facetAspectRatio,
  aggregationFun,
  referenceScaleVector
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{onePlotConfig}{configuration table for this plot}

\item{dataObserved}{`data.table` format corresponds to tables  generated by  function `readDataByDictionary`}

\item{nFacetColumns}{Maximal number of facet columns (default 2) used for facet type "by Order".}

\item{nMaxFacetRows}{maximal number of facet rows}

\item{rmdContainer}{Object of class `RmdContainer`.}

\item{facetAspectRatio}{Aspect ratio for the facets (default 0.5).}

\item{aggregationFun}{Function to aggregate simulated data.}

\item{referenceScaleVector}{scale vector to scale aesthetics color and fill for scenarios with reference scenario}
}
\value{
Object of class `rmdContainer` with added figures.
}
\description{
Generate one plot as facet panels
}
