% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{createPkPlotForPlotName}
\alias{createPkPlotForPlotName}
\title{Create PK Ratio Plot}
\usage{
createPkPlotForPlotName(
  onePlotConfig,
  rmdContainer,
  pkParameterDT,
  pkParameterObserved,
  ratioCalculationMethod,
  ratioCalculationInputs,
  vlineIntercept,
  scaleVectors,
  digitsToRound,
  digitsToShow,
  xlabel,
  tableLabels
)
}
\arguments{
\item{onePlotConfig}{A data.table containing the plot configuration.}

\item{rmdContainer}{An RmdContainer object.}

\item{pkParameterDT}{A data.table with PK parameters.}

\item{vlineIntercept}{Intercept for vertical line on plots.}

\item{scaleVectors}{A list of aesthetic parameters for plots.}

\item{digitsToRound}{Number of digits to round in output.}

\item{digitsToShow}{Number of digits to display.}

\item{xlabel}{Label for x-axis.}

\item{coefficientOfVariation}{Coefficient of variation for simulations.}

\item{nObservationDefault}{Default number of observations.}

\item{nBootstrap}{Number of bootstrap samples.}

\item{confLevel}{Confidence level for intervals.}

\item{statFun}{Function to compute statistics.}

\item{seed}{Random seed for reproducibility.}
}
\value{
An updated RmdContainer object with the added plot.
}
\description{
Creates a plot for the PK ratio based on the provided configuration.
}
