% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-initialization.R
\name{createProjectConfiguration}
\alias{createProjectConfiguration}
\title{#' Create a `ProjectConfiguration`}
\usage{
createProjectConfiguration(
  path = file.path("ProjectConfiguration.xlsx"),
  ignoreVersionCheck = FALSE
)
}
\arguments{
\item{path}{path to the `ProjectConfiguration.xlsx` file. default to the `ProjectConfiguration.xlsx` file located in the working directory.}

\item{ignoreVersionCheck}{If `TRUE`, skip version mismatch checks for esqlabsR and the
reporting framework when loading the configuration file. Defaults to `FALSE`.}
}
\value{
Object of type `ProjectConfigurationRF`
}
\description{
Create a `ProjectConfigurationRF` based on the `"ProjectConfiguration.xlsx"`

based on esqlabsR::ProjectConfiguration but with additional file information for PK Parameter definitions
}
\seealso{
Other project initialization:
\code{\link{ProjectConfigurationRF}},
\code{\link[=getQCpassedEnvironmentVariable]{getQCpassedEnvironmentVariable()}},
\code{\link[=initProject]{initProject()}},
\code{\link[=restoreProjectConfigurationRF]{restoreProjectConfigurationRF()}},
\code{\link[=setWorkflowOptions]{setWorkflowOptions()}},
\code{\link[=snapshotProjectConfigurationRF]{snapshotProjectConfigurationRF()}},
\code{\link[=upgradeToReportingFramework]{upgradeToReportingFramework()}}
}
\concept{project initialization}
