% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{createProjectConfiguration}
\alias{createProjectConfiguration}
\title{#' Create a `ProjectConfiguration`}
\usage{
createProjectConfiguration(path = file.path("ProjectConfiguration.xlsx"))
}
\arguments{
\item{path}{path to the `ProjectConfiguration.xlsx` file. default to the `ProjectConfiguration.xlsx` file located in the working directory.}
}
\value{
Object of type `ProjectConfigurationRF`
}
\description{
Create a `ProjectConfigurationRF` based on the `"ProjectConfiguration.xlsx"`

based on esqlabsR::ProjectConfiguration but with additional file information for PK Parameter definitions
}
\seealso{
Other project initialisation: 
\code{\link{ProjectConfigurationRF}},
\code{\link{getQCpassedEnvironmentVariable}()},
\code{\link{initProject}()},
\code{\link{setWorkflowOptions}()}
}
\concept{project initialisation}
