% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{createProjectConfiguration}
\alias{createProjectConfiguration}
\title{Create a default `ProjectConfiguration`}
\usage{
createProjectConfiguration(path)
}
\arguments{
\item{path}{Full path of an XLS/XLSX file}
}
\value{
Object of type ProjectConfiguration
}
\description{
wraps (esqlabsR::createDefaultProjectConfiguration())
}
