% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{createWorkflowTemplate}
\alias{createWorkflowTemplate}
\title{Opens the workflow template as new document}
\usage{
createWorkflowTemplate()
}
\description{
Opens the workflow template as new document
}
\seealso{
Other Addins Support: 
\code{\link{createDocumentFromTemplate}()},
\code{\link{openEPackageTemplate}()},
\code{\link{openFigureTemplate}()},
\code{\link{openWorkflowTemplate}()}
}
\concept{Addins Support}
