% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{DATACLASS}
\alias{DATACLASS}
\title{enumeration keys for `DataClass`}
\format{
An object of class \code{list} of length 5.
}
\usage{
DATACLASS
}
\description{
enumeration keys for `DataClass`
}
\seealso{
Other enumerations: 
\code{\link{BIOMETRICUNITS}},
\code{\link{EXPORTDIR}},
\code{\link{TIMERANGE}}
}
\concept{enumerations}
\keyword{datasets}
