% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{DATACLASS}
\alias{DATACLASS}
\title{enumeration keys for DataClass}
\format{
An object of class \code{list} of length 3.
}
\usage{
DATACLASS
}
\description{
enumeration keys for DataClass
}
\keyword{datasets}
