% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{.getMappingForTimeprofiles}
\alias{.getMappingForTimeprofiles}
\title{creates mapping for plotData}
\usage{
.getMappingForTimeprofiles(plotData, metaData, userMapping)
}
\arguments{
\item{plotData}{to map}

\item{metaData}{list with metaData for plotData}
}
\value{
mapping
}
\description{
creates mapping for plotData
}
