% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{.validateAndConvertData}
\alias{.validateAndConvertData}
\title{Validates observed data and converts it to appropriate format.}
\usage{
.validateAndConvertData(plotData, predictedIsNeeded)
}
\arguments{
\item{plotData}{Either a data.table with columns 'xValues', 'yValues', 'group' or an object of class 'DataCombined'.}

\item{predictedIsNeeded}{If TRUE, only observed data are returned. If FALSE and the "predicted" column does not exist,
predicted values are calculated.}
}
\value{
A `data.table` with data formatted for plotting.
}
\description{
This function checks the input data for required columns and formats it for plotting. If predicted values are needed,
it calculates them based on the observed and simulated data.
}
\keyword{internal}
