% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowScriptExporter.R
\name{excelToListStructure}
\alias{excelToListStructure}
\title{Convert Excel file to list structure for JSON serialization}
\usage{
excelToListStructure(df)
}
\arguments{
\item{df}{content of an Excel sheet}
}
\value{
List structure ready for JSON serialization
}
\description{
Convert Excel file to list structure for JSON serialization
}
\keyword{internal}
