% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{EXPORTDIR}
\alias{EXPORTDIR}
\title{enumeration keys for exportDirectories}
\format{
An object of class \code{list} of length 3.
}
\usage{
EXPORTDIR
}
\description{
enumeration keys for exportDirectories
}
\seealso{
Other enumerations: 
\code{\link{BIOMETRICUNITS}},
\code{\link{DATACLASS}},
\code{\link{TIMERANGE}}
}
\concept{enumerations}
\keyword{datasets}
