% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-populations.R
\name{exportVirtualTwinPopulations}
\alias{exportVirtualTwinPopulations}
\title{Generate Virtual Twin Population}
\usage{
exportVirtualTwinPopulations(
  projectConfiguration,
  modelFile,
  overwrite = FALSE,
  populationNames = NULL
)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration details, including file paths for populations and scenarios.}

\item{modelFile}{A string representing the name of the model file to be loaded. This file is used for unit conversion}

\item{overwrite}{A logical indicating whether to overwrite existing files. Defaults to FALSE}

\item{populationNames}{a character vector defining the population which should be exported. If NULL (default) all will be exported}
}
\description{
This function generates virtual twin populations based on the provided model and project configuration.
}
