% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{filterIndividualID}
\alias{filterIndividualID}
\title{Filter Time Profile by Individual IDs}
\usage{
filterIndividualID(timeprofile, individualList)
}
\arguments{
\item{timeprofile}{A data frame containing time profile data with an `individualId` column.}

\item{individualList}{A character string of individual IDs to filter by. If the string
contains '*', all individuals will be included.}
}
\value{
A data frame containing only the rows from `timeprofile` that match the specified
        individual IDs. If no IDs are provided, the original `timeprofile` is returned.
}
\description{
This function filters a given time profile data frame to include only the rows
corresponding to specified individual IDs. If the individual list is not provided
othe the timeprofile has no column "individualId" the original time profile is returned.
}
