% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{filterParameterSimulated}
\alias{filterParameterSimulated}
\title{Filter Simulated Parameters}
\usage{
filterParameterSimulated(
  projectConfiguration,
  pkParameterDT,
  onePlotConfig,
  ratioMode,
  coefficientOfVariation,
  asCI
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing the
necessary settings and file paths for the project.}

\item{pkParameterDT}{A data.table containing simulated PK parameter data,
including columns for scenario names, parameters, values, and output paths.}

\item{onePlotConfig}{A data.table containing configuration settings for a
single plot, including plot name, x-axis scale, and other aesthetic settings.}

\item{ratioMode}{Mode for ratio calculations.}

\item{coefficientOfVariation}{A numeric value representing the coefficient of
 variation, which may be used to assess variability in the data. Only relevant for crossover studies (see details)
(for `plotPKForestAbsoluteValuesWithCI` and `plotPKForestRatiosWithCI`)}

\item{asCI}{Logical indicating if confidence intervals should be calculated.}
}
\value{
Filtered data table of simulated PK parameters.
}
\description{
Filters simulated PK parameters based on the provided configuration.
}
\keyword{internal}
