% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{generatePlotForPlotType}
\alias{generatePlotForPlotType}
\title{Generate plots for a specific plot type}
\usage{
generatePlotForPlotType(plotData, facetAspectRatio, plotType, ...)
}
\arguments{
\item{plotData}{Object containing the data for the plot.}

\item{facetAspectRatio}{Aspect ratio for the facets.}

\item{plotType}{Type of the plot to generate.}

\item{rmdPlotManager}{Object of class `RmdPlotManager`.}
}
\value{
Updated `rmdPlotManager` with the generated plot.
}
\description{
Generate plots for a specific plot type
}
\keyword{internal}
