% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{generatePlotForPlotType}
\alias{generatePlotForPlotType}
\title{Title}
\usage{
generatePlotForPlotType(plotData, rmdContainer, facetAspectRatio, plotType)
}
\arguments{
\item{plotType}{}
}
\description{
Title
}
