% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aggregation.R
\name{getAggregatedVariance}
\alias{getAggregatedVariance}
\title{Get Aggregated Variance}
\usage{
getAggregatedVariance(
  dt,
  aggregationFun,
  valueColumn,
  identifier,
  direction = c("y", "x")
)
}
\arguments{
\item{dt}{A data.table containing the data to aggregate.
It must include the column specified in `valueColumn`.}

\item{aggregationFun}{A function to aggregate the data.
This function should accept a numeric vector and return a list with aggregated values.}

\item{valueColumn}{A string indicating the column name containing the values to aggregate.}

\item{identifier}{A character vector specifying the columns to group by.}

\item{direction}{A character string specifying the direction of aggregation,
either 'y' or 'x'.}
}
\value{
A data.table containing the aggregated variance results, including the aggregated values and error types.
}
\description{
This function calculates the aggregated variance based on the specified aggregation function and identifier.
}
\keyword{internal}
