% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aggregation.R
\name{getAggregationFunction}
\alias{getAggregationFunction}
\title{Get Aggregation Function}
\usage{
getAggregationFunction(aggregationFlag, percentiles, customFunction)
}
\arguments{
\item{aggregationFlag}{A character string indicating the aggregation method.}

\item{percentiles}{A numeric vector of percentiles to calculate if aggregationFlag is "Percentiles".}

\item{customFunction}{A custom function for aggregation if aggregationFlag is "Custom".}
}
\value{
the aggregation function
}
\description{
This function retrieves the appropriate aggregation function based on the specified aggregation flag.
}
