% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getCaptionTimeProfile}
\alias{getCaptionTimeProfile}
\title{Title}
\usage{
getCaptionTimeProfile(dtPlotId, yScale, timeRangeTag, PlotCaptionAddon)
}
\arguments{
\item{dtPlotId}{data.table with one row per plotID}

\item{panelAddon}{}
}
\description{
Title
}
