% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{getColumnsForColumnType}
\alias{getColumnsForColumnType}
\title{select all columns where the attribute `columnType` matches the requirement}
\usage{
getColumnsForColumnType(dt, columnTypes)
}
\arguments{
\item{dt}{`data.table` with attributes (e.g. imported by `readObservedDataByDictionary`)}

\item{columnTypes}{vector with required types}
}
\value{
vector with column names
}
\description{
select all columns where the attribute `columnType` matches the requirement
}
