% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{getDataGroups}
\alias{getDataGroups}
\title{Load the properties for data groups}
\usage{
getDataGroups(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
A `data.table` with data group IDs.
}
\description{
Load the properties for data groups
}
