% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configurationEnvironment.R
\name{getDataGroups}
\alias{getDataGroups}
\title{Load the Properties for Data Groups}
\usage{
getDataGroups(wbPlots)
}
\arguments{
\item{wbPlots}{The path to the workbook containing the data groups.}
}
\value{
A `data.table` with data group IDs.
}
\description{
This function loads the properties for data groups from the specified workbook.
}
