% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getDefaultShapesForScaleVector}
\alias{getDefaultShapesForScaleVector}
\title{returns default shapes}
\usage{
getDefaultShapesForScaleVector(n)
}
\arguments{
\item{n}{}
}
\description{
returns default shapes
}
