% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_aggregation.R
\name{getErrorLabels}
\alias{getErrorLabels}
\title{extracts error labels}
\usage{
getErrorLabels(yErrorType)
}
\arguments{
\item{yErrorType}{descriptor saved in data}
}
\value{
character vector length 2 with descriptor for mean and range
}
\description{
extracts error labels
}
