% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_aggregation.R
\name{getErrorTypeForPercentiles}
\alias{getErrorTypeForPercentiles}
\title{creates text for percentiles}
\usage{
getErrorTypeForPercentiles(percentiles)
}
\arguments{
\item{percentiles}{numeric vectors with percentiles}
}
\description{
creates text for percentiles
}
