% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getFootNoteLines}
\alias{getFootNoteLines}
\title{Constructs footnote lines for aggregated data and data references}
\usage{
getFootNoteLines(dataObserved, dtDataReference)
}
\arguments{
\item{dataObserved}{Data observed for the plot.}

\item{dtDataReference}{Data reference table.}
}
\value{
Vector of characters, each entry is one footnote line.
}
\description{
Constructs footnote lines for aggregated data and data references
}
\keyword{internal}
