% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getFunctionByKey}
\alias{getFunctionByKey}
\title{select function by key}
\usage{
getFunctionByKey(key)
}
\arguments{
\item{key}{character for function selection}
}
\value{
selected `function` for key
}
\description{
select function by key
}
\keyword{internal}
