% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getGeomLLOQAttributesForTP}
\alias{getGeomLLOQAttributesForTP}
\title{get additional inputs for TP plotType to enable lines in shape legend}
\usage{
getGeomLLOQAttributesForTP(plotData)
}
\arguments{
\item{plotData}{}
}
\value{
list with additional attributes for lloq line
}
\description{
get additional inputs for TP plotType to enable lines in shape legend
}
\keyword{internal}
