% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getGroupAesthetics}
\alias{getGroupAesthetics}
\title{get groupAesthetics out of scaleVactors}
\usage{
getGroupAesthetics(plotData)
}
\arguments{
\item{plotData}{}
}
\value{
vector with aesthtics to group by
}
\description{
get groupAesthetics out of scaleVactors
}
