% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getGroupbyMapping}
\alias{getGroupbyMapping}
\title{get Mapping for "groupby" out of column Namesitle}
\usage{
getGroupbyMapping(plotData)
}
\arguments{
\item{plotData}{}
}
\description{
get Mapping for "groupby" out of column Namesitle
}
