% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getGroupbyMapping}
\alias{getGroupbyMapping}
\title{Get mapping for "groupby" aesthetics}
\usage{
getGroupbyMapping(plotData, plotType)
}
\arguments{
\item{plotData}{Object containing the data for the plot.}

\item{plotType}{Type of the plot.}
}
\value{
Aesthetic mappings for grouping.
}
\description{
Get mapping for "groupby" aesthetics
}
\keyword{internal}
