% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getObservedDataForTimeProfilePanelPlot}
\alias{getObservedDataForTimeProfilePanelPlot}
\title{load observed data used in this panel}
\usage{
getObservedDataForTimeProfilePanelPlot(observedData, panelConfig, dtUnit)
}
\arguments{
\item{dtUnit}{}
}
\description{
load observed data used in this panel
}
