% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{getObservedDataForTimeProfilePanelPlot}
\alias{getObservedDataForTimeProfilePanelPlot}
\title{load observed data used in this panel}
\usage{
getObservedDataForTimeProfilePanelPlot(dataObserved, onePlotConfig, dtUnit)
}
\arguments{
\item{dtUnit}{`data.table` with unit information}

\item{observedData}{`data.table` format corresponds to tables  generated by \code{\link{readDataByDictionary}}}
}
\value{
`data.table` with observed data used in this plot
}
\description{
load observed data used in this panel
}
