% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{getOutputPathIds}
\alias{getOutputPathIds}
\title{load the output configurations}
\usage{
getOutputPathIds(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
`data.table` with output configurations
}
\description{
load the output configurations
}
