% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{getOutputPathIds}
\alias{getOutputPathIds}
\title{Load the output configurations}
\usage{
getOutputPathIds(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
A `data.table` with output configurations.
}
\description{
Load the output configurations
}
