% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-workflow.R
\name{getQCpassedEnvironmentVariable}
\alias{getQCpassedEnvironmentVariable}
\title{Get QC Passed Environment Variable}
\usage{
getQCpassedEnvironmentVariable()
}
\value{
A logical value indicating whether the QC passed (`TRUE` or `FALSE`).
}
\description{
This function retrieves the value of the environment variable `QCpassed`.
It attempts to convert the value to a logical type. If the environment variable
is not set, is empty, or is non-logical, a warning is issued and the function
defaults the value to `FALSE`.
}
\examples{
\dontrun{
# Set the environment variable for testing
Sys.setenv(QCpassed = "TRUE")
getQCpassedEnvironmentVariable()  # Should return TRUE

# Unset the environment variable for testing
Sys.unsetenv("QCpassed")
getQCpassedEnvironmentVariable()  # Should return FALSE and issue a warning
}
}
