% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{getScalevector}
\alias{getScalevector}
\title{Returns a Scale Vector Usable for Manual Scaling in ggplot}
\usage{
getScalevector(namesOfScaleVector, listOfValues)
}
\arguments{
\item{namesOfScaleVector}{Names of the vector.}

\item{listOfValues}{List of possible entries. Takes the first where all values are not NA.}
}
\value{
A scale vector.
}
\description{
Returns a Scale Vector Usable for Manual Scaling in ggplot
}
\keyword{internal}
