% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configurationEnvironment.R
\name{getScenarioDefinitions}
\alias{getScenarioDefinitions}
\title{Load the Scenario Definitions}
\usage{
getScenarioDefinitions(wbScenarios, wbPlots = NULL)
}
\arguments{
\item{wbScenarios}{The path to the workbook containing additional scenario definitions.}

\item{wbPlots}{The path to the workbook containing the scenario definitions.}
}
\value{
A `data.table` with scenario definitions.
}
\description{
This function loads the scenario definitions from the specified workbooks.
}
