% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{getScenarioDefinitions}
\alias{getScenarioDefinitions}
\title{load the scenario definitions}
\usage{
getScenarioDefinitions(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
`data.table` with scenario definitions
}
\description{
load the scenario definitions
}
