% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{getSimulatedTimeprofile}
\alias{getSimulatedTimeprofile}
\title{Get Simulated Time Profile}
\usage{
getSimulatedTimeprofile(
  simulatedResult,
  outputPaths,
  aggregationFun,
  individualMatch
)
}
\arguments{
\item{simulatedResult}{The simulation results to be processed.}

\item{outputPaths}{A vector of output paths to be included in the time profile.}

\item{aggregationFun}{A function to aggregate the simulation data if necessary.}

\item{individualMatch}{data.table with matches simulated individual id with individual id of observed data,
is only fille for individual poulations otherwise NULL}
}
\value{
A data.table with the processed time profile data.
}
\description{
This helper function processes simulation results into a data table format for time profiles.
}
\keyword{internal}
