% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKBoxwhisker.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{Add Summary Table from Plot Data}
\usage{
getSummaryTable(plotDataPk, asRatio, onePlotConfig, percentiles)
}
\arguments{
\item{plotDataPk}{A data.table containing plot data.}

\item{asRatio}{boolean indicating if the plot is for ratios.}

\item{onePlotConfig}{Configuration for a single plot.}

\item{percentiles}{A vector of percentiles to calculate.}
}
\value{
A data.table summarizing the plot data.
}
\description{
Generates a summary table from the box-and-whisker plot data.
}
\keyword{internal}
