% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{getTableLabelsForPKForest}
\alias{getTableLabelsForPKForest}
\title{Get Table Labels for PK Forest}
\usage{
getTableLabelsForPKForest(
  plotData,
  percentiles,
  asPointeEstimate,
  confLevel,
  ratioMode,
  aggregationFun
)
}
\arguments{
\item{plotData}{Data table containing plot data.}

\item{percentiles}{Percentiles for the plot.}

\item{asPointeEstimate}{Logical indicating if confidence intervals should be calculated.}

\item{confLevel}{Confidence level for intervals.}

\item{ratioMode}{Mode for ratio calculations.}
}
\value{
A named vector of table labels.
}
\description{
Generates table labels for the PK forest plot.
}
\keyword{internal}
