% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-configurationEnvironment.R
\name{getTimeRangeTags}
\alias{getTimeRangeTags}
\title{Load the Time Range Tags}
\usage{
getTimeRangeTags(wbPlots)
}
\arguments{
\item{wbPlots}{The path to the workbook containing the time range tags.}
}
\value{
A `data.table` with time range tags.
}
\description{
This function loads the time range tags from the specified workbook.
}
\seealso{
Other get identifier: 
\code{\link{getDataGroups}()},
\code{\link{getModelParameterDefinitions}()},
\code{\link{getOutputPathIds}()},
\code{\link{getScenarioDefinitions}()},
\code{\link{loadConfigTableEnvironment}()}
}
\concept{get identifier}
