% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{getTimeRangeTags}
\alias{getTimeRangeTags}
\title{Load the time range tags}
\usage{
getTimeRangeTags(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
A `data.table` with time range tags.
}
\description{
Load the time range tags
}
