% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-timeprofile.R
\name{getUnitConversionDT}
\alias{getUnitConversionDT}
\title{Get Unit Conversion Data Table}
\usage{
getUnitConversionDT(dtSimulated, dtOutputs)
}
\arguments{
\item{dtSimulated}{A data.table containing simulated results.}

\item{dtOutputs}{A data.table containing output specifications.}
}
\value{
A data.table with unit conversion factors and unique paths.
}
\description{
This function creates a data table for unit conversion based on simulated data and output specifications.
}
\keyword{internal}
