% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{groupDataByIdentifier}
\alias{groupDataByIdentifier}
\title{Groups the data by identifier}
\usage{
groupDataByIdentifier(dataDT)
}
\arguments{
\item{dataDT}{A `data.table` to be grouped.}
}
\value{
A list with grouped data.
}
\description{
Groups the data by identifier
}
\keyword{internal}
