% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{groupDataByIdentifier}
\alias{groupDataByIdentifier}
\title{groups the data by identifier}
\usage{
groupDataByIdentifier(dataDT)
}
\arguments{
\item{dataDT}{`data.table` to be grouped}
}
\value{
list with grouped data
}
\description{
groups the data by identifier
}
