% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{handleNoConfigTable}
\alias{handleNoConfigTable}
\title{Handle Case with No Configuration Table}
\usage{
handleNoConfigTable(
  rmdPlotManager,
  projectConfiguration,
  inputs,
  suppressExport
)
}
\arguments{
\item{rmdPlotManager}{An RmdPlotManager object responsible for managing Rmd file generation and plot exports.}

\item{projectConfiguration}{A ProjectConfiguration object containing the project configuration settings.}

\item{inputs}{A list of additional inputs for the plot function.}

\item{suppressExport}{A logical value indicating whether to suppress the export of the Rmd file.}
}
\description{
This helper function processes the plotfunctions where no configuration table is provided,
generating plots based on default settings and exporting them.
}
\keyword{internal}
