% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{initializeParametersOfSheets}
\alias{initializeParametersOfSheets}
\title{Initialize Parameters of defined in given Sheets}
\usage{
initializeParametersOfSheets(projectConfiguration, pkParameterSheets)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration, including the PK parameter file.}

\item{pkParameterSheets}{A vector of sheet names to update.}
}
\value{
NULL (updates are made in-place).
}
\description{
This function initialize pharmacokinetic (PK) parameters in specified sheets based on user-defined parameters.
}
