% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{initializePlotManager}
\alias{initializePlotManager}
\title{Initialize the Plot Manager}
\usage{
initializePlotManager(
  projectConfiguration,
  qmdName,
  nameOfplotFunction,
  digitsOfSignificanceCSVDisplay,
  suppressExport
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing the project configuration settings.}

\item{qmdName}{A character string specifying the name of the resulting Qmd and the subfolder where results will be saved.}

\item{nameOfplotFunction}{The name of the plot function as character, indicating which plotting function to use.}

\item{digitsOfSignificanceCSVDisplay}{digits Of significance used for the display in the .qmd for tables, which are exported as .csv}

\item{suppressExport}{A logical value indicating whether to suppress the export of the Qmd file.}
}
\value{
A QmdPlotManager object initialized with the specified parameters.
}
\description{
This helper function creates an instance of the QmdPlotManager for managing Qmd file generation and plot exports.
}
\details{
The `initializePlotManager` function sets up the QmdPlotManager, which is responsible for generating the Qmd files
and managing the export of plots. It takes various parameters to configure the manager according to user needs.
}
\keyword{internal}
