% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{initLogfunction}
\alias{initLogfunction}
\title{Initializing a Log Function}
\usage{
initLogfunction(
  projectConfiguration,
  warningsNotDisplayed = c("introduced infinite values",
    "Each group consists of only one observation", "rows containing non-finite values",
    "rows containing missing values", "Ignoring unknown parameters",
    "was deprecated in ggplot2", "font family not found in Windows font database",
    "mbcsToSbcs"),
  messagesNotDisplayed = c("Each group consists of only one observation"),
  verbose = TRUE,
  loggingFolder = NULL
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and file names}

\item{warningsNotDisplayed}{A list of warnings that should not be logged.}

\item{messagesNotDisplayed}{A list of messages that should not be logged.}

\item{verbose}{boolean, if true log messages will be shown on the console}

\item{loggingFolder}{A character string specifying the folder where log files should be stored. If NULL, defaults to a "Logs" sub-folder within the output folder of the project configuration.}
}
\description{
Initializes logging for a workflow run. Creates a timestamped subfolder under the log
directory and configures warning/message masking via `ospsuite.utils`.
}
\examples{
\dontrun{
initLogfunction(projectConfiguration = myProjectConfig)
}

}
\seealso{
Other log file management:
\code{\link[=addMessageToLog]{addMessageToLog()}},
\code{\link[=captureLog]{captureLog()}},
\code{\link[=saveSessionInfo]{saveSessionInfo()}},
\code{\link[=setShowLogMessages]{setShowLogMessages()}},
\code{\link[=writeTableToLog]{writeTableToLog()}}
}
\concept{log file management}
