% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{initLogfunction}
\alias{initLogfunction}
\title{Initializing a Log Function}
\usage{
initLogfunction(
  projectConfiguration,
  warningsNotDisplayed = c("introduced infinite values",
    "Each group consists of only one observation", "rows containing non-finite values",
    "rows containing missing values", "Ignoring unknown parameters",
    "was deprecated in ggplot2", "font family not found in Windows font database",
    "mbcsToSbcs"),
  messagesNotDisplayed = c("Each group consists of only one observation"),
  verbose = TRUE
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{warningsNotDisplayed}{A list of warnings that should not be logged.}

\item{messagesNotDisplayed}{A list of messages that should not be logged.}

\item{verbose}{boolean, if true log message will be shown on the console}
}
\description{
This function initialize the logging during a workflow. It is called at the start of the workflow script.
It is used to configure options for the log file folder, warnings which should not logged and messages which should not logged.
}
\examples{
\dontrun{
# Initialize the log function
logFunction <- initLogfunction(rootDirectory = "path/to/project")
}

}
