% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_logging.R
\name{initLogfunction}
\alias{initLogfunction}
\title{Initializing a Log Function}
\usage{
initLogfunction(
  projectPath,
  logFileFolder = NULL,
  warningsNotDisplayed = c("introduced infinite values",
    "Each group consists of only one observation", "rows containing non-finite values",
    "rows containing missing values", "Ignoring unknown parameters",
    "was deprecated in ggplot2", "font family not found in Windows font database",
    "mbcsToSbcs"),
  messagesNotDisplayed = c("Each group consists of only one observation"),
  verbose = TRUE
)
}
\arguments{
\item{projectPath}{The path where the default log file folder is generated.}

\item{logFileFolder}{Optional. If NULL, a default log file folder is generated in the `projectPath/logs/timestamp`.}

\item{warningsNotDisplayed}{A list of warnings that should not be logged.}

\item{messagesNotDisplayed}{A list of messages that should not be logged.}

\item{verbose}{boolean, if true log message will be shown on the console}
}
\description{
This function initialize the logging during a workflow. It is called at the start of the workflow script.
It is used to configure options for the log file folder, warningswhich should not logged and messages which should not logged.
}
\examples{
\dontrun{
# Initialize the log function
logFunction <- initLogfunction(projectPath = "path/to/project")
}

}
