% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{initProject}
\alias{initProject}
\title{Initialize Project Directory}
\usage{
initProject(
  configurationDirectory = ".",
  sourceConfigurationXlsx = system.file("templates", "ProjectConfiguration.xlsx", package
    = "ospsuite.reportingframework"),
  templatePath = system.file("templates", package = "ospsuite.reportingframework"),
  overwrite = FALSE
)
}
\arguments{
\item{configurationDirectory}{A character string specifying the path to the project
directory to be initialized. Defaults to the current working directory ('.').}

\item{sourceConfigurationXlsx}{A character string representing the path to the source
Excel file containing the project configuration. By default, it uses the template
provided by the 'ospsuite.reportingframework' package.}

\item{templatePath}{path of all template files}

\item{overwrite}{A logical value indicating whether to overwrite existing files in the
project directory. Defaults to FALSE, meaning existing files will not be overwritten.}
}
\value{
This function returns an invisible NULL. It is used for its side effects of
creating directories and copying files rather than producing a value.
}
\description{
This function initializes a project directory by creating necessary subdirectories
and copying configuration files from a source Excel file. It reads the directory
structure and files specified in the provided Excel template and sets up the project
environment accordingly.
}
\seealso{
Other project initialization: 
\code{\link{ProjectConfigurationRF}},
\code{\link{createProjectConfiguration}()},
\code{\link{getQCpassedEnvironmentVariable}()},
\code{\link{setWorkflowOptions}()}
}
\concept{project initialization}
