% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{initProject}
\alias{initProject}
\title{Initialize Project}
\usage{
initProject(projectPath = ".", sourceFolder, overwrite = FALSE)
}
\arguments{
\item{projectPath}{A string defining the path where to initialize the project.
default to current working directory.}

\item{sourceFolder}{path of template directory available is
`templateDirectory()` default path of `OSPSuite.ReportingFramework`
`esqlabsR:::example_directory("TestProject")`  default path for Esqlabs-projects}

\item{overwrite}{A boolean, if TRUE existing files will be overwritten}
}
\value{
path of project
}
\description{
Creates the default project folder structure with excels file templates in
the working directory.
}
