% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{loadPkAnalysis}
\alias{loadPkAnalysis}
\title{Load PK Analysis from CSV}
\usage{
loadPkAnalysis(
  projectConfiguration,
  scenarioName,
  scenarioResult,
  withRecalculation
)
}
\arguments{
\item{scenarioResult}{A list containing the results of the scenario.}

\item{withRecalculation}{A boolean indicating whether to recalculate PK parameters if results already exist.}

\item{fileName}{The path to the CSV file containing PK analysis results.}
}
\value{
A data.table containing the PK analyses.
}
\description{
This function loads PK analysis results from a CSV file or recalculates them if specified.
}
