% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{loadPKAnalysisPerScenario}
\alias{loadPKAnalysisPerScenario}
\title{Load or Calculate PK Analysis for a Scenario}
\usage{
loadPKAnalysisPerScenario(
  scenarioName,
  scenarioResult,
  pkParameterSheets,
  withRecalculation,
  projectConfiguration
)
}
\arguments{
\item{scenarioName}{The name of the scenario to process.}

\item{scenarioResult}{A list containing the results of the scenario.}

\item{pkParameterSheets}{A vector with names of the sheets in the PK parameter file to read.}

\item{withRecalculation}{A boolean indicating whether to recalculate PK parameters if results already exist.}

\item{projectConfiguration}{A list containing project configuration, including the PK parameter file.}

\item{outputFolder}{The folder where output files are stored.}
}
\value{
A data.table containing the processed PK analyses.
}
\description{
This function loads PK analysis results from a CSV file or recalculates them if specified.
It also processes each specified PK parameter sheet for the scenario.
}
