% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{loadPkAnalysisRawData}
\alias{loadPkAnalysisRawData}
\title{Load PK Analysis from CSV}
\usage{
loadPkAnalysisRawData(projectConfiguration, scenarioName, scenario)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration, including output folder and PK parameter file.}

\item{scenarioName}{name of scenario}
}
\value{
A data.table containing the PK analyses.
}
\description{
This function loads PK analysis results from a CSV file or recalculates them if specified.
}
