% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pkParameter.R
\name{loadPKParameter}
\alias{loadPKParameter}
\title{Load PK Parameter}
\usage{
loadPKParameter(projectConfiguration, scenarioList)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration, including the PK parameter file.}

\item{scenarioList}{A list of scenarios for which PK parameters are to be loaded.}
}
\value{
A data.table containing the processed PK analyses.
}
\description{
This function loads PK parameters for specified scenarios based on project configuration and scenario list.
}
