% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{loadScenarioResultsToFramework}
\alias{loadScenarioResultsToFramework}
\title{Load existing scenario results}
\usage{
loadScenarioResultsToFramework(projectConfiguration, scenarioNames)
}
\arguments{
\item{projectConfiguration}{Configuration for the project, containing paths and settings necessary
to load the results.}

\item{scenarioNames}{Character vector of the names of the scenarios whose results are to be loaded.}
}
\value{
A list containing the loaded scenario results, including population data if available.
throws Error if the scenario results do not exist.
}
\description{
This function loads the results of specified scenarios. If the results do not exist,
it returns an error.
}
\seealso{
Other scenario management: 
\code{\link{calculatePKParameterForScenarios}()},
\code{\link{createScenarios.wrapped}()},
\code{\link{loadPKParameter}()},
\code{\link{runAndSaveScenarios}()},
\code{\link{runOrLoadScenarios}()}
}
\concept{scenario management}
