% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{logCatch}
\alias{logCatch}
\title{function that catches messages, warnings, and errors.
This function has to be initialized by  `initLogfunction` function}
\usage{
logCatch(expr)
}
\arguments{
\item{expr}{The expression to evaluate.}
}
\description{
function that catches messages, warnings, and errors.
This function has to be initialized by  `initLogfunction` function
}
