% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdBullet}
\alias{mdBullet}
\title{Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list}
\usage{
mdBullet(..., bullet = "-", level = 1)
}
\arguments{
\item{...}{passed to `mdPaste`}

\item{bullet}{`character` type of bullet}

\item{level}{`integer`, specifies the indentation level. `level=1` has no indentation}
}
\description{
Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet0]{mdBullet0()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
