% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdBullet}
\alias{mdBullet}
\title{Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list}
\usage{
mdBullet(..., bullet = "-", level = 1)
}
\arguments{
\item{...}{passed to `mdPaste`}

\item{bullet}{`character` type of bullet}

\item{level}{`integer`, specifies the indentation level. `level=1` has no indentation}
}
\description{
Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list
}
