% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdBullet0}
\alias{mdBullet0}
\title{Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list}
\usage{
mdBullet0(..., bullet = "-", level = 1)
}
\arguments{
\item{...}{passed to `mdPaste0`}

\item{bullet}{`character` type of bullet}

\item{level}{`integer`, specifies the indentation level. `level=1` has no indentation}
}
\description{
Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list
}
\seealso{
Other markdown helper function: 
\code{\link{addFiguresAndTables}()},
\code{\link{mdBullet}()},
\code{\link{mdCaption}()},
\code{\link{mdFigure}()},
\code{\link{mdFootNote}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdNewpage}()},
\code{\link{mdPaste}()},
\code{\link{mdPaste0}()},
\code{\link{mdTable}()}
}
\concept{markdown helper function}
