% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markdown.R
\name{mdBullet0}
\alias{mdBullet0}
\title{Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list}
\usage{
mdBullet0(..., bullet = "-", level = 1)
}
\arguments{
\item{...}{passed to `mdPaste0`}

\item{bullet}{`character` type of bullet}

\item{level}{`integer`, specifies the indentation level. `level=1` has no indentation}
}
\description{
Insert a bulleted item, making sure that there are sufficient newlines to build a bulleted list
}
