% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdCaption}
\alias{mdCaption}
\title{Add caption with optional custom style.}
\usage{
mdCaption(subfolder, captionFile, captionPrefix, captionStyle = NULL)
}
\arguments{
\item{subfolder}{Folder location.}

\item{captionFile}{Caption file.}

\item{captionPrefix}{Prefix (e.g., "Figure 1:").}

\item{captionStyle}{Custom style (optional).}
}
\description{
Add caption with optional custom style.
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
