% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdCaption}
\alias{mdCaption}
\title{add Caption to .Rmd}
\usage{
mdCaption(subfolder, captionFile, captionPrefix, captionStyle = NULL)
}
\arguments{
\item{subfolder}{The folder where the file is located relative to Rmd}

\item{captionFile}{file containing caption}

\item{captionPrefix}{'a character which starts the caption like 'Figure 1:'}

\item{captionStyle}{'custom-style for captions'}
}
\description{
add Caption to .Rmd
}
\seealso{
Other markdown helper function: 
\code{\link{addFiguresAndTables}()},
\code{\link{mdBullet}()},
\code{\link{mdBullet0}()},
\code{\link{mdFigure}()},
\code{\link{mdFootNote}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdNewpage}()},
\code{\link{mdPaste}()},
\code{\link{mdPaste0}()},
\code{\link{mdTable}()}
}
\concept{markdown helper function}
