% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markdown.R
\name{mdCaption}
\alias{mdCaption}
\title{add Caption to .Rmd}
\usage{
mdCaption(subfolder, captionFile, captionPrefix, captionStyle = NULL)
}
\arguments{
\item{subfolder}{The folder where the file is located relative to Rmd}

\item{captionFile}{file containing caption}

\item{captionPrefix}{'a character which starts the caption like 'Figure 1:'}

\item{captionStyle}{'custom-style for captions'}
}
\description{
add Caption to .Rmd
}
