% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdFigure}
\alias{mdFigure}
\title{Include figure with caption and footnote.}
\usage{
mdFigure(
  figureNumber,
  figureFile,
  captionFile,
  footNoteFile = NULL,
  subfolder,
  addNewPage = TRUE,
  customStyles = list()
)
}
\arguments{
\item{figureNumber}{Figure number for caption prefix.}

\item{figureFile}{Figure file name.}

\item{captionFile}{Caption file.}

\item{footNoteFile}{Footnote file (optional).}

\item{subfolder}{Folder relative to Rmd.}

\item{addNewPage}{Add page break after (default TRUE).}

\item{customStyles}{Custom styles list.}
}
\description{
Include figure with caption and footnote.
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
