% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markdown.R
\name{mdFigure}
\alias{mdFigure}
\title{mdFigure}
\usage{
mdFigure(
  figureNumber,
  figureFile,
  captionFile,
  footNoteFile = NULL,
  subfolder,
  addNewPage = TRUE,
  customStyles = list()
)
}
\arguments{
\item{figureNumber}{number of figure used in caption prefix}

\item{figureFile}{file of Figure}

\item{captionFile}{file containing caption}

\item{footNoteFile}{file containing footnotes}

\item{subfolder}{The folder where the file is located relative to Rmd}

\item{addNewPage}{boolean if TRUE (default) new page is added after}

\item{customStyles}{list of custom styles usable for figure and table captions and footnotes}
}
\description{
Include a figure file with caption
}
