% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdFootNote}
\alias{mdFootNote}
\title{Add footnote lines from file.}
\usage{
mdFootNote(subfolder, footNoteFile, footNoteCustomStyle = NULL)
}
\arguments{
\item{subfolder}{Folder location.}

\item{footNoteFile}{Footnote file.}

\item{footNoteCustomStyle}{Custom style (optional).}
}
\description{
Add footnote lines from file.
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
