% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdFootNote}
\alias{mdFootNote}
\title{add footnote lines}
\usage{
mdFootNote(subfolder, footNoteFile, footNoteCustomStyle = NULL)
}
\arguments{
\item{subfolder}{The folder where the file is located relative to Rmd}

\item{footNoteFile}{file containing footnotes}

\item{footNoteCustomStyle}{a character describing custom style footnotes}
}
\description{
add footnote lines
}
\seealso{
Other markdown helper function: 
\code{\link{addFiguresAndTables}()},
\code{\link{mdBullet}()},
\code{\link{mdBullet0}()},
\code{\link{mdCaption}()},
\code{\link{mdFigure}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdNewpage}()},
\code{\link{mdPaste}()},
\code{\link{mdPaste0}()},
\code{\link{mdTable}()}
}
\concept{markdown helper function}
