% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_utilities.R
\name{mdFootNote}
\alias{mdFootNote}
\title{add footnote lines}
\usage{
mdFootNote(subfolder, footNoteFile, footNoteCustomStyle = NULL)
}
\arguments{
\item{subfolder}{The folder where the file is located relative to Rmd}

\item{footNoteFile}{file containing footnotes}

\item{footNoteCustomStyle}{a character describing custom style footnotes}
}
\description{
add footnote lines
}
