% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdHeading}
\alias{mdHeading}
\title{mdHeading
Insert heading with specified level (1-6).}
\usage{
mdHeading(..., level = 1, newlines = 2)
}
\arguments{
\item{...}{Content passed to mdPaste.}

\item{level}{Header level (default 1).}

\item{newlines}{Newlines after heading (default 2).}
}
\description{
mdHeading
Insert heading with specified level (1-6).
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
