% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdHeading}
\alias{mdHeading}
\title{mdHeading}
\usage{
mdHeading(..., level = 1, newlines = 2)
}
\arguments{
\item{...}{passed to mdPaste}

\item{level}{The header level, i.e. the number of `#`s. Defaults to 1.}

\item{newlines}{The number of newlines inserted after the heading. Defaults to 2.}
}
\description{
Insert a heading with a specified level.
}
