% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdLink}
\alias{mdLink}
\title{mdLink}
\usage{
mdLink(label, filename, folder, prefix = "")
}
\arguments{
\item{label}{The text to display as the link.}

\item{filename}{The file to link to.}

\item{folder}{The folder where the file is located.}

\item{prefix}{An optional prefix to place in front of the link. Defaults to "".}
}
\description{
Insert a link to another file, and a newline.
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdBullet0]{mdBullet0()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}},
\code{\link[=mdTable]{mdTable()}}
}
\concept{markdown helper function}
