% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdLink}
\alias{mdLink}
\title{mdLink}
\usage{
mdLink(label, filename, folder, prefix = "")
}
\arguments{
\item{label}{The text to display as the link.}

\item{filename}{The file to link to.}

\item{folder}{The folder where the file is located.}

\item{prefix}{An optional prefix to place in front of the link. Defaults to "".}
}
\description{
Insert a link to another file, and a newline.
}
