% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdNewpage}
\alias{mdNewpage}
\title{mdNewpage}
\usage{
mdNewpage()
}
\description{
Insert a page break and a newline.
}
\seealso{
Other markdown helper function: 
\code{\link{addFiguresAndTables}()},
\code{\link{mdBullet}()},
\code{\link{mdBullet0}()},
\code{\link{mdCaption}()},
\code{\link{mdFigure}()},
\code{\link{mdFootNote}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdPaste}()},
\code{\link{mdPaste0}()},
\code{\link{mdTable}()}
}
\concept{markdown helper function}
