% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdPaste}
\alias{mdPaste}
\title{mdPaste}
\usage{
mdPaste(..., sep = " ", collapse = NULL, newlines = 1)
}
\arguments{
\item{...}{Elements to be `paste`d.}

\item{sep}{Passed to `paste`. Defaults to `" "`.}

\item{collapse}{Passed to `paste`. Defaults to `NULL`.}

\item{newlines}{Number of newlines to insert afterwards. Defaults to 1.}
}
\description{
`cat` text using `paste` and optionally end with newlines.
}
\seealso{
Other markdown helper function: 
\code{\link{addFiguresAndTables}()},
\code{\link{mdBullet}()},
\code{\link{mdBullet0}()},
\code{\link{mdCaption}()},
\code{\link{mdFigure}()},
\code{\link{mdFootNote}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdNewpage}()},
\code{\link{mdPaste0}()},
\code{\link{mdTable}()}
}
\concept{markdown helper function}
