% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markdown.R
\name{mdPaste0}
\alias{mdPaste0}
\title{mdPaste0}
\usage{
mdPaste0(..., collapse = NULL, newlines = 1)
}
\arguments{
\item{...}{Elements to be `paste0`d.}

\item{collapse}{Passed to `paste0`. Defaults to `NULL`.}

\item{newlines}{Number of newlines to insert afterwards. Defaults to 1.}
}
\description{
`cat` text using `paste0` and optionally end with newlines.
}
