% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdTable}
\alias{mdTable}
\title{Insert table from CSV with caption.}
\usage{
mdTable(
  tableNumber,
  tableCsv,
  captionFile,
  footNoteFile,
  subfolder,
  customStyles,
  digitsOfSignificance = 3,
  addNewPage = TRUE,
  ...
)
}
\arguments{
\item{tableNumber}{Table number for prefix.}

\item{tableCsv}{CSV file path.}

\item{captionFile}{Caption file.}

\item{footNoteFile}{Footnote file.}

\item{subfolder}{Folder location.}

\item{customStyles}{Custom styles list.}

\item{digitsOfSignificance}{Significant digits (default 3).}

\item{addNewPage}{Add page break after (default TRUE).}

\item{...}{Passed to `knitr::kable`.}
}
\description{
Insert table from CSV with caption.
}
\seealso{
Other markdown helper function:
\code{\link[=addFiguresAndTables]{addFiguresAndTables()}},
\code{\link[=mdBullet]{mdBullet()}},
\code{\link[=mdCaption]{mdCaption()}},
\code{\link[=mdFigure]{mdFigure()}},
\code{\link[=mdFootNote]{mdFootNote()}},
\code{\link[=mdHeading]{mdHeading()}},
\code{\link[=mdLink]{mdLink()}},
\code{\link[=mdNewline]{mdNewline()}},
\code{\link[=mdNewpage]{mdNewpage()}},
\code{\link[=mdPaste]{mdPaste()}},
\code{\link[=mdPaste0]{mdPaste0()}}
}
\concept{markdown helper function}
