% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mdTable}
\alias{mdTable}
\title{mdTable}
\usage{
mdTable(
  tableNumber,
  tableCsv,
  captionFile,
  footNoteFile,
  subfolder,
  customStyles,
  digitsOfSignificance = 3,
  addNewPage = TRUE,
  ...
)
}
\arguments{
\item{tableNumber}{number of table for caption prefix}

\item{tableCsv}{path of .csv file}

\item{captionFile}{file containing caption}

\item{footNoteFile}{file containing footnotes}

\item{subfolder}{The folder where the file is located relative to Rmd}

\item{customStyles}{list of custom styles usable for figure and table captions and footnotes}

\item{digitsOfSignificance}{significance digits to display (default 3)}

\item{addNewPage}{boolean if TRUE (default) new page is added after}

\item{...}{passed to `knitr::kable`}
}
\value{
`x`, invisibly
}
\description{
Insert a table into a markdown document. Essentially a wrapper for `knitr::kable` with `format = "markdown"` and newlines.
}
\seealso{
Other markdown helper function: 
\code{\link{addFiguresAndTables}()},
\code{\link{mdBullet}()},
\code{\link{mdBullet0}()},
\code{\link{mdCaption}()},
\code{\link{mdFigure}()},
\code{\link{mdFootNote}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdNewpage}()},
\code{\link{mdPaste}()},
\code{\link{mdPaste0}()}
}
\concept{markdown helper function}
