% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_markdown.R
\name{mdTable}
\alias{mdTable}
\title{mdTable}
\usage{
mdTable(
  tableNumber,
  tableCsv,
  captionFile,
  footNoteFile,
  subfolder,
  customStyles,
  digitsOfSignificance = 3,
  ...
)
}
\arguments{
\item{tableNumber}{number of table for caption prefix}

\item{tableCsv}{path of .csv file}

\item{captionFile}{file containing caption}

\item{footNoteFile}{file containing footnotes}

\item{subfolder}{The folder where the file is located relative to Rmd}

\item{customStyles}{list of custom styles usable for figure and table captions and footnotes}

\item{digitsOfSignificance}{significance digits to display (default 3)}

\item{...}{passed to `knitr::kable`}
}
\value{
`x`, invisibly
}
\description{
Insert a table into a markdown document. Essentially a wrapper for `knitr::kable` with `format = "markdown"` and newlines.
}
