% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{mergePkData}
\alias{mergePkData}
\title{Merge Pharmacokinetic PK Parameter to config Table}
\usage{
mergePkData(
  configTable,
  pkParameterObserved,
  columns = c("x", "xMin", "xMax", "numberOfObservations")
)
}
\arguments{
\item{configTable}{A data.table containing the configuration data.}

\item{pkParameterObserved}{A data.table containing observed pharmacokinetic parameters.}

\item{columns}{A character vector of additional columns to include in the merge.}
}
\value{
A data.table resulting from the merge operation with missing columns added as NA.
}
\description{
This function merges a configuration table with observed pharmacokinetic parameters.
}
\keyword{internal}
