% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{mergeRmds}
\alias{mergeRmds}
\title{Creates Rmd file which include specified  .Rmd s}
\usage{
mergeRmds(
  newName = "appendix",
  title = "Appendix",
  sourceRmds = c("Demographics", "TimeProfile", "PKParameter", "DDIRatio", "myFigures"),
  projectConfiguration
)
}
\arguments{
\item{newName}{new Name of .Rmd}

\item{title}{Title of new Rmd}

\item{sourceRmds}{list of .rmds to include}

\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\description{
Creates Rmd file which include specified  .Rmd s
}
