% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{openEPackageTemplate}
\alias{openEPackageTemplate}
\title{Opens the  template for figure creation as new document}
\usage{
openEPackageTemplate()
}
\description{
Opens the  template for figure creation as new document
}
