% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{openFigureTemplate}
\alias{openFigureTemplate}
\title{Opens the  template for figure creation as new document}
\usage{
openFigureTemplate()
}
\description{
Opens the  template for figure creation as new document
}
\seealso{
Other Addins Support:
\code{\link[=createDocumentFromTemplate]{createDocumentFromTemplate()}},
\code{\link[=createWorkflowTemplate]{createWorkflowTemplate()}},
\code{\link[=openEPackageTemplate]{openEPackageTemplate()}},
\code{\link[=openWorkflowTemplate]{openWorkflowTemplate()}}
}
\concept{Addins Support}
