% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-templates.R
\name{openFigureTemplate}
\alias{openFigureTemplate}
\title{Opens the  template for figure creation as new document}
\usage{
openFigureTemplate()
}
\description{
Opens the  template for figure creation as new document
}
\seealso{
Other Addins Support: 
\code{\link{createDocumentFromTemplate}()},
\code{\link{createWorkflowTemplate}()},
\code{\link{openEPackageTemplate}()},
\code{\link{openWorkflowTemplate}()}
}
\concept{Addins Support}
