% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSPSuite.ReportingFramework-package.R
\docType{package}
\name{ospsuite.reportingframework-package}
\alias{ospsuite.reportingframework}
\alias{ospsuite.reportingframework-package}
\title{ospsuite.reportingframework: Provides a framework for report generation with the ospsuite}
\description{
This package uses the EsqlabsR package to set up simulations and scenerios. It provides functions to generate default graphics and tables. This figures are included in .Rmd files which are then merged to the selected output format e.g. a "Report.docx".
}
\author{
\strong{Maintainer}: Katrin Coboeken \email{katrin.coboeken@bayer.com}

Other contributors:
\itemize{
  \item Open-Systems-Pharmacology Community [copyright holder]
}

}
\keyword{internal}
