% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotPredictedVsObserved}
\alias{ospsuite_plotPredictedVsObserved}
\title{Plots predicted vs observed data, grouped by "group".}
\usage{
ospsuite_plotPredictedVsObserved(plotData, metaData = NULL, ...)
}
\arguments{
\item{plotData}{An object of class DataCombined or a data.frame generated by a DataCombined object.
Units should be consistent; if not, use `convertUnits(myDataCombined)` to standardize.}

\item{metaData}{A list containing metadata for the plot. If NULL, a default list is constructed from the data.}

\item{...}{Additional arguments passed to `ospsuite.plots::plotPredVsObs`.}
}
\value{
A `ggplot2` plot object representing predicted vs observed values.
}
\description{
This function visualizes the relationship between predicted and observed values, allowing for easy identification of discrepancies.
}
\examples{
\dontrun{
# Generate a predicted vs observed plot for the provided data
plotPredictedVsObserved(convertUnits(
  myDataCombined,
  xUnit = ospUnits$Time$h,
  yUnit = ospUnits$`Concentration [mass]`$`µg/l`
))
}
}
