% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotResidualsVsTime}
\alias{ospsuite_plotResidualsVsTime}
\title{Plots residuals vs Time, grouped by "group".}
\usage{
ospsuite_plotResidualsVsTime(plotData, metaData = NULL, ...)
}
\arguments{
\item{plotData}{An object of class DataCombined or a data.frame generated by a DataCombined object.
Units should be consistent; if not, use `convertUnits(myDataCombined)` to standardize.}

\item{metaData}{A list containing metadata for the plot. If NULL, a default list is constructed from the data.}

\item{...}{Additional arguments passed to `ospsuite.plots::plotResVsCov`.}
}
\value{
A `ggplot2` plot object representing residuals vs time.
}
\description{
This function visualizes the residuals over time, providing insights into the accuracy of the model predictions.
}
\examples{
\dontrun{
# Generate a residuals vs time plot for the provided data
plotResidualsVsTime(convertUnits(
  myDataCombined,
  xUnit = ospUnits$Time$h,
  yUnit = ospUnits$`Concentration [mass]`$`µg/l`
))
}
}
