% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite_plotTimeprofile.R
\name{ospsuite_plotTimeProfile}
\alias{ospsuite_plotTimeProfile}
\title{Creates a time profile plot for given data.}
\usage{
ospsuite_plotTimeProfile(plotData, metaData = NULL, ...)
}
\arguments{
\item{plotData}{An object of class DataCombined or a data.frame generated by a DataCombined object.
It is essential that the units are consistent; if not, use `convertUnits(myDataCombined)` to standardize.}

\item{metaData}{A list containing metadata for the plot. If NULL, a default list is constructed from the data.}

\item{...}{Additional arguments passed to `ospsuite.plots::plotTimeProfile`.}
}
\value{
A `ggplot2` plot object representing the time profile.
}
\description{
This function generates a time profile plot using ggplot2, where the data is grouped by a column named "group".
}
\examples{
\dontrun{
# Generate a time profile plot for the provided data
plotTimeProfile(convertUnits(
  myDataCombined,
  xUnit = ospUnits$Time$h,
  yUnit = ospUnits$`Concentration [mass]`$`µg/l`
))
}
}
