% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-aggregation.R
\name{performAggregation}
\alias{performAggregation}
\title{Perform Aggregation}
\usage{
performAggregation(dataToAggregate, aggregationFun, aggrCriteria)
}
\arguments{
\item{dataToAggregate}{A data.table containing the data to be aggregated.}

\item{aggregationFun}{A function to aggregate the data.}

\item{aggrCriteria}{A character vector specifying the columns to group by.}
}
\value{
A data.table containing aggregated results with counts and aggregated values.
}
\description{
This function performs the aggregation of observed data based on the specified criteria.
}
