% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\docType{class}
\name{PlotDataTimeProfile}
\alias{PlotDataTimeProfile}
\title{plotData for Timeprofile plotData}
\value{
RmdContainer object
Collects all observed and simulated data needed for this plots
}
\description{
An object to collect plotData and other informations to create a timeprofile plot
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{RmdContainer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{digitsOfSignificance}}{digits for significance in table display
load observed data used in this panel}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{data For Plot}

\item{\code{digitsOfSignificance}}{digits for significance in table display
load observed data used in this panel}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RmdContainer-new}{\code{PlotDataTimeProfile$new()}}
\item \href{#method-RmdContainer-collectData}{\code{PlotDataTimeProfile$collectData()}}
\item \href{#method-RmdContainer-clone}{\code{PlotDataTimeProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$new(
  projectConfiguration = projectConfiguration,
  onePlotConfig = onePlotConfig
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{\code{projectConfiguration}}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{\code{onePlotConfig}}{plotconfiguration for one plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-collectData"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-collectData}{}}}
\subsection{Method \code{collectData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$collectData(projectConfiguration, dataObserved)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{\code{projectConfiguration}}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
