% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\docType{class}
\name{PlotDataTimeProfile}
\alias{PlotDataTimeProfile}
\title{plotData for Timeprofile plotData}
\value{
RmdContainer object
Collects all observed and simulated data needed for this plots
}
\description{
An object to collect plotData and other informations to create a timeprofile plot
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{RmdContainer}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{data For Plot}

\item{\code{configTable}}{configuration table for one plot}

\item{\code{dtCaption}}{`digits for significance in table display`data.table`with caption information`}

\item{\code{scaleVectors}}{list with scaling vectors to manually scale aestehtics}

\item{\code{tpLabelSimulatedMean}}{label for simulated mean}

\item{\code{tpLabelSimulatedRange}}{label for simulated range}

\item{\code{tpLabelObserved}}{label for observed data}

\item{\code{dtCaption}}{list of time profile labels to add on ggplot}

\item{\code{dtCaption}}{`digits for significance in table display`data.table`with caption information`
loads simulated result and convert for plotting

if population scenario, time profiles are aggregated
Units are converted to desired units
load observed data used in this panel}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RmdContainer-new}{\code{PlotDataTimeProfile$new()}}
\item \href{#method-RmdContainer-collectData}{\code{PlotDataTimeProfile$collectData()}}
\item \href{#method-RmdContainer-setOrderAndFactors}{\code{PlotDataTimeProfile$setOrderAndFactors()}}
\item \href{#method-RmdContainer-addDtCaption}{\code{PlotDataTimeProfile$addDtCaption()}}
\item \href{#method-RmdContainer-setColorIndex}{\code{PlotDataTimeProfile$setColorIndex()}}
\item \href{#method-RmdContainer-getCaptionTimeProfile}{\code{PlotDataTimeProfile$getCaptionTimeProfile()}}
\item \href{#method-RmdContainer-getFootNoteLines}{\code{PlotDataTimeProfile$getFootNoteLines()}}
\item \href{#method-RmdContainer-clone}{\code{PlotDataTimeProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$new(
  projectConfiguration = projectConfiguration,
  onePlotConfig = onePlotConfig
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{\code{onePlotConfig}}{plotconfiguration for one plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-collectData"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-collectData}{}}}
\subsection{Method \code{collectData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$collectData(projectConfiguration, dataObserved)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{projectConfiguration}

\item{\code{observedDataDT}}{'data.table'  with observed data
setfactors to plot}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-setOrderAndFactors"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-setOrderAndFactors}{}}}
\subsection{Method \code{setOrderAndFactors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$setOrderAndFactors(identifierColumns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifierColumns}}{identifier columns which define sort order
Creates PlotId for each scenario outputs group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-addDtCaption"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-addDtCaption}{}}}
\subsection{Method \code{addDtCaption()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$addDtCaption(nFacetColumns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nFacetColumns}}{number of maximal facet columns
add colorIndex column
creates a caption text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-setColorIndex"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-setColorIndex}{}}}
\subsection{Method \code{setColorIndex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$setColorIndex()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-getCaptionTimeProfile"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-getCaptionTimeProfile}{}}}
\subsection{Method \code{getCaptionTimeProfile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$getCaptionTimeProfile(yScale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yScale}}{scale of Y axis}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`character` with caption text
constructs footnote lines for aggregated data and data references
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-getFootNoteLines"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-getFootNoteLines}{}}}
\subsection{Method \code{getFootNoteLines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$getFootNoteLines()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
vector of characters, each entry is one footnote line
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
