% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\docType{class}
\name{PlotDataTimeProfile}
\alias{PlotDataTimeProfile}
\title{plotData for Timeprofile plotData}
\value{
RmdContainer object
get the timelabel for the filtered timerange
}
\description{
An object to collect plotData and other informations to create a timeprofile plot
}
\section{Super class}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{RmdContainer}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{data For Plot}

\item{\code{dataReference}}{data with references for data}

\item{\code{configTable}}{configuration table for one plot}

\item{\code{dtCaption}}{`digits for significance in table display`data.table`with caption information`}

\item{\code{scaleVectors}}{list with scaling vectors to manually scale aesthetics}

\item{\code{scaleVectorsObserved}}{list with scaling vectors to manually scale aesthetics of observed data}

\item{\code{tpLabelSimulatedMean}}{label for simulated mean}

\item{\code{tpLabelSimulatedRange}}{label for simulated range}

\item{\code{tpLabelObserved}}{label for observed data}

\item{\code{tpLabels}}{list of time profile labels to add on ggplot}

\item{\code{nFacetColumns}}{number of facet columns}

\item{\code{timeRangeTagFilter}}{list with filters for time range tags
Collects all observed and simulated data needed for this plots
helper function to extract time profiles for results
setfactors to plot
Creates PlotId for each scenario outputs group}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RmdContainer-new}{\code{PlotDataTimeProfile$new()}}
\item \href{#method-RmdContainer-getTimeLabelForTimeRange}{\code{PlotDataTimeProfile$getTimeLabelForTimeRange()}}
\item \href{#method-RmdContainer-getDataForTimeRange}{\code{PlotDataTimeProfile$getDataForTimeRange()}}
\item \href{#method-RmdContainer-useColorIndex}{\code{PlotDataTimeProfile$useColorIndex()}}
\item \href{#method-RmdContainer-useShapeIndex}{\code{PlotDataTimeProfile$useShapeIndex()}}
\item \href{#method-RmdContainer-hasSimulatedPop}{\code{PlotDataTimeProfile$hasSimulatedPop()}}
\item \href{#method-RmdContainer-hasObservedData}{\code{PlotDataTimeProfile$hasObservedData()}}
\item \href{#method-RmdContainer-clone}{\code{PlotDataTimeProfile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-new"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$new(
  projectConfiguration = projectConfiguration,
  onePlotConfig = onePlotConfig,
  dataObserved = dataObserved,
  aggregationFun = aggregationFun,
  nFacetColumns = nFacetColumns
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfiguration}}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{\code{onePlotConfig}}{plotconfiguration for one plot}

\item{\code{aggregationFun}}{function to aggregate simulated data}

\item{\code{default}}{number of facet columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-getTimeLabelForTimeRange"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-getTimeLabelForTimeRange}{}}}
\subsection{Method \code{getTimeLabelForTimeRange()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$getTimeLabelForTimeRange(filterName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filterName}}{character with name of time range filter
get the data for the filtered time range}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-getDataForTimeRange"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-getDataForTimeRange}{}}}
\subsection{Method \code{getDataForTimeRange()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$getDataForTimeRange(filterName, typeFilter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filterName}}{name of timeRange filter}

\item{\code{typeFilter}}{filter for data type}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`data.table` with filtered plotData
function to determine if color legend is needed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-useColorIndex"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-useColorIndex}{}}}
\subsection{Method \code{useColorIndex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$useColorIndex()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical
function to determine if shape legend is needed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-useShapeIndex"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-useShapeIndex}{}}}
\subsection{Method \code{useShapeIndex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$useShapeIndex()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical
function to determine if simulation is aggregated
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-hasSimulatedPop"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-hasSimulatedPop}{}}}
\subsection{Method \code{hasSimulatedPop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$hasSimulatedPop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical
function to determine if data contains oberved data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-hasObservedData"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-hasObservedData}{}}}
\subsection{Method \code{hasObservedData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$hasObservedData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RmdContainer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RmdContainer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotDataTimeProfile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
