% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{plotHistograms}
\alias{plotHistograms}
\title{Plot Histograms}
\usage{
plotHistograms(
  projectConfiguration,
  onePlotConfig,
  pkParameterDT = NULL,
  scenarioList = NULL,
  facetAspectRatio = 0.5,
  colorVector = c(scenario = NA, referenceScenario = NA),
  nMaxFacetRows = 2,
  ...
)
}
\arguments{
\item{projectConfiguration}{A configuration object for the project.}

\item{onePlotConfig}{A configuration for the specific plot.}

\item{pkParameterDT}{A data.table containing PK parameter data (optional).}

\item{scenarioList}{A list of scenarios to consider (optional).}

\item{facetAspectRatio}{A numeric value for the aspect ratio of the facets (default is 0.5).}

\item{colorVector}{A named vector for colors corresponding to scenarios (default is c(scenario = NA, referenceScenario = NA)).}

\item{nMaxFacetRows}{Maximum number of facet rows (default is 2).}

\item{...}{Additional arguments passed to the histogram plotting function.}
}
\value{
A list of histogram plot objects generated based on the input configuration.
}
\description{
This function generates histogram plots based on the provided configuration and data.
}
