% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotOfOneTimeProfilePanel}
\alias{plotOfOneTimeProfilePanel}
\title{Title}
\usage{
plotOfOneTimeProfilePanel(
  projectConfiguration,
  panelConfig,
  observedData,
  rmdContainer
)
}
\arguments{
\item{rmdContainer}{}
}
\description{
Title
}
