% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKBoxwhisker.R
\name{plotPKBoxwhisker}
\alias{plotPKBoxwhisker}
\title{Plot PK Box-and-Whisker}
\usage{
plotPKBoxwhisker(
  projectConfiguration,
  onePlotConfig,
  pkParameterDT,
  percentiles = getOspsuite.plots.option(optionKey = OptionKeys$Percentiles),
  xAxisTextAngle = 0,
  colorVector = c(scenario = NA, referenceScenario = NA),
  facetAspectRatio = 0.5,
  ...
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object that contains
settings and paths relevant to the project.}

\item{onePlotConfig}{A data.table containing configuration settings for a single plot,
including details about which plots to generate and their formatting.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{percentiles}{A numeric vector specifying the percentiles to calculate
for the box-and-whisker plots. Default is retrieved from project options.}

\item{xAxisTextAngle}{An integer specifying the angle (in degrees) for
rotating the x-axis text labels. Default is 0 (no rotation).}

\item{colorVector}{A named vector specifying colors for the plots. Names should correspond to the characters
defined in the configtable column colorLegend. If your color legend is for example 'DDI|control',
your color vector could be colorVector = c(DDI = 'red',control = 'blue'), if no color is defined default values are used.}

\item{facetAspectRatio}{A numeric value specifying the aspect ratio for
faceting the plots. Default is 0.5, which may need adjustment based on the
number of facets and plot dimensions.}

\item{...}{Additional arguments passed to plotting functions for further
customization.}
}
\value{
A list of ggplot objects representing the generated plots. Each plot
  can be rendered using ggplot2 or similar plotting systems. The list may include both absolute
  and ratio plots depending on the configuration.
}
\description{
Generates box-and-whisker plots for pharmacokinetic (PK) parameters
based on a provided project configuration and plot configuration. This function creates
visual representations of the distribution of PK parameters across different
scenarios, allowing for comparison of absolute values and ratios.
}
\details{
The function is intended to be used in combination with the `runPlot` function.
See the vignette `Plot and Report Generation` for more details.

There exists a helper function `addDefaultConfigForPKBoxwhsikerPlots` which is a helpful utility designed to
streamline the process of setting up your Excel configuration sheet for box whisker plots.
This function automatically populates a new sheet with default values that you can
customize according to your needs.
}
\examples{
# Example usage of plotPKBoxwhisker
\dontrun{
plotList <- plotPKBoxwhisker(
  projectConfiguration = myProjectConfig,
  onePlotConfig = myPlotConfig,
  pkParameterDT = myPKData,
  percentiles = c(0.05, 0.25, 0.5, 0.75, 0.95),
  xAxisTextAngle = 45,
  colorVector = c(DDI = "red", control = "blue"),
  facetAspectRatio = 1,
)
}
}
\seealso{
Other plot functions:
\code{\link[=addDefaultConfigForPKForestPlots]{addDefaultConfigForPKForestPlots()}},
\code{\link[=addDefaultConfigForTimeProfilePlots]{addDefaultConfigForTimeProfilePlots()}},
\code{\link[=plotDistributionVsDemographics]{plotDistributionVsDemographics()}},
\code{\link[=plotHistograms]{plotHistograms()}},
\code{\link[=plotPKForestAggregatedAbsoluteValues]{plotPKForestAggregatedAbsoluteValues()}},
\code{\link[=plotPKForestAggregatedRatios]{plotPKForestAggregatedRatios()}},
\code{\link[=plotPKForestPointEstimateOfAbsoluteValues]{plotPKForestPointEstimateOfAbsoluteValues()}},
\code{\link[=plotPKForestPointEstimateOfRatios]{plotPKForestPointEstimateOfRatios()}},
\code{\link[=plotSensitivity]{plotSensitivity()}},
\code{\link[=plotTimeProfiles]{plotTimeProfiles()}}

Other functions to generate box whisker plots:
\code{\link[=addDefaultConfigForPKBoxwhsikerPlots]{addDefaultConfigForPKBoxwhsikerPlots()}},
\code{\link[=validatePKBoxwhiskerConfig]{validatePKBoxwhiskerConfig()}}
}
\concept{functions to generate box whisker plots}
\concept{plot functions}
