% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForestPlot.R
\name{plotPKRatioForestPlotByRatioAggregation}
\alias{plotPKRatioForestPlotByRatioAggregation}
\title{Plot PK Ratio by Ratio Aggregation}
\usage{
plotPKRatioForestPlotByRatioAggregation(
  projectConfiguration,
  subfolder,
  configTableSheet,
  pkParameterDT,
  pkParameterObserved = NULL,
  aggregationFlag = c("GeometricStdDev", "ArithmeticStdDev", "Percentiles", "Custom"),
  percentiles = c(5, 50, 95),
  customFunction = NULL,
  vlineIntercept = 1,
  scaleVectors = list(simulated = list(color = "darkgrey", fill = "darkgrey", shape =
    "circle"), observed = list(color = "darkgrey", fill = "lightgrey", shape =
    "triangle filled")),
  digitsToRound = 2,
  digitsToShow = 2,
  xlabel = "Ratio",
  tableLabels = NULL
)
}
\arguments{
\item{projectConfiguration}{A ProjectConfiguration object containing project settings.}

\item{subfolder}{A string specifying the output directory.}

\item{configTableSheet}{Name of the sheet containing configuration tables.}

\item{pkParameterDT}{A data.table with PK parameters.}

\item{pkParameterObserved}{A data.table with observed PK parameters (optional).}

\item{aggregationFlag}{A character string indicating the aggregation method. Options include:
- "GeometricStdDev"
- "ArithmeticStdDev"
- "Percentiles"
- "Custom"}

\item{percentiles}{A numeric vector of percentiles to calculate if aggregationFlag is "Percentiles". Default is c(5, 50, 95).}

\item{customFunction}{A custom function for aggregation if aggregationFlag is "Custom". Default is NULL.
 The custom function should take a numeric vector `y` as input and return a list containing:
- `yValues`: The aggregated value (e.g., mean).
- `yMin`: The lower value of the aggregated data, (e.g. mean - sd).
- `yMax`: The upper value of the aggregated data, (e.g. mean + sd).}

\item{vlineIntercept}{Intercept for vertical line on plots.}

\item{scaleVectors}{A list of aesthetic parameters for plots.}

\item{digitsToRound}{Number of digits to round in output.}

\item{digitsToShow}{Number of digits to display.}

\item{xlabel}{Label for the x-axis.}

\item{tableLabels}{Labels for tables generated in the plot.}
}
\value{
An RmdContainer object containing the generated plots and related information.
}
\description{
This function generates plots for pharmacokinetic (PK) ratios using various aggregation methods.
}
\details{
The function allows for different aggregation techniques, including:
- Geometric and Arithmetic Standard Deviations
- Percentiles
- A user-defined custom function for aggregation
}
