% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{plotSensitivity}
\alias{plotSensitivity}
\title{Plot Sensitivity Function}
\usage{
plotSensitivity(projectConfiguration, onePlotConfig, scenarioList)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration settings, including file paths and parameters.}

\item{scenarioList}{A list of scenarios for which the sensitivity analysis will be performed.}

\item{subfolder}{A string specifying the subfolder where the plots will be saved.}

\item{configTableSheet}{A string indicating the name of the configuration table sheet to be read.}
}
\value{
An RmdPlotManager object containing the generated sensitivity plots.
}
\description{
This function generates sensitivity plots based on the provided project configuration and scenario list.
}
