% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensitivity.R
\name{plotSensitivity}
\alias{plotSensitivity}
\title{Plot Sensitivity Function}
\usage{
plotSensitivity(projectConfiguration, onePlotConfig, scenarioList)
}
\arguments{
\item{projectConfiguration}{A list containing project configuration settings, including file paths and parameters.}

\item{onePlotConfig}{A data frame containing the configuration for a single plot.}

\item{scenarioList}{A list of scenarios for which the sensitivity analysis will be performed.}
}
\value{
A list of plots and tables, where each entry corresponds to a generated sensitivity plot or its data.
}
\description{
Generates sensitivity plots based on the provided project configuration and scenario list.
}
