% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotTimeProfilePanels}
\alias{plotTimeProfilePanels}
\title{Generate a list of time profiles as facet panels}
\usage{
plotTimeProfilePanels(
  projectConfiguration,
  subfolder,
  configTableSheet,
  dataObserved,
  nFacetColumns = 2,
  nMaxFacetRows = 4,
  facetAspectRatio = 0.5,
  aggregationFlag = c("GeometricStdDev", "ArithmeticStdDev", "Percentiles", "Custom"),
  percentiles = c(5, 50, 95),
  customFunction = NULL,
  referenceScaleVector = list(Control = c(NA, NA), Reference = c(NA, NA))
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{subfolder}{Name of the subfolder in the projectconfiguration$outputFolder directory.}

\item{configTableSheet}{Name of the sheet in the plot configuration xlsx that defines the plot.}

\item{nFacetColumns}{Maximal number of facet columns (default 2) used for facet type "by Order".}

\item{facetAspectRatio}{Aspect ratio for the facets (default 0.5).}

\item{aggregationFlag}{Type of aggregation function to use (default is "GeometricStdDev").}

\item{percentiles}{Percentiles to consider (default is c(5, 50, 95)).}

\item{customFunction}{Optional custom function for aggregation.}

\item{referenceScaleVector}{list to configure labels and colors for the display of reference scenarios}

\item{observedData}{either `data.table` which was generated by \code{\link{readDataByDictionary}} or object of class `DataCombined`,
which will internally transfer to the data.table format by \code{\link{convertDataCombinedToDataTable}}}
}
\value{
Object of class `rmdContainer`.
}
\description{
Generate a list of time profiles as facet panels
}
