% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotTimeProfilePanels}
\alias{plotTimeProfilePanels}
\title{Generate Time Profile Panels}
\usage{
plotTimeProfilePanels(
  projectConfiguration,
  subfolder,
  configTableSheet,
  dataObserved,
  scenarioResults = list(),
  nFacetColumns = 2,
  nMaxFacetRows = 4,
  facetAspectRatio = 0.5,
  aggregationFlag = c("GeometricStdDev", "ArithmeticStdDev", "Percentiles", "Custom"),
  percentiles = c(5, 50, 95),
  customFunction = NULL,
  referenceScaleVector = list(Control = c(NA, NA), Reference = c(NA, NA))
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing the project settings,
including file paths and scenario definitions. This is essential for the function to access the necessary data.}

\item{subfolder}{A character string specifying the subfolder within the output directory where
the plots will be saved.}

\item{configTableSheet}{The name of the sheet in the Excel plot configuration file that defines
the plot settings.}

\item{dataObserved}{A data.table or DataCombined object containing the observed data to be plotted.}

\item{scenarioResults}{A list with simulated scenario results. This parameter may be an empty list or
contain only some of all scenarios defined in the project configuration, in which
case the function will attempt to load missing scenario results from the specified folder.}

\item{nFacetColumns}{An integer specifying the maximum number of facet columns (default is 2)
used for the facet type "by Order".}

\item{nMaxFacetRows}{An integer specifying the maximum number of facet rows (default is 4)
used for all facet types.}

\item{facetAspectRatio}{A numeric value specifying the aspect ratio for the facets (default is 0.5).}

\item{aggregationFlag}{A character vector indicating the type of aggregation function to use
for simulated data. Options include "GeometricStdDev", "ArithmeticStdDev", "Percentiles",
and "Custom" (default is "GeometricStdDev").}

\item{percentiles}{A numeric vector of percentiles to consider if the aggregation method is set
to "Percentiles" (default is c(5, 50, 95)).}

\item{customFunction}{An optional custom function for aggregation. This function should take a
numeric vector as input and return a list containing aggregated values and error types.}

\item{referenceScaleVector}{A named list that configures labels and colors for the display of
reference scenarios. The names of the list are displayed as labels in the legend.}
}
\value{
Generates time profile plots and saves them to the specified output directory.
The function does not return any value.
}
\description{
The `plotTimeProfilePanels` function creates a series of time profile plots as facet panels
from the provided observed data. This function is designed to visualize complex time-dependent
data by allowing for multiple plots organized in a user-defined layout.
}
\details{
This function is particularly useful for researchers and analysts who need to visualize
concentration-time profiles of various scenarios, compare predicted versus observed data,
and analyze residuals cohesively. For more detailed information, please refer to the
accompanying vignettes:
- **TimeProfilePlots**
- **Time Profile Plotting Tutorial**

The function primarily focuses on simulated scenarios, with each panel displaying the
result of one scenario. Plots containing more than one scenario or only observed data are not
supported.
}
