% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotTimeProfiles}
\alias{plotTimeProfiles}
\title{Generate Time Profile Panels}
\usage{
plotTimeProfiles(
  projectConfiguration,
  onePlotConfig,
  dataObserved,
  scenarioResults = list(),
  nFacetColumns = 2,
  nMaxFacetRows = 4,
  facetAspectRatio = 0.5,
  aggregationFlag = c("GeometricStdDev", "ArithmeticStdDev", "Percentiles", "Custom"),
  percentiles = getOspsuite.plots.option(optionKey = OptionKeys$Percentiles)[c(1, 3, 5)],
  customFunction = NULL,
  referenceScaleVector = list(Simulation = c(NA, NA), Reference = c(NA, NA)),
  ...
)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing the project settings,
including file paths and scenario definitions. This is essential for the function to access the necessary data.}

\item{dataObserved}{A data.table or DataCombined object containing the observed data to be plotted.}

\item{scenarioResults}{A list with simulated scenario results. This parameter may be an empty list or
contain only some of all scenarios defined in the project configuration, in which
case the function will attempt to load missing scenario results from the specified folder.}

\item{nFacetColumns}{An integer specifying the maximum number of facet columns (default is 2)
used for the facet type "by Order".}

\item{nMaxFacetRows}{An integer specifying the maximum number of facet rows (default is 4)
used for all facet types.}

\item{facetAspectRatio}{A numeric value specifying the aspect ratio for the facets (default is 0.5).}

\item{aggregationFlag}{A character vector indicating the type of aggregation function to use
for simulated data. Options include "GeometricStdDev" (Default), "ArithmeticStdDev", "Percentiles",
and "Custom".}

\item{percentiles}{A numeric vector of percentiles to consider if the aggregation method is set
to "Percentiles" (default is c(5, 50, 95)).}

\item{customFunction}{An optional custom function for aggregation.
A custom function should take a numeric vector `y` as input and return a list containing:

- `yValues`: The aggregated value (e.g., mean).

- `yMin`: The lower value of the aggregated data, (e.g. mean - sd).

- `yMax`: The upper value of the aggregated data, (e.g. mean + sd).

- `yErrorType`: A string indicating the type of error associated with the aggregation,
it is used in plot legends and captions. It must be a concatenation of the descriptor of yValues and the descriptor of yMin - yMax range
separated by "|" (e.g., "mean | standard deviation" or "median | 5th - 95th percentile").}

\item{referenceScaleVector}{A named list that configures labels and colors for the display of
reference scenarios. The names of the list are displayed as labels in the legend.

Default:
referenceScaleVector = list(
  Simulation = c(NA, NA),
  Reference = c(NA, NA)
)}

\item{configTableSheet}{The name of the sheet in the Excel plot configuration file that defines
the plot settings.}

\item{subfolder}{A character string specifying the subfolder within the output directory where
the plots will be saved. Default is name of configTableSheet}
}
\value{
Generates time profile plots and saves them to the specified output directory.
The function does not return any value.
}
\description{
The `plotTimeProfilePanels` function creates a series of time profile plots as facet panels
from the provided observed data. This function is designed to visualize complex time-dependent
data by allowing for multiple plots organized in a user-defined layout.
}
\details{
This function is particularly useful for researchers who need to visualize
concentration-time profiles of various scenarios, compare predicted versus observed data,
and analyze residuals cohesively. For more detailed information, please refer to the
accompanying vignettes:
- **TimeProfilePlots**
- **Time Profile Plotting Tutorial**

The function primarily focuses on simulated scenarios, with each panel displaying the
result of one scenario. Plots containing more than one scenario or only observed data are not
supported.
}
\examples{
\dontrun{
# Example of using the referenceScaleVector parameter for a DDI Analysis
referenceScaleVector <- list(
  DDI = c("#843C39FF", "#AD494AFF"),
  Control = c("#3182BDFF", "#6BAED6FF")
)

# The scenario would be labeled as 'DDI', with aesthetic color as dark red ("#843C39FF") and
# aesthetic fill as light red ("#AD494AFF").
# The reference scenario would be labeled 'Control', with aesthetic color as dark blue ("#3182BDFF") and
# aesthetic fill as light blue ("#6BAED6FF").

# Example of using the referenceScaleVector parameter for a comparison between a pediatric
# and an adult simulation using the default colors
referenceScaleVector <- list(
  "Pediatric population" = c(NA, NA),
  "Adult population" = c(NA, NA)
)

# The scenario would be labeled as 'Pediatric population', with aesthetic color as dark blue ("#3182BDFF") and
# aesthetic fill as light blue ("#6BAED6FF").
# The reference scenario would be labeled 'Adult population', with both aesthetic color and fill as 'grey'.

# Example of using the aggregationFlag parameter "Custom"
plotTimeProfilePanels(
  projectConfiguration = myProjectConfig,
  configTableSheet = "PlotSettings",
  dataObserved = myObservedData,
  scenarioResults = myScenarioResults,
  aggregationFlag = c("Custom"),
  customFunction = function(y) {
    list(yValues = mean(y), yMin = mean(y) - sd(y), yMax = mean(y) + sd(y), yErrorType = "mean | standard deviation")
  }
)

# Example of using the aggregationFlag parameter "Percentiles"
plotTimeProfilePanels(
  projectConfiguration = myProjectConfig,
  configTableSheet = "PlotSettings",
  dataObserved = myObservedData,
  scenarioResults = myScenarioResults,
  aggregationFlag = c("Percentiles"),
  percentiles = c(0.025, 0.5, 0.975)
)
}

}
