% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{plotTPPanel}
\alias{plotTPPanel}
\title{create the TimeProfile panels}
\usage{
plotTPPanel(plotData, rmdContainer, facetAspectRatio)
}
\arguments{
\item{plotData}{object of class PlotDataTimeprofile which contains data and informations neede for plotting}

\item{rmdContainer}{object of class `rmdContainer`}
}
\description{
create the TimeProfile panels
}
