% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{prepareDataForPKForest}
\alias{prepareDataForPKForest}
\title{Prepare Data for PK Forest}
\usage{
prepareDataForPKForest(
  onePlotConfig,
  pkParameterFiltered,
  pkParameterObserved,
  ratioMode,
  asPointeEstimate,
  nBootstrap,
  confLevel,
  aggregationFun,
  aggregationFlag
)
}
\arguments{
\item{onePlotConfig}{A configuration list for the plot.}

\item{pkParameterFiltered}{A data table containing filtered PK parameter data.}

\item{pkParameterObserved}{An optional data table for observed PK parameters.}

\item{ratioMode}{A string indicating the mode for ratio calculations.}

\item{asPointeEstimate}{A logical indicating if point estimates should be calculated.}

\item{nBootstrap}{An integer specifying the number of bootstrap samples.}

\item{confLevel}{A numeric value representing the confidence level for intervals.}

\item{aggregationFun}{A function used for data aggregation.}

\item{aggregationFlag}{An optional string indicating the aggregation method.}
}
\value{
A list of prepared data for the PK forest plot.
}
\description{
Prepares data for generating a pharmacokinetic (PK) forest plot.
}
\keyword{internal}
