% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKBoxwhisker.R
\name{prepareTableForExport}
\alias{prepareTableForExport}
\title{Prepare Data for PK Boxplot}
\usage{
prepareTableForExport(dtExport, asRatio, plotCaptionAddon, plotDataPk)
}
\arguments{
\item{asRatio}{Logical indicating if the plot is for ratios.}

\item{onePlotConfig}{Configuration for a single plot.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{colorVector}{A named vector for colors.}
}
\value{
A data.table prepared for plotting, including merged configuration and parameter data.
}
\description{
Prepares and cleans data for box-and-whisker plotting by merging the PK parameter data
with the configuration settings. This function ensures that the data is structured
correctly for plotting.
}
\keyword{internal}
