% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKBoxwhisker.R
\name{prepareTableForExport}
\alias{prepareTableForExport}
\title{Prepare Data for PK Boxplot}
\usage{
prepareTableForExport(dtExport, asRatio, plotCaptionAddon, plotDataPk)
}
\arguments{
\item{dtExport}{A data.table containing the data to be exported for plotting.}

\item{asRatio}{A logical indicating if the plot is for ratios.}

\item{plotCaptionAddon}{An optional string to be added to the plot caption.}

\item{plotDataPk}{A data.table containing PK parameter data for generating the plot caption.}
}
\value{
A data.table prepared for plotting, including merged configuration and parameter data.
}
\description{
Prepares and cleans data for box-and-whisker plotting by merging the PK parameter data
with the configuration settings. This function ensures that the data is structured
correctly for plotting.
}
\keyword{internal}
