% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-testProject.R
\name{prepareTestProject}
\alias{prepareTestProject}
\title{Run Example 2 of the Reporting Framework}
\usage{
prepareTestProject(rootDirectory)
}
\arguments{
\item{rootDirectory}{A character string specifying the root directory for the test project.}
}
\value{
A list containing:
  - `projectConfiguration`: The configuration of the initialized project.
  - `scenarioList`: A list of scenarios initialized for the project.
  - `scenarioResults`: The results from running the scenarios.
}
\description{
Build Test Project Directory
}
\details{
This function creates a test project directory for users to experiment with examples
in the `ospsuite.reportingframework` package.
}
\examples{
\dontrun{
prepareTestProject("path/to/my/testDirectory")
}
}
