% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-configuration.R
\docType{class}
\name{ProjectConfigurationRF}
\alias{ProjectConfigurationRF}
\title{ProjectConfiguration}
\description{
An object storing configuration used project-wide
}
\seealso{
Other project initialization: 
\code{\link{createProjectConfiguration}()},
\code{\link{getQCpassedEnvironmentVariable}()},
\code{\link{initProject}()},
\code{\link{setWorkflowOptions}()}
}
\concept{project initialization}
\section{Super class}{
\code{\link[esqlabsR:ProjectConfiguration]{esqlabsR::ProjectConfiguration}} -> \code{ProjectConfigurationRF}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{addOns}}{list with non default configurations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectConfigurationRF-new}{\code{ProjectConfigurationRF$new()}}
\item \href{#method-ProjectConfigurationRF-print}{\code{ProjectConfigurationRF$print()}}
\item \href{#method-ProjectConfigurationRF-addAddOnFileToConfiguration}{\code{ProjectConfigurationRF$addAddOnFileToConfiguration()}}
\item \href{#method-ProjectConfigurationRF-addAddOnFolderToConfiguration}{\code{ProjectConfigurationRF$addAddOnFolderToConfiguration()}}
\item \href{#method-ProjectConfigurationRF-clone}{\code{ProjectConfigurationRF$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="esqlabsR" data-topic="ProjectConfiguration" data-id="save"><a href='../../esqlabsR/html/ProjectConfiguration.html#method-ProjectConfiguration-save'><code>esqlabsR::ProjectConfiguration$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-new}{}}}
\subsection{Method \code{new()}}{
Initializes the ProjectConfiguration object with a specified configuration file path.


Adds new line to configuration xlsx


Read configuration from file
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$new(projectConfigurationFilePath = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfigurationFilePath}}{A string representing the path to the
project configuration file.
Print}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-print}{}}}
\subsection{Method \code{print()}}{
print prints a summary of the Project Configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$print(className = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{Whether to print the name of the class at the beginning. default to TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-addAddOnFileToConfiguration"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-addAddOnFileToConfiguration}{}}}
\subsection{Method \code{addAddOnFileToConfiguration()}}{
Adds an add-on file to the project configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$addAddOnFileToConfiguration(
  property,
  value,
  description,
  templatePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{property}}{A string representing the name of the property to add.}

\item{\code{value}}{A string representing the basename of the file to add.}

\item{\code{description}}{A string providing a description of the property.}

\item{\code{templatePath}}{A string representing the path of the file to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-addAddOnFolderToConfiguration"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-addAddOnFolderToConfiguration}{}}}
\subsection{Method \code{addAddOnFolderToConfiguration()}}{
Adds an add-on file to the project configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$addAddOnFolderToConfiguration(
  property,
  value,
  description
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{property}}{A string representing the name of the property to add.}

\item{\code{value}}{A string representing the path of the value to add.}

\item{\code{description}}{A string providing a description of the property.}

\item{\code{templatePath}}{A string representing the path of the template file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
