% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-configuration.R
\docType{class}
\name{ProjectConfigurationRF}
\alias{ProjectConfigurationRF}
\title{ProjectConfigurationRF}
\description{
An object storing configuration used project-wide, composed of
  an esqlabsR project configuration and reporting-framework add-ons.
}
\seealso{
Other project initialization:
\code{\link[=createProjectConfiguration]{createProjectConfiguration()}},
\code{\link[=getQCpassedEnvironmentVariable]{getQCpassedEnvironmentVariable()}},
\code{\link[=initProject]{initProject()}},
\code{\link[=restoreProjectConfigurationRF]{restoreProjectConfigurationRF()}},
\code{\link[=setWorkflowOptions]{setWorkflowOptions()}},
\code{\link[=snapshotProjectConfigurationRF]{snapshotProjectConfigurationRF()}},
\code{\link[=upgradeToReportingFramework]{upgradeToReportingFramework()}}
}
\concept{project initialization}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{baseProjectconfiguration}}{esqlabsR project configuration object
Initialize}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{projectConfigurationFilePath}}{Path to the project configuration file.}

    \item{\code{addOns}}{list with non default configurations}

    \item{\code{ospsuiteReportingFrameworkVersion}}{Version of the reporting framework
package stored in the project configuration. Read-only.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ProjectConfigurationRF-initialize}{\code{ProjectConfigurationRF$new()}}
    \item \href{#method-ProjectConfigurationRF-print}{\code{ProjectConfigurationRF$print()}}
    \item \href{#method-ProjectConfigurationRF-save}{\code{ProjectConfigurationRF$save()}}
    \item \href{#method-ProjectConfigurationRF-addAddOnFileToConfiguration}{\code{ProjectConfigurationRF$addAddOnFileToConfiguration()}}
    \item \href{#method-ProjectConfigurationRF-addAddOnFolderToConfiguration}{\code{ProjectConfigurationRF$addAddOnFolderToConfiguration()}}
    \item \href{#method-ProjectConfigurationRF-copy}{\code{ProjectConfigurationRF$copy()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-initialize}{}}}
\subsection{\code{ProjectConfigurationRF$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProjectConfigurationRF$new(
  projectConfigurationFilePath = character(),
  ignoreVersionCheck = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{projectConfigurationFilePath}}{A string representing the path to the
project configuration file.}
      \item{\code{ignoreVersionCheck}}{If `TRUE`, skip the version mismatch check when
  loading the configuration file. Defaults to `FALSE`.
Print}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-print}{}}}
\subsection{\code{ProjectConfigurationRF$print()}}{
  print prints a summary of the Project Configuration.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProjectConfigurationRF$print(className = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{className}}{Whether to print the name of the class at the beginning. default to TRUE.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-save"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-save}{}}}
\subsection{\code{ProjectConfigurationRF$save()}}{
  Export ProjectConfiguration object to ProjectConfiguration.xlsx
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProjectConfigurationRF$save(path = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{path}}{a string representing the path or file name where to save the file.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-addAddOnFileToConfiguration"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-addAddOnFileToConfiguration}{}}}
\subsection{\code{ProjectConfigurationRF$addAddOnFileToConfiguration()}}{
  Adds an add-on file to the project configuration.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProjectConfigurationRF$addAddOnFileToConfiguration(
  property,
  value,
  description,
  templatePath
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{property}}{A string representing the name of the property to add.}
      \item{\code{value}}{A string representing the basename of the file to add.}
      \item{\code{description}}{A string providing a description of the property.}
      \item{\code{templatePath}}{A string representing the path of the file to add.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-addAddOnFolderToConfiguration"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-addAddOnFolderToConfiguration}{}}}
\subsection{\code{ProjectConfigurationRF$addAddOnFolderToConfiguration()}}{
  Adds an add-on folder to the project configuration.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProjectConfigurationRF$addAddOnFolderToConfiguration(
  property,
  value,
  description
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{property}}{A string representing the name of the property to add.}
      \item{\code{value}}{A string representing the path of the value to add.}
      \item{\code{description}}{A string providing a description of the property.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfigurationRF-copy"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfigurationRF-copy}{}}}
\subsection{\code{ProjectConfigurationRF$copy()}}{
  Creates a copy of this object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProjectConfigurationRF$copy(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
