% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-configuration.R
\docType{class}
\name{ProjectConfigurationRF}
\alias{ProjectConfigurationRF}
\title{ProjectConfiguration}
\description{
An object storing configuration used project-wide
}
\section{Super classes}{
\code{\link[ospsuite.utils:Printable]{ospsuite.utils::Printable}} -> \code{\link[esqlabsR:ProjectConfiguration]{esqlabsR::ProjectConfiguration}} -> \code{ProjectConfiguration}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pKParameterFile}}{Path to the file containing BMLM Identification}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{pKParameterFile}}{Path to the file containing BMLM Identification}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProjectConfiguration-new}{\code{ProjectConfigurationRF$new()}}
\item \href{#method-ProjectConfiguration-print}{\code{ProjectConfigurationRF$print()}}
\item \href{#method-ProjectConfiguration-addAddOnfileToConfiguration}{\code{ProjectConfigurationRF$addAddOnfileToConfiguration()}}
\item \href{#method-ProjectConfiguration-clone}{\code{ProjectConfigurationRF$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="esqlabsR" data-topic="ProjectConfiguration" data-id="save"><a href='../../esqlabsR/html/ProjectConfiguration.html#method-ProjectConfiguration-save'><code>esqlabsR::ProjectConfiguration$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfiguration-new}{}}}
\subsection{Method \code{new()}}{
Initializes the ProjectConfiguration object with a specified configuration file path.


Read configuration from file
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$new(projectConfigurationFilePath = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{projectConfigurationFilePath}}{A string representing the path to the
project configuration file.
Print}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfiguration-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfiguration-print}{}}}
\subsection{Method \code{print()}}{
print prints a summary of the Project Configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfiguration-addAddOnfileToConfiguration"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfiguration-addAddOnfileToConfiguration}{}}}
\subsection{Method \code{addAddOnfileToConfiguration()}}{
Adds an add-on file to the project configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$addAddOnfileToConfiguration(
  property,
  value,
  description,
  templatePath
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{property}}{A string representing the name of the property to add.}

\item{\code{value}}{A string representing the path of the value to add.}

\item{\code{description}}{A string providing a description of the property.}

\item{\code{templatePath}}{A string representing the path of the template file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectConfigurationRF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
