% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{readObservedDataByDictionary}
\alias{readObservedDataByDictionary}
\title{Read data by dictionary}
\usage{
readObservedDataByDictionary(
  projectConfiguration,
  addBiometricsToConfigFlag = TRUE
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{addBiometricsToConfigFlag}{if TRUE inividual biometrics are added  to Individualxlsx,}
}
\value{
Processed data based on the dictionary
}
\description{
This function reads and processes data based on the provided project configuration.
}
