% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{readObservedDataByDictionary}
\alias{readObservedDataByDictionary}
\title{Read data by dictionary}
\usage{
readObservedDataByDictionary(projectConfiguration, spreadData = TRUE)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{spreadData}{If TRUE, information derived from observed data, such as identifiers
and biometrics, is spread to other tables.}
}
\value{
Processed data based on the dictionary.
}
\description{
This function reads and processes data based on the provided project configuration.
}
