% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{readTimeprofileConfigTable}
\alias{readTimeprofileConfigTable}
\title{Read the configtable for the timeprofile}
\usage{
readTimeprofileConfigTable(sheetName)
}
\arguments{
\item{sheetName}{}
}
\value{
`data.table` with plot configurations
}
\description{
Read the configtable for the timeprofile
}
