% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{readTimeprofileConfigTable}
\alias{readTimeprofileConfigTable}
\title{Read the configuration table for time profiles}
\usage{
readTimeprofileConfigTable(projectConfiguration, sheetName, dataObserved)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\value{
`data.table` with plot configurations.
}
\description{
Read the configuration table for time profiles
}
\keyword{internal}
