% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{runAndSaveScenarios}
\alias{runAndSaveScenarios}
\title{Run a set of scenarios.}
\usage{
runAndSaveScenarios(
  projectConfiguration,
  scenarioList,
  simulationRunOptions = NULL,
  ...
)
}
\arguments{
\item{projectConfiguration}{An instance of ProjectConfiguration}

\item{scenarioList}{Named list of Scenario objects.}

\item{simulationRunOptions}{Object of type SimulationRunOptions that will be passed to simulation runs.
If NULL, default options are used.}

\item{...}{arguments passed to esqlabsR::saveScenarioResults}
}
\description{
uses `esqlabsR::runScenarios` and `esqlabsR::saveScenarioResults
runs the simulations and save the result`
}
