% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{runOrLoadScenarios}
\alias{runOrLoadScenarios}
\title{Run or load scenarios}
\usage{
runOrLoadScenarios(
  projectConfiguration,
  scenarioList,
  simulationRunOptions = NULL,
  ...
)
}
\arguments{
\item{projectConfiguration}{Configuration for the project, containing paths and settings necessary
to run the simulations and load the results.}

\item{scenarioList}{Named list of Scenario objects to be managed.}

\item{simulationRunOptions}{Object of type SimulationRunOptions that will be passed to simulation runs.
If `NULL`, default options are used.}

\item{...}{Additional arguments passed to `runAndSaveScenarios`.}
}
\value{
A list containing the simulation results for each scenario that was loaded or run.
}
\description{
This function checks if the simulation results for scenarios already exist.
If they do, it loads them; otherwise, it runs the scenarios and saves the results.
}
