% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{runPlot}
\alias{runPlot}
\title{Title}
\usage{
runPlot(
  projectConfiguration,
  functionKey = NULL,
  plotFunction = NULL,
  subfolder = functionKey,
  inputs = list()
)
}
\arguments{
\item{inputs}{}
}
\description{
Title
}
