% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{runPlot}
\alias{runPlot}
\title{Title}
\usage{
runPlot(
  projectConfiguration,
  functionKey = c("TimeProfile"),
  plotFunction = NULL,
  subfolder = NULL,
  inputs = list()
)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{functionKey}{keyword which select plot function, if NULL plotfunction is needed}

\item{plotFunction}{function which is used for plotting, if not null, it overwrites functionKey selection}

\item{subfolder}{subfolder where results are filed, if functionKey is used, as default value the sheetname of the plotconfiguartion is used}

\item{inputs}{additionally inputs for the}
}
\description{
Title
}
