% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esqlabsR_bracket.R
\name{saveScenarioResults.wrapped}
\alias{saveScenarioResults.wrapped}
\title{Save results of scenario simulations to .csv file}
\usage{
saveScenarioResults.wrapped(
  simulatedScenariosResults,
  projectConfiguration,
  ...
)
}
\arguments{
\item{simulatedScenariosResults}{Named list with `simulation`, `results`, `outputValues`, and `population` as produced by runScenarios()}

\item{projectConfiguration}{An instance of ProjectConfiguration}

\item{...}{arguments passed to esqlabsR::saveScenarioResults}
}
\description{
wrap of `esqlabsR::saveScenarioResults`
}
