% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_logging.R
\name{saveSessionInfo}
\alias{saveSessionInfo}
\title{Save Session Info}
\usage{
saveSessionInfo()
}
\description{
This function can be called at the end of your script to save the session information,
including the loaded packages and R version, into a log file.
The path for the log file has to be initialized by  the `initLogfunction`
}
\examples{
\dontrun{
# Save session info to log file
saveSessionInfo()
}

}
