% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{setDataTypeAttributes}
\alias{setDataTypeAttributes}
\title{sets comulnType attribute according to dictionary}
\usage{
setDataTypeAttributes(dataDT, dict = NULL)
}
\arguments{
\item{dataDT}{`data.table` with observed data}

\item{dict}{named list with columnNames and columntypes, if NULL,
list is produced by template saved in package installation}
}
\value{
`data.table` with attributes
}
\description{
sets comulnType attribute according to dictionary
}
