% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{setHeadersToLowerCase}
\alias{setHeadersToLowerCase}
\title{Convert Data Table Column Names to Lowercase}
\usage{
setHeadersToLowerCase(dt)
}
\arguments{
\item{dt}{A data.table object whose column names need to be converted to lowercase.}
}
\value{
The modified data.table with updated column names.
}
\description{
This function takes a data.table and converts the first letter of all column names to lowercase.
}
\examples{
dt <- data.table::data.table(FirstName = c("John", "Jane"), LastName = c("Doe", "Smith"))
dtLower <- setHeadersToLowerCase(dt)
print(names(dtLower)) # Result: c("firstname", "lastname")

}
\seealso{
Other function to read from and write to xlsx:
\code{\link[=splitInputs]{splitInputs()}},
\code{\link[=xlsxAddDataUsingTemplate]{xlsxAddDataUsingTemplate()}},
\code{\link[=xlsxAddSheet]{xlsxAddSheet()}},
\code{\link[=xlsxCloneAndSet]{xlsxCloneAndSet()}},
\code{\link[=xlsxReadData]{xlsxReadData()}},
\code{\link[=xlsxWriteData]{xlsxWriteData()}}
}
\concept{function to read from and write to xlsx}
