% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{setHeadersToLowerCase}
\alias{setHeadersToLowerCase}
\title{Convert Data Table Column Names to Lowercase}
\usage{
setHeadersToLowerCase(dt)
}
\arguments{
\item{dt}{A data.table object whose column names need to be converted to lowercase.}
}
\description{
This function takes a data.table and converts the first letter of all column names to lowercase,
}
