% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{setShowLogMessages}
\alias{setShowLogMessages}
\title{Function to switch the display of log messages on the console on and off}
\usage{
setShowLogMessages(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean, if true log message will be shown}
}
\description{
Function to switch the display of log messages on the console on and off
}
\seealso{
Other log file management: 
\code{\link{addMessageToLog}()},
\code{\link{captureLog}()},
\code{\link{initLogfunction}()},
\code{\link{saveSessionInfo}()},
\code{\link{writeTableToLog}()},
\code{\link{writeToLog}()}
}
\concept{log file management}
