% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDataTimeProfile.R
\name{setTimeRangeFilter}
\alias{setTimeRangeFilter}
\title{Set Time Range Filter}
\usage{
setTimeRangeFilter(splitPlotsPerTimeRange, timeTags)
}
\arguments{
\item{splitPlotsPerTimeRange}{Boolean, if TRUE for each Timerange a plot is created}

\item{timeTags}{A vector of time tags.}
}
\value{
A list of time range filters.
}
\description{
This function sets the time range filter based on the facet type and time tags.
}
\keyword{internal}
