% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-workflow.R
\name{setWorkflowOptions}
\alias{setWorkflowOptions}
\title{Sets options for a reporting frame work workflow.}
\usage{
setWorkflowOptions(
  isValidRun,
  ePackageGeneration = c("None", "withEPackage", "onlyEPackage")
)
}
\arguments{
\item{isValidRun}{A logical value indicating if the run is valid. If TRUE,
options are set for a valid run; if FALSE, options are set for an invalid run.}

\item{ePackageGeneration}{A character string indicating the type of ePackage generation.
       Options include:

       - 'None': No ePackage generation.

       - 'withEPackage': Generate ePackage alongside the run.

       - 'onlyEPackage': Only generate ePackage without exporting figures.}
}
\description{
Typically a workflow is applied in two different cases.
1. As valid run, the result of a workflow is intended as a package,
   which can be used to generate final reports, in this case figures
   should not have watermarks, functions which manipulate inputs are not allowed,
   and the workflow stops if an error occurs during.
2. For exploratory analysis or preparation of a valid run.
}
\details{
A workflow can also produce a ePackage (see vignette XXX)
This function configures several options that affect the behavior of the
reporting frame work workflow. It allows for the enabling or disabling of watermarks,
controls the handling of failing plots, manages the execution of helper functions,
and specifies ePackage generation and plot export options.

Relevant options include:

- `ospsuite.plots.watermark_enabled`: Set to TRUE when `isValidRun` is FALSE
  to display watermarks on figures, and FALSE when `isValidRun` is TRUE.

- `OSPSuite.RF.skipFailingPlots`: Set to TRUE when `isValidRun` is FALSE
  to skip plots that fail to generate, and FALSE when `isValidRun` is TRUE.

- `OSPSuite.RF.stopHelperFunction`: Set to TRUE when `isValidRun` is TRUE
  to stop the execution of helper functions during valid runs.

- `OSPSuite.RF.withEPackage`: Set to TRUE if `ePackageGeneration` is either
  'withEpackage' or 'onlyEpackage', indicating that EPackage generation should occur.

- `OSPSuite.RF.withPlotExport`: Set to TRUE if `ePackageGeneration` is either
  'None' or 'withEpackage', indicating that plot export should occur.
}
