% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{splitInputs}
\alias{splitInputs}
\title{Split the elements of a vector by comma}
\usage{
splitInputs(originalVector)
}
\arguments{
\item{originalVector}{The original vector to be split.}
}
\value{
A vector containing the split elements.
}
\description{
This function takes an original vector as input and splits its elements by comma.
}
\examples{
\dontrun{
originalVector <- c("group1, group2", "group3")
splitInputs(originalVector)
# Result: c('group1', 'group2', 'group3')
}

}
\seealso{
Other function to read from and write to xlsx:
\code{\link[=setHeadersToLowerCase]{setHeadersToLowerCase()}},
\code{\link[=xlsxAddDataUsingTemplate]{xlsxAddDataUsingTemplate()}},
\code{\link[=xlsxAddSheet]{xlsxAddSheet()}},
\code{\link[=xlsxCloneAndSet]{xlsxCloneAndSet()}},
\code{\link[=xlsxReadData]{xlsxReadData()}},
\code{\link[=xlsxWriteData]{xlsxWriteData()}}
}
\concept{function to read from and write to xlsx}
