% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{startRmd}
\alias{startRmd}
\title{creates startlines of Rmd file}
\usage{
startRmd(title = "Report")
}
\arguments{
\item{title}{title of report}
}
\value{
character with startlines
}
\description{
creates startlines of Rmd file
}
\keyword{internal}
