% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{synchronizeScenariosOutputsWithPlots}
\alias{synchronizeScenariosOutputsWithPlots}
\title{Synchronize Scenario Outputs with Plot Outputs}
\usage{
synchronizeScenariosOutputsWithPlots(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing the file paths for scenariosFile and plotsFile.}
}
\value{
Returns invisibly.
}
\description{
This function synchronizes output paths between two Excel files: one containing
scenario outputs and the other containing plot outputs. It only writes back to the
Excel files if changes are detected in the output paths.
}
\keyword{internal}
