% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{synchronizeScenariosWithPlots}
\alias{synchronizeScenariosWithPlots}
\title{Synchronize Scenarios Between Scenario and Plot Files}
\usage{
synchronizeScenariosWithPlots(projectConfiguration)
}
\arguments{
\item{projectConfiguration}{An object of class ProjectConfiguration containing the file paths for scenariosFile and plotsFile.}
}
\value{
Returns invisibly.
}
\description{
This function synchronizes scenarios between two Excel files: one containing
scenarios for a project and the other containing scenarios related to plots.
It adds any missing scenarios from the scenarios file to the plots file.
}
\keyword{internal}
