% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_enum.R
\docType{data}
\name{TIMERANGE}
\alias{TIMERANGE}
\title{enumeration keys for Timerange shortcut}
\format{
An object of class \code{list} of length 3.
}
\usage{
TIMERANGE
}
\description{
enumeration keys for Timerange shortcut
}
\keyword{datasets}
