% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-enum.R
\docType{data}
\name{TIMERANGE}
\alias{TIMERANGE}
\title{enumeration keys for Time range shortcut}
\format{
An object of class \code{list} of length 3.
}
\usage{
TIMERANGE
}
\description{
enumeration keys for Time range shortcut
}
\seealso{
Other enumerations: 
\code{\link{BIOMETRICUNITS}},
\code{\link{DATACLASS}},
\code{\link{EXPORTDIR}}
}
\concept{enumerations}
\keyword{datasets}
