% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{updateDataGroupId}
\alias{updateDataGroupId}
\title{Update Data Group IDs in Project Configuration}
\usage{
updateDataGroupId(projectConfiguration, dataDT)
}
\arguments{
\item{projectConfiguration}{An object of class `ProjectConfiguration`
containing configuration details}

\item{dataDT}{A `data.table` containing the new data to be added to the Data Groups sheet.
It must have columns that can be matched with existing identifiers, including "group", "studyId", and "studyArm".}
}
\value{
NULL This function updates the Excel workbook in place and does not return a value. It is called for its side effects.
}
\description{
This function updates the Data Groups sheet in an Excel workbook based on a provided data table.
It reads existing data group IDs, merges them with new data, and ensures that the identifiers remain unique.
}
\details{
The function loads the existing Data Groups from the specified Excel workbook, selects relevant columns from the input data,
and appends new entries while maintaining the uniqueness of the identifiers. The updated data is then written back to the workbook.
}
\seealso{
Other observed data processing:
\code{\link[=addBiometricsToConfig]{addBiometricsToConfig()}},
\code{\link[=aggregateObservedDataGroups]{aggregateObservedDataGroups()}},
\code{\link[=convertDataCombinedToDataTable]{convertDataCombinedToDataTable()}},
\code{\link[=convertDataTableToDataCombined]{convertDataTableToDataCombined()}},
\code{\link[=readObservedDataByDictionary]{readObservedDataByDictionary()}},
\code{\link[=updateOutputPathId]{updateOutputPathId()}},
\code{\link[=validateObservedData]{validateObservedData()}}
}
\concept{observed data processing}
