% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{updateDataGroupId}
\alias{updateDataGroupId}
\title{Add group ids of observed data to configration sheet}
\usage{
updateDataGroupId(projectConfiguration, data)
}
\arguments{
\item{projectConfiguration}{The project configuration}

\item{data}{`data.table`with observed Data}
}
\description{
Add group ids of observed data to configration sheet
}
