% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{updateDataGroupId}
\alias{updateDataGroupId}
\title{Add group IDs of observed data to configuration sheet}
\usage{
updateDataGroupId(projectConfiguration, dataDT)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{data}{`data.table` with observed data.}
}
\description{
Add group IDs of observed data to configuration sheet
}
