% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{updateOutputPathId}
\alias{updateOutputPathId}
\title{Add outputpath ids of observed data to configuration sheet}
\usage{
updateOutputPathId(projectConfiguration, dataDT)
}
\arguments{
\item{projectConfiguration}{The project configuration}

\item{data}{`data.table`with observed Data}
}
\description{
Add outputpath ids of observed data to configuration sheet
}
