% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{updateOutputPathId}
\alias{updateOutputPathId}
\title{Add output path IDs of observed data to configuration sheet}
\usage{
updateOutputPathId(projectConfiguration, dataDT)
}
\arguments{
\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{data}{`data.table` with observed data.}
}
\description{
Add output path IDs of observed data to configuration sheet
}
