% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{updateOutputPathId}
\alias{updateOutputPathId}
\title{Update Output Path IDs in Project Configuration}
\usage{
updateOutputPathId(projectConfiguration, dataDT)
}
\arguments{
\item{projectConfiguration}{An object of class `ProjectConfiguration` containing configuration details.}

\item{dataDT}{A `data.table` containing new output path IDs to be added to the Outputs sheet.
It must include a column named "outputPathId".}
}
\value{
NULL This function updates the Excel workbook in place and does not return a value. It is called for its side effects.
}
\description{
This function updates the Outputs sheet in an Excel workbook based on a provided data table,
ensuring that output path IDs remain unique.
}
\details{
The function loads the existing Output Paths from the specified Excel workbook, extracts unique output path IDs from the input data,
and appends them while maintaining uniqueness. The updated data is then written back to the workbook.
}
