% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateColorLegend}
\alias{validateColorLegend}
\title{This function checks if the `colorLegend` column in a data.table contains
valid entries. Each entry must be a character string concatenated from
two characters separated by a pipe (`|`). If any entry does not meet this
requirement, the function will print the invalid entries and stop execution.}
\usage{
validateColorLegend(dt)
}
\arguments{
\item{dt}{A data.table containing a column named `colorLegend` and
optionally a `plotName` column.}
}
\value{
NULL The function stops execution if invalid entries are found.
}
\description{
This function checks if the `colorLegend` column in a data.table contains
valid entries. Each entry must be a character string concatenated from
two characters separated by a pipe (`|`). If any entry does not meet this
requirement, the function will print the invalid entries and stop execution.
}
\keyword{internal}
