% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateConfigTableForPlots}
\alias{validateConfigTableForPlots}
\title{Validate Configuration Table for Plots}
\usage{
validateConfigTableForPlots(configTable, ...)
}
\arguments{
\item{configTable}{A data.table containing the plot configuration to be validated.}
}
\value{
NULL. The function will throw an error if the validation fails.
}
\description{
This function validates the structure and content of the configuration table for plots.
}
\keyword{internal}
