% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateConfigTableForTimeProfiles}
\alias{validateConfigTableForTimeProfiles}
\title{Validation of config table for time profiles plots}
\usage{
validateConfigTableForTimeProfiles(configTable, dataObserved, ...)
}
\arguments{
\item{configTable}{Plot configuration table.}

\item{dataObserved}{`data.table` format corresponds to tables  generated by  function `readDataByDictionary`}

\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}
}
\description{
Validation of config table for time profiles plots
}
\keyword{internal}
