% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateConfigTableForTimeProfiles}
\alias{validateConfigTableForTimeProfiles}
\title{Validation of config table fo time profiles plots}
\usage{
validateConfigTableForTimeProfiles(
  configTable,
  observedData,
  projectConfiguration
)
}
\arguments{
\item{configTable}{plot Configuration table}

\item{observedData}{observed data as data.table}
}
\description{
Validation of config table fo time profiles plots
}
