% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateConfigTableForTimeProfiles}
\alias{validateConfigTableForTimeProfiles}
\title{Validation of config table fo time profiles plots}
\usage{
validateConfigTableForTimeProfiles(
  configTable,
  dataObserved,
  projectConfiguration
)
}
\arguments{
\item{configTable}{plot Configuration table}

\item{projectConfiguration}{Object of class `ProjectConfiguration` containing information on paths and filenames}

\item{observedData}{`data.table` format corresponds to tables  generated by \code{\link{readDataByDictionary}}}
}
\description{
Validation of config table fo time profiles plots
}
