% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateDistributionVsDemographicsConfig}
\alias{validateDistributionVsDemographicsConfig}
\title{Validate Distribution vs Demographics Configuration}
\usage{
validateDistributionVsDemographicsConfig(configTable, scenarioList, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{scenarioList}{List of scenarios}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration table for distribution vs demographics plots.
}
\seealso{
Other plot configuration validation function: 
\code{\link{validateHistogramsConfig}()},
\code{\link{validatePKBoxwhiskerConfig}()},
\code{\link{validatePKForestAggregatedAbsoluteValuesConfig}()},
\code{\link{validatePKForestAggregatedRatiosConfig}()},
\code{\link{validatePKForestPointEstimateOfAbsoluteValuesConfig}()},
\code{\link{validatePKForestPointEstimateOfRatiosConfig}()},
\code{\link{validateSensitivityConfig}()},
\code{\link{validateTimeProfilesConfig}()}

Other functions to generate plots displaying distribution vs demographics: 
\code{\link{addDefaultConfigForHistograms}()},
\code{\link{plotDistributionVsDemographics}()}
}
\concept{functions to generate plots displaying distribution vs demographics}
\concept{plot configuration validation function}
