% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateDistributionVsDemographicsConfig}
\alias{validateDistributionVsDemographicsConfig}
\title{Validate Distribution vs Demographics Configuration}
\usage{
validateDistributionVsDemographicsConfig(configTable, scenarioList, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{scenarioList}{List of scenarios}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration table for distribution vs demographics plots.
}
