% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateHistogramsConfig}
\alias{validateHistogramsConfig}
\title{Validate Histograms Configuration}
\usage{
validateHistogramsConfig(configTable, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration table for histogram plots.
}
\seealso{
Other plot configuration validation function: 
\code{\link{validateDistributionVsDemographicsConfig}()},
\code{\link{validatePKBoxwhiskerConfig}()},
\code{\link{validatePKForestAggregatedAbsoluteValuesConfig}()},
\code{\link{validatePKForestAggregatedRatiosConfig}()},
\code{\link{validatePKForestPointEstimateOfAbsoluteValuesConfig}()},
\code{\link{validatePKForestPointEstimateOfRatiosConfig}()},
\code{\link{validateSensitivityConfig}()},
\code{\link{validateTimeProfilesConfig}()}

Other functions to generate histograms: 
\code{\link{addDefaultConfigForDistributionsVsDemographics}()},
\code{\link{plotHistograms}()}
}
\concept{functions to generate histograms}
\concept{plot configuration validation function}
