% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDemographics.R
\name{validateHistogramsConfig}
\alias{validateHistogramsConfig}
\title{Validate Histograms Configuration}
\usage{
validateHistogramsConfig(configTable, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration table for histogram plots.
}
