% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{validateMdFigureTableInputs}
\alias{validateMdFigureTableInputs}
\title{Validate figure/table input files and styles.}
\usage{
validateMdFigureTableInputs(subfolder, importFile, captionFile, customStyles)
}
\arguments{
\item{subfolder}{Folder location.}

\item{importFile}{Figure/table file.}

\item{captionFile}{Caption file.}

\item{customStyles}{Custom styles list.}
}
\description{
Validate figure/table input files and styles.
}
\keyword{internal}
