% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-markdown.R
\name{validateMdFigureTableInputs}
\alias{validateMdFigureTableInputs}
\title{validates common inputs for `mdFigure` and `mdTable`}
\usage{
validateMdFigureTableInputs(subfolder, importFile, captionFile, customStyles)
}
\arguments{
\item{subfolder}{The folder where the file is located relative to Rmd}

\item{importFile}{figure file or table .csv file}

\item{captionFile}{file containing caption}

\item{customStyles}{list of custom styles usable for figure and table captions and footnotes}
}
\description{
Include a figure file with caption
}
\seealso{
Other markdown helper function: 
\code{\link{addFiguresAndTables}()},
\code{\link{mdBullet}()},
\code{\link{mdBullet0}()},
\code{\link{mdCaption}()},
\code{\link{mdFootNote}()},
\code{\link{mdHeading}()},
\code{\link{mdLink}()},
\code{\link{mdNewline}()},
\code{\link{mdNewpage}()},
\code{\link{mdPaste}()},
\code{\link{mdPaste0}()},
\code{\link{mdTable}()}
}
\keyword{internal}
