% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{validateObservedData}
\alias{validateObservedData}
\title{Validate Observed Data}
\usage{
validateObservedData(dataDT, dataClassType)
}
\arguments{
\item{dataDT}{A `data.table` containing observed data with the following relevant columns:
- `individualId`: Unique identifier for individuals.
- `group`: Group identifier.
- `outputPathId`: Identifier for output paths.
- `xValues`: Values for the x-axis.
- `yUnit`: Unit for the Y values.
- `yErrorType`: Type of error for Y values (optional).
- `yErrorValues`: Values representing errors (optional).
- `yMin`, `yMax`: Minimum and maximum Y values (optional).
- `lloq`: Lower limit of quantification (optional).
- `nBelowLLOQ`: Count of values below the lower limit of quantification (optional).}

\item{dataClassType}{A string indicating the type of data class (e.g., "timeprofile" or "pkParameter").}
}
\value{
NULL This function performs checks and stops execution if any validation fails.
It does not return a value.
}
\description{
Checks the integrity and validity of observed data in a `data.table`.
This function verifies the presence of required attributes, checks for duplicates,
and ensures that there are no missing or empty values in the relevant columns.
Additionally, it checks for ambiguities in the Y unit and validates error type columns.
}
\details{
The function performs several checks, including:
- Ensuring all data columns have the appropriate attributes.
- Verifying that the data is unique based on specified identifier columns.
- Checking for NA or empty values in all relevant columns.
- Ensuring that the Y unit is consistent across output paths.
- Validating the presence of necessary columns based on the Y error type.
}
\seealso{
Other observed data processing:
\code{\link[=addBiometricsToConfig]{addBiometricsToConfig()}},
\code{\link[=aggregateObservedDataGroups]{aggregateObservedDataGroups()}},
\code{\link[=convertDataCombinedToDataTable]{convertDataCombinedToDataTable()}},
\code{\link[=convertDataTableToDataCombined]{convertDataTableToDataCombined()}},
\code{\link[=readObservedDataByDictionary]{readObservedDataByDictionary()}},
\code{\link[=updateDataGroupId]{updateDataGroupId()}},
\code{\link[=updateOutputPathId]{updateOutputPathId()}}
}
\concept{observed data processing}
