% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-data.R
\name{validateObservedData}
\alias{validateObservedData}
\title{Validate observed data}
\usage{
validateObservedData(data, stopIfValidationFails = TRUE)
}
\arguments{
\item{data}{The observed dataset}

\item{stopIfValidationFails}{Flag to indicate whether to stop if validation fails}
}
\description{
This function checks the validity of the observed dataset.
}
