% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-plot.R
\name{validateOutputIdsForPlot}
\alias{validateOutputIdsForPlot}
\title{Validation of data.table with outputPath Ids}
\usage{
validateOutputIdsForPlot()
}
\arguments{
\item{dtOutputPaths}{data.table with outputPath Ids}
}
\description{
Validation of data.table with outputPath Ids
}
