% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validatePanelConsistency}
\alias{validatePanelConsistency}
\title{check if panel columns ar filled consistently}
\usage{
validatePanelConsistency(configTablePlots, panelColumns, dtOutputPaths)
}
\arguments{
\item{configTablePlots}{`data.table` ConfigurationTable without header lines}

\item{panelcolumns}{vector of columns which should be consistent}
}
\description{
check if panel columns ar filled consistently
}
