% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKBoxwhisker.R
\name{validatePKBoxwhiskerConfig}
\alias{validatePKBoxwhiskerConfig}
\title{Validate PK Box-and-Whisker Configuration Table}
\usage{
validatePKBoxwhiskerConfig(configTable, pkParameterDT, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration table for PK box-and-whisker plots.
}
\seealso{
Other plot configuration validation function: 
\code{\link{validateDistributionVsDemographicsConfig}()},
\code{\link{validateHistogramsConfig}()},
\code{\link{validatePKForestAggregatedAbsoluteValuesConfig}()},
\code{\link{validatePKForestAggregatedRatiosConfig}()},
\code{\link{validatePKForestPointEstimateOfAbsoluteValuesConfig}()},
\code{\link{validatePKForestPointEstimateOfRatiosConfig}()},
\code{\link{validateSensitivityConfig}()},
\code{\link{validateTimeProfilesConfig}()}

Other functions to generate box whisker plots: 
\code{\link{addDefaultConfigForPKBoxwhsikerPlots}()},
\code{\link{plotPKBoxwhisker}()}
}
\concept{functions to generate box whisker plots}
\concept{plot configuration validation function}
