% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{validatePKForestAggregatedAbsoluteValuesConfig}
\alias{validatePKForestAggregatedAbsoluteValuesConfig}
\title{Validate PK Forest Absolute Values with Variance Configuration}
\usage{
validatePKForestAggregatedAbsoluteValuesConfig(configTable, pkParameterDT, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration for PK forest plots with absolute values and variance.
}
\seealso{
Other plot configuration validation function: 
\code{\link{validateDistributionVsDemographicsConfig}()},
\code{\link{validateHistogramsConfig}()},
\code{\link{validatePKBoxwhiskerConfig}()},
\code{\link{validatePKForestAggregatedRatiosConfig}()},
\code{\link{validatePKForestPointEstimateOfAbsoluteValuesConfig}()},
\code{\link{validatePKForestPointEstimateOfRatiosConfig}()},
\code{\link{validateSensitivityConfig}()},
\code{\link{validateTimeProfilesConfig}()}

Other functions to generate forest plots: 
\code{\link{addDefaultConfigForPKForestPlots}()},
\code{\link{plotPKForestAggregatedAbsoluteValues}()},
\code{\link{plotPKForestAggregatedRatios}()},
\code{\link{plotPKForestPointEstimateOfAbsoluteValues}()},
\code{\link{plotPKForestPointEstimateOfRatios}()},
\code{\link{validatePKForestAggregatedRatiosConfig}()},
\code{\link{validatePKForestPointEstimateOfAbsoluteValuesConfig}()},
\code{\link{validatePKForestPointEstimateOfRatiosConfig}()}
}
\concept{functions to generate forest plots}
\concept{plot configuration validation function}
