% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPKForest.R
\name{validatePKForestAggregatedAbsoluteValuesConfig}
\alias{validatePKForestAggregatedAbsoluteValuesConfig}
\title{Validate PK Forest Absolute Values with Variance Configuration}
\usage{
validatePKForestAggregatedAbsoluteValuesConfig(configTable, pkParameterDT, ...)
}
\arguments{
\item{configTable}{A data.table containing the configuration table.}

\item{pkParameterDT}{A data.table containing PK parameter data.}

\item{...}{Additional arguments for validation.}
}
\value{
NULL (invisible).
}
\description{
Validates the configuration for PK forest plots with absolute values and variance.
}
