% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateTimeProfilesConfig}
\alias{validateTimeProfilesConfig}
\title{Validation of config table for time profiles plots}
\usage{
validateTimeProfilesConfig(
  configTable,
  dataObserved = NULL,
  scenarioResults,
  ...
)
}
\arguments{
\item{configTable}{Plot configuration table.}

\item{dataObserved}{A `data.table` (formatted as produced by `readObservedDataByDictionary`)
or `DataCombined` object containing the observed data to be plotted.}

\item{scenarioResults}{A list containing simulated scenario results.}

\item{...}{Additional arguments passed to `ospsuite.plots::plotTimeprofile`.}
}
\description{
Validation of config table for time profiles plots
}
\seealso{
Other plot configuration validation function: 
\code{\link{validateDistributionVsDemographicsConfig}()},
\code{\link{validateHistogramsConfig}()},
\code{\link{validatePKBoxwhiskerConfig}()},
\code{\link{validatePKForestAggregatedAbsoluteValuesConfig}()},
\code{\link{validatePKForestAggregatedRatiosConfig}()},
\code{\link{validatePKForestPointEstimateOfAbsoluteValuesConfig}()},
\code{\link{validatePKForestPointEstimateOfRatiosConfig}()},
\code{\link{validateSensitivityConfig}()}
}
\concept{plot configuration validation function}
