% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateTimeProfilesConfig}
\alias{validateTimeProfilesConfig}
\title{Validation of config table for time profiles plots}
\usage{
validateTimeProfilesConfig(
  configTable,
  dataObserved = NULL,
  scenarioResults,
  ...
)
}
\arguments{
\item{configTable}{Plot configuration table.}

\item{dataObserved}{A data.table or DataCombined object containing the observed data to be plotted.}

\item{scenarioResults}{A list with simulated scenario results.}

\item{...}{argumnets passed to ospsuite.plots::plotTimeprofile}
}
\description{
Validation of config table for time profiles plots
}
