% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateTimeRangeColumns}
\alias{validateTimeRangeColumns}
\title{validates timerange columns}
\usage{
validateTimeRangeColumns(configTablePlots)
}
\arguments{
\item{configTablePlots}{`data.table` ConfigurationTable without header lines}
}
\description{
Timerange columns must be character and mast contain NA
'total','firstApplication','lastApplication'
or a string which evaluates in r to a numeric vector length 2  (e.g. 'c(2,3)' or 'c(2,NA)'
}
