% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeProfilePanels.R
\name{validateUnitConsistency}
\alias{validateUnitConsistency}
\title{Check if panel columns are filled consistently}
\usage{
validateUnitConsistency(configTablePlots)
}
\arguments{
\item{configTablePlots}{`data.table` ConfigurationTable without header lines}

\item{panelColumns}{Vector of columns which should be consistent.}
}
\description{
Check if panel columns are filled consistently
}
\keyword{internal}
