% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{writeTableToLog}
\alias{writeTableToLog}
\title{write a table to the logfile}
\usage{
writeTableToLog(dt, filename = "run.log")
}
\arguments{
\item{dt}{table to log}

\item{filename}{filename of log}
}
\description{
write a table to the logfile
}
\seealso{
Other log file management: 
\code{\link{addMessageToLog}()},
\code{\link{captureLog}()},
\code{\link{initLogfunction}()},
\code{\link{saveSessionInfo}()},
\code{\link{setShowLogMessages}()},
\code{\link{writeToLog}()}
}
\concept{log file management}
