% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxAddDataUsingTemplate}
\alias{xlsxAddDataUsingTemplate}
\title{Add new data to a config file using a template sheet}
\usage{
xlsxAddDataUsingTemplate(
  wb,
  templateSheet,
  sheetName,
  dtNewData,
  templateXlsx = "Plots.xlsx"
)
}
\arguments{
\item{wb}{Current configuration file.}

\item{templateSheet}{Name of the template sheet.}

\item{sheetName}{Name of the new sheet.}

\item{dtNewData}{A `data.table` with new data.}

\item{templateXlsx}{A character string specifying the template xlsx file name (default is "Plots.xlsx").}
}
\value{
The current configuration file with the added sheet.
}
\description{
If the template does not exist in the configuration file in the project directory,
it is taken from the configuration file of the package installation.
In this case, formats are not preserved.
}
\examples{
\dontrun{
wb <- loadWorkbook("config.xlsx")
newData <- data.table(Name = c("Gina", "Hank"), Age = c(29, 33))
xlsxAddDataUsingTemplate(wb, "TemplateSheet", "NewDataSheet", newData)
}

}
\seealso{
Other function to read from and write to xlsx: 
\code{\link{setHeadersToLowerCase}()},
\code{\link{splitInputs}()},
\code{\link{xlsxAddSheet}()},
\code{\link{xlsxCloneAndSet}()},
\code{\link{xlsxReadData}()},
\code{\link{xlsxWriteData}()}
}
\concept{function to read from and write to xlsx}
