% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxAddSheet}
\alias{xlsxAddSheet}
\title{Add a new sheet to a workbook with data}
\usage{
xlsxAddSheet(wb, sheetName, dt)
}
\arguments{
\item{wb}{A workbook object created by `openxlsx::loadWorkbook()`.}

\item{sheetName}{A character string specifying the name of the sheet to add.}

\item{dt}{A `data.table` containing the data to be written to the new sheet.}
}
\value{
An invisible NULL value.
}
\description{
This function wraps `openxlsx::addWorksheet` and adds data to the newly created sheet.
If the sheet already exists, it issues a warning and clears the existing content.
}
