% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxAddSheet}
\alias{xlsxAddSheet}
\title{adds a new sheet to a workbook}
\usage{
xlsxAddSheet(wb, sheetName, dt)
}
\arguments{
\item{wb}{A Workbook object to attach the new worksheet}

\item{sheetName}{A name for the new worksheet}

\item{dt}{`data.table`  to attach}
}
\description{
wrapps `openxlsx::addWorksheet`, and adds new data to the sheet
}
