% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxCloneAndSet}
\alias{xlsxCloneAndSet}
\title{Clone a sheet and set new content}
\usage{
xlsxCloneAndSet(wb, clonedSheet, sheetName, dt)
}
\arguments{
\item{wb}{A workbook object created by `openxlsx::loadWorkbook()`.}

\item{clonedSheet}{A character string specifying the name of the sheet to clone.}

\item{sheetName}{A character string specifying the name of the new sheet.}

\item{dt}{A `data.table` with new content.}
}
\value{
An invisible NULL value. The function performs an action (clone a sheet)
}
\description{
This function wraps `openxlsx::cloneWorksheet` but sets new content in the cloned sheet.
}
\examples{
\dontrun{
wb <- loadWorkbook("example.xlsx")
data <- data.table(Name = c("Eve", "Frank"), Age = c(22, 35))
xlsxCloneAndSet(wb, "ExistingSheet", "ClonedSheet", data)
}

}
\seealso{
Other function to read from and write to xlsx: 
\code{\link{copyConfigSheet}()},
\code{\link{separateAndTrimColumn}()},
\code{\link{setHeadersToLowerCase}()},
\code{\link{splitInputs}()},
\code{\link{synchronizeScenariosOutputsWithPlots}()},
\code{\link{synchronizeScenariosWithPlots}()},
\code{\link{xlsxAddDataUsingTemplate}()},
\code{\link{xlsxAddSheet}()},
\code{\link{xlsxReadData}()},
\code{\link{xlsxWriteData}()}
}
\concept{function to read from and write to xlsx}
