% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxCloneAndSet}
\alias{xlsxCloneAndSet}
\title{create a clone of a sheet and set new contet}
\usage{
xlsxCloneAndSet(wb, clonedSheet, sheetName, dt)
}
\arguments{
\item{wb}{A Workbook object to attach the new worksheet}

\item{clonedSheet}{The name of the existing worksheet to be cloned.}

\item{sheetName}{A name for the new worksheet}

\item{dt}{`data.table` with new content}
}
\value{
workbook with cloned sheet
}
\description{
wrappes openxlsx::cloneWorksheet but set new content
}
