% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxCloneAndSet}
\alias{xlsxCloneAndSet}
\title{Clone a sheet and set new content}
\usage{
xlsxCloneAndSet(wb, clonedSheet, sheetName, dt)
}
\arguments{
\item{wb}{A workbook object created by `openxlsx::loadWorkbook()`.}

\item{clonedSheet}{A character string specifying the name of the sheet to clone.}

\item{sheetName}{A character string specifying the name of the new sheet.}

\item{dt}{A `data.table` with new content.}
}
\description{
This function wraps `openxlsx::cloneWorksheet` but sets new content in the cloned sheet.
}
