% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxReadData}
\alias{xlsxReadData}
\title{Read data from a worksheet}
\usage{
xlsxReadData(
  wb,
  sheetName = 1,
  skipDescriptionRow = FALSE,
  alwaysCharacter = c("Group", "Id$", "Ids$"),
  emptyAsNA = TRUE,
  convertHeaders = TRUE
)
}
\arguments{
\item{wb}{A workbook object or a character string specifying the path to the xlsx file.}

\item{sheetName}{A character string specifying the name of the sheet to read.}

\item{skipDescriptionRow}{A logical value or an integer indicating whether the first row
(or rows, if an integer) should be treated as description rows and skipped during reading.
 When set to TRUE, the first row is skipped; when set to a positive integer, that number of rows will be skipped.
The 'Comment' column is also excluded from the resulting data table.}

\item{alwaysCharacter}{A character vector with column names or regex patterns that should be returned as character (typically identifiers).}

\item{emptyAsNA}{A logical value. If TRUE, empty strings in character columns are converted to NA.
If FALSE NA is returned as empty string. Numeric columns Return always NA}

\item{convertHeaders}{A logical value. If TRUE, column names are converted to start with a lowercase letter.}
}
\value{
A `data.table` containing the sheet data.
}
\description{
This function wraps `openxlsx::read.xlsx` and returns the data as a `data.table`.
}
\details{
The function accepts `skipDescriptionRow` either as a flag (`TRUE`/`FALSE`) or
  as a non-negative integer indicating exactly how many leading rows to skip.
}
\examples{
\dontrun{
wb <- loadWorkbook("example.xlsx")
data <- xlsxReadData(wb, "DataSheet", skipDescriptionRow = 1)
print(data)
}

}
\seealso{
Other function to read from and write to xlsx:
\code{\link[=setHeadersToLowerCase]{setHeadersToLowerCase()}},
\code{\link[=splitInputs]{splitInputs()}},
\code{\link[=xlsxAddDataUsingTemplate]{xlsxAddDataUsingTemplate()}},
\code{\link[=xlsxAddSheet]{xlsxAddSheet()}},
\code{\link[=xlsxCloneAndSet]{xlsxCloneAndSet()}},
\code{\link[=xlsxWriteData]{xlsxWriteData()}}
}
\concept{function to read from and write to xlsx}
