% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxReadData}
\alias{xlsxReadData}
\title{Read data from a worksheet}
\usage{
xlsxReadData(
  wb,
  sheetName,
  skipDescriptionRow = FALSE,
  alwaysCharacter = c("IndividualId", "IndividualIds", "StudyId", "group",
    "outputPathId", "DataGroupIds", "DataGroupId"),
  emptyAsNA = TRUE
)
}
\arguments{
\item{wb}{A workbook or xlsx file.}

\item{sheetName}{A character string specifying the name of the sheet to read.}

\item{skipDescriptionRow}{A boolean indicating if the first line should be interpreted as a description and skipped.}

\item{alwaysCharacter}{A vector with column names that should be returned as character (typically identifiers).}

\item{emptyAsNA}{If TRUE empty strings are converted to NA}
}
\value{
A `data.table` containing the sheet data.
}
\description{
This function wraps `openxlsx::read.xlsx` but returns data as a `data.table`.
}
