% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxReadData}
\alias{xlsxReadData}
\title{reads data}
\usage{
xlsxReadData(wb, sheetName)
}
\arguments{
\item{wb}{workbook or xlsxfile}

\item{sheetName}{sheet name}
}
\value{
`data.table` with sheet data
}
\description{
wrappes openxlsx::read.xlsx but returns data as data.table()
}
