% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxReadData}
\alias{xlsxReadData}
\title{reads data}
\usage{
xlsxReadData(
  wb,
  sheetName,
  skipDescriptionRow = FALSE,
  alwaysCharacter = c("IndividualId", "StudyId", "group", "OutputPathId", "DataGroupIds")
)
}
\arguments{
\item{wb}{workbook or xlsxfile}

\item{sheetName}{sheet name}

\item{skipDescriptionRow}{`boolean`if TRUE first line is interpreted as a description and skipped}

\item{alwaysCharacter}{vector with column names which should be returned as character (typically identifier)}
}
\value{
`data.table` with sheet data
}
\description{
wrapps `openxlsx::read.xlsx` but returns data as data.table()
}
