% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxWriteData}
\alias{xlsxWriteData}
\title{adds data to worksheet}
\usage{
xlsxWriteData(wb, sheetName, dt)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{dt}{`data.table` to write}
}
\description{
wrapps `openxlsx::writeData`, but delete before all current content
}
