% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{xlsxWriteData}
\alias{xlsxWriteData}
\title{Write data to a worksheet, clearing existing content}
\usage{
xlsxWriteData(wb, sheetName, dt)
}
\arguments{
\item{wb}{A workbook object created by `openxlsx::loadWorkbook()`.}

\item{sheetName}{A character string specifying the name of the sheet where data will be written.}

\item{dt}{A `data.table` to write.}
}
\value{
An invisible NULL value.
}
\description{
This function wraps `openxlsx::writeData`, but deletes all current content before writing new data.
}
